Module godot::bind

Export user-defined classes and methods to be called by the engine.

Structs

ExportInfo

Info needed for godot to understand how to export a type to the editor.

Traits

Export

Trait implemented for types that can be used as #[export] fields. This creates a copy of the value, for some type-specific definition of “copy”. For example, Array and Gd are returned via Share::share() instead of copying the actual data.

TypeStringHint

Trait for types that can be represented as a type string for use with PropertyHint::PROPERTY_HINT_TYPE_STRING.

Attribute Macros

godot_api

Derive Macros

FromVariant

GodotClass

Derive macro for GodotClass on structs. You should normally use this macro, rather than implement GodotClass manually for your type.

ToVariant