Derive Macro godot::bind::GodotClass

#[derive(GodotClass)]
{
    // Attributes available to this derive:
    #[class]
    #[base]
    #[export]
    #[init]
    #[signal]
}

Derive macro for GodotClass on structs. You should normally use this macro, rather than implement GodotClass manually for your type.

Construction

To generate a constructor that will let you call MyStruct.new() from GDScript, annotate your struct with #[class(init)]:

#[derive(GodotClass)]
#[class(init)]
struct MyStruct {
    // ...
}

The generated init function will initialize each struct field (except the field annotated with #[base], if any) using Default::default(). To assign some other value, annotate the field with #[init(default = ...)]:

#[derive(GodotClass)]
#[class(init)]
struct MyStruct {
    #[init(default = 42)]
    my_field: i64
}

The given value can be any Rust expression that can be evaluated in the scope where you write the attribute. However, due to limitations in the parser, some complex expressions must be surrounded by parentheses. This is the case if the expression includes a , that is not inside any pair of (...), [...] or {...} (even if it is, for example, inside <...> or |...|). A contrived example:

    #[init(default = (HashMap::<i64, i64>::new()))]
    //                             ^ parentheses needed due to this comma

Inheritance

Unlike C++, Rust doesn’t really have inheritance, but the GDExtension API lets us “inherit” from a built-in engine class.

By default, classes created with this library inherit from RefCounted.

To specify a different class to inherit from, add #[class(base = Base)] as an annotation on your struct:

use godot::prelude::*;

#[derive(GodotClass)]
#[class(base = Node2D)]
struct MyStruct {
    // ...
}

If you need a reference to the base class, you can add a field of type Gd<Base> and annotate it with #[base]:

use godot::prelude::*;

#[derive(GodotClass)]
#[class(base = Node2D)]
struct MyStruct {
    #[base]
    base: Gd<Node2D>,
}

Exported properties

In GDScript, there is a distinction between properties (fields with a get or set declaration) and exports (fields annotated with @export). In the GDExtension API, these two concepts are merged into one.

You can export fields of your struct using the #[export] annotation:

use godot::prelude::*;

#[derive(GodotClass)]
struct MyStruct {
    #[export]
    my_field: i64,
}

This makes the field accessible in GDScript using my_struct.my_field syntax. Additionally, it generates a trivial getter and setter named get_my_field and set_my_field, respectively. These are pub in Rust, since they’re exposed from GDScript anyway.

If you want to implement your own getter and/or setter, write those as a function on your Rust type, expose it using #[func], and annotate the field with #[export(get = ..., set = ...)]:

use godot::prelude::*;

#[derive(GodotClass)]
struct MyStruct {
    #[export(get = get_my_field, set = set_my_field)]
    my_field: i64,
}

#[godot_api]
impl MyStruct {
    #[func]
    pub fn get_my_field(&self) -> i64 {
        self.my_field
    }

    #[func]
    pub fn set_my_field(&mut self, value: i64) {
        self.my_field = value;
    }
}

If you specify only get, no setter is generated, making the field read-only. If you specify only set, no getter is generated, making the field write-only (rarely useful). To add a generated getter or setter in these cases anyway, use get or set without a value:

use godot::prelude::*;

#[derive(GodotClass)]
struct MyStruct {
    // Default getter, custom setter.
    #[export(get, set = set_my_field)]
    my_field: i64,
}

#[godot_api]
impl MyStruct {
    #[func]
    pub fn set_my_field(&mut self, value: i64) {
        self.my_field = value;
    }
}

Signals

The #[signal] attribute is accepted, but not yet implemented. See issue #8.