Module godot::bind::property

Modules

export_info_functions

Functions used to translate user-provided arguments into export hints.

Structs

PropertyHintInfo

Info needed for godot to understand how to export a type to the editor.

Traits

Export

Trait implemented for types that can be used as #[export] fields.

Property

Trait implemented for types that can be used as #[var] fields. This creates a copy of the value, for some type-specific definition of “copy”. For example, Array, Dictionary and Gd are returned by shared reference instead of copying the actual data.

TypeStringHint

Trait for types that can be represented as a type string for use with PropertyHint::PROPERTY_HINT_TYPE_STRING.