Trait godot::bind::property::Cannot_export_without_godot_api_impl

pub trait Cannot_export_without_godot_api_impl {
    const EXISTS: () = ();
}

To export properties to Godot, you must have an impl-block with the #[godot_api] attribute, even if it is empty.

This trait is automatically implemented when such an impl-block is present. If Rust complains that it is not implemented, then you can usually fix this by adding:

#[godot_api]
impl MyClass {}

Where you replace MyClass with the name of your class.

Provided Associated Constants

const EXISTS: () = ()

Object Safety

This trait is not object safe.