Enum godot::builtin::Rid

pub enum Rid {
    Valid(NonZeroU64),
    Invalid,
}

A RID (“resource ID”) is an opaque handle that refers to a Godot Resource.

RIDs do not grant access to the resource itself. Instead, they can be used in lower-level resource APIs such as the servers. See also Godot API docs for RID.

RIDs should be largely safe to work with. Certain calls to servers may fail, however doing so will trigger an error from Godot, and will not cause any UB.

Safety Caveat:

In Godot 3, RID was not as safe as described here. We believe that this is fixed in Godot 4, but this has not been extensively tested as of yet. Some confirmed UB from Godot 3 does not occur anymore, but if you find anything suspicious or outright UB please open an issue.

Variants

Valid(NonZeroU64)

A valid RID may refer to some resource, but is not guaranteed to do so.

Invalid

An invalid RID will never refer to a resource. Internally it is represented as a 0.

Implementations

impl Rid

pub const fn new(id: u64) -> Rid

Create a new RID.

pub const fn to_u64(self) -> u64

Convert this RID into a u64. Returns 0 if it is invalid.

Godot equivalent: Rid.get_id()

pub const fn to_valid_u64(self) -> Option<u64>

Convert this RID into a u64 if it is valid. Otherwise return None.

pub const fn is_valid(&self) -> bool

Returns true if this is a valid RID.

pub const fn is_invalid(&self) -> bool

Returns true if this is an invalid RID.