Trait godot::builtin::meta::GodotCompatible

pub trait GodotCompatible {
    type Via: GodotType;
}

Indicates that a type has some canonical Godot type that can represent it.

The type specified here is what will be used to pass this type across to ffi-boundary to/from Godot. Generally ToGodot needs to be implemented to pass a type to Godot, and FromGodot to receive this type from Godot.

Required Associated Types

type Via: GodotType

The type used for ffi-passing.