Trait godot::builtin::meta::GodotType

pub trait GodotType: GodotConvert<Via = Self> + ToGodot + FromGodot + Sealed {
    type Ffi: GodotFfiVariant;

    // Required methods
    fn to_ffi(&self) -> Self::Ffi;
    fn into_ffi(self) -> Self::Ffi;
    fn try_from_ffi(ffi: Self::Ffi) -> Option<Self>;

    // Provided methods
    fn from_ffi(ffi: Self::Ffi) -> Self { ... }
    fn param_metadata() -> u32 { ... }
    fn class_name() -> ClassName { ... }
    fn property_info(property_name: &str) -> PropertyInfo { ... }
    fn argument_info(property_name: &str) -> MethodParamOrReturnInfo { ... }
    fn return_info() -> Option<MethodParamOrReturnInfo> { ... }
}

Types that can represent some Godot type.

This trait cannot be implemented for custom user types, for that you should see GodotConvert instead. A type implements GodotType when it can directly represent some primitive type exposed by Godot. For instance, i64 implements GodotType, since it can be directly represented by Godot’s int type. But VariantType does not implement GodotType. Since while it is an enum Godot uses, we have no native way to indicate to Godot that a value should be one of the variants of VariantType.

Unlike [GodotFfi], types implementing this trait don’t need to fully represent its corresponding Godot type. For instance i32 does not implement [GodotFfi] because it cannot represent all values of Godot’s int type, however it does implement GodotType because we can set the metadata of values with this type to indicate that they are 32 bits large.

Required Associated Types

type Ffi: GodotFfiVariant

Required Methods

fn to_ffi(&self) -> Self::Ffi

fn into_ffi(self) -> Self::Ffi

fn try_from_ffi(ffi: Self::Ffi) -> Option<Self>

Provided Methods

fn from_ffi(ffi: Self::Ffi) -> Self

fn param_metadata() -> u32

fn class_name() -> ClassName

fn property_info(property_name: &str) -> PropertyInfo

fn argument_info(property_name: &str) -> MethodParamOrReturnInfo

fn return_info() -> Option<MethodParamOrReturnInfo>