Struct godot::builtin::GodotString

pub struct GodotString { /* private fields */ }

Godotโs reference counted string type.

Implementations

impl GodotString

pub fn new() -> GodotString

Construct a new empty GodotString.

pub fn hash(&self) -> u32

Returns a 32-bit integer hash value representing the string.

pub unsafe fn move_string_ptr(self, dst: *mut __GdextString)

Move self into a system pointer. This transfers ownership and thus does not call the destructor.

Safety

dst must be a pointer to a GodotString which is suitable for ffi with Godot.

pub fn chars_checked(&self) -> &[char]

Gets the internal chars slice from a GodotString.

Note: This operation is O(n). Consider using [chars_unchecked] if you can make sure the string is a valid UTF-32.

pub unsafe fn chars_unchecked(&self) -> &[char]

Gets the internal chars slice from a GodotString.

Safety

Make sure the string only contains valid unicode scalar values, currently Godot allows for unpaired surrogates and out of range code points to be appended into the string.