godot::builtin

Struct Quaternion

#[repr(C)]
pub struct Quaternion {
    pub x: f32,
    pub y: f32,
    pub z: f32,
    pub w: f32,
}

Unit quaternion to represent 3D rotations.

See also Quaternion in the Godot documentation.

Godot docs

Quaternion (stable)

Fields
x: f32y: f32z: f32w: f32
Implementations

impl Quaternion

pub const IDENTITY: Quaternion

The identity quaternion, representing no rotation. This has the same rotation as Basis::IDENTITY.

If a Vector3 is rotated (multiplied) by this quaternion, it does not change.

pub fn new(x: f32, y: f32, z: f32, w: f32) -> Quaternion

pub fn from_axis_angle(axis: Vector3, angle: f32) -> Quaternion

Creates a quaternion from a Vector3 and an angle.

Panics

If the vector3 is not normalized.

pub fn from_rotation_arc(arc_from: Vector3, arc_to: Vector3) -> Quaternion

Constructs a Quaternion representing the shortest arc between arc_from and arc_to.

These can be imagined as two points intersecting a unit sphereโs surface, with a radius of 1.0.

The inputs must be unit vectors.

For near-singular cases (arc_fromโarc_to or arc_fromโ-arc_to) the current implementation is only accurate to about 0.001, or better if double-precision is enabled.

Godot equivalent: Quaternion(arc_from: Vector3, arc_to: Vector3)

pub fn angle_to(self, to: Quaternion) -> f32

pub fn dot(self, with: Quaternion) -> f32

pub fn exp(self) -> Quaternion

pub fn to_exp(self) -> Quaternion

๐Deprecated: Moved to Quaternion::exp()

pub fn from_euler(euler: Vector3) -> Quaternion

pub fn get_angle(self) -> f32

pub fn get_axis(self) -> Vector3

pub fn to_euler(self, order: EulerOrder) -> Vector3

pub fn inverse(self) -> Quaternion

pub fn is_finite(self) -> bool

pub fn is_normalized(self) -> bool

pub fn length(self) -> f32

pub fn length_squared(self) -> f32

pub fn log(self) -> Quaternion

pub fn normalized(self) -> Quaternion

Panics

If the quaternion has length of 0.

pub fn slerp(self, to: Quaternion, weight: f32) -> Quaternion

Panics

If either quaternion is not normalized.

pub fn slerpni(self, to: Quaternion, weight: f32) -> Quaternion

Panics

If either quaternion is not normalized.

pub fn spherical_cubic_interpolate( self, b: Quaternion, pre_a: Quaternion, post_b: Quaternion, weight: f32, ) -> Quaternion

Panics

If any quaternions are not normalized.

pub fn spherical_cubic_interpolate_in_time( self, b: Quaternion, pre_a: Quaternion, post_b: Quaternion, weight: f32, b_t: f32, pre_a_t: f32, post_b_t: f32, ) -> Quaternion

Panics

If any quaternions are not normalized.