Struct godot::builtin::Signal

pub struct Signal { /* private fields */ }

A Signal represents a signal of an Object instance in Godot.

Signals are composed of a reference to an Object and the name of the signal on this object.

Godot docs

Signal (stable)

Implementations

impl Signal

pub fn from_object_signal<T, S>(object: &Gd<T>, signal_name: S) -> Signal
where T: GodotClass, S: Into<StringName>,

Create a signal for the signal object::signal_name.

Godot equivalent: Signal(Object object, StringName signal)

pub fn invalid() -> Signal

Creates an invalid/empty signal that cannot be called.

Godot equivalent: Signal()

pub fn connect(&self, callable: Callable, flags: i64) -> Error

Connects this signal to the specified callable.

Optional flags can be also added to configure the connectionโs behavior (see ConnectFlags constants). You can provide additional arguments to the connected callable by using Callable::bind.

A signal can only be connected once to the same Callable. If the signal is already connected, returns Error::ERR_INVALID_PARAMETER and pushes an error message, unless the signal is connected with ConnectFlags::REFERENCE_COUNTED. To prevent this, use Self::is_connected first to check for existing connections.

pub fn disconnect(&self, callable: Callable)

Disconnects this signal from the specified Callable.

If the connection does not exist, generates an error. Use Self::is_connected to make sure that the connection exists.

pub fn emit(&self, varargs: &[Variant])

Emits this signal.

All Callables connected to this signal will be triggered.

pub fn connections(&self) -> Array<Dictionary>

Returns an Array of connections for this signal.

Each connection is represented as a Dictionary that contains three entries:

signal is a reference to this Signal;

callable is a reference to the connected Callable;

flags is a combination of ConnectFlags.

Godot equivalent: get_connections

pub fn name(&self) -> StringName

Returns the name of the signal.

pub fn object(&self) -> Option<Gd<Object>>

Returns the object to which this signal belongs.

Returns None when this signal doesnโt have any object.

Godot equivalent: get_object

pub fn object_id(&self) -> Option<InstanceId>

Returns the ID of this signalโs object, see also Gd::instance_id.

Returns None when this signal doesnโt have any object.

Godot equivalent: get_object_id

pub fn is_connected(&self, callable: Callable) -> bool

Returns true if the specified Callable is connected to this signal.

pub fn is_null(&self) -> bool

Returns true if the signalโs name does not exist in its object, or the object is not valid.