Struct godot::builtin::Variant

pub struct Variant { /* private fields */ }

Implementations

impl Variant

pub fn nil() -> Variant

Create an empty variant (null value in GDScript).

pub fn from<T>(value: T) -> Variantwhere T: ToVariant,

Create a variant holding a non-nil value.

Equivalent to value.to_variant().

pub fn to<T>(&self) -> Twhere T: FromVariant,

⚠️ Convert to type T, panicking on failure.

Equivalent to T::from_variant(&self).

Panics

When this variant holds a different type.

pub fn try_to<T>(&self) -> Result<T, VariantConversionError>where T: FromVariant,

Convert to type T, returning Err on failure.

Equivalent to T::try_from_variant(&self).

pub fn is_nil(&self) -> bool

Checks whether the variant is empty (null value in GDScript).

See also Self::get_type.

pub fn get_type(&self) -> VariantType

Returns the type that is currently held by this variant.

If this variant holds a type Object but no instance (represented as a null object pointer), then Nil will be returned for consistency. This may deviate from Godot behavior – for example, calling Node::get_node_or_null() with an invalid path returns a variant that has type Object but acts like Nil for all practical purposes.

pub fn call(&self, method: impl Into<StringName>, args: &[Variant]) -> Variant

⚠️ Calls the specified method with the given args.

Supports Object as well as built-ins with methods (e.g. Array, Vector3, GodotString, etc).

Panics

If self is not a variant type which supports method calls.

If the method does not exist or the signature is not compatible with the passed arguments.

If the call causes an error.

pub fn evaluate(&self, rhs: &Variant, op: VariantOperator) -> Option<Variant>

pub fn stringify(&self) -> GodotString

return a GodotString representing the variant

pub fn hash(&self) -> i64

return the hash value of the variant.

Godot equivalent : @GlobalScope.hash()

pub fn booleanize(&self) -> bool

return a false only if the variant is Variant::NIL or an empty TypedArray or Dictionary.