Struct godot::builtin::Vector2

#[repr(C)]
pub struct Vector2 {
    pub x: f32,
    pub y: f32,
}

Vector used for 2D math using floating point coordinates.

2-element structure that can be used to represent positions in 2D space or any other pair of numeric values.

It uses floating-point coordinates of 32-bit precision, unlike the engine’s float type which is always 64-bit. The engine can be compiled with the option precision=double to use 64-bit vectors; use the gdext library with the double-precision feature in that case.

See Vector2i for its integer counterpart.

Fields
x: f32

The vector’s X component.

y: f32

The vector’s Y component.

Implementations

impl Vector2

pub const ZERO: Vector2 = _

Zero vector, a vector with all components set to 0.

pub const ONE: Vector2 = _

One vector, a vector with all components set to 1.

impl Vector2

pub const INF: Vector2 = _

Infinity vector, a vector with all components set to real::INFINITY.

impl Vector2

pub const LEFT: Vector2 = _

Left unit vector. Represents the direction of left.

pub const RIGHT: Vector2 = _

Right unit vector. Represents the direction of right.

pub const UP: Vector2 = _

Up unit vector. Y is down in 2D, so this vector points -Y.

pub const DOWN: Vector2 = _

Down unit vector. Y is down in 2D, so this vector points +Y.

impl Vector2

pub const fn new(x: f32, y: f32) -> Vector2

Returns a vector with the given components.

pub const fn splat(v: f32) -> Vector2

Returns a new vector with all components set to v.

pub fn abs(self) -> Vector2

Returns a new vector with all components in absolute values (i.e. positive or zero).

pub fn new_abs(self) -> Vector2

Returns a new vector with all components in absolute values (i.e. positive or zero).

pub fn clamped(self, min: Vector2, max: Vector2) -> Vector2

Returns a new vector with all components clamped between the components of min and max.

Panics Panics if min > max, min is NaN, or max is NaN.

pub fn dot(self, with: Vector2) -> f32

Returns the dot product of this vector and with.

pub fn new_length(self) -> f32

Returns the length (magnitude) of this vector.

pub fn new_length_squared(self) -> f32

Squared length (squared magnitude) of this vector.

Runs faster than Self::length, so prefer it if you need to compare vectors or need the squared distance for some formula.

pub fn coord_min(self, other: Vector2) -> Vector2

Returns a new vector containing the minimum of the two vectors, component-wise.

pub fn new_coord_min(self, other: Vector2) -> Vector2

Returns a new vector containing the minimum of the two vectors, component-wise.

pub fn coord_max(self, other: Vector2) -> Vector2

Returns a new vector containing the maximum of the two vectors, component-wise.

pub fn new_coord_max(self, other: Vector2) -> Vector2

Returns a new vector containing the maximum of the two vectors, component-wise.

pub fn new_max_axis_index(self)

Returns the axis of the vector’s highest value. See AXIS_* constants. If all components are equal, this method returns AXIS_X.

pub fn new_min_axis_index(self)

Returns the axis of the vector’s lowest value. See AXIS_* constants. If all components are equal, this method returns AXIS_W.

pub fn new_sign(self) -> Vector2

Returns a new vector with each component set to 1 if it’s positive, -1 if it’s negative, and 0 if it’s zero.

pub fn snapped(self, step: Vector2) -> Vector2

A new vector with each component snapped to the closest multiple of the corresponding component in step.

impl Vector2

pub fn new_ceil(self) -> Vector2

Returns a new vector with all components rounded up (towards positive infinity).

pub fn cubic_interpolate( self, b: Vector2, pre_a: Vector2, post_b: Vector2, weight: f32 ) -> Vector2

Performs a cubic interpolation between this vector and b using pre_a and post_b as handles, and returns the result at position weight. weight is on the range of 0.0 to 1.0, representing the amount of interpolation.

pub fn cubic_interpolate_in_time( self, b: Vector2, pre_a: Vector2, post_b: Vector2, weight: f32, b_t: f32, pre_a_t: f32, post_b_t: f32 ) -> Vector2

Performs a cubic interpolation between this vector and b using pre_a and post_b as handles, and returns the result at position weight. weight is on the range of 0.0 to 1.0, representing the amount of interpolation. It can perform smoother interpolation than Self::cubic_interpolate by the time values.

pub fn new_direction_to(self, to: Vector2) -> Vector2

Returns the normalized vector pointing from this vector to to. This is equivalent to using (b - a).normalized().

pub fn new_distance_squared_to(self, to: Vector2) -> f32

Returns the squared distance between this vector and to.

This method runs faster than Self::distance_to, so prefer it if you need to compare vectors or need the squared distance for some formula.

pub fn new_distance_to(self, to: Vector2) -> f32

Returns the distance between this vector and to.

pub fn new_dot(self, with: Vector2) -> f32

Returns the dot product of this vector and with.

pub fn new_floor(self) -> Vector2

Returns a new vector with all components rounded down (towards negative infinity).

pub fn new_is_finite(self) -> bool

Returns true if each component of this vector is finite.

pub fn new_is_normalized(self) -> bool

Returns true if the vector is normalized, i.e. its length is approximately equal to 1.

pub fn is_zero_approx(self) -> bool

Returns true if this vector’s values are approximately zero. This method is faster than using [Self::is_equal_approx] with one value as a zero vector.

pub fn lerp(self, other: Vector2, weight: f32) -> Vector2

Returns the result of the linear interpolation between this vector and to by amount weight. weight is on the range of 0.0 to 1.0, representing the amount of interpolation.

pub fn new_normalized(self) -> Vector2

Returns the result of scaling the vector to unit length. Equivalent to v / v.length(). See also Self::is_normalized.

Note: This function may return incorrect values if the input vector length is near zero.

pub fn posmod(self, pmod: f32) -> Vector2

Returns a vector composed of the FloatExt::fposmod of this vector’s components and pmod.

pub fn posmodv(self, modv: Vector2) -> Vector2

Returns a vector composed of the FloatExt::fposmod of this vector’s components and modv’s components.

pub fn new_round(self) -> Vector2

Returns a new vector with all components rounded to the nearest integer, with halfway cases rounded away from zero.

pub fn sign(self) -> Vector2

Returns a new vector with each component set to 1.0 if it’s positive, -1.0 if it’s negative, and 0.0 if it’s zero.

impl Vector2

pub fn aspect(self) -> f32

Returns the aspect ratio of this vector, the ratio of Self::x to Self::y.

pub fn max_axis(self) -> Vector2Axis

Returns the axis of the vector’s highest value. See Vector2Axis enum. If all components are equal, this method returns Vector2Axis::X.

pub fn min_axis(self) -> Vector2Axis

Returns the axis of the vector’s lowest value. See Vector2Axis enum. If all components are equal, this method returns Vector2Axis::Y.

impl Vector2

pub fn new_angle_to(self, _to: Vector2) -> f32

Returns the angle to the given vector, in radians.

pub fn bezier_derivative( self, control_1: Vector2, control_2: Vector2, end: Vector2, t: f32 ) -> Vector2

Returns the derivative at the given t on the Bézier curve defined by this vector and the given control_1, control_2, and end points.

pub fn bezier_interpolate( self, control_1: Vector2, control_2: Vector2, end: Vector2, t: f32 ) -> Vector2

Returns the point at the given t on the Bézier curve defined by this vector and the given control_1, control_2, and end points.

pub fn bounce(self, normal: Vector2) -> Vector2

Returns a new vector “bounced off” from a plane defined by the given normal.

pub fn new_cross(self, _with: Vector2) -> Vector2

Returns the cross product of this vector and with.

This returns a vector perpendicular to both this and with, which would be the normal vector of the plane defined by the two vectors. As there are two such vectors, in opposite directions, this method returns the vector defined by a right-handed coordinate system. If the two vectors are parallel this returns an empty vector, making it useful for testing if two vectors are parallel.

pub fn new_limit_length(self, length: Option<f32>) -> Vector2

Returns the vector with a maximum length by limiting its length to length.

pub fn move_toward(self, to: Vector2, delta: f32) -> Vector2

Returns a new vector moved toward to by the fixed delta amount. Will not go past the final value.

pub fn new_project(self, b: Vector2) -> Vector2

Returns the result of projecting the vector onto the given vector b.

pub fn reflect(self, n: Vector2) -> Vector2

Returns the result of reflecting the vector defined by the given direction vector n.

pub fn new_slerp(self, to: Vector2, weight: f32) -> Vector2

Returns the result of spherical linear interpolation between this vector and to, by amount weight. weight is on the range of 0.0 to 1.0, representing the amount of interpolation.

This method also handles interpolating the lengths if the input vectors have different lengths. For the special case of one or both input vectors having zero length, this method behaves like Self::lerp.

pub fn slide(self, n: Vector2) -> Vector2

Returns a new vector slid along a plane defined by the given normal.

impl Vector2

pub fn length(self) -> f32

Returns the length (magnitude) of this vector.

pub fn normalized(self) -> Vector2

Returns the vector scaled to unit length. Equivalent to self / self.length(). See also is_normalized().

If the vector is zero, the result is also zero.

impl Vector2

pub const fn from_vector2i(v: Vector2i) -> Vector2

Constructs a new Vector2 from a Vector2i.

pub fn angle(self) -> f32

pub fn angle_to(self, to: Vector2) -> f32

pub fn angle_to_point(self, to: Vector2) -> f32

pub fn ceil(self) -> Vector2

pub fn clamp(self, min: Vector2, max: Vector2) -> Vector2

pub fn cross(self, with: Vector2) -> f32

pub fn direction_to(self, to: Vector2) -> Vector2

pub fn distance_squared_to(self, to: Vector2) -> f32

pub fn distance_to(self, to: Vector2) -> f32

pub fn floor(self) -> Vector2

pub fn from_angle(angle: f32) -> Vector2

pub fn is_finite(self) -> bool

pub fn is_normalized(self) -> bool

pub fn length_squared(self) -> f32

pub fn limit_length(self, length: Option<f32>) -> Vector2

pub fn max_axis_index(self) -> Vector2Axis

pub fn min_axis_index(self) -> Vector2Axis

pub fn orthogonal(self) -> Vector2

pub fn project(self, b: Vector2) -> Vector2

pub fn round(self) -> Vector2

pub fn slerp(self, to: Vector2, weight: f32) -> Vector2

pub fn rotated(self, angle: f32) -> Vector2

Returns the result of rotating this vector by angle (in radians).

pub fn coords(&self) -> (f32, f32)