Struct godot::builtin::Vector4

pub struct Vector4 {
    pub x: f32,
    pub y: f32,
    pub z: f32,
    pub w: f32,
}

Vector used for 4D math using floating point coordinates.

4-element structure that can be used to represent any quadruplet of numeric values.

It uses floating-point coordinates of 32-bit precision, unlike the engine’s float type which is always 64-bit. The engine can be compiled with the option precision=double to use 64-bit vectors; use the gdext library with the double-precision feature in that case.

See Vector4i for its integer counterpart.

Fields
x: f32

The vector’s X component.

y: f32

The vector’s Y component.

z: f32

The vector’s Z component.

w: f32

The vector’s W component.

Implementations

impl Vector4

pub fn abs(self) -> Vector4

Returns a new vector with all components in absolute values (i.e. positive or zero).

pub fn coord_min(self, other: Vector4) -> Vector4

Returns a new vector containing the minimum of the two vectors, component-wise.

pub fn coord_max(self, other: Vector4) -> Vector4

Returns a new vector containing the maximum of the two vectors, component-wise.

impl Vector4

pub fn length(self) -> f32

Returns the length (magnitude) of this vector.

pub fn normalized(self) -> Vector4

Returns the vector scaled to unit length. Equivalent to self / self.length(). See also is_normalized().

If the vector is zero, the result is also zero.

impl Vector4

pub const fn new(x: f32, y: f32, z: f32, w: f32) -> Vector4

Returns a Vector4 with the given components.

pub const fn splat(v: f32) -> Vector4

Returns a new Vector4 with all components set to v.

pub const fn from_vector4i(v: Vector4i) -> Vector4

Constructs a new Vector3 from a [Vector3i].

pub const ZERO: Vector4 = Self::splat(0.0)

Zero vector, a vector with all components set to 0.0.

pub const ONE: Vector4 = Self::splat(1.0)

One vector, a vector with all components set to 1.0.

pub const INF: Vector4 = Self::splat(real::INFINITY)

Infinity vector, a vector with all components set to real::INFINITY.

pub fn is_equal_approx(self, to: Vector4) -> bool