Trait godot::builtin::FromVariant

pub trait FromVariant: Sized {
    // Required method
    fn try_from_variant(
        variant: &Variant
    ) -> Result<Self, VariantConversionError>;

    // Provided method
    fn from_variant(variant: &Variant) -> Self { ... }
}

Trait to enable conversions of types from the Variant type.

Required Methods

fn try_from_variant(variant: &Variant) -> Result<Self, VariantConversionError>

Tries to convert a Variant to Self, allowing to check the success or failure.

Provided Methods

fn from_variant(variant: &Variant) -> Self

⚠️ Converts from Variant to Self, panicking on error.

This method should generally not be overridden by trait impls, even if conversions are infallible. Implementing Self::try_from_variant suffices.