godot::builtin

Type Alias real

pub type real = f32;

Floating point type used for many structs and functions in Godot.

This type is f32 by default, and f64 when the Cargo feature double-precision is enabled.

This is not the float type in GDScript; that type is always 64-bits. Rather, many structs in Godot may use either 32-bit or 64-bit floats, for example Vector2. To convert between real and f32 or f64, see RealConv.

See also the Godot docs on float.