godot::classes
Module native

Support for Godot native structures.

Native structures are a niche API in Godot. These are low-level data types that are passed as pointers to/from the engine. In Rust, they are represented as #[repr(C)] structs.

There is unfortunately not much official documentation available; you may need to look at Godot source code. Most users will not need native structures, as they are very specialized.

Structs

AudioFrame

Native structure; can be passed via pointer in APIs that are not exposed to GDScript.

CaretInfo

Native structure; can be passed via pointer in APIs that are not exposed to GDScript.

Glyph

Native structure; can be passed via pointer in APIs that are not exposed to GDScript.

ObjectId

Native structure; can be passed via pointer in APIs that are not exposed to GDScript.

PhysicsServer2DExtensionMotionResult

Native structure; can be passed via pointer in APIs that are not exposed to GDScript.

PhysicsServer2DExtensionRayResult

Native structure; can be passed via pointer in APIs that are not exposed to GDScript.

PhysicsServer2DExtensionShapeRestInfo

Native structure; can be passed via pointer in APIs that are not exposed to GDScript.

PhysicsServer2DExtensionShapeResult

Native structure; can be passed via pointer in APIs that are not exposed to GDScript.

PhysicsServer3DExtensionMotionCollision

Native structure; can be passed via pointer in APIs that are not exposed to GDScript.

PhysicsServer3DExtensionMotionResult

Native structure; can be passed via pointer in APIs that are not exposed to GDScript.

PhysicsServer3DExtensionRayResult

Native structure; can be passed via pointer in APIs that are not exposed to GDScript.

PhysicsServer3DExtensionShapeRestInfo

Native structure; can be passed via pointer in APIs that are not exposed to GDScript.

PhysicsServer3DExtensionShapeResult

Native structure; can be passed via pointer in APIs that are not exposed to GDScript.

ScriptLanguageExtensionProfilingInfo

Native structure; can be passed via pointer in APIs that are not exposed to GDScript.