godot::classes

Struct Area2D

pub struct Area2D { /* private fields */ }

Godot class Area2D.

Inherits CollisionObject2D.

Related symbols:

area_2d: sidecar module with related enum/flag types

IArea2D: virtual methods

SignalsOfArea2D: signal collection

See also Godot docs for Area2D.

Construction

This class is manually managed. You can create a new instance using Area2D::new_alloc().

Do not forget to call free() or hand over ownership to Godot.

Godot docs

Area2D is a region of 2D space defined by one or multiple CollisionShape2D or CollisionPolygon2D child nodes. It detects when other CollisionObject2Ds enter or exit it, and it also keeps track of which collision objects haven’t exited it yet (i.e. which one are overlapping it).

This node can also locally alter or override physics parameters (gravity, damping) and route audio to custom audio buses.

Note: Areas and bodies created with PhysicsServer2D might not interact as expected with Area2Ds, and might not emit signals or track objects correctly.

Implementations

impl Area2D

pub fn set_gravity_space_override_mode( &mut self, space_override_mode: SpaceOverride, )

pub fn get_gravity_space_override_mode(&self) -> SpaceOverride

pub fn set_gravity_is_point(&mut self, enable: bool)

pub fn is_gravity_a_point(&self) -> bool

pub fn set_gravity_point_unit_distance(&mut self, distance_scale: f32)

pub fn get_gravity_point_unit_distance(&self) -> f32

pub fn set_gravity_point_center(&mut self, center: Vector2)

pub fn get_gravity_point_center(&self) -> Vector2

pub fn set_gravity_direction(&mut self, direction: Vector2)

pub fn get_gravity_direction(&self) -> Vector2

pub fn set_gravity(&mut self, gravity: f32)

pub fn get_gravity(&self) -> f32

pub fn set_linear_damp_space_override_mode( &mut self, space_override_mode: SpaceOverride, )

pub fn get_linear_damp_space_override_mode(&self) -> SpaceOverride

pub fn set_angular_damp_space_override_mode( &mut self, space_override_mode: SpaceOverride, )

pub fn get_angular_damp_space_override_mode(&self) -> SpaceOverride

pub fn set_linear_damp(&mut self, linear_damp: f32)

pub fn get_linear_damp(&self) -> f32

pub fn set_angular_damp(&mut self, angular_damp: f32)

pub fn get_angular_damp(&self) -> f32

pub fn set_priority(&mut self, priority: i32)

pub fn get_priority(&self) -> i32

pub fn set_monitoring(&mut self, enable: bool)

pub fn is_monitoring(&self) -> bool

pub fn set_monitorable(&mut self, enable: bool)

pub fn is_monitorable(&self) -> bool

pub fn get_overlapping_bodies(&self) -> Array<Gd<Node2D>>

Godot docs

Returns a list of intersecting PhysicsBody2Ds and TileMaps. The overlapping body’s [member CollisionObject2D.collision_layer] must be part of this area’s [member CollisionObject2D.collision_mask] in order to be detected.

For performance reasons (collisions are all processed at the same time) this list is modified once during the physics step, not immediately after objects are moved. Consider using signals instead.

pub fn get_overlapping_areas(&self) -> Array<Gd<Area2D>>

Godot docs

Returns a list of intersecting Area2Ds. The overlapping area’s [member CollisionObject2D.collision_layer] must be part of this area’s [member CollisionObject2D.collision_mask] in order to be detected.

For performance reasons (collisions are all processed at the same time) this list is modified once during the physics step, not immediately after objects are moved. Consider using signals instead.

pub fn has_overlapping_bodies(&self) -> bool

Godot docs

Returns true if intersecting any PhysicsBody2Ds or TileMaps, otherwise returns false. The overlapping body’s [member CollisionObject2D.collision_layer] must be part of this area’s [member CollisionObject2D.collision_mask] in order to be detected.

For performance reasons (collisions are all processed at the same time) the list of overlapping bodies is modified once during the physics step, not immediately after objects are moved. Consider using signals instead.

pub fn has_overlapping_areas(&self) -> bool

Godot docs

Returns true if intersecting any Area2Ds, otherwise returns false. The overlapping area’s [member CollisionObject2D.collision_layer] must be part of this area’s [member CollisionObject2D.collision_mask] in order to be detected.

For performance reasons (collisions are all processed at the same time) the list of overlapping areas is modified once during the physics step, not immediately after objects are moved. Consider using signals instead.

pub fn overlaps_body(&self, body: impl AsArg<Gd<Node>>) -> bool

Godot docs

Returns true if the given physics body intersects or overlaps this Area2D, false otherwise.

Note: The result of this test is not immediate after moving objects. For performance, list of overlaps is updated once per frame and before the physics step. Consider using signals instead.

The body argument can either be a PhysicsBody2D or a TileMap instance. While TileMaps are not physics bodies themselves, they register their tiles with collision shapes as a virtual physics body.

pub fn overlaps_area(&self, area: impl AsArg<Gd<Node>>) -> bool

Godot docs

Returns true if the given Area2D intersects or overlaps this Area2D, false otherwise.

Note: The result of this test is not immediate after moving objects. For performance, the list of overlaps is updated once per frame and before the physics step. Consider using signals instead.

pub fn set_audio_bus_name(&mut self, name: impl AsArg<StringName>)

pub fn get_audio_bus_name(&self) -> StringName

pub fn set_audio_bus_override(&mut self, enable: bool)

pub fn is_overriding_audio_bus(&self) -> bool