godot::classes
Struct Container

#[repr(C)]
pub struct Container { /* private fields */ }

Godot class Container.

Inherits Control.

Related symbols:

IContainer: virtual methods

ContainerNotification: notification type

See also Godot docs for Container.

Construction

This class is manually managed. You can create a new instance using Container::new_alloc().

Do not forget to call free() or hand over ownership to Godot.

Implementations

impl Container

pub fn queue_sort(&mut self)

pub fn fit_child_in_rect( &mut self, child: impl AsObjectArg<Control>, rect: Rect2, )

pub fn notify(&mut self, what: ContainerNotification)

Sends a Godot notification to all classes inherited by the object.

Triggers calls to on_notification(), and depending on the notification, also to Godots lifecycle callbacks such as ready().

Starts from the highest ancestor (the Object class) and goes down the hierarchy. See also Godot docs for Object::notification().

Panics

If you call this method on a user-defined object while holding a GdRef or GdMut guard on the instance, you will encounter a panic. The reason is that the receiving virtual method on_notification() acquires a GdMut lock dynamically, which must be exclusive.

pub fn notify_reversed(&mut self, what: ContainerNotification)

Like Self::notify(), but starts at the most-derived class and goes up the hierarchy.

See docs of that method, including the panics.