godot::classes

Struct DirAccess

pub struct DirAccess { /* private fields */ }

Godot class DirAccess.

Inherits RefCounted.

Related symbols:

dir_access: sidecar module with related enum/flag types

See also Godot docs for DirAccess.

Not instantiable

This class cannot be constructed. Obtain Gd<DirAccess> instances via Godot APIs.

Final class

This class is final, meaning you cannot inherit from it, and it comes without I* interface trait. It is still possible that other Godot classes inherit from it, but that is limited to the engine itself.

Godot docs

This class is used to manage directories and their content, even outside of the project folder.

DirAccess can’t be instantiated directly. Instead it is created with a static method that takes a path for which it will be opened.

Most of the methods have a static alternative that can be used without creating a DirAccess. Static methods only support absolute paths (including res:// and user://).

# Standard
var dir = DirAccess.open("user://levels")
dir.make_dir("world1")
# Static
DirAccess.make_dir_absolute("user://levels/world1")

Note: Accessing project (“res://”) directories once exported may behave unexpectedly as some files are converted to engine-specific formats and their original source files may not be present in the expected PCK package. Because of this, to access resources in an exported project, it is recommended to use ResourceLoader instead of FileAccess.

Here is an example on how to iterate through the files of a directory:

func dir_contents(path):
	var dir = DirAccess.open(path)
	if dir:
		dir.list_dir_begin()
		var file_name = dir.get_next()
		while file_name != "":
			if dir.current_is_dir():
				print("Found directory: " + file_name)
			else:
				print("Found file: " + file_name)
			file_name = dir.get_next()
	else:
		print("An error occurred when trying to access the path.")

Keep in mind that file names may change or be remapped after export. If you want to see the actual resource file list as it appears in the editor, use list_directory instead.

Implementations

impl DirAccess

pub fn open(path: impl AsArg<GString>) -> Option<Gd<DirAccess>>

Creates a new DirAccess object and opens an existing directory of the filesystem. The path argument can be within the project tree (res://folder), the user directory (user://folder) or an absolute path of the user filesystem (e.g. /tmp/folder or C:\tmp\folder).

Returns null if opening the directory failed. You can use get_open_error to check the error that occurred.

pub fn get_open_error() -> Error

Returns the result of the last open call in the current thread.

pub fn create_temp() -> Option<Gd<DirAccess>>

To set the default parameters, use create_temp_ex and its builder methods. See the book for detailed usage instructions. Creates a temporary directory. This directory will be freed when the returned DirAccess is freed.

If prefix is not empty, it will be prefixed to the directory name, separated by a -.

If keep is true, the directory is not deleted when the returned DirAccess is freed.

Returns null if opening the directory failed. You can use get_open_error to check the error that occurred.

pub fn create_temp_ex<'ex>() -> ExCreateTemp<'ex>

Creates a temporary directory. This directory will be freed when the returned DirAccess is freed.

If prefix is not empty, it will be prefixed to the directory name, separated by a -.

If keep is true, the directory is not deleted when the returned DirAccess is freed.

Returns null if opening the directory failed. You can use get_open_error to check the error that occurred.

pub fn list_dir_begin(&mut self) -> Error

Initializes the stream used to list all files and directories using the get_next function, closing the currently opened stream if needed. Once the stream has been processed, it should typically be closed with list_dir_end.

Affected by [member include_hidden] and [member include_navigational].

Note: The order of files and directories returned by this method is not deterministic, and can vary between operating systems. If you want a list of all files or folders sorted alphabetically, use get_files or get_directories.

pub fn get_next(&mut self) -> GString

Returns the next element (file or directory) in the current directory.

The name of the file or directory is returned (and not its full path). Once the stream has been fully processed, the method returns an empty String and closes the stream automatically (i.e. list_dir_end would not be mandatory in such a case).

pub fn current_is_dir(&self) -> bool

Returns whether the current item processed with the last get_next call is a directory (. and .. are considered directories).

pub fn list_dir_end(&mut self)

Closes the current stream opened with list_dir_begin (whether it has been fully processed with get_next does not matter).

pub fn get_files(&self) -> PackedArray<GString>

Returns a PackedStringArray containing filenames of the directory contents, excluding directories. The array is sorted alphabetically.

Affected by [member include_hidden].

Note: When used on a res:// path in an exported project, only the files actually included in the PCK at the given folder level are returned. In practice, this means that since imported resources are stored in a top-level .godot/ folder, only paths to *.gd and *.import files are returned (plus a few files such as project.godot or project.binary and the project icon). In an exported project, the list of returned files will also vary depending on whether [member ProjectSettings.editor/export/convert_text_resources_to_binary] is true.

pub fn get_files_at(path: impl AsArg<GString>) -> PackedArray<GString>

Returns a PackedStringArray containing filenames of the directory contents, excluding directories, at the given path. The array is sorted alphabetically.

Use get_files if you want more control of what gets included.

Note: When used on a res:// path in an exported project, only the files included in the PCK at the given folder level are returned. In practice, this means that since imported resources are stored in a top-level .godot/ folder, only paths to .gd and .import files are returned (plus a few other files, such as project.godot or project.binary and the project icon). In an exported project, the list of returned files will also vary depending on [member ProjectSettings.editor/export/convert_text_resources_to_binary].

pub fn get_directories(&self) -> PackedArray<GString>

Returns a PackedStringArray containing filenames of the directory contents, excluding files. The array is sorted alphabetically.

Affected by [member include_hidden] and [member include_navigational].

Note: The returned directories in the editor and after exporting in the res:// directory may differ as some files are converted to engine-specific formats when exported.

pub fn get_directories_at(path: impl AsArg<GString>) -> PackedArray<GString>

Returns a PackedStringArray containing filenames of the directory contents, excluding files, at the given path. The array is sorted alphabetically.

Use get_directories if you want more control of what gets included.

Note: The returned directories in the editor and after exporting in the res:// directory may differ as some files are converted to engine-specific formats when exported.

pub fn get_drive_count() -> i32

On Windows, returns the number of drives (partitions) mounted on the current filesystem.

On macOS and Android, returns the number of mounted volumes.

On Linux, returns the number of mounted volumes and GTK 3 bookmarks.

On other platforms, the method returns 0.

pub fn get_drive_name(idx: i32) -> GString

On Windows, returns the name of the drive (partition) passed as an argument (e.g. C:).

On macOS, returns the path to the mounted volume passed as an argument.

On Linux, returns the path to the mounted volume or GTK 3 bookmark passed as an argument.

On Android (API level 30+), returns the path to the mounted volume as an argument.

On other platforms, or if the requested drive does not exist, the method returns an empty String.

pub fn get_current_drive(&self) -> i32

Returns the currently opened directory’s drive index. See get_drive_name to convert returned index to the name of the drive.

pub fn change_dir(&mut self, to_dir: impl AsArg<GString>) -> Error

Changes the currently opened directory to the one passed as an argument. The argument can be relative to the current directory (e.g. newdir or ../newdir), or an absolute path (e.g. /tmp/newdir or res://somedir/newdir).

Returns one of the [enum Error] code constants (Error::OK on success).

Note: The new directory must be within the same scope, e.g. when you had opened a directory inside res://, you can’t change it to user:// directory. If you need to open a directory in another access scope, use open to create a new instance instead.

pub fn get_current_dir(&self) -> GString

To set the default parameters, use get_current_dir_ex and its builder methods. See the book for detailed usage instructions. Returns the absolute path to the currently opened directory (e.g. res://folder or C:\tmp\folder).

pub fn get_current_dir_ex<'ex>(&'ex self) -> ExGetCurrentDir<'ex>

Returns the absolute path to the currently opened directory (e.g. res://folder or C:\tmp\folder).

pub fn make_dir(&mut self, path: impl AsArg<GString>) -> Error

Creates a directory. The argument can be relative to the current directory, or an absolute path. The target directory should be placed in an already existing directory (to create the full path recursively, see make_dir_recursive).

Returns one of the [enum Error] code constants (Error::OK on success).

pub fn make_dir_absolute(path: impl AsArg<GString>) -> Error

Static version of make_dir. Supports only absolute paths.

pub fn make_dir_recursive(&mut self, path: impl AsArg<GString>) -> Error

Creates a target directory and all necessary intermediate directories in its path, by calling make_dir recursively. The argument can be relative to the current directory, or an absolute path.

Returns one of the [enum Error] code constants (Error::OK on success).

pub fn make_dir_recursive_absolute(path: impl AsArg<GString>) -> Error

Static version of make_dir_recursive. Supports only absolute paths.

pub fn file_exists(&mut self, path: impl AsArg<GString>) -> bool

Returns whether the target file exists. The argument can be relative to the current directory, or an absolute path.

For a static equivalent, use file_exists.

Note: Many resources types are imported (e.g. textures or sound files), and their source asset will not be included in the exported game, as only the imported version is used. See exists for an alternative approach that takes resource remapping into account.

pub fn dir_exists(&mut self, path: impl AsArg<GString>) -> bool

Returns whether the target directory exists. The argument can be relative to the current directory, or an absolute path.

Note: The returned bool in the editor and after exporting when used on a path in the res:// directory may be different. Some files are converted to engine-specific formats when exported, potentially changing the directory structure.

pub fn dir_exists_absolute(path: impl AsArg<GString>) -> bool

Static version of dir_exists. Supports only absolute paths.

Note: The returned bool in the editor and after exporting when used on a path in the res:// directory may be different. Some files are converted to engine-specific formats when exported, potentially changing the directory structure.

pub fn get_space_left(&self) -> u64

Returns the available space on the current directory’s disk, in bytes. Returns 0 if the platform-specific method to query the available space fails.

pub fn copy( &mut self, from: impl AsArg<GString>, to: impl AsArg<GString>, ) -> Error

To set the default parameters, use copy_ex and its builder methods. See the book for detailed usage instructions. Copies the from file to the to destination. Both arguments should be paths to files, either relative or absolute. If the destination file exists and is not access-protected, it will be overwritten.

If chmod_flags is different than -1, the Unix permissions for the destination path will be set to the provided value, if available on the current operating system.

Returns one of the [enum Error] code constants (Error::OK on success).

pub fn copy_ex<'ex>( &'ex mut self, from: impl AsArg<GString> + 'ex, to: impl AsArg<GString> + 'ex, ) -> ExCopy<'ex>

Copies the from file to the to destination. Both arguments should be paths to files, either relative or absolute. If the destination file exists and is not access-protected, it will be overwritten.

If chmod_flags is different than -1, the Unix permissions for the destination path will be set to the provided value, if available on the current operating system.

Returns one of the [enum Error] code constants (Error::OK on success).

pub fn copy_absolute( from: impl AsArg<GString>, to: impl AsArg<GString>, ) -> Error

To set the default parameters, use copy_absolute_ex and its builder methods. See the book for detailed usage instructions. Static version of copy. Supports only absolute paths.

pub fn copy_absolute_ex<'ex>( from: impl AsArg<GString> + 'ex, to: impl AsArg<GString> + 'ex, ) -> ExCopyAbsolute<'ex>

Static version of copy. Supports only absolute paths.

pub fn rename( &mut self, from: impl AsArg<GString>, to: impl AsArg<GString>, ) -> Error

Renames (move) the from file or directory to the to destination. Both arguments should be paths to files or directories, either relative or absolute. If the destination file or directory exists and is not access-protected, it will be overwritten.

Returns one of the [enum Error] code constants (Error::OK on success).

pub fn rename_absolute( from: impl AsArg<GString>, to: impl AsArg<GString>, ) -> Error

Static version of rename. Supports only absolute paths.

pub fn remove(&mut self, path: impl AsArg<GString>) -> Error

Permanently deletes the target file or an empty directory. The argument can be relative to the current directory, or an absolute path. If the target directory is not empty, the operation will fail.

If you don’t want to delete the file/directory permanently, use move_to_trash instead.

Returns one of the [enum Error] code constants (Error::OK on success).

pub fn remove_absolute(path: impl AsArg<GString>) -> Error

Static version of remove. Supports only absolute paths.

pub fn is_link(&self, path: impl AsArg<GString>) -> bool

Returns true if the file or directory is a symbolic link, directory junction, or other reparse point.

Note: This method is implemented on macOS, Linux, and Windows.

pub fn read_link(&mut self, path: impl AsArg<GString>) -> GString

Returns target of the symbolic link.

Note: This method is implemented on macOS, Linux, and Windows.

pub fn create_link( &mut self, source: impl AsArg<GString>, target: impl AsArg<GString>, ) -> Error

Creates symbolic link between files or folders.

Note: On Windows, this method works only if the application is running with elevated privileges or Developer Mode is enabled.

Note: This method is implemented on macOS, Linux, and Windows.

pub fn is_bundle(&self, path: impl AsArg<GString>) -> bool

Returns true if the directory is a macOS bundle.

Note: This method is implemented on macOS.

pub fn set_include_navigational(&mut self, enable: bool)

pub fn get_include_navigational(&self) -> bool

pub fn set_include_hidden(&mut self, enable: bool)

pub fn get_include_hidden(&self) -> bool

pub fn get_filesystem_type(&self) -> GString

Returns file system type name of the current directory’s disk. Returned values are uppercase strings like NTFS, FAT32, EXFAT, APFS, EXT4, BTRFS, and so on.

Note: This method is implemented on macOS, Linux, Windows and for PCK virtual file system.

pub fn is_case_sensitive(&self, path: impl AsArg<GString>) -> bool

Returns true if the file system or directory use case sensitive file names.

Note: This method is implemented on macOS, Linux (for EXT4 and F2FS filesystems only) and Windows. On other platforms, it always returns true.

pub fn is_equivalent( &self, path_a: impl AsArg<GString>, path_b: impl AsArg<GString>, ) -> bool

Returns true if paths path_a and path_b resolve to the same file system object. Returns false otherwise, even if the files are bit-for-bit identical (e.g., identical copies of the file that are not symbolic links).