godot::classes

Struct NavigationAgent2D

pub struct NavigationAgent2D { /* private fields */ }

Available on crate feature experimental-godot-api only.

Godot class NavigationAgent2D.

Inherits Node.

Related symbols:

navigation_agent_2d: sidecar module with related enum/flag types

INavigationAgent2D: virtual methods

SignalsOfNavigationAgent2D: signal collection

See also Godot docs for NavigationAgent2D.

Construction

This class is manually managed. You can create a new instance using NavigationAgent2D::new_alloc().

Do not forget to call free() or hand over ownership to Godot.

Godot docs

A 2D agent used to pathfind to a position while avoiding static and dynamic obstacles. The calculation can be used by the parent node to dynamically move it along the path. Requires navigation data to work correctly.

Dynamic obstacles are avoided using RVO collision avoidance. Avoidance is computed before physics, so the pathfinding information can be used safely in the physics step.

Note: After setting the [member target_position] property, the get_next_path_position method must be used once every physics frame to update the internal path logic of the navigation agent. The vector position it returns should be used as the next movement position for the agent’s parent node.

Note: Several methods of this class, such as get_next_path_position, can trigger a new path calculation. Calling these in your callback to an agent’s signal, such as waypoint_reached, can cause infinite recursion. It is recommended to call these methods in the physics step or, alternatively, delay their call until the end of the frame (see call_deferred or ConnectFlags::DEFERRED).

Implementations

impl NavigationAgent2D

pub fn get_rid(&self) -> Rid

Returns the RID of this agent on the NavigationServer2D.

pub fn set_avoidance_enabled(&mut self, enabled: bool)

pub fn get_avoidance_enabled(&self) -> bool

pub fn set_path_desired_distance(&mut self, desired_distance: f32)

pub fn get_path_desired_distance(&self) -> f32

pub fn set_target_desired_distance(&mut self, desired_distance: f32)

pub fn get_target_desired_distance(&self) -> f32

pub fn set_radius(&mut self, radius: f32)

pub fn get_radius(&self) -> f32

pub fn set_neighbor_distance(&mut self, neighbor_distance: f32)

pub fn get_neighbor_distance(&self) -> f32

pub fn set_max_neighbors(&mut self, max_neighbors: i32)

pub fn get_max_neighbors(&self) -> i32

pub fn set_time_horizon_agents(&mut self, time_horizon: f32)

pub fn get_time_horizon_agents(&self) -> f32

pub fn set_time_horizon_obstacles(&mut self, time_horizon: f32)

pub fn get_time_horizon_obstacles(&self) -> f32

pub fn set_max_speed(&mut self, max_speed: f32)

pub fn get_max_speed(&self) -> f32

pub fn set_path_max_distance(&mut self, max_speed: f32)

pub fn get_path_max_distance(&self) -> f32

pub fn set_navigation_layers(&mut self, navigation_layers: u32)

pub fn get_navigation_layers(&self) -> u32

pub fn set_navigation_layer_value(&mut self, layer_number: i32, value: bool)

Based on value, enables or disables the specified layer in the [member navigation_layers] bitmask, given a layer_number between 1 and 32.

pub fn get_navigation_layer_value(&self, layer_number: i32) -> bool

Returns whether or not the specified layer of the [member navigation_layers] bitmask is enabled, given a layer_number between 1 and 32.

pub fn set_pathfinding_algorithm( &mut self, pathfinding_algorithm: PathfindingAlgorithm, )

pub fn get_pathfinding_algorithm(&self) -> PathfindingAlgorithm

pub fn set_path_postprocessing( &mut self, path_postprocessing: PathPostProcessing, )

pub fn get_path_postprocessing(&self) -> PathPostProcessing

pub fn set_path_metadata_flags(&mut self, flags: PathMetadataFlags)

pub fn get_path_metadata_flags(&self) -> PathMetadataFlags

pub fn set_navigation_map(&mut self, navigation_map: Rid)

Sets the RID of the navigation map this NavigationAgent node should use and also updates the agent on the NavigationServer.

pub fn get_navigation_map(&self) -> Rid

Returns the RID of the navigation map for this NavigationAgent node. This function returns always the map set on the NavigationAgent node and not the map of the abstract agent on the NavigationServer. If the agent map is changed directly with the NavigationServer API the NavigationAgent node will not be aware of the map change. Use set_navigation_map to change the navigation map for the NavigationAgent and also update the agent on the NavigationServer.

pub fn set_target_position(&mut self, position: Vector2)

pub fn get_target_position(&self) -> Vector2

pub fn set_simplify_path(&mut self, enabled: bool)

pub fn get_simplify_path(&self) -> bool

pub fn set_simplify_epsilon(&mut self, epsilon: f32)

pub fn get_simplify_epsilon(&self) -> f32

pub fn set_path_return_max_length(&mut self, length: f32)

pub fn get_path_return_max_length(&self) -> f32

pub fn set_path_return_max_radius(&mut self, radius: f32)

pub fn get_path_return_max_radius(&self) -> f32

pub fn set_path_search_max_polygons(&mut self, max_polygons: i32)

pub fn get_path_search_max_polygons(&self) -> i32

pub fn set_path_search_max_distance(&mut self, distance: f32)

pub fn get_path_search_max_distance(&self) -> f32

pub fn get_path_length(&self) -> f32

Returns the length of the currently calculated path. The returned value is 0.0, if the path is still calculating or no calculation has been requested yet.

pub fn get_next_path_position(&mut self) -> Vector2

Returns the next position in global coordinates that can be moved to, making sure that there are no static objects in the way. If the agent does not have a navigation path, it will return the position of the agent’s parent. The use of this function once every physics frame is required to update the internal path logic of the NavigationAgent.

pub fn set_velocity_forced(&mut self, velocity: Vector2)

Replaces the internal velocity in the collision avoidance simulation with velocity. When an agent is teleported to a new position this function should be used in the same frame. If called frequently this function can get agents stuck.

pub fn set_velocity(&mut self, velocity: Vector2)

pub fn get_velocity(&self) -> Vector2

pub fn distance_to_target(&self) -> f32

Returns the distance to the target position, using the agent’s global position. The user must set [member target_position] in order for this to be accurate.

pub fn get_current_navigation_result( &self, ) -> Option<Gd<NavigationPathQueryResult2D>>

Returns the path query result for the path the agent is currently following.

pub fn get_current_navigation_path(&self) -> PackedArray<Vector2>

Returns this agent’s current path from start to finish in global coordinates. The path only updates when the target position is changed or the agent requires a repath. The path array is not intended to be used in direct path movement as the agent has its own internal path logic that would get corrupted by changing the path array manually. Use the intended get_next_path_position once every physics frame to receive the next path point for the agents movement as this function also updates the internal path logic.

pub fn get_current_navigation_path_index(&self) -> i32

Returns which index the agent is currently on in the navigation path’s PackedVector2Array.

pub fn is_target_reached(&self) -> bool

Returns true if the agent reached the target, i.e. the agent moved within [member target_desired_distance] of the [member target_position]. It may not always be possible to reach the target but it should always be possible to reach the final position. See get_final_position.

pub fn is_target_reachable(&self) -> bool

Returns true if get_final_position is within [member target_desired_distance] of the [member target_position].

pub fn is_navigation_finished(&self) -> bool

Returns true if the agent’s navigation has finished. If the target is reachable, navigation ends when the target is reached. If the target is unreachable, navigation ends when the last waypoint of the path is reached.

Note: While true prefer to stop calling update functions like get_next_path_position. This avoids jittering the standing agent due to calling repeated path updates.

pub fn get_final_position(&self) -> Vector2

Returns the reachable final position of the current navigation path in global coordinates. This position can change if the agent needs to update the navigation path which makes the agent emit the path_changed signal.

pub fn set_avoidance_layers(&mut self, layers: u32)

pub fn get_avoidance_layers(&self) -> u32

pub fn set_avoidance_mask(&mut self, mask: u32)

pub fn get_avoidance_mask(&self) -> u32

pub fn set_avoidance_layer_value(&mut self, layer_number: i32, value: bool)

Based on value, enables or disables the specified layer in the [member avoidance_layers] bitmask, given a layer_number between 1 and 32.

pub fn get_avoidance_layer_value(&self, layer_number: i32) -> bool

Returns whether or not the specified layer of the [member avoidance_layers] bitmask is enabled, given a layer_number between 1 and 32.

pub fn set_avoidance_mask_value(&mut self, mask_number: i32, value: bool)

Based on value, enables or disables the specified mask in the [member avoidance_mask] bitmask, given a mask_number between 1 and 32.

pub fn get_avoidance_mask_value(&self, mask_number: i32) -> bool

Returns whether or not the specified mask of the [member avoidance_mask] bitmask is enabled, given a mask_number between 1 and 32.

pub fn set_avoidance_priority(&mut self, priority: f32)

pub fn get_avoidance_priority(&self) -> f32

pub fn set_debug_enabled(&mut self, enabled: bool)

pub fn get_debug_enabled(&self) -> bool

pub fn set_debug_use_custom(&mut self, enabled: bool)

pub fn get_debug_use_custom(&self) -> bool

pub fn set_debug_path_custom_color(&mut self, color: Color)

pub fn get_debug_path_custom_color(&self) -> Color

pub fn set_debug_path_custom_point_size(&mut self, point_size: f32)

pub fn get_debug_path_custom_point_size(&self) -> f32

pub fn set_debug_path_custom_line_width(&mut self, line_width: f32)

pub fn get_debug_path_custom_line_width(&self) -> f32