godot::classes

Struct NavigationServer3D

pub struct NavigationServer3D { /* private fields */ }

Available on crate feature experimental-godot-api only.

Godot class NavigationServer3D.

Inherits Object.

Related symbols:

navigation_server_3d: sidecar module with related enum/flag types

SignalsOfNavigationServer3D: signal collection

See also Godot docs for NavigationServer3D.

Singleton

This class is a singleton. You can get the one instance using Singleton::singleton().

Final class

This class is final, meaning you cannot inherit from it, and it comes without I* interface trait. It is still possible that other Godot classes inherit from it, but that is limited to the engine itself.

Godot docs

NavigationServer3D is the server that handles navigation maps, regions and agents. It does not handle A* navigation from AStar3D.

Maps are divided into regions, which are composed of navigation meshes. Together, they define the navigable areas in the 3D world.

Note: Most NavigationServer3D changes take effect after the next physics frame and not immediately. This includes all changes made to maps, regions or agents by navigation-related nodes in the scene tree or made through scripts.

For two regions to be connected to each other, they must share a similar edge. An edge is considered connected to another if both of its two vertices are at a distance less than edge_connection_margin to the respective other edge’s vertex.

You may assign navigation layers to regions with region_set_navigation_layers, which then can be checked upon when requesting a path with map_get_path. This can be used to allow or deny certain areas for some objects.

To use the collision avoidance system, you may use agents. You can set an agent’s target velocity, then the servers will emit a callback with a modified velocity.

Note: The collision avoidance system ignores regions. Using the modified velocity directly may move an agent outside of the traversable area. This is a limitation of the collision avoidance system, any more complex situation may require the use of the physics engine.

This server keeps tracks of any call and executes them during the sync phase. This means that you can request any change to the map, using any thread, without worrying.

Implementations

impl NavigationServer3D

pub fn get_maps(&self) -> Array<Rid>

Returns all created navigation map RIDs on the NavigationServer. This returns both 2D and 3D created navigation maps as there is technically no distinction between them.

pub fn map_create(&mut self) -> Rid

Create a new map.

pub fn map_set_active(&mut self, map: Rid, active: bool)

Sets the map active.

pub fn map_is_active(&self, map: Rid) -> bool

Returns true if the map is active.

pub fn map_set_up(&mut self, map: Rid, up: Vector3)

Sets the map up direction.

pub fn map_get_up(&self, map: Rid) -> Vector3

Returns the map’s up direction.

pub fn map_set_cell_size(&mut self, map: Rid, cell_size: f32)

Sets the map cell size used to rasterize the navigation mesh vertices on the XZ plane. Must match with the cell size of the used navigation meshes.

pub fn map_get_cell_size(&self, map: Rid) -> f32

Returns the map cell size used to rasterize the navigation mesh vertices on the XZ plane.

pub fn map_set_cell_height(&mut self, map: Rid, cell_height: f32)

Sets the map cell height used to rasterize the navigation mesh vertices on the Y axis. Must match with the cell height of the used navigation meshes.

pub fn map_get_cell_height(&self, map: Rid) -> f32

Returns the map cell height used to rasterize the navigation mesh vertices on the Y axis.

pub fn map_set_merge_rasterizer_cell_scale(&mut self, map: Rid, scale: f32)

Set the map’s internal merge rasterizer cell scale used to control merging sensitivity.

pub fn map_get_merge_rasterizer_cell_scale(&self, map: Rid) -> f32

Returns map’s internal merge rasterizer cell scale.

pub fn map_set_use_edge_connections(&mut self, map: Rid, enabled: bool)

Set the navigation map edge connection use. If enabled is true, the navigation map allows navigation regions to use edge connections to connect with other navigation regions within proximity of the navigation map edge connection margin.

pub fn map_get_use_edge_connections(&self, map: Rid) -> bool

Returns true if the navigation map allows navigation regions to use edge connections to connect with other navigation regions within proximity of the navigation map edge connection margin.

pub fn map_set_edge_connection_margin(&mut self, map: Rid, margin: f32)

Set the map edge connection margin used to weld the compatible region edges.

pub fn map_get_edge_connection_margin(&self, map: Rid) -> f32

Returns the edge connection margin of the map. This distance is the minimum vertex distance needed to connect two edges from different regions.

pub fn map_set_link_connection_radius(&mut self, map: Rid, radius: f32)

Set the map’s link connection radius used to connect links to navigation polygons.

pub fn map_get_link_connection_radius(&self, map: Rid) -> f32

Returns the link connection radius of the map. This distance is the maximum range any link will search for navigation mesh polygons to connect to.

pub fn map_get_path( &mut self, map: Rid, origin: Vector3, destination: Vector3, optimize: bool, ) -> PackedArray<Vector3>

To set the default parameters, use map_get_path_ex and its builder methods. See the book for detailed usage instructions. Returns the navigation path to reach the destination from the origin. navigation_layers is a bitmask of all region navigation layers that are allowed to be in the path.

pub fn map_get_path_ex<'ex>( &'ex mut self, map: Rid, origin: Vector3, destination: Vector3, optimize: bool, ) -> ExMapGetPath<'ex>

Returns the navigation path to reach the destination from the origin. navigation_layers is a bitmask of all region navigation layers that are allowed to be in the path.

pub fn map_get_closest_point_to_segment( &self, map: Rid, start: Vector3, end: Vector3, ) -> Vector3

To set the default parameters, use map_get_closest_point_to_segment_ex and its builder methods. See the book for detailed usage instructions. Returns the navigation mesh surface point closest to the provided start and end segment on the navigation map.

If use_collision is true, a closest point test is only done when the segment intersects with the navigation mesh surface.

pub fn map_get_closest_point_to_segment_ex<'ex>( &'ex self, map: Rid, start: Vector3, end: Vector3, ) -> ExMapGetClosestPointToSegment<'ex>

Returns the navigation mesh surface point closest to the provided start and end segment on the navigation map.

If use_collision is true, a closest point test is only done when the segment intersects with the navigation mesh surface.

pub fn map_get_closest_point(&self, map: Rid, to_point: Vector3) -> Vector3

Returns the navigation mesh surface point closest to the provided to_point on the navigation map.

pub fn map_get_closest_point_normal( &self, map: Rid, to_point: Vector3, ) -> Vector3

Returns the navigation mesh surface normal closest to the provided to_point on the navigation map.

pub fn map_get_closest_point_owner(&self, map: Rid, to_point: Vector3) -> Rid

Returns the owner region RID for the navigation mesh surface point closest to the provided to_point on the navigation map.

pub fn map_get_links(&self, map: Rid) -> Array<Rid>

Returns all navigation link RIDs that are currently assigned to the requested navigation map.

pub fn map_get_regions(&self, map: Rid) -> Array<Rid>

Returns all navigation regions RIDs that are currently assigned to the requested navigation map.

pub fn map_get_agents(&self, map: Rid) -> Array<Rid>

Returns all navigation agents RIDs that are currently assigned to the requested navigation map.

pub fn map_get_obstacles(&self, map: Rid) -> Array<Rid>

Returns all navigation obstacle RIDs that are currently assigned to the requested navigation map.

pub fn map_force_update(&mut self, map: Rid)

This function immediately forces synchronization of the specified navigation map RID. By default navigation maps are only synchronized at the end of each physics frame. This function can be used to immediately (re)calculate all the navigation meshes and region connections of the navigation map. This makes it possible to query a navigation path for a changed map immediately and in the same frame (multiple times if needed).

Due to technical restrictions the current NavigationServer command queue will be flushed. This means all already queued update commands for this physics frame will be executed, even those intended for other maps, regions and agents not part of the specified map. The expensive computation of the navigation meshes and region connections of a map will only be done for the specified map. Other maps will receive the normal synchronization at the end of the physics frame. Should the specified map receive changes after the forced update it will update again as well when the other maps receive their update.

Avoidance processing and dispatch of the safe_velocity signals is unaffected by this function and continues to happen for all maps and agents at the end of the physics frame.

Note: With great power comes great responsibility. This function should only be used by users that really know what they are doing and have a good reason for it. Forcing an immediate update of a navigation map requires locking the NavigationServer and flushing the entire NavigationServer command queue. Not only can this severely impact the performance of a game but it can also introduce bugs if used inappropriately without much foresight.

pub fn map_get_iteration_id(&self, map: Rid) -> u32

Returns the current iteration id of the navigation map. Every time the navigation map changes and synchronizes the iteration id increases. An iteration id of 0 means the navigation map has never synchronized.

Note: The iteration id will wrap back to 1 after reaching its range limit.

pub fn map_set_use_async_iterations(&mut self, map: Rid, enabled: bool)

If enabled is true the map synchronization uses an async process that runs on a background thread.

pub fn map_get_use_async_iterations(&self, map: Rid) -> bool

Returns true if the map synchronization uses an async process that runs on a background thread.

pub fn map_get_random_point( &self, map: Rid, navigation_layers: u32, uniformly: bool, ) -> Vector3

Returns a random position picked from all map region polygons with matching navigation_layers.

If uniformly is true, all map regions, polygons, and faces are weighted by their surface area (slower).

If uniformly is false, just a random region and a random polygon are picked (faster).

pub fn query_path( &mut self, parameters: impl AsArg<Option<Gd<NavigationPathQueryParameters3D>>>, result: impl AsArg<Option<Gd<NavigationPathQueryResult3D>>>, )

To set the default parameters, use query_path_ex and its builder methods. See the book for detailed usage instructions. Queries a path in a given navigation map. Start and target position and other parameters are defined through NavigationPathQueryParameters3D. Updates the provided NavigationPathQueryResult3D result object with the path among other results requested by the query. After the process is finished the optional callback will be called.

pub fn query_path_ex<'ex>( &'ex mut self, parameters: impl AsArg<Option<Gd<NavigationPathQueryParameters3D>>> + 'ex, result: impl AsArg<Option<Gd<NavigationPathQueryResult3D>>> + 'ex, ) -> ExQueryPath<'ex>

Queries a path in a given navigation map. Start and target position and other parameters are defined through NavigationPathQueryParameters3D. Updates the provided NavigationPathQueryResult3D result object with the path among other results requested by the query. After the process is finished the optional callback will be called.

pub fn region_create(&mut self) -> Rid

Creates a new region.

pub fn region_get_iteration_id(&self, region: Rid) -> u32

Returns the current iteration ID of the navigation region. Every time the navigation region changes and synchronizes, the iteration ID increases. An iteration ID of 0 means the navigation region has never synchronized.

Note: The iteration ID will wrap around to 1 after reaching its range limit.

pub fn region_set_use_async_iterations(&mut self, region: Rid, enabled: bool)

If enabled is true the region uses an async synchronization process that runs on a background thread.

pub fn region_get_use_async_iterations(&self, region: Rid) -> bool

Returns true if the region uses an async synchronization process that runs on a background thread.

pub fn region_set_enabled(&mut self, region: Rid, enabled: bool)

If enabled is true, the specified region will contribute to its current navigation map.

pub fn region_get_enabled(&self, region: Rid) -> bool

Returns true if the specified region is enabled.

pub fn region_set_use_edge_connections(&mut self, region: Rid, enabled: bool)

If enabled is true, the navigation region will use edge connections to connect with other navigation regions within proximity of the navigation map edge connection margin.

pub fn region_get_use_edge_connections(&self, region: Rid) -> bool

Returns true if the navigation region is set to use edge connections to connect with other navigation regions within proximity of the navigation map edge connection margin.

pub fn region_set_enter_cost(&mut self, region: Rid, enter_cost: f32)

Sets the enter_cost for this region.

pub fn region_get_enter_cost(&self, region: Rid) -> f32

Returns the enter cost of this region.

pub fn region_set_travel_cost(&mut self, region: Rid, travel_cost: f32)

Sets the travel_cost for this region.

pub fn region_get_travel_cost(&self, region: Rid) -> f32

Returns the travel cost of this region.

pub fn region_set_owner_id(&mut self, region: Rid, owner_id: u64)

Set the ObjectID of the object which manages this region.

pub fn region_get_owner_id(&self, region: Rid) -> u64

Returns the ObjectID of the object which manages this region.

pub fn region_owns_point(&self, region: Rid, point: Vector3) -> bool

Returns true if the provided point in world space is currently owned by the provided navigation region. Owned in this context means that one of the region’s navigation mesh polygon faces has a possible position at the closest distance to this point compared to all other navigation meshes from other navigation regions that are also registered on the navigation map of the provided region.

If multiple navigation meshes have positions at equal distance the navigation region whose polygons are processed first wins the ownership. Polygons are processed in the same order that navigation regions were registered on the NavigationServer.

Note: If navigation meshes from different navigation regions overlap (which should be avoided in general) the result might not be what is expected.

pub fn region_set_map(&mut self, region: Rid, map: Rid)

Sets the map for the region.

pub fn region_get_map(&self, region: Rid) -> Rid

Returns the navigation map RID the requested region is currently assigned to.

pub fn region_set_navigation_layers( &mut self, region: Rid, navigation_layers: u32, )

Set the region’s navigation layers. This allows selecting regions from a path request (when using map_get_path).

pub fn region_get_navigation_layers(&self, region: Rid) -> u32

Returns the region’s navigation layers.

pub fn region_set_transform(&mut self, region: Rid, transform: Transform3D)

Sets the global transformation for the region.

pub fn region_get_transform(&self, region: Rid) -> Transform3D

Returns the global transformation of this region.

pub fn region_set_navigation_mesh( &mut self, region: Rid, navigation_mesh: impl AsArg<Option<Gd<NavigationMesh>>>, )

Sets the navigation mesh for the region.

pub fn region_bake_navigation_mesh( &mut self, navigation_mesh: impl AsArg<Option<Gd<NavigationMesh>>>, root_node: impl AsArg<Option<Gd<Node>>>, )

Bakes the navigation_mesh with bake source geometry collected starting from the root_node.

pub fn region_get_connections_count(&self, region: Rid) -> i32

Returns how many connections this region has with other regions in the map.

pub fn region_get_connection_pathway_start( &self, region: Rid, connection: i32, ) -> Vector3

Returns the starting point of a connection door. connection is an index between 0 and the return value of region_get_connections_count.

pub fn region_get_connection_pathway_end( &self, region: Rid, connection: i32, ) -> Vector3

Returns the ending point of a connection door. connection is an index between 0 and the return value of region_get_connections_count.

pub fn region_get_closest_point_to_segment( &self, region: Rid, start: Vector3, end: Vector3, ) -> Vector3

To set the default parameters, use region_get_closest_point_to_segment_ex and its builder methods. See the book for detailed usage instructions. Returns the navigation mesh surface point closest to the provided start and end segment on the navigation region.

If use_collision is true, a closest point test is only done when the segment intersects with the navigation mesh surface.

pub fn region_get_closest_point_to_segment_ex<'ex>( &'ex self, region: Rid, start: Vector3, end: Vector3, ) -> ExRegionGetClosestPointToSegment<'ex>

Returns the navigation mesh surface point closest to the provided start and end segment on the navigation region.

If use_collision is true, a closest point test is only done when the segment intersects with the navigation mesh surface.

pub fn region_get_closest_point( &self, region: Rid, to_point: Vector3, ) -> Vector3

Returns the navigation mesh surface point closest to the provided to_point on the navigation region.

pub fn region_get_closest_point_normal( &self, region: Rid, to_point: Vector3, ) -> Vector3

Returns the navigation mesh surface normal closest to the provided to_point on the navigation region.

pub fn region_get_random_point( &self, region: Rid, navigation_layers: u32, uniformly: bool, ) -> Vector3

Returns a random position picked from all region polygons with matching navigation_layers.

If uniformly is true, all region polygons and faces are weighted by their surface area (slower).

If uniformly is false, just a random polygon and face is picked (faster).

pub fn region_get_bounds(&self, region: Rid) -> Aabb

Returns the axis-aligned bounding box for the region’s transformed navigation mesh.

pub fn link_create(&mut self) -> Rid

Create a new link between two positions on a map.

pub fn link_get_iteration_id(&self, link: Rid) -> u32

Returns the current iteration ID of the navigation link. Every time the navigation link changes and synchronizes, the iteration ID increases. An iteration ID of 0 means the navigation link has never synchronized.

Note: The iteration ID will wrap around to 1 after reaching its range limit.

pub fn link_set_map(&mut self, link: Rid, map: Rid)

Sets the navigation map RID for the link.

pub fn link_get_map(&self, link: Rid) -> Rid

Returns the navigation map RID the requested link is currently assigned to.

pub fn link_set_enabled(&mut self, link: Rid, enabled: bool)

If enabled is true, the specified link will contribute to its current navigation map.

pub fn link_get_enabled(&self, link: Rid) -> bool

Returns true if the specified link is enabled.

pub fn link_set_bidirectional(&mut self, link: Rid, bidirectional: bool)

Sets whether this link can be travelled in both directions.

pub fn link_is_bidirectional(&self, link: Rid) -> bool

Returns whether this link can be travelled in both directions.

pub fn link_set_navigation_layers(&mut self, link: Rid, navigation_layers: u32)

Set the links’s navigation layers. This allows selecting links from a path request (when using map_get_path).

pub fn link_get_navigation_layers(&self, link: Rid) -> u32

Returns the navigation layers for this link.

pub fn link_set_start_position(&mut self, link: Rid, position: Vector3)

Sets the entry position for this link.

pub fn link_get_start_position(&self, link: Rid) -> Vector3

Returns the starting position of this link.

pub fn link_set_end_position(&mut self, link: Rid, position: Vector3)

Sets the exit position for the link.

pub fn link_get_end_position(&self, link: Rid) -> Vector3

Returns the ending position of this link.

pub fn link_set_enter_cost(&mut self, link: Rid, enter_cost: f32)

Sets the enter_cost for this link.

pub fn link_get_enter_cost(&self, link: Rid) -> f32

Returns the enter cost of this link.

pub fn link_set_travel_cost(&mut self, link: Rid, travel_cost: f32)

Sets the travel_cost for this link.

pub fn link_get_travel_cost(&self, link: Rid) -> f32

Returns the travel cost of this link.

pub fn link_set_owner_id(&mut self, link: Rid, owner_id: u64)

Set the ObjectID of the object which manages this link.

pub fn link_get_owner_id(&self, link: Rid) -> u64

Returns the ObjectID of the object which manages this link.

pub fn agent_create(&mut self) -> Rid

Creates the agent.

pub fn agent_set_avoidance_enabled(&mut self, agent: Rid, enabled: bool)

If enabled is true, the provided agent calculates avoidance.

pub fn agent_get_avoidance_enabled(&self, agent: Rid) -> bool

Returns true if the provided agent has avoidance enabled.

pub fn agent_set_use_3d_avoidance(&mut self, agent: Rid, enabled: bool)

Sets if the agent uses the 2D avoidance or the 3D avoidance while avoidance is enabled.

If true the agent calculates avoidance velocities in 3D for the xyz-axis, e.g. for games that take place in air, underwater or space. The 3D using agent only avoids other 3D avoidance using agent’s. The 3D using agent only reacts to radius based avoidance obstacles. The 3D using agent ignores any vertices based obstacles. The 3D using agent only avoids other 3D using agent’s.

If false the agent calculates avoidance velocities in 2D along the xz-axis ignoring the y-axis. The 2D using agent only avoids other 2D avoidance using agent’s. The 2D using agent reacts to radius avoidance obstacles. The 2D using agent reacts to vertices based avoidance obstacles. The 2D using agent only avoids other 2D using agent’s. 2D using agents will ignore other 2D using agents or obstacles that are below their current position or above their current position including the agents height in 2D avoidance.

pub fn agent_get_use_3d_avoidance(&self, agent: Rid) -> bool

Returns true if the provided agent uses avoidance in 3D space Vector3(x,y,z) instead of horizontal 2D Vector2(x,y) / Vector3(x,0.0,z).

pub fn agent_set_map(&mut self, agent: Rid, map: Rid)

Puts the agent in the map.

pub fn agent_get_map(&self, agent: Rid) -> Rid

Returns the navigation map RID the requested agent is currently assigned to.

pub fn agent_set_paused(&mut self, agent: Rid, paused: bool)

If paused is true the specified agent will not be processed. For example, it will not calculate avoidance velocities or receive avoidance callbacks.

pub fn agent_get_paused(&self, agent: Rid) -> bool

Returns true if the specified agent is paused.

pub fn agent_set_neighbor_distance(&mut self, agent: Rid, distance: f32)

Sets the maximum distance to other agents this agent takes into account in the navigation. The larger this number, the longer the running time of the simulation. If the number is too low, the simulation will not be safe.

pub fn agent_get_neighbor_distance(&self, agent: Rid) -> f32

Returns the maximum distance to other agents the specified agent takes into account in the navigation.

pub fn agent_set_max_neighbors(&mut self, agent: Rid, count: i32)

Sets the maximum number of other agents the agent takes into account in the navigation. The larger this number, the longer the running time of the simulation. If the number is too low, the simulation will not be safe.

pub fn agent_get_max_neighbors(&self, agent: Rid) -> i32

Returns the maximum number of other agents the specified agent takes into account in the navigation.

pub fn agent_set_time_horizon_agents(&mut self, agent: Rid, time_horizon: f32)

The minimal amount of time for which the agent’s velocities that are computed by the simulation are safe with respect to other agents. The larger this number, the sooner this agent will respond to the presence of other agents, but the less freedom this agent has in choosing its velocities. A too high value will slow down agents movement considerably. Must be positive.

pub fn agent_get_time_horizon_agents(&self, agent: Rid) -> f32

Returns the minimal amount of time for which the specified agent’s velocities that are computed by the simulation are safe with respect to other agents.

pub fn agent_set_time_horizon_obstacles( &mut self, agent: Rid, time_horizon: f32, )

The minimal amount of time for which the agent’s velocities that are computed by the simulation are safe with respect to static avoidance obstacles. The larger this number, the sooner this agent will respond to the presence of static avoidance obstacles, but the less freedom this agent has in choosing its velocities. A too high value will slow down agents movement considerably. Must be positive.

pub fn agent_get_time_horizon_obstacles(&self, agent: Rid) -> f32

Returns the minimal amount of time for which the specified agent’s velocities that are computed by the simulation are safe with respect to static avoidance obstacles.

pub fn agent_set_radius(&mut self, agent: Rid, radius: f32)

Sets the radius of the agent.

pub fn agent_get_radius(&self, agent: Rid) -> f32

Returns the radius of the specified agent.

pub fn agent_set_height(&mut self, agent: Rid, height: f32)

Updates the provided agent height.

pub fn agent_get_height(&self, agent: Rid) -> f32

Returns the height of the specified agent.

pub fn agent_set_max_speed(&mut self, agent: Rid, max_speed: f32)

Sets the maximum speed of the agent. Must be positive.

pub fn agent_get_max_speed(&self, agent: Rid) -> f32

Returns the maximum speed of the specified agent.

pub fn agent_set_velocity_forced(&mut self, agent: Rid, velocity: Vector3)

Replaces the internal velocity in the collision avoidance simulation with velocity for the specified agent. When an agent is teleported to a new position this function should be used in the same frame. If called frequently this function can get agents stuck.

pub fn agent_set_velocity(&mut self, agent: Rid, velocity: Vector3)

Sets velocity as the new wanted velocity for the specified agent. The avoidance simulation will try to fulfill this velocity if possible but will modify it to avoid collision with other agent’s and obstacles. When an agent is teleported to a new position use agent_set_velocity_forced as well to reset the internal simulation velocity.

pub fn agent_get_velocity(&self, agent: Rid) -> Vector3

Returns the velocity of the specified agent.

pub fn agent_set_position(&mut self, agent: Rid, position: Vector3)

Sets the position of the agent in world space.

pub fn agent_get_position(&self, agent: Rid) -> Vector3

Returns the position of the specified agent in world space.

pub fn agent_is_map_changed(&self, agent: Rid) -> bool

Returns true if the map got changed the previous frame.

pub fn agent_set_avoidance_callback(&mut self, agent: Rid, callback: &Callable)

Sets the callback Callable that gets called after each avoidance processing step for the agent. The calculated safe_velocity will be dispatched with a signal to the object just before the physics calculations.

Note: Created callbacks are always processed independently of the SceneTree state as long as the agent is on a navigation map and not freed. To disable the dispatch of a callback from an agent use agent_set_avoidance_callback again with an empty Callable.

pub fn agent_has_avoidance_callback(&self, agent: Rid) -> bool

Return true if the specified agent has an avoidance callback.

pub fn agent_set_avoidance_layers(&mut self, agent: Rid, layers: u32)

Set the agent’s avoidance_layers bitmask.

pub fn agent_get_avoidance_layers(&self, agent: Rid) -> u32

Returns the avoidance_layers bitmask of the specified agent.

pub fn agent_set_avoidance_mask(&mut self, agent: Rid, mask: u32)

Set the agent’s avoidance_mask bitmask.

pub fn agent_get_avoidance_mask(&self, agent: Rid) -> u32

Returns the avoidance_mask bitmask of the specified agent.

pub fn agent_set_avoidance_priority(&mut self, agent: Rid, priority: f32)

Set the agent’s avoidance_priority with a priority between 0.0 (lowest priority) to 1.0 (highest priority).

The specified agent does not adjust the velocity for other agents that would match the avoidance_mask but have a lower avoidance_priority. This in turn makes the other agents with lower priority adjust their velocities even more to avoid collision with this agent.

pub fn agent_get_avoidance_priority(&self, agent: Rid) -> f32

Returns the avoidance_priority of the specified agent.

pub fn obstacle_create(&mut self) -> Rid

Creates a new obstacle.

pub fn obstacle_set_avoidance_enabled(&mut self, obstacle: Rid, enabled: bool)

If enabled is true, the provided obstacle affects avoidance using agents.

pub fn obstacle_get_avoidance_enabled(&self, obstacle: Rid) -> bool

Returns true if the provided obstacle has avoidance enabled.

pub fn obstacle_set_use_3d_avoidance(&mut self, obstacle: Rid, enabled: bool)

Sets if the obstacle uses the 2D avoidance or the 3D avoidance while avoidance is enabled.

pub fn obstacle_get_use_3d_avoidance(&self, obstacle: Rid) -> bool

Returns true if the provided obstacle uses avoidance in 3D space Vector3(x,y,z) instead of horizontal 2D Vector2(x,y) / Vector3(x,0.0,z).

pub fn obstacle_set_map(&mut self, obstacle: Rid, map: Rid)

Assigns the obstacle to a navigation map.

pub fn obstacle_get_map(&self, obstacle: Rid) -> Rid

Returns the navigation map RID the requested obstacle is currently assigned to.

pub fn obstacle_set_paused(&mut self, obstacle: Rid, paused: bool)

If paused is true the specified obstacle will not be processed. For example, it will no longer affect avoidance velocities.

pub fn obstacle_get_paused(&self, obstacle: Rid) -> bool

Returns true if the specified obstacle is paused.

pub fn obstacle_set_radius(&mut self, obstacle: Rid, radius: f32)

Sets the radius of the dynamic obstacle.

pub fn obstacle_get_radius(&self, obstacle: Rid) -> f32

Returns the radius of the specified dynamic obstacle.

pub fn obstacle_set_height(&mut self, obstacle: Rid, height: f32)

Sets the height for the obstacle. In 3D agents will ignore obstacles that are above or below them while using 2D avoidance.

pub fn obstacle_get_height(&self, obstacle: Rid) -> f32

Returns the height of the specified obstacle.

pub fn obstacle_set_velocity(&mut self, obstacle: Rid, velocity: Vector3)

Sets velocity of the dynamic obstacle. Allows other agents to better predict the movement of the dynamic obstacle. Only works in combination with the radius of the obstacle.

pub fn obstacle_get_velocity(&self, obstacle: Rid) -> Vector3

Returns the velocity of the specified dynamic obstacle.

pub fn obstacle_set_position(&mut self, obstacle: Rid, position: Vector3)

Updates the position in world space for the obstacle.

pub fn obstacle_get_position(&self, obstacle: Rid) -> Vector3

Returns the position of the specified obstacle in world space.

pub fn obstacle_set_vertices( &mut self, obstacle: Rid, vertices: &PackedArray<Vector3>, )

Sets the outline vertices for the obstacle. If the vertices are winded in clockwise order agents will be pushed in by the obstacle, else they will be pushed out.

pub fn obstacle_get_vertices(&self, obstacle: Rid) -> PackedArray<Vector3>

Returns the outline vertices for the specified obstacle.

pub fn obstacle_set_avoidance_layers(&mut self, obstacle: Rid, layers: u32)

Set the obstacles’s avoidance_layers bitmask.

pub fn obstacle_get_avoidance_layers(&self, obstacle: Rid) -> u32

Returns the avoidance_layers bitmask of the specified obstacle.

pub fn parse_source_geometry_data( &mut self, navigation_mesh: impl AsArg<Option<Gd<NavigationMesh>>>, source_geometry_data: impl AsArg<Option<Gd<NavigationMeshSourceGeometryData3D>>>, root_node: impl AsArg<Option<Gd<Node>>>, )

To set the default parameters, use parse_source_geometry_data_ex and its builder methods. See the book for detailed usage instructions. Parses the SceneTree for source geometry according to the properties of navigation_mesh. Updates the provided source_geometry_data resource with the resulting data. The resource can then be used to bake a navigation mesh with bake_from_source_geometry_data. After the process is finished the optional callback will be called.

Note: This function needs to run on the main thread or with a deferred call as the SceneTree is not thread-safe.

Performance: While convenient, reading data arrays from Mesh resources can affect the frame rate negatively. The data needs to be received from the GPU, stalling the RenderingServer in the process. For performance prefer the use of e.g. collision shapes or creating the data arrays entirely in code.

pub fn parse_source_geometry_data_ex<'ex>( &'ex mut self, navigation_mesh: impl AsArg<Option<Gd<NavigationMesh>>> + 'ex, source_geometry_data: impl AsArg<Option<Gd<NavigationMeshSourceGeometryData3D>>> + 'ex, root_node: impl AsArg<Option<Gd<Node>>> + 'ex, ) -> ExParseSourceGeometryData<'ex>

Parses the SceneTree for source geometry according to the properties of navigation_mesh. Updates the provided source_geometry_data resource with the resulting data. The resource can then be used to bake a navigation mesh with bake_from_source_geometry_data. After the process is finished the optional callback will be called.

Note: This function needs to run on the main thread or with a deferred call as the SceneTree is not thread-safe.

Performance: While convenient, reading data arrays from Mesh resources can affect the frame rate negatively. The data needs to be received from the GPU, stalling the RenderingServer in the process. For performance prefer the use of e.g. collision shapes or creating the data arrays entirely in code.

pub fn bake_from_source_geometry_data( &mut self, navigation_mesh: impl AsArg<Option<Gd<NavigationMesh>>>, source_geometry_data: impl AsArg<Option<Gd<NavigationMeshSourceGeometryData3D>>>, )

To set the default parameters, use bake_from_source_geometry_data_ex and its builder methods. See the book for detailed usage instructions. Bakes the provided navigation_mesh with the data from the provided source_geometry_data. After the process is finished the optional callback will be called.

pub fn bake_from_source_geometry_data_ex<'ex>( &'ex mut self, navigation_mesh: impl AsArg<Option<Gd<NavigationMesh>>> + 'ex, source_geometry_data: impl AsArg<Option<Gd<NavigationMeshSourceGeometryData3D>>> + 'ex, ) -> ExBakeFromSourceGeometryData<'ex>

Bakes the provided navigation_mesh with the data from the provided source_geometry_data. After the process is finished the optional callback will be called.

pub fn bake_from_source_geometry_data_async( &mut self, navigation_mesh: impl AsArg<Option<Gd<NavigationMesh>>>, source_geometry_data: impl AsArg<Option<Gd<NavigationMeshSourceGeometryData3D>>>, )

To set the default parameters, use bake_from_source_geometry_data_async_ex and its builder methods. See the book for detailed usage instructions. Bakes the provided navigation_mesh with the data from the provided source_geometry_data as an async task running on a background thread. After the process is finished the optional callback will be called.

pub fn bake_from_source_geometry_data_async_ex<'ex>( &'ex mut self, navigation_mesh: impl AsArg<Option<Gd<NavigationMesh>>> + 'ex, source_geometry_data: impl AsArg<Option<Gd<NavigationMeshSourceGeometryData3D>>> + 'ex, ) -> ExBakeFromSourceGeometryDataAsync<'ex>

Bakes the provided navigation_mesh with the data from the provided source_geometry_data as an async task running on a background thread. After the process is finished the optional callback will be called.

pub fn is_baking_navigation_mesh( &self, navigation_mesh: impl AsArg<Option<Gd<NavigationMesh>>>, ) -> bool

Returns true when the provided navigation mesh is being baked on a background thread.

pub fn source_geometry_parser_create(&mut self) -> Rid

Creates a new source geometry parser. If a Callable is set for the parser with source_geometry_parser_set_callback the callback will be called for every single node that gets parsed whenever parse_source_geometry_data is used.

pub fn source_geometry_parser_set_callback( &mut self, parser: Rid, callback: &Callable, )

Sets the callback Callable for the specific source geometry parser. The Callable will receive a call with the following parameters:

navigation_mesh - The NavigationMesh reference used to define the parse settings. Do NOT edit or add directly to the navigation mesh.

source_geometry_data - The NavigationMeshSourceGeometryData3D reference. Add custom source geometry for navigation mesh baking to this object.

node - The Node that is parsed.

pub fn simplify_path( &mut self, path: &PackedArray<Vector3>, epsilon: f32, ) -> PackedArray<Vector3>

Returns a simplified version of path with less critical path points removed. The simplification amount is in worlds units and controlled by epsilon. The simplification uses a variant of Ramer-Douglas-Peucker algorithm for curve point decimation.

Path simplification can be helpful to mitigate various path following issues that can arise with certain agent types and script behaviors. E.g. “steering” agents or avoidance in “open fields”.

pub fn free_rid(&mut self, rid: Rid)

Destroys the given RID.

pub fn set_active(&mut self, active: bool)

Control activation of this server.

pub fn set_debug_enabled(&mut self, enabled: bool)

If true enables debug mode on the NavigationServer.

pub fn get_debug_enabled(&self) -> bool

Returns true when the NavigationServer has debug enabled.

pub fn get_process_info(&self, process_info: ProcessInfo) -> i32

Returns information about the current state of the NavigationServer.