godot::classes

Struct PhysicsDirectBodyState2DExtension

pub struct PhysicsDirectBodyState2DExtension { /* private fields */ }

Godot class PhysicsDirectBodyState2DExtension.

Inherits PhysicsDirectBodyState2D.

Related symbols:

IPhysicsDirectBodyState2DExtension: virtual methods

See also Godot docs for PhysicsDirectBodyState2DExtension.

Construction

This class is manually managed. You can create a new instance using PhysicsDirectBodyState2DExtension::new_alloc().

Do not forget to call free() or hand over ownership to Godot.

Godot docs

This class extends PhysicsDirectBodyState2D by providing additional virtual methods that can be overridden. When these methods are overridden, they will be called instead of the internal methods of the physics server.

Intended for use with GDExtension to create custom implementations of PhysicsDirectBodyState2D.