godot::classes

Struct PhysicsServer3DExtension

pub struct PhysicsServer3DExtension { /* private fields */ }

Godot class PhysicsServer3DExtension.

Inherits PhysicsServer3D.

Related symbols:

IPhysicsServer3DExtension: virtual methods

See also Godot docs for PhysicsServer3DExtension.

Construction

This class is manually managed. You can create a new instance using PhysicsServer3DExtension::new_alloc().

Do not forget to call free() or hand over ownership to Godot.

Godot docs

This class extends PhysicsServer3D by providing additional virtual methods that can be overridden. When these methods are overridden, they will be called instead of the internal methods of the physics server.

Intended for use with GDExtension to create custom implementations of PhysicsServer3D.

Implementations

impl PhysicsServer3DExtension

pub fn body_test_motion_is_excluding_body(&self, body: Rid) -> bool

pub fn body_test_motion_is_excluding_object(&self, object: u64) -> bool