godot::classes

Struct Popup

pub struct Popup { /* private fields */ }

Godot class Popup.

Inherits Window.

Related symbols:

popup: sidecar module with related enum/flag types

IPopup: virtual methods

SignalsOfPopup: signal collection

See also Godot docs for Popup.

Construction

This class is manually managed. You can create a new instance using Popup::new_alloc().

Do not forget to call free() or hand over ownership to Godot.

Godot docs

Popup is a base class for contextual windows and panels with fixed position. It’s a modal by default (see [member Window.popup_window]) and provides methods for implementing custom popup behavior.

Note: Popup is invisible by default. To make it visible, call one of the popup_* methods from Window on the node, such as popup_centered_clamped.