godot::classes

Struct TreeItem

#[repr(C)]
pub struct TreeItem { /* private fields */ }

Godot class TreeItem.

Inherits Object.

Related symbols:

tree_item: sidecar module with related enum/flag types

ITreeItem: virtual methods

See also Godot docs for TreeItem.

Not instantiable

This class cannot be constructed. Obtain Gd<TreeItem> instances via Godot APIs.

Implementations

impl TreeItem

pub fn set_cell_mode(&mut self, column: i32, mode: TreeCellMode)

pub fn get_cell_mode(&self, column: i32) -> TreeCellMode

pub fn set_edit_multiline(&mut self, column: i32, multiline: bool)

pub fn is_edit_multiline(&self, column: i32) -> bool

pub fn set_checked(&mut self, column: i32, checked: bool)

pub fn set_indeterminate(&mut self, column: i32, indeterminate: bool)

pub fn is_checked(&self, column: i32) -> bool

pub fn is_indeterminate(&self, column: i32) -> bool

pub fn propagate_check(&mut self, column: i32)

To set the default parameters, use Self::propagate_check_ex and its builder methods. See the book for detailed usage instructions.

pub fn propagate_check_ex<'a>(&'a mut self, column: i32) -> ExPropagateCheck<'a>

pub fn set_text(&mut self, column: i32, text: impl AsArg<GString>)

pub fn get_text(&self, column: i32) -> GString

pub fn set_text_direction(&mut self, column: i32, direction: TextDirection)

pub fn get_text_direction(&self, column: i32) -> TextDirection

pub fn set_autowrap_mode(&mut self, column: i32, autowrap_mode: AutowrapMode)

pub fn get_autowrap_mode(&self, column: i32) -> AutowrapMode

pub fn set_text_overrun_behavior( &mut self, column: i32, overrun_behavior: OverrunBehavior, )

pub fn get_text_overrun_behavior(&self, column: i32) -> OverrunBehavior

pub fn set_structured_text_bidi_override( &mut self, column: i32, parser: StructuredTextParser, )

pub fn get_structured_text_bidi_override( &self, column: i32, ) -> StructuredTextParser

pub fn set_structured_text_bidi_override_options( &mut self, column: i32, args: &Array<Variant>, )

pub fn get_structured_text_bidi_override_options( &self, column: i32, ) -> Array<Variant>

pub fn set_language(&mut self, column: i32, language: impl AsArg<GString>)

pub fn get_language(&self, column: i32) -> GString

pub fn set_suffix(&mut self, column: i32, text: impl AsArg<GString>)

pub fn get_suffix(&self, column: i32) -> GString

pub fn set_icon(&mut self, column: i32, texture: impl AsObjectArg<Texture2D>)

pub fn get_icon(&self, column: i32) -> Option<Gd<Texture2D>>

pub fn set_icon_region(&mut self, column: i32, region: Rect2)

pub fn get_icon_region(&self, column: i32) -> Rect2

pub fn set_icon_max_width(&mut self, column: i32, width: i32)

pub fn get_icon_max_width(&self, column: i32) -> i32

pub fn set_icon_modulate(&mut self, column: i32, modulate: Color)

pub fn get_icon_modulate(&self, column: i32) -> Color

pub fn set_range(&mut self, column: i32, value: f64)

pub fn get_range(&self, column: i32) -> f64

pub fn set_range_config(&mut self, column: i32, min: f64, max: f64, step: f64)

To set the default parameters, use Self::set_range_config_ex and its builder methods. See the book for detailed usage instructions.

pub fn set_range_config_ex<'a>( &'a mut self, column: i32, min: f64, max: f64, step: f64, ) -> ExSetRangeConfig<'a>

pub fn get_range_config(&mut self, column: i32) -> Dictionary

pub fn set_metadata(&mut self, column: i32, meta: &Variant)

pub fn get_metadata(&self, column: i32) -> Variant

pub fn set_custom_draw( &mut self, column: i32, object: impl AsObjectArg<Object>, callback: impl AsArg<StringName>, )

pub fn set_custom_draw_callback(&mut self, column: i32, callback: &Callable)

pub fn get_custom_draw_callback(&self, column: i32) -> Callable

pub fn set_collapsed(&mut self, enable: bool)

pub fn is_collapsed(&mut self) -> bool

pub fn set_collapsed_recursive(&mut self, enable: bool)

pub fn is_any_collapsed(&mut self) -> bool

To set the default parameters, use Self::is_any_collapsed_ex and its builder methods. See the book for detailed usage instructions.

pub fn is_any_collapsed_ex<'a>(&'a mut self) -> ExIsAnyCollapsed<'a>

pub fn set_visible(&mut self, enable: bool)

pub fn is_visible(&mut self) -> bool

pub fn is_visible_in_tree(&self) -> bool

pub fn uncollapse_tree(&mut self)

pub fn set_custom_minimum_height(&mut self, height: i32)

pub fn get_custom_minimum_height(&self) -> i32

pub fn set_selectable(&mut self, column: i32, selectable: bool)

pub fn is_selectable(&self, column: i32) -> bool

pub fn is_selected(&mut self, column: i32) -> bool

pub fn select(&mut self, column: i32)

pub fn deselect(&mut self, column: i32)

pub fn set_editable(&mut self, column: i32, enabled: bool)

pub fn is_editable(&mut self, column: i32) -> bool

pub fn set_custom_color(&mut self, column: i32, color: Color)

pub fn get_custom_color(&self, column: i32) -> Color

pub fn clear_custom_color(&mut self, column: i32)

pub fn set_custom_font(&mut self, column: i32, font: impl AsObjectArg<Font>)

pub fn get_custom_font(&self, column: i32) -> Option<Gd<Font>>

pub fn set_custom_font_size(&mut self, column: i32, font_size: i32)

pub fn get_custom_font_size(&self, column: i32) -> i32

pub fn set_custom_bg_color(&mut self, column: i32, color: Color)

To set the default parameters, use Self::set_custom_bg_color_ex and its builder methods. See the book for detailed usage instructions.

pub fn set_custom_bg_color_ex<'a>( &'a mut self, column: i32, color: Color, ) -> ExSetCustomBgColor<'a>

pub fn clear_custom_bg_color(&mut self, column: i32)

pub fn get_custom_bg_color(&self, column: i32) -> Color

pub fn set_custom_as_button(&mut self, column: i32, enable: bool)

pub fn is_custom_set_as_button(&self, column: i32) -> bool

pub fn add_button(&mut self, column: i32, button: impl AsObjectArg<Texture2D>)

To set the default parameters, use Self::add_button_ex and its builder methods. See the book for detailed usage instructions.

pub fn add_button_ex<'a>( &'a mut self, column: i32, button: impl AsObjectArg<Texture2D>, ) -> ExAddButton<'a>

pub fn get_button_count(&self, column: i32) -> i32

pub fn get_button_tooltip_text(&self, column: i32, button_index: i32) -> GString

pub fn get_button_id(&self, column: i32, button_index: i32) -> i32

pub fn get_button_by_id(&self, column: i32, id: i32) -> i32

pub fn get_button_color(&self, column: i32, id: i32) -> Color

pub fn get_button( &self, column: i32, button_index: i32, ) -> Option<Gd<Texture2D>>

pub fn set_button_tooltip_text( &mut self, column: i32, button_index: i32, tooltip: impl AsArg<GString>, )

pub fn set_button( &mut self, column: i32, button_index: i32, button: impl AsObjectArg<Texture2D>, )

pub fn erase_button(&mut self, column: i32, button_index: i32)

pub fn set_button_disabled( &mut self, column: i32, button_index: i32, disabled: bool, )

pub fn set_button_color(&mut self, column: i32, button_index: i32, color: Color)

pub fn is_button_disabled(&self, column: i32, button_index: i32) -> bool

pub fn set_tooltip_text(&mut self, column: i32, tooltip: impl AsArg<GString>)

pub fn get_tooltip_text(&self, column: i32) -> GString

pub fn set_text_alignment( &mut self, column: i32, text_alignment: HorizontalAlignment, )

pub fn get_text_alignment(&self, column: i32) -> HorizontalAlignment

pub fn set_expand_right(&mut self, column: i32, enable: bool)

pub fn get_expand_right(&self, column: i32) -> bool

pub fn set_disable_folding(&mut self, disable: bool)

pub fn is_folding_disabled(&self) -> bool

pub fn create_child(&mut self) -> Option<Gd<TreeItem>>

To set the default parameters, use Self::create_child_ex and its builder methods. See the book for detailed usage instructions.

pub fn create_child_ex<'a>(&'a mut self) -> ExCreateChild<'a>

pub fn add_child(&mut self, child: impl AsObjectArg<TreeItem>)

pub fn remove_child(&mut self, child: impl AsObjectArg<TreeItem>)

pub fn get_tree(&self) -> Option<Gd<Tree>>

pub fn get_next(&self) -> Option<Gd<TreeItem>>

pub fn get_prev(&mut self) -> Option<Gd<TreeItem>>

pub fn get_parent(&self) -> Option<Gd<TreeItem>>

pub fn get_first_child(&self) -> Option<Gd<TreeItem>>

pub fn get_next_in_tree(&mut self) -> Option<Gd<TreeItem>>

To set the default parameters, use Self::get_next_in_tree_ex and its builder methods. See the book for detailed usage instructions.

pub fn get_next_in_tree_ex<'a>(&'a mut self) -> ExGetNextInTree<'a>

pub fn get_prev_in_tree(&mut self) -> Option<Gd<TreeItem>>

To set the default parameters, use Self::get_prev_in_tree_ex and its builder methods. See the book for detailed usage instructions.

pub fn get_prev_in_tree_ex<'a>(&'a mut self) -> ExGetPrevInTree<'a>

pub fn get_next_visible(&mut self) -> Option<Gd<TreeItem>>

To set the default parameters, use Self::get_next_visible_ex and its builder methods. See the book for detailed usage instructions.

pub fn get_next_visible_ex<'a>(&'a mut self) -> ExGetNextVisible<'a>

pub fn get_prev_visible(&mut self) -> Option<Gd<TreeItem>>

To set the default parameters, use Self::get_prev_visible_ex and its builder methods. See the book for detailed usage instructions.

pub fn get_prev_visible_ex<'a>(&'a mut self) -> ExGetPrevVisible<'a>

pub fn get_child(&mut self, index: i32) -> Option<Gd<TreeItem>>

pub fn get_child_count(&mut self) -> i32

pub fn get_children(&mut self) -> Array<Gd<TreeItem>>

pub fn get_index(&mut self) -> i32

pub fn move_before(&mut self, item: impl AsObjectArg<TreeItem>)

pub fn move_after(&mut self, item: impl AsObjectArg<TreeItem>)

pub fn call_recursive( &mut self, method: impl AsArg<StringName>, varargs: &[Variant], )

Panics

This is a varcall method, meaning parameters and return values are passed as Variant. It can detect call failures and will panic in such a case.

pub fn try_call_recursive( &mut self, method: impl AsArg<StringName>, varargs: &[Variant], ) -> Result<(), CallError>

Return type

This is a varcall method, meaning parameters and return values are passed as Variant. It can detect call failures and will return Err in such a case.