godot::classes

Struct VideoStreamPlayer

pub struct VideoStreamPlayer { /* private fields */ }

Godot class VideoStreamPlayer.

Inherits Control.

Related symbols:

video_stream_player: sidecar module with related enum/flag types

IVideoStreamPlayer: virtual methods

SignalsOfVideoStreamPlayer: signal collection

See also Godot docs for VideoStreamPlayer.

Construction

This class is manually managed. You can create a new instance using VideoStreamPlayer::new_alloc().

Do not forget to call free() or hand over ownership to Godot.

Godot docs

A control used for playback of VideoStream resources.

Supported video formats are Ogg Theora (.ogv, VideoStreamTheora) and any format exposed via a GDExtension plugin.

Warning: On Web, video playback will perform poorly due to missing architecture-specific assembly optimizations.

Implementations

impl VideoStreamPlayer

pub fn set_stream(&mut self, stream: impl AsArg<Option<Gd<VideoStream>>>)

pub fn get_stream(&self) -> Option<Gd<VideoStream>>

pub fn play(&mut self)

Starts the video playback from the beginning. If the video is paused, this will not unpause the video.

pub fn stop(&mut self)

Stops the video playback and sets the stream position to 0.

Note: Although the stream position will be set to 0, the first frame of the video stream won’t become the current frame.

pub fn is_playing(&self) -> bool

Returns true if the video is playing.

Note: The video is still considered playing if paused during playback.

pub fn set_paused(&mut self, paused: bool)

pub fn is_paused(&self) -> bool

pub fn set_loop(&mut self, loop_: bool)

pub fn has_loop(&self) -> bool

pub fn set_volume(&mut self, volume: f32)

pub fn get_volume(&self) -> f32

pub fn set_volume_db(&mut self, db: f32)

pub fn get_volume_db(&self) -> f32

pub fn set_speed_scale(&mut self, speed_scale: f32)

pub fn get_speed_scale(&self) -> f32

pub fn set_audio_track(&mut self, track: i32)

pub fn get_audio_track(&self) -> i32

pub fn get_stream_name(&self) -> GString

Returns the video stream’s name, or "<No Stream>" if no video stream is assigned.

pub fn get_stream_length(&self) -> f64

The length of the current stream, in seconds.

pub fn set_stream_position(&mut self, position: f64)

pub fn get_stream_position(&self) -> f64

pub fn set_autoplay(&mut self, enabled: bool)

pub fn has_autoplay(&self) -> bool

pub fn set_expand(&mut self, enable: bool)

pub fn has_expand(&self) -> bool

pub fn set_buffering_msec(&mut self, msec: i32)

pub fn get_buffering_msec(&self) -> i32

pub fn set_bus(&mut self, bus: impl AsArg<StringName>)

pub fn get_bus(&self) -> StringName

pub fn get_video_texture(&self) -> Option<Gd<Texture2D>>

Returns the current frame as a Texture2D.