godot::classes
Trait IXrInterfaceExtension

pub trait IXrInterfaceExtension: GodotClass<Base = XrInterfaceExtension> + You_forgot_the_attribute__godot_api {

    // Provided methods
    fn init(base: Base<Self::Base>) -> Self { ... }
    fn to_string(&self) -> GString { ... }
    fn on_notification(&mut self, what: ObjectNotification) { ... }
    fn get_property(&self, property: StringName) -> Option<Variant> { ... }
    fn set_property(&mut self, property: StringName, value: Variant) -> bool { ... }
    fn get_property_list(&mut self) -> Vec<PropertyInfo> { ... }
    fn property_get_revert(&self, property: StringName) -> Option<Variant> { ... }
    fn get_name(&self) -> StringName { ... }
    fn get_capabilities(&self) -> u32 { ... }
    fn is_initialized(&self) -> bool { ... }
    fn initialize(&mut self) -> bool { ... }
    fn uninitialize(&mut self) { ... }
    fn get_system_info(&self) -> Dictionary { ... }
    fn supports_play_area_mode(&self, mode: PlayAreaMode) -> bool { ... }
    fn get_play_area_mode(&self) -> PlayAreaMode { ... }
    fn set_play_area_mode(&self, mode: PlayAreaMode) -> bool { ... }
    fn get_play_area(&self) -> PackedVector3Array { ... }
    fn get_render_target_size(&mut self) -> Vector2 { ... }
    fn get_view_count(&mut self) -> u32 { ... }
    fn get_camera_transform(&mut self) -> Transform3D { ... }
    fn get_transform_for_view(
        &mut self,
        view: u32,
        cam_transform: Transform3D,
    ) -> Transform3D { ... }
    fn get_projection_for_view(
        &mut self,
        view: u32,
        aspect: f64,
        z_near: f64,
        z_far: f64,
    ) -> PackedFloat64Array { ... }
    fn get_vrs_texture(&mut self) -> Rid { ... }
    fn process(&mut self) { ... }
    fn pre_render(&mut self) { ... }
    fn pre_draw_viewport(&mut self, render_target: Rid) -> bool { ... }
    fn post_draw_viewport(&mut self, render_target: Rid, screen_rect: Rect2) { ... }
    fn end_frame(&mut self) { ... }
    fn get_suggested_tracker_names(&self) -> PackedStringArray { ... }
    fn get_suggested_pose_names(
        &self,
        tracker_name: StringName,
    ) -> PackedStringArray { ... }
    fn get_tracking_status(&self) -> TrackingStatus { ... }
    fn trigger_haptic_pulse(
        &mut self,
        action_name: GString,
        tracker_name: StringName,
        frequency: f64,
        amplitude: f64,
        duration_sec: f64,
        delay_sec: f64,
    ) { ... }
    fn get_anchor_detection_is_enabled(&self) -> bool { ... }
    fn set_anchor_detection_is_enabled(&mut self, enabled: bool) { ... }
    fn get_camera_feed_id(&self) -> i32 { ... }
    fn get_color_texture(&mut self) -> Rid { ... }
    fn get_depth_texture(&mut self) -> Rid { ... }
    fn get_velocity_texture(&mut self) -> Rid { ... }

}

Virtual methods for class XrInterfaceExtension.

These methods represent constructors (init) or callbacks invoked by the engine.

See also Godot docs for XRInterfaceExtension methods.

Provided Methods

fn init(base: Base<Self::Base>) -> Self

Godot constructor, accepting an injected base object.

base refers to the base instance of the class, which can either be stored in a Base<T> field or discarded. This method returns a fully-constructed instance, which will then be moved into a Gd<T> pointer.

If the class has a #[class(init)] attribute, this method will be auto-generated and must not be overridden.

fn to_string(&self) -> GString

String representation of the Godot instance.

Override this method to define how the instance is represented as a string. Used by impl Display for Gd<T>, as well as str() and print() in GDScript.

fn on_notification(&mut self, what: ObjectNotification)

Called when the object receives a Godot notification.

The type of notification can be identified through what. The enum is designed to hold all possible NOTIFICATION_* constants that the current class can handle. However, this is not validated in Godot, so an enum variant Unknown exists to represent integers out of known constants (mistakes or future additions).

This method is named _notification in Godot, but on_notification in Rust. To send notifications, use the Object::notify method.

See also in Godot docs:

Object::_notification.

Notifications tutorial.

fn get_property(&self, property: StringName) -> Option<Variant>

Called whenever get() is called or Godot gets the value of a property.

Should return the given property’s value as Some(value), or None if the property should be handled normally.

See also in Godot docs:

Object::_get.

fn set_property(&mut self, property: StringName, value: Variant) -> bool

Called whenever Godot set() is called or Godot sets the value of a property.

Should set property to the given value and return true, or return false to indicate the property should be handled normally.

See also in Godot docs:

Object::_set.

fn get_property_list(&mut self) -> Vec<PropertyInfo>

Available on since_api="4.3" only.

Called whenever Godot get_property_list() is called, the returned vector here is appended to the existing list of properties.

This should mainly be used for advanced purposes, such as dynamically updating the property list in the editor.

See also in Godot docs:

Object::_get_property_list

fn property_get_revert(&self, property: StringName) -> Option<Variant>

Called by Godot to tell if a property has a custom revert or not.

Return None for no custom revert, and return Some(value) to specify the custom revert.

This is a combination of Godot’s Object::_property_get_revert and Object::_property_can_revert. This means that this function will usually be called twice by Godot to find the revert.

Note that this should be a pure function. That is, it should always return the same value for a property as long as self remains unchanged. Otherwise, this may lead to unexpected (safe) behavior.

fn get_name(&self) -> StringName

fn get_capabilities(&self) -> u32

fn is_initialized(&self) -> bool

fn initialize(&mut self) -> bool

fn uninitialize(&mut self)

fn get_system_info(&self) -> Dictionary

fn supports_play_area_mode(&self, mode: PlayAreaMode) -> bool

fn get_play_area_mode(&self) -> PlayAreaMode

fn set_play_area_mode(&self, mode: PlayAreaMode) -> bool

fn get_play_area(&self) -> PackedVector3Array

fn get_render_target_size(&mut self) -> Vector2

fn get_view_count(&mut self) -> u32

fn get_camera_transform(&mut self) -> Transform3D

fn get_transform_for_view( &mut self, view: u32, cam_transform: Transform3D, ) -> Transform3D

fn get_projection_for_view( &mut self, view: u32, aspect: f64, z_near: f64, z_far: f64, ) -> PackedFloat64Array

fn get_vrs_texture(&mut self) -> Rid

fn process(&mut self)

fn pre_render(&mut self)

fn pre_draw_viewport(&mut self, render_target: Rid) -> bool

fn post_draw_viewport(&mut self, render_target: Rid, screen_rect: Rect2)

fn end_frame(&mut self)

fn get_suggested_tracker_names(&self) -> PackedStringArray

fn get_suggested_pose_names( &self, tracker_name: StringName, ) -> PackedStringArray

fn get_tracking_status(&self) -> TrackingStatus

fn trigger_haptic_pulse( &mut self, action_name: GString, tracker_name: StringName, frequency: f64, amplitude: f64, duration_sec: f64, delay_sec: f64, )

fn get_anchor_detection_is_enabled(&self) -> bool

fn set_anchor_detection_is_enabled(&mut self, enabled: bool)

fn get_camera_feed_id(&self) -> i32

fn get_color_texture(&mut self) -> Rid

fn get_depth_texture(&mut self) -> Rid

fn get_velocity_texture(&mut self) -> Rid

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.