Module godot::engine

Maps the Godot class API to Rust.

This module contains the following symbols:

Classes: CanvasItem, etc.

Virtual traits: CanvasItemVirtual, etc.

Enum/flag modules: canvas_item, etc.

Noteworthy sub-modules are:

notify: all notification types, used when working with the virtual callback to handle lifecycle notifications.

global: global enums not belonging to a specific class.

utilities: utility methods that are global in Godot. Godot engine classes and methods.

Modules

a_star_grid_2d

Sidecar module for class AStarGrid2D.

aes_context

Sidecar module for class AesContext.

animation

Sidecar module for class Animation.

animation_node

Sidecar module for class AnimationNode.

animation_node_animation

Sidecar module for class AnimationNodeAnimation.

animation_node_blend_space1_d

Sidecar module for class AnimationNodeBlendSpace1D.

animation_node_blend_space_2d

Sidecar module for class AnimationNodeBlendSpace2D.

animation_node_one_shot

Sidecar module for class AnimationNodeOneShot.

animation_node_state_machine_transition

Sidecar module for class AnimationNodeStateMachineTransition.

animation_player

Sidecar module for class AnimationPlayer.

animation_tree

Sidecar module for class AnimationTree.

area_2d

Sidecar module for class Area2D.

area_3d

Sidecar module for class Area3D.

aspect_ratio_container

Sidecar module for class AspectRatioContainer.

audio_effect_distortion

Sidecar module for class AudioEffectDistortion.

audio_effect_filter

Sidecar module for class AudioEffectFilter.

audio_effect_pitch_shift

Sidecar module for class AudioEffectPitchShift.

audio_effect_spectrum_analyzer

Sidecar module for class AudioEffectSpectrumAnalyzer.

audio_effect_spectrum_analyzer_instance

Sidecar module for class AudioEffectSpectrumAnalyzerInstance.

audio_server

Sidecar module for class AudioServer.

audio_stream_player

Sidecar module for class AudioStreamPlayer.

audio_stream_player_3d

Sidecar module for class AudioStreamPlayer3D.

audio_stream_randomizer

Sidecar module for class AudioStreamRandomizer.

audio_stream_wav

Sidecar module for class AudioStreamWav.

back_buffer_copy

Sidecar module for class BackBufferCopy.

base_button

Sidecar module for class BaseButton.

base_material_3d

Sidecar module for class BaseMaterial3D.

box_container

Sidecar module for class BoxContainer.

camera_2d

Sidecar module for class Camera2D.

camera_3d

Sidecar module for class Camera3D.

camera_feed

Sidecar module for class CameraFeed.

camera_server

Sidecar module for class CameraServer.

canvas_item

Sidecar module for class CanvasItem.

canvas_item_material

Sidecar module for class CanvasItemMaterial.

character_body_2d

Sidecar module for class CharacterBody2D.

character_body_3d

Sidecar module for class CharacterBody3D.

class_methods

class_virtuals

classes

code_edit

Sidecar module for class CodeEdit.

collision_object_2d

Sidecar module for class CollisionObject2D.

collision_object_3d

Sidecar module for class CollisionObject3D.

collision_polygon_2d

Sidecar module for class CollisionPolygon2D.

color_picker

Sidecar module for class ColorPicker.

cone_twist_joint_3d

Sidecar module for class ConeTwistJoint3D.

control

Sidecar module for class Control.

cpu_particles_2d

Sidecar module for class CpuParticles2D.

cpu_particles_3d

Sidecar module for class CpuParticles3D.

csg_polygon_3d

Sidecar module for class CsgPolygon3D.

csg_shape_3d

Sidecar module for class CsgShape3D.

curve

Sidecar module for class Curve.

curve_texture

Sidecar module for class CurveTexture.

decal

Sidecar module for class Decal.

directional_light_3d

Sidecar module for class DirectionalLight3D.

display_server

Sidecar module for class DisplayServer.

e_net_connection

Sidecar module for class ENetConnection.

e_net_packet_peer

Sidecar module for class ENetPacketPeer.

editor_feature_profile

Sidecar module for class EditorFeatureProfile.

editor_file_dialog

Sidecar module for class EditorFileDialog.

editor_plugin

Sidecar module for class EditorPlugin.

editor_scene_post_import_plugin

Sidecar module for class EditorScenePostImportPlugin.

editor_undo_redo_manager

Sidecar module for class EditorUndoRedoManager.

editor_vcs_interface

Sidecar module for class EditorVcsInterface.

environment

Sidecar module for class Environment.

fast_noise_lite

Sidecar module for class FastNoiseLite.

file_access

Sidecar module for class FileAccess.

file_dialog

Sidecar module for class FileDialog.

flow_container

Sidecar module for class FlowContainer.

gdextension

Sidecar module for class GDExtension.

gdextension_manager

Sidecar module for class GDExtensionManager.

generic6_dof_joint_3d

Sidecar module for class Generic6DofJoint3D.

geometry_2d

Sidecar module for class Geometry2D.

geometry_instance_3d

Sidecar module for class GeometryInstance3D.

global

Global enums and constants.

gpu_particles_2d

Sidecar module for class GpuParticles2D.

gpu_particles_3d

Sidecar module for class GpuParticles3D.

gpu_particles_collision_height_field_3d

Sidecar module for class GpuParticlesCollisionHeightField3D.

gpu_particles_collision_sdf_3d

Sidecar module for class GpuParticlesCollisionSdf3d.

gradient

Sidecar module for class Gradient.

gradient_texture_2d

Sidecar module for class GradientTexture2D.

graph_edit

Sidecar module for class GraphEdit.

graph_node

Sidecar module for class GraphNode.

hashing_context

Sidecar module for class HashingContext.

hinge_joint_3d

Sidecar module for class HingeJoint3D.

http_client

Sidecar module for class HttpClient.

http_request

Sidecar module for class HttpRequest.

image

Sidecar module for class Image.

image_format_loader

Sidecar module for class ImageFormatLoader.

input

Sidecar module for class Input.

ip

Sidecar module for class Ip.

item_list

Sidecar module for class ItemList.

json_rpc

Sidecar module for class JsonRpc.

label_3d

Sidecar module for class Label3D.

light_2d

Sidecar module for class Light2D.

light_3d

Sidecar module for class Light3D.

lightmap_gi

Sidecar module for class LightmapGi.

line_2d

Sidecar module for class Line2D.

line_edit

Sidecar module for class LineEdit.

link_button

Sidecar module for class LinkButton.

mesh

Sidecar module for class Mesh.

multi_mesh

Sidecar module for class MultiMesh.

multiplayer_api

Sidecar module for class MultiplayerApi.

multiplayer_peer

Sidecar module for class MultiplayerPeer.

multiplayer_synchronizer

Sidecar module for class MultiplayerSynchronizer.

navigation_mesh

Sidecar module for class NavigationMesh.

navigation_path_query_parameters_2d

Sidecar module for class NavigationPathQueryParameters2D.

navigation_path_query_parameters_3d

Sidecar module for class NavigationPathQueryParameters3D.

navigation_path_query_result_2d

Sidecar module for class NavigationPathQueryResult2D.

navigation_path_query_result_3d

Sidecar module for class NavigationPathQueryResult3D.

navigation_server_3d

Sidecar module for class NavigationServer3D.

nine_patch_rect

Sidecar module for class NinePatchRect.

node

Sidecar module for class Node.

node_3d

Sidecar module for class Node3D.

notify

object

Sidecar module for class Object.

occluder_polygon_2d

Sidecar module for class OccluderPolygon2D.

omni_light_3d

Sidecar module for class OmniLight3D.

open_xr_action

Sidecar module for class OpenXrAction.

open_xr_hand

Sidecar module for class OpenXrHand.

os

Sidecar module for class Os.

packed_scene

Sidecar module for class PackedScene.

packet_peer_dtls

Sidecar module for class PacketPeerDtls.

particle_process_material

Sidecar module for class ParticleProcessMaterial.

path_follow_3d

Sidecar module for class PathFollow3D.

performance

Sidecar module for class Performance.

physical_bone_3d

Sidecar module for class PhysicalBone3D.

physics_server_2d

Sidecar module for class PhysicsServer2D.

physics_server_3d

Sidecar module for class PhysicsServer3D.

pin_joint_3d

Sidecar module for class PinJoint3D.

plane_mesh

Sidecar module for class PlaneMesh.

portable_compressed_texture_2d

Sidecar module for class PortableCompressedTexture2D.

progress_bar

Sidecar module for class ProgressBar.

reflection_probe

Sidecar module for class ReflectionProbe.

rendering_device

Sidecar module for class RenderingDevice.

rendering_server

Sidecar module for class RenderingServer.

resource_format_loader

Sidecar module for class ResourceFormatLoader.

resource_importer

Sidecar module for class ResourceImporter.

resource_loader

Sidecar module for class ResourceLoader.

resource_saver

Sidecar module for class ResourceSaver.

ribbon_trail_mesh

Sidecar module for class RibbonTrailMesh.

rich_text_label

Sidecar module for class RichTextLabel.

rigid_body_2d

Sidecar module for class RigidBody2D.

rigid_body_3d

Sidecar module for class RigidBody3D.

scene_state

Sidecar module for class SceneState.

scene_tree

Sidecar module for class SceneTree.

script_language_extension

Sidecar module for class ScriptLanguageExtension.

scroll_container

Sidecar module for class ScrollContainer.

shader

Sidecar module for class Shader.

skeleton_profile

Sidecar module for class SkeletonProfile.

sky

Sidecar module for class Sky.

slider_joint_3d

Sidecar module for class SliderJoint3D.

soft_body_3d

Sidecar module for class SoftBody3D.

split_container

Sidecar module for class SplitContainer.

sprite_base_3d

Sidecar module for class SpriteBase3D.

stream_peer_tcp

Sidecar module for class StreamPeerTcp.

stream_peer_tls

Sidecar module for class StreamPeerTls.

style_box_texture

Sidecar module for class StyleBoxTexture.

sub_viewport

Sidecar module for class SubViewport.

surface_tool

Sidecar module for class SurfaceTool.

tab_bar

Sidecar module for class TabBar.

text_edit

Sidecar module for class TextEdit.

text_server

Sidecar module for class TextServer.

texture_button

Sidecar module for class TextureButton.

texture_layered

Sidecar module for class TextureLayered.

texture_progress_bar

Sidecar module for class TextureProgressBar.

texture_rect

Sidecar module for class TextureRect.

theme

Sidecar module for class Theme.

tile_map

Sidecar module for class TileMap.

tile_set

Sidecar module for class TileSet.

time

Sidecar module for class Time.

timer

Sidecar module for class Timer.

touch_screen_button

Sidecar module for class TouchScreenButton.

tree

Sidecar module for class Tree.

tree_item

Sidecar module for class TreeItem.

tween

Sidecar module for class Tween.

undo_redo

Sidecar module for class UndoRedo.

upnp

Sidecar module for class Upnp.

upnp_device

Sidecar module for class UpnpDevice.

utilities

Global utility functions.

viewport

Sidecar module for class Viewport.

visible_on_screen_enabler_2d

Sidecar module for class VisibleOnScreenEnabler2D.

visible_on_screen_enabler_3d

Sidecar module for class VisibleOnScreenEnabler3D.

visual_shader

Sidecar module for class VisualShader.

visual_shader_node

Sidecar module for class VisualShaderNode.

visual_shader_node_billboard

Sidecar module for class VisualShaderNodeBillboard.

visual_shader_node_clamp

Sidecar module for class VisualShaderNodeClamp.

visual_shader_node_color_func

Sidecar module for class VisualShaderNodeColorFunc.

visual_shader_node_color_op

Sidecar module for class VisualShaderNodeColorOp.

visual_shader_node_compare

Sidecar module for class VisualShaderNodeCompare.

visual_shader_node_cubemap

Sidecar module for class VisualShaderNodeCubemap.

visual_shader_node_derivative_func

Sidecar module for class VisualShaderNodeDerivativeFunc.

visual_shader_node_float_func

Sidecar module for class VisualShaderNodeFloatFunc.

visual_shader_node_float_op

Sidecar module for class VisualShaderNodeFloatOp.

visual_shader_node_float_parameter

Sidecar module for class VisualShaderNodeFloatParameter.

visual_shader_node_int_func

Sidecar module for class VisualShaderNodeIntFunc.

visual_shader_node_int_op

Sidecar module for class VisualShaderNodeIntOp.

visual_shader_node_int_parameter

Sidecar module for class VisualShaderNodeIntParameter.

visual_shader_node_is

Sidecar module for class VisualShaderNodeIs.

visual_shader_node_mix

Sidecar module for class VisualShaderNodeMix.

visual_shader_node_multiply_add

Sidecar module for class VisualShaderNodeMultiplyAdd.

visual_shader_node_parameter

Sidecar module for class VisualShaderNodeParameter.

visual_shader_node_particle_accelerator

Sidecar module for class VisualShaderNodeParticleAccelerator.

visual_shader_node_particle_emit

Sidecar module for class VisualShaderNodeParticleEmit.

visual_shader_node_particle_randomness

Sidecar module for class VisualShaderNodeParticleRandomness.

visual_shader_node_sample_3d

Sidecar module for class VisualShaderNodeSample3D.

visual_shader_node_smooth_step

Sidecar module for class VisualShaderNodeSmoothStep.

visual_shader_node_step

Sidecar module for class VisualShaderNodeStep.

visual_shader_node_switch

Sidecar module for class VisualShaderNodeSwitch.

visual_shader_node_texture

Sidecar module for class VisualShaderNodeTexture.

visual_shader_node_texture_parameter

Sidecar module for class VisualShaderNodeTextureParameter.

visual_shader_node_transform_func

Sidecar module for class VisualShaderNodeTransformFunc.

visual_shader_node_transform_op

Sidecar module for class VisualShaderNodeTransformOp.

visual_shader_node_transform_vec_mult

Sidecar module for class VisualShaderNodeTransformVecMult.

visual_shader_node_u_int_func

Sidecar module for class VisualShaderNodeUIntFunc.

visual_shader_node_u_int_op

Sidecar module for class VisualShaderNodeUIntOp.

visual_shader_node_uv_func

Sidecar module for class VisualShaderNodeUvFunc.

visual_shader_node_vector_base

Sidecar module for class VisualShaderNodeVectorBase.

visual_shader_node_vector_func

Sidecar module for class VisualShaderNodeVectorFunc.

visual_shader_node_vector_op

Sidecar module for class VisualShaderNodeVectorOp.

voxel_gi

Sidecar module for class VoxelGi.

web_rtc_data_channel

Sidecar module for class WebRtcDataChannel.

web_rtc_peer_connection

Sidecar module for class WebRtcPeerConnection.

web_socket_peer

Sidecar module for class WebSocketPeer.

web_xr_interface

Sidecar module for class WebXrInterface.

window

Sidecar module for class Window.

xml_parser

Sidecar module for class XmlParser.

xr_interface

Sidecar module for class XrInterface.

xr_pose

Sidecar module for class XrPose.

xr_positional_tracker

Sidecar module for class XrPositionalTracker.

xr_server

Sidecar module for class XrServer.

zip_packer

Sidecar module for class ZipPacker.

Traits

NodeExt

Extension trait with convenience functions for the node tree.

Functions

load

Loads a resource from the filesystem located at path, panicking on error.

try_load

Loads a resource from the filesystem located at path.