Struct godot::engine::classes::DisplayServer

pub struct DisplayServer { /* private fields */ }

Godot class DisplayServer.

Inherits Object.

Related symbols:

display_server: sidecar module with related enum/flag types

DisplayServerVirtual: virtual methods

See also Godot docs for DisplayServer.

Implementations

impl DisplayServer

pub fn singleton() -> Gd<DisplayServer>

pub fn notify(&mut self, what: ObjectNotification)

⚠️ Sends a Godot notification to all classes inherited by the object.

Triggers calls to on_notification(), and depending on the notification, also to Godot’s lifecycle callbacks such as ready().

Starts from the highest ancestor (the Object class) and goes down the hierarchy. See also Godot docs for Object::notification().

Panics

If you call this method on a user-defined object while holding a GdRef or GdMut guard on the instance, you will encounter a panic. The reason is that the receiving virtual method on_notification() acquires a GdMut lock dynamically, which must be exclusive.

pub fn notify_reversed(&mut self, what: ObjectNotification)

⚠️ Like Self::notify(), but starts at the most-derived class and goes up the hierarchy.

See docs of that method, including the panics.

pub const SCREEN_PRIMARY: i32 = -2i32

pub const SCREEN_OF_MAIN_WINDOW: i32 = -1i32

pub const MAIN_WINDOW_ID: i32 = 0i32

pub const INVALID_WINDOW_ID: i32 = -1i32