Enum godot::engine::notify::NodeNotification

#[repr(i32)]
pub enum NodeNotification {

    EnterTree = 10,
    ExitTree = 11,
    MovedInParent = 12,
    Ready = 13,
    Paused = 14,
    Unpaused = 15,
    PhysicsProcess = 16,
    Process = 17,
    Parented = 18,
    Unparented = 19,
    SceneInstantiated = 20,
    DragBegin = 21,
    DragEnd = 22,
    PathRenamed = 23,
    ChildOrderChanged = 24,
    InternalProcess = 25,
    InternalPhysicsProcess = 26,
    PostEnterTree = 27,
    Disabled = 28,
    Enabled = 29,
    ResetPhysicsInterpolation = 2_001,
    EditorPreSave = 9_001,
    EditorPostSave = 9_002,
    WmMouseEnter = 1_002,
    WmMouseExit = 1_003,
    WmWindowFocusIn = 1_004,
    WmWindowFocusOut = 1_005,
    WmCloseRequest = 1_006,
    WmGoBackRequest = 1_007,
    WmSizeChanged = 1_008,
    WmDpiChange = 1_009,
    VpMouseEnter = 1_010,
    VpMouseExit = 1_011,
    OsMemoryWarning = 2_009,
    TranslationChanged = 2_010,
    WmAbout = 2_011,
    Crash = 2_012,
    OsImeUpdate = 2_013,
    ApplicationResumed = 2_014,
    ApplicationPaused = 2_015,
    ApplicationFocusIn = 2_016,
    ApplicationFocusOut = 2_017,
    TextServerChanged = 2_018,
    Postinitialize = 0,
    Predelete = 1,
    ExtensionReloaded = 2,
    Unknown(i32),

}

Notification type for class Node.

Makes it easier to keep an overview all possible notification variants for a given class, including notifications defined in base classes.

Variants

EnterTree = 10

ExitTree = 11

MovedInParent = 12

Ready = 13

Paused = 14

Unpaused = 15

PhysicsProcess = 16

Process = 17

Parented = 18

Unparented = 19

SceneInstantiated = 20

DragBegin = 21

DragEnd = 22

PathRenamed = 23

ChildOrderChanged = 24

InternalProcess = 25

InternalPhysicsProcess = 26

PostEnterTree = 27

Disabled = 28

Enabled = 29

ResetPhysicsInterpolation = 2_001

EditorPreSave = 9_001

EditorPostSave = 9_002

WmMouseEnter = 1_002

WmMouseExit = 1_003

WmWindowFocusIn = 1_004

WmWindowFocusOut = 1_005

WmCloseRequest = 1_006

WmGoBackRequest = 1_007

WmSizeChanged = 1_008

WmDpiChange = 1_009

VpMouseEnter = 1_010

VpMouseExit = 1_011

OsMemoryWarning = 2_009

TranslationChanged = 2_010

WmAbout = 2_011

Crash = 2_012

OsImeUpdate = 2_013

ApplicationResumed = 2_014

ApplicationPaused = 2_015

ApplicationFocusIn = 2_016

ApplicationFocusOut = 2_017

TextServerChanged = 2_018

Postinitialize = 0

Predelete = 1

ExtensionReloaded = 2

Unknown(i32)

Since Godot represents notifications as integers, it’s always possible that a notification outside the known types is received. For example, the user can manually issue notifications through Object::notify().