Enum godot::engine::notify::WindowNotification

pub enum WindowNotification {

    VisibilityChangedOrNodeRecacheRequested,
    ThemeChanged,
    EnterTree,
    ExitTree,
    MovedInParent,
    Ready,
    Paused,
    Unpaused,
    PhysicsProcess,
    Process,
    Parented,
    Unparented,
    SceneInstantiated,
    DragBegin,
    DragEnd,
    PathRenamed,
    InternalProcess,
    InternalPhysicsProcess,
    PostEnterTree,
    Disabled,
    Enabled,
    EditorPreSave,
    EditorPostSave,
    WmMouseEnter,
    WmMouseExit,
    WmWindowFocusIn,
    WmWindowFocusOut,
    WmCloseRequest,
    WmGoBackRequest,
    WmSizeChanged,
    WmDpiChange,
    VpMouseEnter,
    VpMouseExit,
    OsMemoryWarning,
    TranslationChanged,
    WmAbout,
    Crash,
    OsImeUpdate,
    ApplicationResumed,
    ApplicationPaused,
    ApplicationFocusIn,
    ApplicationFocusOut,
    TextServerChanged,
    Postinitialize,
    Predelete,
    Unknown(i32),

}

Notification type for class Window.

Makes it easier to keep an overview all possible notification variants for a given class, including notifications defined in base classes.

Variants

VisibilityChangedOrNodeRecacheRequested

ThemeChanged

EnterTree

ExitTree

MovedInParent

Ready

Paused

Unpaused

PhysicsProcess

Process

Parented

Unparented

SceneInstantiated

DragBegin

DragEnd

PathRenamed

InternalProcess

InternalPhysicsProcess

PostEnterTree

Disabled

Enabled

EditorPreSave

EditorPostSave

WmMouseEnter

WmMouseExit

WmWindowFocusIn

WmWindowFocusOut

WmCloseRequest

WmGoBackRequest

WmSizeChanged

WmDpiChange

VpMouseEnter

VpMouseExit

OsMemoryWarning

TranslationChanged

WmAbout

Crash

OsImeUpdate

ApplicationResumed

ApplicationPaused

ApplicationFocusIn

ApplicationFocusOut

TextServerChanged

Postinitialize

Predelete

Unknown(i32)

Since Godot represents notifications as integers, it’s always possible that a notification outside the known types is received. For example, the user can manually issue notifications through Object.notification().