Struct godot::engine::AudioStreamPlayer2D

pub struct AudioStreamPlayer2D { /* private fields */ }

Godot class AudioStreamPlayer2D.

Inherits Node2D.

Related symbols:

AudioStreamPlayer2DVirtual: virtual methods

See also Godot docs for AudioStreamPlayer2D.

Implementations

impl AudioStreamPlayer2D

pub fn new_alloc() -> Gd<AudioStreamPlayer2D>

pub fn notify(&mut self, what: CanvasItemNotification)

⚠️ Sends a Godot notification to all classes inherited by the object.

Triggers calls to on_notification(), and depending on the notification, also to Godot’s lifecycle callbacks such as ready().

Starts from the highest ancestor (the Object class) and goes down the hierarchy. See also Godot docs for Object::notification().

Panics

If you call this method on a user-defined object while holding a GdRef or GdMut guard on the instance, you will encounter a panic. The reason is that the receiving virtual method on_notification() acquires a GdMut lock dynamically, which must be exclusive.

pub fn notify_reversed(&mut self, what: CanvasItemNotification)

⚠️ Like Self::notify(), but starts at the most-derived class and goes up the hierarchy.

See docs of that method, including the panics.

pub fn set_stream(&mut self, stream: Gd<AudioStream>)

pub fn get_stream(&self) -> Option<Gd<AudioStream>>

pub fn set_volume_db(&mut self, volume_db: f64)

pub fn get_volume_db(&self) -> f64

pub fn set_pitch_scale(&mut self, pitch_scale: f64)

pub fn get_pitch_scale(&self) -> f64

pub fn play(&mut self, from_position: f64)

pub fn seek(&mut self, to_position: f64)

pub fn stop(&mut self)

pub fn is_playing(&self) -> bool

pub fn get_playback_position(&mut self) -> f64

pub fn set_bus(&mut self, bus: StringName)

pub fn get_bus(&self) -> StringName

pub fn set_autoplay(&mut self, enable: bool)

pub fn is_autoplay_enabled(&mut self) -> bool

pub fn set_max_distance(&mut self, pixels: f64)

pub fn get_max_distance(&self) -> f64

pub fn set_attenuation(&mut self, curve: f64)

pub fn get_attenuation(&self) -> f64

pub fn set_area_mask(&mut self, mask: i64)

pub fn get_area_mask(&self) -> i64

pub fn set_stream_paused(&mut self, pause: bool)

pub fn get_stream_paused(&self) -> bool

pub fn set_max_polyphony(&mut self, max_polyphony: i64)

pub fn get_max_polyphony(&self) -> i64

pub fn set_panning_strength(&mut self, panning_strength: f64)

pub fn get_panning_strength(&self) -> f64

pub fn has_stream_playback(&mut self) -> bool

pub fn get_stream_playback(&mut self) -> Option<Gd<AudioStreamPlayback>>