Struct godot::engine::DisplayServer

pub struct DisplayServer { /* private fields */ }

Godot class DisplayServer.

Inherits Object.

Related symbols:

display_server: sidecar module with related enum/flag types

DisplayServerVirtual: virtual methods

See also Godot docs for DisplayServer.

Implementations

impl DisplayServer

pub fn singleton() -> Gd<DisplayServer>

pub fn notify(&mut self, what: ObjectNotification)

⚠️ Sends a Godot notification to all classes inherited by the object.

Triggers calls to on_notification(), and depending on the notification, also to Godot’s lifecycle callbacks such as ready().

Starts from the highest ancestor (the Object class) and goes down the hierarchy. See also Godot docs for Object::notification().

Panics

If you call this method on a user-defined object while holding a GdRef or GdMut guard on the instance, you will encounter a panic. The reason is that the receiving virtual method on_notification() acquires a GdMut lock dynamically, which must be exclusive.

pub fn notify_reversed(&mut self, what: ObjectNotification)

⚠️ Like Self::notify(), but starts at the most-derived class and goes up the hierarchy.

See docs of that method, including the panics.

pub fn has_feature(&self, feature: Feature) -> bool

pub fn get_name(&self) -> GodotString

pub fn global_menu_add_submenu_item( &mut self, menu_root: GodotString, label: GodotString, submenu: GodotString, index: i64 ) -> i64

pub fn global_menu_add_item( &mut self, menu_root: GodotString, label: GodotString, callback: Callable, key_callback: Callable, tag: Variant, accelerator: Key, index: i64 ) -> i64

pub fn global_menu_add_check_item( &mut self, menu_root: GodotString, label: GodotString, callback: Callable, key_callback: Callable, tag: Variant, accelerator: Key, index: i64 ) -> i64

pub fn global_menu_add_icon_item( &mut self, menu_root: GodotString, icon: Gd<Texture2D>, label: GodotString, callback: Callable, key_callback: Callable, tag: Variant, accelerator: Key, index: i64 ) -> i64

pub fn global_menu_add_icon_check_item( &mut self, menu_root: GodotString, icon: Gd<Texture2D>, label: GodotString, callback: Callable, key_callback: Callable, tag: Variant, accelerator: Key, index: i64 ) -> i64

pub fn global_menu_add_radio_check_item( &mut self, menu_root: GodotString, label: GodotString, callback: Callable, key_callback: Callable, tag: Variant, accelerator: Key, index: i64 ) -> i64

pub fn global_menu_add_icon_radio_check_item( &mut self, menu_root: GodotString, icon: Gd<Texture2D>, label: GodotString, callback: Callable, key_callback: Callable, tag: Variant, accelerator: Key, index: i64 ) -> i64

pub fn global_menu_add_multistate_item( &mut self, menu_root: GodotString, label: GodotString, max_states: i64, default_state: i64, callback: Callable, key_callback: Callable, tag: Variant, accelerator: Key, index: i64 ) -> i64

pub fn global_menu_add_separator( &mut self, menu_root: GodotString, index: i64 ) -> i64

pub fn global_menu_get_item_index_from_text( &self, menu_root: GodotString, text: GodotString ) -> i64

pub fn global_menu_get_item_index_from_tag( &self, menu_root: GodotString, tag: Variant ) -> i64

pub fn global_menu_is_item_checked( &self, menu_root: GodotString, idx: i64 ) -> bool

pub fn global_menu_is_item_checkable( &self, menu_root: GodotString, idx: i64 ) -> bool

pub fn global_menu_is_item_radio_checkable( &self, menu_root: GodotString, idx: i64 ) -> bool

pub fn global_menu_get_item_callback( &self, menu_root: GodotString, idx: i64 ) -> Callable

pub fn global_menu_get_item_key_callback( &self, menu_root: GodotString, idx: i64 ) -> Callable

pub fn global_menu_get_item_tag( &self, menu_root: GodotString, idx: i64 ) -> Variant

pub fn global_menu_get_item_text( &self, menu_root: GodotString, idx: i64 ) -> GodotString

pub fn global_menu_get_item_submenu( &self, menu_root: GodotString, idx: i64 ) -> GodotString

pub fn global_menu_get_item_accelerator( &self, menu_root: GodotString, idx: i64 ) -> Key

pub fn global_menu_is_item_disabled( &self, menu_root: GodotString, idx: i64 ) -> bool

pub fn global_menu_get_item_tooltip( &self, menu_root: GodotString, idx: i64 ) -> GodotString

pub fn global_menu_get_item_state(&self, menu_root: GodotString, idx: i64) -> i64

pub fn global_menu_get_item_max_states( &self, menu_root: GodotString, idx: i64 ) -> i64

pub fn global_menu_get_item_icon( &self, menu_root: GodotString, idx: i64 ) -> Option<Gd<Texture2D>>

pub fn global_menu_get_item_indentation_level( &self, menu_root: GodotString, idx: i64 ) -> i64

pub fn global_menu_set_item_checked( &mut self, menu_root: GodotString, idx: i64, checked: bool )

pub fn global_menu_set_item_checkable( &mut self, menu_root: GodotString, idx: i64, checkable: bool )

pub fn global_menu_set_item_radio_checkable( &mut self, menu_root: GodotString, idx: i64, checkable: bool )

pub fn global_menu_set_item_callback( &mut self, menu_root: GodotString, idx: i64, callback: Callable )

pub fn global_menu_set_item_key_callback( &mut self, menu_root: GodotString, idx: i64, key_callback: Callable )

pub fn global_menu_set_item_tag( &mut self, menu_root: GodotString, idx: i64, tag: Variant )

pub fn global_menu_set_item_text( &mut self, menu_root: GodotString, idx: i64, text: GodotString )

pub fn global_menu_set_item_submenu( &mut self, menu_root: GodotString, idx: i64, submenu: GodotString )

pub fn global_menu_set_item_accelerator( &mut self, menu_root: GodotString, idx: i64, keycode: Key )

pub fn global_menu_set_item_disabled( &mut self, menu_root: GodotString, idx: i64, disabled: bool )

pub fn global_menu_set_item_tooltip( &mut self, menu_root: GodotString, idx: i64, tooltip: GodotString )

pub fn global_menu_set_item_state( &mut self, menu_root: GodotString, idx: i64, state: i64 )

pub fn global_menu_set_item_max_states( &mut self, menu_root: GodotString, idx: i64, max_states: i64 )

pub fn global_menu_set_item_icon( &mut self, menu_root: GodotString, idx: i64, icon: Gd<Texture2D> )

pub fn global_menu_set_item_indentation_level( &mut self, menu_root: GodotString, idx: i64, level: i64 )

pub fn global_menu_get_item_count(&self, menu_root: GodotString) -> i64

pub fn global_menu_remove_item(&mut self, menu_root: GodotString, idx: i64)

pub fn global_menu_clear(&mut self, menu_root: GodotString)

pub fn tts_is_speaking(&self) -> bool

pub fn tts_is_paused(&self) -> bool

pub fn tts_get_voices(&self) -> Array<Dictionary>

pub fn tts_get_voices_for_language( &self, language: GodotString ) -> PackedStringArray

pub fn tts_speak( &mut self, text: GodotString, voice: GodotString, volume: i64, pitch: f64, rate: f64, utterance_id: i64, interrupt: bool )

pub fn tts_pause(&mut self)

pub fn tts_resume(&mut self)

pub fn tts_stop(&mut self)

pub fn tts_set_utterance_callback( &mut self, event: TTSUtteranceEvent, callable: Callable )

pub fn is_dark_mode_supported(&self) -> bool

pub fn is_dark_mode(&self) -> bool

pub fn get_accent_color(&self) -> Color

pub fn mouse_set_mode(&mut self, mouse_mode: MouseMode)

pub fn mouse_get_mode(&self) -> MouseMode

pub fn warp_mouse(&mut self, position: Vector2i)

pub fn mouse_get_position(&self) -> Vector2i

pub fn mouse_get_button_state(&self) -> MouseButtonMask

pub fn clipboard_set(&mut self, clipboard: GodotString)

pub fn clipboard_get(&self) -> GodotString

pub fn clipboard_has(&self) -> bool

pub fn clipboard_set_primary(&mut self, clipboard_primary: GodotString)

pub fn clipboard_get_primary(&self) -> GodotString

pub fn get_display_cutouts(&self) -> Array<Rect2>

pub fn get_display_safe_area(&self) -> Rect2i

pub fn get_screen_count(&self) -> i64

pub fn get_primary_screen(&self) -> i64

pub fn get_screen_from_rect(&self, rect: Rect2) -> i64

pub fn screen_get_position(&self, screen: i64) -> Vector2i

pub fn screen_get_size(&self, screen: i64) -> Vector2i

pub fn screen_get_usable_rect(&self, screen: i64) -> Rect2i

pub fn screen_get_dpi(&self, screen: i64) -> i64

pub fn screen_get_scale(&self, screen: i64) -> f64

pub fn is_touchscreen_available(&self) -> bool

pub fn screen_get_max_scale(&self) -> f64

pub fn screen_get_refresh_rate(&self, screen: i64) -> f64

pub fn screen_set_orientation( &mut self, orientation: ScreenOrientation, screen: i64 )

pub fn screen_get_orientation(&self, screen: i64) -> ScreenOrientation

pub fn screen_set_keep_on(&mut self, enable: bool)

pub fn screen_is_kept_on(&self) -> bool

pub fn get_window_list(&self) -> PackedInt32Array

pub fn get_window_at_screen_position(&self, position: Vector2i) -> i64

pub fn window_get_native_handle( &self, handle_type: HandleType, window_id: i64 ) -> i64

pub fn window_get_active_popup(&self) -> i64

pub fn window_set_popup_safe_rect(&mut self, window: i64, rect: Rect2i)

pub fn window_get_popup_safe_rect(&self, window: i64) -> Rect2i

pub fn window_set_title(&mut self, title: GodotString, window_id: i64)

pub fn window_set_mouse_passthrough( &mut self, region: PackedVector2Array, window_id: i64 )

pub fn window_get_current_screen(&self, window_id: i64) -> i64

pub fn window_set_current_screen(&mut self, screen: i64, window_id: i64)

pub fn window_get_position(&self, window_id: i64) -> Vector2i

pub fn window_get_position_with_decorations(&self, window_id: i64) -> Vector2i

pub fn window_set_position(&mut self, position: Vector2i, window_id: i64)

pub fn window_get_size(&self, window_id: i64) -> Vector2i

pub fn window_set_size(&mut self, size: Vector2i, window_id: i64)

pub fn window_set_rect_changed_callback( &mut self, callback: Callable, window_id: i64 )

pub fn window_set_window_event_callback( &mut self, callback: Callable, window_id: i64 )

pub fn window_set_input_event_callback( &mut self, callback: Callable, window_id: i64 )

pub fn window_set_input_text_callback( &mut self, callback: Callable, window_id: i64 )

pub fn window_set_drop_files_callback( &mut self, callback: Callable, window_id: i64 )

pub fn window_get_attached_instance_id(&self, window_id: i64) -> i64

pub fn window_get_max_size(&self, window_id: i64) -> Vector2i

pub fn window_set_max_size(&mut self, max_size: Vector2i, window_id: i64)

pub fn window_get_min_size(&self, window_id: i64) -> Vector2i

pub fn window_set_min_size(&mut self, min_size: Vector2i, window_id: i64)

pub fn window_get_size_with_decorations(&self, window_id: i64) -> Vector2i

pub fn window_get_mode(&self, window_id: i64) -> WindowMode

pub fn window_set_mode(&mut self, mode: WindowMode, window_id: i64)

pub fn window_set_flag(&mut self, flag: WindowFlags, enabled: bool, window_id: i64)

pub fn window_get_flag(&self, flag: WindowFlags, window_id: i64) -> bool

pub fn window_set_window_buttons_offset( &mut self, offset: Vector2i, window_id: i64 )

pub fn window_get_safe_title_margins(&self, window_id: i64) -> Vector3i

pub fn window_request_attention(&mut self, window_id: i64)

pub fn window_move_to_foreground(&mut self, window_id: i64)

pub fn window_can_draw(&self, window_id: i64) -> bool

pub fn window_set_transient(&mut self, window_id: i64, parent_window_id: i64)

pub fn window_set_exclusive(&mut self, window_id: i64, exclusive: bool)

pub fn window_set_ime_active(&mut self, active: bool, window_id: i64)

pub fn window_set_ime_position(&mut self, position: Vector2i, window_id: i64)

pub fn window_set_vsync_mode(&mut self, vsync_mode: VSyncMode, window_id: i64)

pub fn window_get_vsync_mode(&self, window_id: i64) -> VSyncMode

pub fn window_is_maximize_allowed(&self, window_id: i64) -> bool

pub fn window_maximize_on_title_dbl_click(&self) -> bool

pub fn window_minimize_on_title_dbl_click(&self) -> bool

pub fn ime_get_selection(&self) -> Vector2i

pub fn ime_get_text(&self) -> GodotString

pub fn virtual_keyboard_show( &mut self, existing_text: GodotString, position: Rect2, type_: VirtualKeyboardType, max_length: i64, cursor_start: i64, cursor_end: i64 )

pub fn virtual_keyboard_hide(&mut self)

pub fn virtual_keyboard_get_height(&self) -> i64

pub fn cursor_set_shape(&mut self, shape: CursorShape)

pub fn cursor_get_shape(&self) -> CursorShape

pub fn cursor_set_custom_image( &mut self, cursor: Gd<Resource>, shape: CursorShape, hotspot: Vector2 )

pub fn get_swap_cancel_ok(&mut self) -> bool

pub fn enable_for_stealing_focus(&mut self, process_id: i64)

pub fn dialog_show( &mut self, title: GodotString, description: GodotString, buttons: PackedStringArray, callback: Callable ) -> Error

pub fn dialog_input_text( &mut self, title: GodotString, description: GodotString, existing_text: GodotString, callback: Callable ) -> Error

pub fn keyboard_get_layout_count(&self) -> i64

pub fn keyboard_get_current_layout(&self) -> i64

pub fn keyboard_set_current_layout(&mut self, index: i64)

pub fn keyboard_get_layout_language(&self, index: i64) -> GodotString

pub fn keyboard_get_layout_name(&self, index: i64) -> GodotString

pub fn keyboard_get_keycode_from_physical(&self, keycode: Key) -> Key

pub fn process_events(&mut self)

pub fn force_process_and_drop_events(&mut self)

pub fn set_native_icon(&mut self, filename: GodotString)

pub fn set_icon(&mut self, image: Gd<Image>)

pub fn tablet_get_driver_count(&self) -> i64

pub fn tablet_get_driver_name(&self, idx: i64) -> GodotString

pub fn tablet_get_current_driver(&self) -> GodotString

pub fn tablet_set_current_driver(&mut self, name: GodotString)

pub const SCREEN_PRIMARY: i32 = -2i32

pub const SCREEN_OF_MAIN_WINDOW: i32 = -1i32

pub const MAIN_WINDOW_ID: i32 = 0i32

pub const INVALID_WINDOW_ID: i32 = -1i32