Struct godot::engine::Environment

pub struct Environment { /* private fields */ }

Godot class Environment.

Inherits Resource.

Related symbols:

environment: sidecar module with related enum/flag types

EnvironmentVirtual: virtual methods

See also Godot docs for Environment.

Implementations

impl Environment

pub fn new() -> Gd<Environment>

pub fn notify(&mut self, what: ObjectNotification)

⚠️ Sends a Godot notification to all classes inherited by the object.

Triggers calls to on_notification(), and depending on the notification, also to Godot’s lifecycle callbacks such as ready().

Starts from the highest ancestor (the Object class) and goes down the hierarchy. See also Godot docs for Object::notification().

Panics

If you call this method on a user-defined object while holding a GdRef or GdMut guard on the instance, you will encounter a panic. The reason is that the receiving virtual method on_notification() acquires a GdMut lock dynamically, which must be exclusive.

pub fn notify_reversed(&mut self, what: ObjectNotification)

⚠️ Like Self::notify(), but starts at the most-derived class and goes up the hierarchy.

See docs of that method, including the panics.

pub fn set_background(&mut self, mode: BGMode)

pub fn get_background(&self) -> BGMode

pub fn set_sky(&mut self, sky: Gd<Sky>)

pub fn get_sky(&self) -> Option<Gd<Sky>>

pub fn set_sky_custom_fov(&mut self, scale: f64)

pub fn get_sky_custom_fov(&self) -> f64

pub fn set_sky_rotation(&mut self, euler_radians: Vector3)

pub fn get_sky_rotation(&self) -> Vector3

pub fn set_bg_color(&mut self, color: Color)

pub fn get_bg_color(&self) -> Color

pub fn set_bg_energy_multiplier(&mut self, energy: f64)

pub fn get_bg_energy_multiplier(&self) -> f64

pub fn set_bg_intensity(&mut self, energy: f64)

pub fn get_bg_intensity(&self) -> f64

pub fn set_canvas_max_layer(&mut self, layer: i64)

pub fn get_canvas_max_layer(&self) -> i64

pub fn set_camera_feed_id(&mut self, id: i64)

pub fn get_camera_feed_id(&self) -> i64

pub fn set_ambient_light_color(&mut self, color: Color)

pub fn get_ambient_light_color(&self) -> Color

pub fn set_ambient_source(&mut self, source: AmbientSource)

pub fn get_ambient_source(&self) -> AmbientSource

pub fn set_ambient_light_energy(&mut self, energy: f64)

pub fn get_ambient_light_energy(&self) -> f64

pub fn set_ambient_light_sky_contribution(&mut self, ratio: f64)

pub fn get_ambient_light_sky_contribution(&self) -> f64

pub fn set_reflection_source(&mut self, source: ReflectionSource)

pub fn get_reflection_source(&self) -> ReflectionSource

pub fn set_tonemapper(&mut self, mode: ToneMapper)

pub fn get_tonemapper(&self) -> ToneMapper

pub fn set_tonemap_exposure(&mut self, exposure: f64)

pub fn get_tonemap_exposure(&self) -> f64

pub fn set_tonemap_white(&mut self, white: f64)

pub fn get_tonemap_white(&self) -> f64

pub fn set_ssr_enabled(&mut self, enabled: bool)

pub fn is_ssr_enabled(&self) -> bool

pub fn set_ssr_max_steps(&mut self, max_steps: i64)

pub fn get_ssr_max_steps(&self) -> i64

pub fn set_ssr_fade_in(&mut self, fade_in: f64)

pub fn get_ssr_fade_in(&self) -> f64

pub fn set_ssr_fade_out(&mut self, fade_out: f64)

pub fn get_ssr_fade_out(&self) -> f64

pub fn set_ssr_depth_tolerance(&mut self, depth_tolerance: f64)

pub fn get_ssr_depth_tolerance(&self) -> f64

pub fn set_ssao_enabled(&mut self, enabled: bool)

pub fn is_ssao_enabled(&self) -> bool

pub fn set_ssao_radius(&mut self, radius: f64)

pub fn get_ssao_radius(&self) -> f64

pub fn set_ssao_intensity(&mut self, intensity: f64)

pub fn get_ssao_intensity(&self) -> f64

pub fn set_ssao_power(&mut self, power: f64)

pub fn get_ssao_power(&self) -> f64

pub fn set_ssao_detail(&mut self, detail: f64)

pub fn get_ssao_detail(&self) -> f64

pub fn set_ssao_horizon(&mut self, horizon: f64)

pub fn get_ssao_horizon(&self) -> f64

pub fn set_ssao_sharpness(&mut self, sharpness: f64)

pub fn get_ssao_sharpness(&self) -> f64

pub fn set_ssao_direct_light_affect(&mut self, amount: f64)

pub fn get_ssao_direct_light_affect(&self) -> f64

pub fn set_ssao_ao_channel_affect(&mut self, amount: f64)

pub fn get_ssao_ao_channel_affect(&self) -> f64

pub fn set_ssil_enabled(&mut self, enabled: bool)

pub fn is_ssil_enabled(&self) -> bool

pub fn set_ssil_radius(&mut self, radius: f64)

pub fn get_ssil_radius(&self) -> f64

pub fn set_ssil_intensity(&mut self, intensity: f64)

pub fn get_ssil_intensity(&self) -> f64

pub fn set_ssil_sharpness(&mut self, sharpness: f64)

pub fn get_ssil_sharpness(&self) -> f64

pub fn set_ssil_normal_rejection(&mut self, normal_rejection: f64)

pub fn get_ssil_normal_rejection(&self) -> f64

pub fn set_sdfgi_enabled(&mut self, enabled: bool)

pub fn is_sdfgi_enabled(&self) -> bool

pub fn set_sdfgi_cascades(&mut self, amount: i64)

pub fn get_sdfgi_cascades(&self) -> i64

pub fn set_sdfgi_min_cell_size(&mut self, size: f64)

pub fn get_sdfgi_min_cell_size(&self) -> f64

pub fn set_sdfgi_max_distance(&mut self, distance: f64)

pub fn get_sdfgi_max_distance(&self) -> f64

pub fn set_sdfgi_cascade0_distance(&mut self, distance: f64)

pub fn get_sdfgi_cascade0_distance(&self) -> f64

pub fn set_sdfgi_y_scale(&mut self, scale: SDFGIYScale)

pub fn get_sdfgi_y_scale(&self) -> SDFGIYScale

pub fn set_sdfgi_use_occlusion(&mut self, enable: bool)

pub fn is_sdfgi_using_occlusion(&self) -> bool

pub fn set_sdfgi_bounce_feedback(&mut self, amount: f64)

pub fn get_sdfgi_bounce_feedback(&self) -> f64

pub fn set_sdfgi_read_sky_light(&mut self, enable: bool)

pub fn is_sdfgi_reading_sky_light(&self) -> bool

pub fn set_sdfgi_energy(&mut self, amount: f64)

pub fn get_sdfgi_energy(&self) -> f64

pub fn set_sdfgi_normal_bias(&mut self, bias: f64)

pub fn get_sdfgi_normal_bias(&self) -> f64

pub fn set_sdfgi_probe_bias(&mut self, bias: f64)

pub fn get_sdfgi_probe_bias(&self) -> f64

pub fn set_glow_enabled(&mut self, enabled: bool)

pub fn is_glow_enabled(&self) -> bool

pub fn set_glow_level(&mut self, idx: i64, intensity: f64)

pub fn get_glow_level(&self, idx: i64) -> f64

pub fn set_glow_normalized(&mut self, normalize: bool)

pub fn is_glow_normalized(&self) -> bool

pub fn set_glow_intensity(&mut self, intensity: f64)

pub fn get_glow_intensity(&self) -> f64

pub fn set_glow_strength(&mut self, strength: f64)

pub fn get_glow_strength(&self) -> f64

pub fn set_glow_mix(&mut self, mix: f64)

pub fn get_glow_mix(&self) -> f64

pub fn set_glow_bloom(&mut self, amount: f64)

pub fn get_glow_bloom(&self) -> f64

pub fn set_glow_blend_mode(&mut self, mode: GlowBlendMode)

pub fn get_glow_blend_mode(&self) -> GlowBlendMode

pub fn set_glow_hdr_bleed_threshold(&mut self, threshold: f64)

pub fn get_glow_hdr_bleed_threshold(&self) -> f64

pub fn set_glow_hdr_bleed_scale(&mut self, scale: f64)

pub fn get_glow_hdr_bleed_scale(&self) -> f64

pub fn set_glow_hdr_luminance_cap(&mut self, amount: f64)

pub fn get_glow_hdr_luminance_cap(&self) -> f64

pub fn set_glow_map_strength(&mut self, strength: f64)

pub fn get_glow_map_strength(&self) -> f64

pub fn set_glow_map(&mut self, mode: Gd<Texture>)

pub fn get_glow_map(&self) -> Option<Gd<Texture>>

pub fn set_fog_enabled(&mut self, enabled: bool)

pub fn is_fog_enabled(&self) -> bool

pub fn set_fog_light_color(&mut self, light_color: Color)

pub fn get_fog_light_color(&self) -> Color

pub fn set_fog_light_energy(&mut self, light_energy: f64)

pub fn get_fog_light_energy(&self) -> f64

pub fn set_fog_sun_scatter(&mut self, sun_scatter: f64)

pub fn get_fog_sun_scatter(&self) -> f64

pub fn set_fog_density(&mut self, density: f64)

pub fn get_fog_density(&self) -> f64

pub fn set_fog_height(&mut self, height: f64)

pub fn get_fog_height(&self) -> f64

pub fn set_fog_height_density(&mut self, height_density: f64)

pub fn get_fog_height_density(&self) -> f64

pub fn set_fog_aerial_perspective(&mut self, aerial_perspective: f64)

pub fn get_fog_aerial_perspective(&self) -> f64

pub fn set_fog_sky_affect(&mut self, sky_affect: f64)

pub fn get_fog_sky_affect(&self) -> f64

pub fn set_volumetric_fog_enabled(&mut self, enabled: bool)

pub fn is_volumetric_fog_enabled(&self) -> bool

pub fn set_volumetric_fog_emission(&mut self, color: Color)

pub fn get_volumetric_fog_emission(&self) -> Color

pub fn set_volumetric_fog_albedo(&mut self, color: Color)

pub fn get_volumetric_fog_albedo(&self) -> Color

pub fn set_volumetric_fog_density(&mut self, density: f64)

pub fn get_volumetric_fog_density(&self) -> f64

pub fn set_volumetric_fog_emission_energy(&mut self, begin: f64)

pub fn get_volumetric_fog_emission_energy(&self) -> f64

pub fn set_volumetric_fog_anisotropy(&mut self, anisotropy: f64)

pub fn get_volumetric_fog_anisotropy(&self) -> f64

pub fn set_volumetric_fog_length(&mut self, length: f64)

pub fn get_volumetric_fog_length(&self) -> f64

pub fn set_volumetric_fog_detail_spread(&mut self, detail_spread: f64)

pub fn get_volumetric_fog_detail_spread(&self) -> f64

pub fn set_volumetric_fog_gi_inject(&mut self, gi_inject: f64)

pub fn get_volumetric_fog_gi_inject(&self) -> f64

pub fn set_volumetric_fog_ambient_inject(&mut self, enabled: f64)

pub fn get_volumetric_fog_ambient_inject(&self) -> f64

pub fn set_volumetric_fog_sky_affect(&mut self, sky_affect: f64)

pub fn get_volumetric_fog_sky_affect(&self) -> f64

pub fn set_volumetric_fog_temporal_reprojection_enabled(&mut self, enabled: bool)

pub fn is_volumetric_fog_temporal_reprojection_enabled(&self) -> bool

pub fn set_volumetric_fog_temporal_reprojection_amount( &mut self, temporal_reprojection_amount: f64 )

pub fn get_volumetric_fog_temporal_reprojection_amount(&self) -> f64

pub fn set_adjustment_enabled(&mut self, enabled: bool)

pub fn is_adjustment_enabled(&self) -> bool

pub fn set_adjustment_brightness(&mut self, brightness: f64)

pub fn get_adjustment_brightness(&self) -> f64

pub fn set_adjustment_contrast(&mut self, contrast: f64)

pub fn get_adjustment_contrast(&self) -> f64

pub fn set_adjustment_saturation(&mut self, saturation: f64)

pub fn get_adjustment_saturation(&self) -> f64

pub fn set_adjustment_color_correction(&mut self, color_correction: Gd<Texture>)

pub fn get_adjustment_color_correction(&self) -> Option<Gd<Texture>>