Struct godot::engine::FileDialog

pub struct FileDialog { /* private fields */ }

Godot class FileDialog.

Inherits ConfirmationDialog.

Related symbols:

file_dialog: sidecar module with related enum/flag types

FileDialogVirtual: virtual methods

See also Godot docs for FileDialog.

Implementations

impl FileDialog

pub fn new_alloc() -> Gd<FileDialog>

pub fn notify(&mut self, what: WindowNotification)

⚠️ Sends a Godot notification to all classes inherited by the object.

Triggers calls to on_notification(), and depending on the notification, also to Godot’s lifecycle callbacks such as ready().

Starts from the highest ancestor (the Object class) and goes down the hierarchy. See also Godot docs for Object::notification().

Panics

If you call this method on a user-defined object while holding a GdRef or GdMut guard on the instance, you will encounter a panic. The reason is that the receiving virtual method on_notification() acquires a GdMut lock dynamically, which must be exclusive.

pub fn notify_reversed(&mut self, what: WindowNotification)

⚠️ Like Self::notify(), but starts at the most-derived class and goes up the hierarchy.

See docs of that method, including the panics.

pub fn clear_filters(&mut self)

pub fn add_filter(&mut self, filter: GodotString, description: GodotString)

pub fn set_filters(&mut self, filters: PackedStringArray)

pub fn get_filters(&self) -> PackedStringArray

pub fn get_current_dir(&self) -> GodotString

pub fn get_current_file(&self) -> GodotString

pub fn get_current_path(&self) -> GodotString

pub fn set_current_dir(&mut self, dir: GodotString)

pub fn set_current_file(&mut self, file: GodotString)

pub fn set_current_path(&mut self, path: GodotString)

pub fn set_mode_overrides_title(&mut self, override_: bool)

pub fn is_mode_overriding_title(&self) -> bool

pub fn set_file_mode(&mut self, mode: FileMode)

pub fn get_file_mode(&self) -> FileMode

pub fn get_vbox(&mut self) -> Option<Gd<VBoxContainer>>

pub fn get_line_edit(&mut self) -> Option<Gd<LineEdit>>

pub fn set_access(&mut self, access: Access)

pub fn get_access(&self) -> Access

pub fn set_root_subfolder(&mut self, dir: GodotString)

pub fn get_root_subfolder(&self) -> GodotString

pub fn set_show_hidden_files(&mut self, show: bool)

pub fn is_showing_hidden_files(&self) -> bool

pub fn deselect_all(&mut self)

pub fn invalidate(&mut self)