Struct godot::engine::NavigationLink3D

pub struct NavigationLink3D { /* private fields */ }

Godot class NavigationLink3D.

Inherits Node3D.

Related symbols:

NavigationLink3DVirtual: virtual methods

See also Godot docs for NavigationLink3D.

Implementations

impl NavigationLink3D

pub fn new_alloc() -> Gd<NavigationLink3D>

pub fn notify(&mut self, what: Node3DNotification)

⚠️ Sends a Godot notification to all classes inherited by the object.

Triggers calls to on_notification(), and depending on the notification, also to Godot’s lifecycle callbacks such as ready().

Starts from the highest ancestor (the Object class) and goes down the hierarchy. See also Godot docs for Object::notification().

Panics

If you call this method on a user-defined object while holding a GdRef or GdMut guard on the instance, you will encounter a panic. The reason is that the receiving virtual method on_notification() acquires a GdMut lock dynamically, which must be exclusive.

pub fn notify_reversed(&mut self, what: Node3DNotification)

⚠️ Like Self::notify(), but starts at the most-derived class and goes up the hierarchy.

See docs of that method, including the panics.

pub fn set_enabled(&mut self, enabled: bool)

pub fn is_enabled(&self) -> bool

pub fn set_bidirectional(&mut self, bidirectional: bool)

pub fn is_bidirectional(&self) -> bool

pub fn set_navigation_layers(&mut self, navigation_layers: i64)

pub fn get_navigation_layers(&self) -> i64

pub fn set_navigation_layer_value(&mut self, layer_number: i64, value: bool)

pub fn get_navigation_layer_value(&self, layer_number: i64) -> bool

pub fn set_start_position(&mut self, position: Vector3)

pub fn get_start_position(&self) -> Vector3

pub fn set_end_position(&mut self, position: Vector3)

pub fn get_end_position(&self) -> Vector3

pub fn set_global_start_position(&mut self, position: Vector3)

pub fn get_global_start_position(&self) -> Vector3

pub fn set_global_end_position(&mut self, position: Vector3)

pub fn get_global_end_position(&self) -> Vector3

pub fn set_enter_cost(&mut self, enter_cost: f64)

pub fn get_enter_cost(&self) -> f64

pub fn set_travel_cost(&mut self, travel_cost: f64)

pub fn get_travel_cost(&self) -> f64