Struct godot::engine::PhysicsServer3D

pub struct PhysicsServer3D { /* private fields */ }

Godot class PhysicsServer3D.

Inherits Object.

Related symbols:

physics_server_3d: sidecar module with related enum/flag types

PhysicsServer3DVirtual: virtual methods

See also Godot docs for PhysicsServer3D.

Implementations

impl PhysicsServer3D

pub fn singleton() -> Gd<PhysicsServer3D>

pub fn notify(&mut self, what: ObjectNotification)

⚠️ Sends a Godot notification to all classes inherited by the object.

Triggers calls to on_notification(), and depending on the notification, also to Godot’s lifecycle callbacks such as ready().

Starts from the highest ancestor (the Object class) and goes down the hierarchy. See also Godot docs for Object::notification().

Panics

If you call this method on a user-defined object while holding a GdRef or GdMut guard on the instance, you will encounter a panic. The reason is that the receiving virtual method on_notification() acquires a GdMut lock dynamically, which must be exclusive.

pub fn notify_reversed(&mut self, what: ObjectNotification)

⚠️ Like Self::notify(), but starts at the most-derived class and goes up the hierarchy.

See docs of that method, including the panics.

pub fn world_boundary_shape_create(&mut self) -> Rid

pub fn separation_ray_shape_create(&mut self) -> Rid

pub fn sphere_shape_create(&mut self) -> Rid

pub fn box_shape_create(&mut self) -> Rid

pub fn capsule_shape_create(&mut self) -> Rid

pub fn cylinder_shape_create(&mut self) -> Rid

pub fn convex_polygon_shape_create(&mut self) -> Rid

pub fn concave_polygon_shape_create(&mut self) -> Rid

pub fn heightmap_shape_create(&mut self) -> Rid

pub fn custom_shape_create(&mut self) -> Rid

pub fn shape_set_data(&mut self, shape: Rid, data: Variant)

pub fn shape_get_type(&self, shape: Rid) -> ShapeType

pub fn shape_get_data(&self, shape: Rid) -> Variant

pub fn space_create(&mut self) -> Rid

pub fn space_set_active(&mut self, space: Rid, active: bool)

pub fn space_is_active(&self, space: Rid) -> bool

pub fn space_set_param(&mut self, space: Rid, param: SpaceParameter, value: f64)

pub fn space_get_param(&self, space: Rid, param: SpaceParameter) -> f64

pub fn space_get_direct_state( &mut self, space: Rid ) -> Option<Gd<PhysicsDirectSpaceState3D>>

pub fn area_create(&mut self) -> Rid

pub fn area_set_space(&mut self, area: Rid, space: Rid)

pub fn area_get_space(&self, area: Rid) -> Rid

pub fn area_add_shape( &mut self, area: Rid, shape: Rid, transform: Transform3D, disabled: bool )

pub fn area_set_shape(&mut self, area: Rid, shape_idx: i64, shape: Rid)

pub fn area_set_shape_transform( &mut self, area: Rid, shape_idx: i64, transform: Transform3D )

pub fn area_set_shape_disabled( &mut self, area: Rid, shape_idx: i64, disabled: bool )

pub fn area_get_shape_count(&self, area: Rid) -> i64

pub fn area_get_shape(&self, area: Rid, shape_idx: i64) -> Rid

pub fn area_get_shape_transform(&self, area: Rid, shape_idx: i64) -> Transform3D

pub fn area_remove_shape(&mut self, area: Rid, shape_idx: i64)

pub fn area_clear_shapes(&mut self, area: Rid)

pub fn area_set_collision_layer(&mut self, area: Rid, layer: i64)

pub fn area_get_collision_layer(&self, area: Rid) -> i64

pub fn area_set_collision_mask(&mut self, area: Rid, mask: i64)

pub fn area_get_collision_mask(&self, area: Rid) -> i64

pub fn area_set_param(&mut self, area: Rid, param: AreaParameter, value: Variant)

pub fn area_set_transform(&mut self, area: Rid, transform: Transform3D)

pub fn area_get_param(&self, area: Rid, param: AreaParameter) -> Variant

pub fn area_get_transform(&self, area: Rid) -> Transform3D

pub fn area_attach_object_instance_id(&mut self, area: Rid, id: i64)

pub fn area_get_object_instance_id(&self, area: Rid) -> i64

pub fn area_set_monitor_callback(&mut self, area: Rid, callback: Callable)

pub fn area_set_area_monitor_callback(&mut self, area: Rid, callback: Callable)

pub fn area_set_monitorable(&mut self, area: Rid, monitorable: bool)

pub fn area_set_ray_pickable(&mut self, area: Rid, enable: bool)

pub fn body_create(&mut self) -> Rid

pub fn body_set_space(&mut self, body: Rid, space: Rid)

pub fn body_get_space(&self, body: Rid) -> Rid

pub fn body_set_mode(&mut self, body: Rid, mode: BodyMode)

pub fn body_get_mode(&self, body: Rid) -> BodyMode

pub fn body_set_collision_layer(&mut self, body: Rid, layer: i64)

pub fn body_get_collision_layer(&self, body: Rid) -> i64

pub fn body_set_collision_mask(&mut self, body: Rid, mask: i64)

pub fn body_get_collision_mask(&self, body: Rid) -> i64

pub fn body_set_collision_priority(&mut self, body: Rid, priority: f64)

pub fn body_get_collision_priority(&self, body: Rid) -> f64

pub fn body_add_shape( &mut self, body: Rid, shape: Rid, transform: Transform3D, disabled: bool )

pub fn body_set_shape(&mut self, body: Rid, shape_idx: i64, shape: Rid)

pub fn body_set_shape_transform( &mut self, body: Rid, shape_idx: i64, transform: Transform3D )

pub fn body_set_shape_disabled( &mut self, body: Rid, shape_idx: i64, disabled: bool )

pub fn body_get_shape_count(&self, body: Rid) -> i64

pub fn body_get_shape(&self, body: Rid, shape_idx: i64) -> Rid

pub fn body_get_shape_transform(&self, body: Rid, shape_idx: i64) -> Transform3D

pub fn body_remove_shape(&mut self, body: Rid, shape_idx: i64)

pub fn body_clear_shapes(&mut self, body: Rid)

pub fn body_attach_object_instance_id(&mut self, body: Rid, id: i64)

pub fn body_get_object_instance_id(&self, body: Rid) -> i64

pub fn body_set_enable_continuous_collision_detection( &mut self, body: Rid, enable: bool )

pub fn body_is_continuous_collision_detection_enabled(&self, body: Rid) -> bool

pub fn body_set_param(&mut self, body: Rid, param: BodyParameter, value: Variant)

pub fn body_get_param(&self, body: Rid, param: BodyParameter) -> Variant

pub fn body_reset_mass_properties(&mut self, body: Rid)

pub fn body_set_state(&mut self, body: Rid, state: BodyState, value: Variant)

pub fn body_get_state(&self, body: Rid, state: BodyState) -> Variant

pub fn body_apply_central_impulse(&mut self, body: Rid, impulse: Vector3)

pub fn body_apply_impulse( &mut self, body: Rid, impulse: Vector3, position: Vector3 )

pub fn body_apply_torque_impulse(&mut self, body: Rid, impulse: Vector3)

pub fn body_apply_central_force(&mut self, body: Rid, force: Vector3)

pub fn body_apply_force(&mut self, body: Rid, force: Vector3, position: Vector3)

pub fn body_apply_torque(&mut self, body: Rid, torque: Vector3)

pub fn body_add_constant_central_force(&mut self, body: Rid, force: Vector3)

pub fn body_add_constant_force( &mut self, body: Rid, force: Vector3, position: Vector3 )

pub fn body_add_constant_torque(&mut self, body: Rid, torque: Vector3)

pub fn body_set_constant_force(&mut self, body: Rid, force: Vector3)

pub fn body_get_constant_force(&self, body: Rid) -> Vector3

pub fn body_set_constant_torque(&mut self, body: Rid, torque: Vector3)

pub fn body_get_constant_torque(&self, body: Rid) -> Vector3

pub fn body_set_axis_velocity(&mut self, body: Rid, axis_velocity: Vector3)

pub fn body_set_axis_lock(&mut self, body: Rid, axis: BodyAxis, lock: bool)

pub fn body_is_axis_locked(&self, body: Rid, axis: BodyAxis) -> bool

pub fn body_add_collision_exception(&mut self, body: Rid, excepted_body: Rid)

pub fn body_remove_collision_exception(&mut self, body: Rid, excepted_body: Rid)

pub fn body_set_max_contacts_reported(&mut self, body: Rid, amount: i64)

pub fn body_get_max_contacts_reported(&self, body: Rid) -> i64

pub fn body_set_omit_force_integration(&mut self, body: Rid, enable: bool)

pub fn body_is_omitting_force_integration(&self, body: Rid) -> bool

pub fn body_set_force_integration_callback( &mut self, body: Rid, callable: Callable, userdata: Variant )

pub fn body_set_ray_pickable(&mut self, body: Rid, enable: bool)

pub fn body_test_motion( &mut self, body: Rid, parameters: Gd<PhysicsTestMotionParameters3D>, result: Gd<PhysicsTestMotionResult3D> ) -> bool

pub fn body_get_direct_state( &mut self, body: Rid ) -> Option<Gd<PhysicsDirectBodyState3D>>

pub fn soft_body_get_bounds(&self, body: Rid) -> Aabb

pub fn joint_create(&mut self) -> Rid

pub fn joint_clear(&mut self, joint: Rid)

pub fn joint_make_pin( &mut self, joint: Rid, body_A: Rid, local_A: Vector3, body_B: Rid, local_B: Vector3 )

pub fn pin_joint_set_param(&mut self, joint: Rid, param: PinJointParam, value: f64)

pub fn pin_joint_get_param(&self, joint: Rid, param: PinJointParam) -> f64

pub fn pin_joint_set_local_a(&mut self, joint: Rid, local_A: Vector3)

pub fn pin_joint_get_local_a(&self, joint: Rid) -> Vector3

pub fn pin_joint_set_local_b(&mut self, joint: Rid, local_B: Vector3)

pub fn pin_joint_get_local_b(&self, joint: Rid) -> Vector3

pub fn joint_make_hinge( &mut self, joint: Rid, body_A: Rid, hinge_A: Transform3D, body_B: Rid, hinge_B: Transform3D )

pub fn hinge_joint_set_param( &mut self, joint: Rid, param: HingeJointParam, value: f64 )

pub fn hinge_joint_get_param(&self, joint: Rid, param: HingeJointParam) -> f64

pub fn hinge_joint_set_flag( &mut self, joint: Rid, flag: HingeJointFlag, enabled: bool )

pub fn hinge_joint_get_flag(&self, joint: Rid, flag: HingeJointFlag) -> bool

pub fn joint_make_slider( &mut self, joint: Rid, body_A: Rid, local_ref_A: Transform3D, body_B: Rid, local_ref_B: Transform3D )

pub fn slider_joint_set_param( &mut self, joint: Rid, param: SliderJointParam, value: f64 )

pub fn slider_joint_get_param(&self, joint: Rid, param: SliderJointParam) -> f64

pub fn joint_make_cone_twist( &mut self, joint: Rid, body_A: Rid, local_ref_A: Transform3D, body_B: Rid, local_ref_B: Transform3D )

pub fn cone_twist_joint_set_param( &mut self, joint: Rid, param: ConeTwistJointParam, value: f64 )

pub fn cone_twist_joint_get_param( &self, joint: Rid, param: ConeTwistJointParam ) -> f64

pub fn joint_get_type(&self, joint: Rid) -> JointType

pub fn joint_set_solver_priority(&mut self, joint: Rid, priority: i64)

pub fn joint_get_solver_priority(&self, joint: Rid) -> i64

pub fn joint_disable_collisions_between_bodies( &mut self, joint: Rid, disable: bool )

pub fn joint_is_disabled_collisions_between_bodies(&self, joint: Rid) -> bool

pub fn joint_make_generic_6dof( &mut self, joint: Rid, body_A: Rid, local_ref_A: Transform3D, body_B: Rid, local_ref_B: Transform3D )

pub fn generic_6dof_joint_set_param( &mut self, joint: Rid, axis: Vector3Axis, param: G6DOFJointAxisParam, value: f64 )

pub fn generic_6dof_joint_get_param( &self, joint: Rid, axis: Vector3Axis, param: G6DOFJointAxisParam ) -> f64

pub fn generic_6dof_joint_set_flag( &mut self, joint: Rid, axis: Vector3Axis, flag: G6DOFJointAxisFlag, enable: bool )

pub fn generic_6dof_joint_get_flag( &self, joint: Rid, axis: Vector3Axis, flag: G6DOFJointAxisFlag ) -> bool

pub fn free_rid(&mut self, rid: Rid)

pub fn set_active(&mut self, active: bool)

pub fn get_process_info(&mut self, process_info: ProcessInfo) -> i64