Struct godot::engine::PopupPanel

pub struct PopupPanel { /* private fields */ }

Godot class PopupPanel.

Inherits Popup.

Related symbols:

PopupPanelVirtual: virtual methods

See also Godot docs for PopupPanel.

Implementations

impl PopupPanel

pub fn new_alloc() -> Gd<PopupPanel>

pub fn notify(&mut self, what: WindowNotification)

⚠️ Sends a Godot notification to all classes inherited by the object.

Triggers calls to on_notification(), and depending on the notification, also to Godot’s lifecycle callbacks such as ready().

Starts from the highest ancestor (the Object class) and goes down the hierarchy. See also Godot docs for Object::notification().

Panics

If you call this method on a user-defined object while holding a GdRef or GdMut guard on the instance, you will encounter a panic. The reason is that the receiving virtual method on_notification() acquires a GdMut lock dynamically, which must be exclusive.

pub fn notify_reversed(&mut self, what: WindowNotification)

⚠️ Like Self::notify(), but starts at the most-derived class and goes up the hierarchy.

See docs of that method, including the panics.