Struct godot::engine::RenderingDevice

pub struct RenderingDevice { /* private fields */ }

Godot class RenderingDevice.

Inherits Object.

Related symbols:

rendering_device: sidecar module with related enum/flag types

RenderingDeviceVirtual: virtual methods

See also Godot docs for RenderingDevice.

Implementations

impl RenderingDevice

pub fn notify(&mut self, what: ObjectNotification)

⚠️ Sends a Godot notification to all classes inherited by the object.

Triggers calls to on_notification(), and depending on the notification, also to Godot’s lifecycle callbacks such as ready().

Starts from the highest ancestor (the Object class) and goes down the hierarchy. See also Godot docs for Object::notification().

Panics

If you call this method on a user-defined object while holding a GdRef or GdMut guard on the instance, you will encounter a panic. The reason is that the receiving virtual method on_notification() acquires a GdMut lock dynamically, which must be exclusive.

pub fn notify_reversed(&mut self, what: ObjectNotification)

⚠️ Like Self::notify(), but starts at the most-derived class and goes up the hierarchy.

See docs of that method, including the panics.

pub fn texture_create( &mut self, format: Gd<RdTextureFormat>, view: Gd<RdTextureView>, data: PackedByteArray ) -> Rid

pub fn texture_create_shared( &mut self, view: Gd<RdTextureView>, with_texture: Rid ) -> Rid

pub fn texture_create_shared_from_slice( &mut self, view: Gd<RdTextureView>, with_texture: Rid, layer: i64, mipmap: i64, mipmaps: i64, slice_type: TextureSliceType ) -> Rid

pub fn texture_update( &mut self, texture: Rid, layer: i64, data: PackedByteArray, post_barrier: BarrierMask ) -> Error

pub fn texture_get_data(&mut self, texture: Rid, layer: i64) -> PackedByteArray

pub fn texture_is_format_supported_for_usage( &self, format: DataFormat, usage_flags: TextureUsageBits ) -> bool

pub fn texture_is_shared(&mut self, texture: Rid) -> bool

pub fn texture_is_valid(&mut self, texture: Rid) -> bool

pub fn texture_copy( &mut self, from_texture: Rid, to_texture: Rid, from_pos: Vector3, to_pos: Vector3, size: Vector3, src_mipmap: i64, dst_mipmap: i64, src_layer: i64, dst_layer: i64, post_barrier: BarrierMask ) -> Error

pub fn texture_clear( &mut self, texture: Rid, color: Color, base_mipmap: i64, mipmap_count: i64, base_layer: i64, layer_count: i64, post_barrier: BarrierMask ) -> Error

pub fn texture_resolve_multisample( &mut self, from_texture: Rid, to_texture: Rid, post_barrier: BarrierMask ) -> Error

pub fn framebuffer_format_create( &mut self, attachments: Array<Gd<RdAttachmentFormat>>, view_count: i64 ) -> i64

pub fn framebuffer_format_create_multipass( &mut self, attachments: Array<Gd<RdAttachmentFormat>>, passes: Array<Gd<RdFramebufferPass>>, view_count: i64 ) -> i64

pub fn framebuffer_format_create_empty(&mut self, samples: TextureSamples) -> i64

pub fn framebuffer_format_get_texture_samples( &mut self, format: i64, render_pass: i64 ) -> TextureSamples

pub fn framebuffer_create( &mut self, textures: Array<Rid>, validate_with_format: i64, view_count: i64 ) -> Rid

pub fn framebuffer_create_multipass( &mut self, textures: Array<Rid>, passes: Array<Gd<RdFramebufferPass>>, validate_with_format: i64, view_count: i64 ) -> Rid

pub fn framebuffer_create_empty( &mut self, size: Vector2i, samples: TextureSamples, validate_with_format: i64 ) -> Rid

pub fn framebuffer_get_format(&mut self, framebuffer: Rid) -> i64

pub fn framebuffer_is_valid(&self, framebuffer: Rid) -> bool

pub fn sampler_create(&mut self, state: Gd<RdSamplerState>) -> Rid

pub fn vertex_buffer_create( &mut self, size_bytes: i64, data: PackedByteArray, use_as_storage: bool ) -> Rid

pub fn vertex_format_create( &mut self, vertex_descriptions: Array<Gd<RdVertexAttribute>> ) -> i64

pub fn vertex_array_create( &mut self, vertex_count: i64, vertex_format: i64, src_buffers: Array<Rid>, offsets: PackedInt64Array ) -> Rid

pub fn index_buffer_create( &mut self, size_indices: i64, format: IndexBufferFormat, data: PackedByteArray, use_restart_indices: bool ) -> Rid

pub fn index_array_create( &mut self, index_buffer: Rid, index_offset: i64, index_count: i64 ) -> Rid

pub fn shader_compile_spirv_from_source( &mut self, shader_source: Gd<RdShaderSource>, allow_cache: bool ) -> Option<Gd<RdShaderSpirv>>

pub fn shader_compile_binary_from_spirv( &mut self, spirv_data: Gd<RdShaderSpirv>, name: GodotString ) -> PackedByteArray

pub fn shader_create_from_spirv( &mut self, spirv_data: Gd<RdShaderSpirv>, name: GodotString ) -> Rid

pub fn shader_create_from_bytecode( &mut self, binary_data: PackedByteArray ) -> Rid

pub fn shader_get_vertex_input_attribute_mask(&mut self, shader: Rid) -> i64

pub fn uniform_buffer_create( &mut self, size_bytes: i64, data: PackedByteArray ) -> Rid

pub fn storage_buffer_create( &mut self, size_bytes: i64, data: PackedByteArray, usage: StorageBufferUsage ) -> Rid

pub fn texture_buffer_create( &mut self, size_bytes: i64, format: DataFormat, data: PackedByteArray ) -> Rid

pub fn uniform_set_create( &mut self, uniforms: Array<Gd<RdUniform>>, shader: Rid, shader_set: i64 ) -> Rid

pub fn uniform_set_is_valid(&mut self, uniform_set: Rid) -> bool

pub fn buffer_update( &mut self, buffer: Rid, offset: i64, size_bytes: i64, data: PackedByteArray, post_barrier: BarrierMask ) -> Error

pub fn buffer_clear( &mut self, buffer: Rid, offset: i64, size_bytes: i64, post_barrier: BarrierMask ) -> Error

pub fn buffer_get_data( &mut self, buffer: Rid, offset_bytes: i64, size_bytes: i64 ) -> PackedByteArray

pub fn render_pipeline_create( &mut self, shader: Rid, framebuffer_format: i64, vertex_format: i64, primitive: RenderPrimitive, rasterization_state: Gd<RdPipelineRasterizationState>, multisample_state: Gd<RdPipelineMultisampleState>, stencil_state: Gd<RdPipelineDepthStencilState>, color_blend_state: Gd<RdPipelineColorBlendState>, dynamic_state_flags: PipelineDynamicStateFlags, for_render_pass: i64, specialization_constants: Array<Gd<RdPipelineSpecializationConstant>> ) -> Rid

pub fn render_pipeline_is_valid(&mut self, render_pipeline: Rid) -> bool

pub fn compute_pipeline_create( &mut self, shader: Rid, specialization_constants: Array<Gd<RdPipelineSpecializationConstant>> ) -> Rid

pub fn compute_pipeline_is_valid(&mut self, compute_pieline: Rid) -> bool

pub fn screen_get_width(&self, screen: i64) -> i64

pub fn screen_get_height(&self, screen: i64) -> i64

pub fn screen_get_framebuffer_format(&self) -> i64

pub fn draw_list_begin_for_screen( &mut self, screen: i64, clear_color: Color ) -> i64

pub fn draw_list_begin( &mut self, framebuffer: Rid, initial_color_action: InitialAction, final_color_action: FinalAction, initial_depth_action: InitialAction, final_depth_action: FinalAction, clear_color_values: PackedColorArray, clear_depth: f64, clear_stencil: i64, region: Rect2, storage_textures: Array<Variant> ) -> i64

pub fn draw_list_begin_split( &mut self, framebuffer: Rid, splits: i64, initial_color_action: InitialAction, final_color_action: FinalAction, initial_depth_action: InitialAction, final_depth_action: FinalAction, clear_color_values: PackedColorArray, clear_depth: f64, clear_stencil: i64, region: Rect2, storage_textures: Array<Rid> ) -> PackedInt64Array

pub fn draw_list_set_blend_constants(&mut self, draw_list: i64, color: Color)

pub fn draw_list_bind_render_pipeline( &mut self, draw_list: i64, render_pipeline: Rid )

pub fn draw_list_bind_uniform_set( &mut self, draw_list: i64, uniform_set: Rid, set_index: i64 )

pub fn draw_list_bind_vertex_array(&mut self, draw_list: i64, vertex_array: Rid)

pub fn draw_list_bind_index_array(&mut self, draw_list: i64, index_array: Rid)

pub fn draw_list_set_push_constant( &mut self, draw_list: i64, buffer: PackedByteArray, size_bytes: i64 )

pub fn draw_list_draw( &mut self, draw_list: i64, use_indices: bool, instances: i64, procedural_vertex_count: i64 )

pub fn draw_list_enable_scissor(&mut self, draw_list: i64, rect: Rect2)

pub fn draw_list_disable_scissor(&mut self, draw_list: i64)

pub fn draw_list_switch_to_next_pass(&mut self) -> i64

pub fn draw_list_switch_to_next_pass_split( &mut self, splits: i64 ) -> PackedInt64Array

pub fn draw_list_end(&mut self, post_barrier: BarrierMask)

pub fn compute_list_begin(&mut self, allow_draw_overlap: bool) -> i64

pub fn compute_list_bind_compute_pipeline( &mut self, compute_list: i64, compute_pipeline: Rid )

pub fn compute_list_set_push_constant( &mut self, compute_list: i64, buffer: PackedByteArray, size_bytes: i64 )

pub fn compute_list_bind_uniform_set( &mut self, compute_list: i64, uniform_set: Rid, set_index: i64 )

pub fn compute_list_dispatch( &mut self, compute_list: i64, x_groups: i64, y_groups: i64, z_groups: i64 )

pub fn compute_list_add_barrier(&mut self, compute_list: i64)

pub fn compute_list_end(&mut self, post_barrier: BarrierMask)

pub fn free_rid(&mut self, rid: Rid)

pub fn capture_timestamp(&mut self, name: GodotString)

pub fn get_captured_timestamps_count(&self) -> i64

pub fn get_captured_timestamps_frame(&self) -> i64

pub fn get_captured_timestamp_gpu_time(&self, index: i64) -> i64

pub fn get_captured_timestamp_cpu_time(&self, index: i64) -> i64

pub fn get_captured_timestamp_name(&self, index: i64) -> GodotString

pub fn limit_get(&self, limit: Limit) -> i64

pub fn get_frame_delay(&self) -> i64

pub fn submit(&mut self)

pub fn sync(&mut self)

pub fn barrier(&mut self, from: BarrierMask, to: BarrierMask)

pub fn full_barrier(&mut self)

pub fn create_local_device(&mut self) -> Option<Gd<RenderingDevice>>

pub fn set_resource_name(&mut self, id: Rid, name: GodotString)

pub fn draw_command_begin_label(&mut self, name: GodotString, color: Color)

pub fn draw_command_insert_label(&mut self, name: GodotString, color: Color)

pub fn draw_command_end_label(&mut self)

pub fn get_device_vendor_name(&self) -> GodotString

pub fn get_device_name(&self) -> GodotString

pub fn get_device_pipeline_cache_uuid(&self) -> GodotString

pub fn get_memory_usage(&self, type_: MemoryType) -> i64

pub fn get_driver_resource( &mut self, resource: DriverResource, rid: Rid, index: i64 ) -> i64

pub const INVALID_ID: i32 = -1i32

pub const INVALID_FORMAT_ID: i32 = -1i32