Struct godot::engine::RenderingServer

pub struct RenderingServer { /* private fields */ }

Godot class RenderingServer.

Inherits Object.

Related symbols:

rendering_server: sidecar module with related enum/flag types

RenderingServerVirtual: virtual methods

See also Godot docs for RenderingServer.

Implementations

impl RenderingServer

pub fn singleton() -> Gd<RenderingServer>

pub fn notify(&mut self, what: ObjectNotification)

⚠️ Sends a Godot notification to all classes inherited by the object.

Triggers calls to on_notification(), and depending on the notification, also to Godot’s lifecycle callbacks such as ready().

Starts from the highest ancestor (the Object class) and goes down the hierarchy. See also Godot docs for Object::notification().

Panics

If you call this method on a user-defined object while holding a GdRef or GdMut guard on the instance, you will encounter a panic. The reason is that the receiving virtual method on_notification() acquires a GdMut lock dynamically, which must be exclusive.

pub fn notify_reversed(&mut self, what: ObjectNotification)

⚠️ Like Self::notify(), but starts at the most-derived class and goes up the hierarchy.

See docs of that method, including the panics.

pub const NO_INDEX_ARRAY: i32 = -1i32

pub const ARRAY_WEIGHTS_SIZE: i32 = 4i32

pub const CANVAS_ITEM_Z_MIN: i32 = -4_096i32

pub const CANVAS_ITEM_Z_MAX: i32 = 4_096i32

pub const MAX_GLOW_LEVELS: i32 = 7i32

pub const MAX_CURSORS: i32 = 8i32

pub const MAX_2D_DIRECTIONAL_LIGHTS: i32 = 8i32

pub const MATERIAL_RENDER_PRIORITY_MIN: i32 = -128i32

pub const MATERIAL_RENDER_PRIORITY_MAX: i32 = 127i32

pub const ARRAY_CUSTOM_COUNT: i32 = 4i32

pub const PARTICLES_EMIT_FLAG_POSITION: i32 = 1i32

pub const PARTICLES_EMIT_FLAG_ROTATION_SCALE: i32 = 2i32

pub const PARTICLES_EMIT_FLAG_VELOCITY: i32 = 4i32

pub const PARTICLES_EMIT_FLAG_COLOR: i32 = 8i32

pub const PARTICLES_EMIT_FLAG_CUSTOM: i32 = 16i32