Struct godot::engine::TextServer

pub struct TextServer { /* private fields */ }

Godot class TextServer.

Inherits RefCounted.

Related symbols:

text_server: sidecar module with related enum/flag types

TextServerVirtual: virtual methods

See also Godot docs for TextServer.

Implementations

impl TextServer

pub fn notify(&mut self, what: ObjectNotification)

⚠️ Sends a Godot notification to all classes inherited by the object.

Triggers calls to on_notification(), and depending on the notification, also to Godot’s lifecycle callbacks such as ready().

Starts from the highest ancestor (the Object class) and goes down the hierarchy. See also Godot docs for Object::notification().

Panics

If you call this method on a user-defined object while holding a GdRef or GdMut guard on the instance, you will encounter a panic. The reason is that the receiving virtual method on_notification() acquires a GdMut lock dynamically, which must be exclusive.

pub fn notify_reversed(&mut self, what: ObjectNotification)

⚠️ Like Self::notify(), but starts at the most-derived class and goes up the hierarchy.

See docs of that method, including the panics.

pub fn has_feature(&self, feature: Feature) -> bool

pub fn get_name(&self) -> GodotString

pub fn get_features(&self) -> i64

pub fn load_support_data(&mut self, filename: GodotString) -> bool

pub fn get_support_data_filename(&self) -> GodotString

pub fn get_support_data_info(&self) -> GodotString

pub fn save_support_data(&self, filename: GodotString) -> bool

pub fn is_locale_right_to_left(&self, locale: GodotString) -> bool

pub fn name_to_tag(&self, name: GodotString) -> i64

pub fn tag_to_name(&self, tag: i64) -> GodotString

pub fn has(&mut self, rid: Rid) -> bool

pub fn free_rid(&mut self, rid: Rid)

pub fn create_font(&mut self) -> Rid

pub fn font_set_data(&mut self, font_rid: Rid, data: PackedByteArray)

pub fn font_set_face_index(&mut self, font_rid: Rid, face_index: i64)

pub fn font_get_face_index(&self, font_rid: Rid) -> i64

pub fn font_get_face_count(&self, font_rid: Rid) -> i64

pub fn font_set_style(&mut self, font_rid: Rid, style: FontStyle)

pub fn font_get_style(&self, font_rid: Rid) -> FontStyle

pub fn font_set_name(&mut self, font_rid: Rid, name: GodotString)

pub fn font_get_name(&self, font_rid: Rid) -> GodotString

pub fn font_set_style_name(&mut self, font_rid: Rid, name: GodotString)

pub fn font_get_style_name(&self, font_rid: Rid) -> GodotString

pub fn font_set_weight(&mut self, font_rid: Rid, weight: i64)

pub fn font_get_weight(&self, font_rid: Rid) -> i64

pub fn font_set_stretch(&mut self, font_rid: Rid, weight: i64)

pub fn font_get_stretch(&self, font_rid: Rid) -> i64

pub fn font_set_antialiasing( &mut self, font_rid: Rid, antialiasing: FontAntialiasing )

pub fn font_get_antialiasing(&self, font_rid: Rid) -> FontAntialiasing

pub fn font_set_generate_mipmaps(&mut self, font_rid: Rid, generate_mipmaps: bool)

pub fn font_get_generate_mipmaps(&self, font_rid: Rid) -> bool

pub fn font_set_multichannel_signed_distance_field( &mut self, font_rid: Rid, msdf: bool )

pub fn font_is_multichannel_signed_distance_field(&self, font_rid: Rid) -> bool

pub fn font_set_msdf_pixel_range(&mut self, font_rid: Rid, msdf_pixel_range: i64)

pub fn font_get_msdf_pixel_range(&self, font_rid: Rid) -> i64

pub fn font_set_msdf_size(&mut self, font_rid: Rid, msdf_size: i64)

pub fn font_get_msdf_size(&self, font_rid: Rid) -> i64

pub fn font_set_fixed_size(&mut self, font_rid: Rid, fixed_size: i64)

pub fn font_get_fixed_size(&self, font_rid: Rid) -> i64

pub fn font_set_allow_system_fallback( &mut self, font_rid: Rid, allow_system_fallback: bool )

pub fn font_is_allow_system_fallback(&self, font_rid: Rid) -> bool

pub fn font_set_force_autohinter(&mut self, font_rid: Rid, force_autohinter: bool)

pub fn font_is_force_autohinter(&self, font_rid: Rid) -> bool

pub fn font_set_hinting(&mut self, font_rid: Rid, hinting: Hinting)

pub fn font_get_hinting(&self, font_rid: Rid) -> Hinting

pub fn font_set_subpixel_positioning( &mut self, font_rid: Rid, subpixel_positioning: SubpixelPositioning )

pub fn font_get_subpixel_positioning(&self, font_rid: Rid) -> SubpixelPositioning

pub fn font_set_embolden(&mut self, font_rid: Rid, strength: f64)

pub fn font_get_embolden(&self, font_rid: Rid) -> f64

pub fn font_set_transform(&mut self, font_rid: Rid, transform: Transform2D)

pub fn font_get_transform(&self, font_rid: Rid) -> Transform2D

pub fn font_set_variation_coordinates( &mut self, font_rid: Rid, variation_coordinates: Dictionary )

pub fn font_get_variation_coordinates(&self, font_rid: Rid) -> Dictionary

pub fn font_set_oversampling(&mut self, font_rid: Rid, oversampling: f64)

pub fn font_get_oversampling(&self, font_rid: Rid) -> f64

pub fn font_get_size_cache_list(&self, font_rid: Rid) -> Array<Vector2i>

pub fn font_clear_size_cache(&mut self, font_rid: Rid)

pub fn font_remove_size_cache(&mut self, font_rid: Rid, size: Vector2i)

pub fn font_set_ascent(&mut self, font_rid: Rid, size: i64, ascent: f64)

pub fn font_get_ascent(&self, font_rid: Rid, size: i64) -> f64

pub fn font_set_descent(&mut self, font_rid: Rid, size: i64, descent: f64)

pub fn font_get_descent(&self, font_rid: Rid, size: i64) -> f64

pub fn font_set_underline_position( &mut self, font_rid: Rid, size: i64, underline_position: f64 )

pub fn font_get_underline_position(&self, font_rid: Rid, size: i64) -> f64

pub fn font_set_underline_thickness( &mut self, font_rid: Rid, size: i64, underline_thickness: f64 )

pub fn font_get_underline_thickness(&self, font_rid: Rid, size: i64) -> f64

pub fn font_set_scale(&mut self, font_rid: Rid, size: i64, scale: f64)

pub fn font_get_scale(&self, font_rid: Rid, size: i64) -> f64

pub fn font_get_texture_count(&self, font_rid: Rid, size: Vector2i) -> i64

pub fn font_clear_textures(&mut self, font_rid: Rid, size: Vector2i)

pub fn font_remove_texture( &mut self, font_rid: Rid, size: Vector2i, texture_index: i64 )

pub fn font_set_texture_image( &mut self, font_rid: Rid, size: Vector2i, texture_index: i64, image: Gd<Image> )

pub fn font_get_texture_image( &self, font_rid: Rid, size: Vector2i, texture_index: i64 ) -> Option<Gd<Image>>

pub fn font_set_texture_offsets( &mut self, font_rid: Rid, size: Vector2i, texture_index: i64, offset: PackedInt32Array )

pub fn font_get_texture_offsets( &self, font_rid: Rid, size: Vector2i, texture_index: i64 ) -> PackedInt32Array

pub fn font_get_glyph_list( &self, font_rid: Rid, size: Vector2i ) -> PackedInt32Array

pub fn font_clear_glyphs(&mut self, font_rid: Rid, size: Vector2i)

pub fn font_remove_glyph(&mut self, font_rid: Rid, size: Vector2i, glyph: i64)

pub fn font_get_glyph_advance( &self, font_rid: Rid, size: i64, glyph: i64 ) -> Vector2

pub fn font_set_glyph_advance( &mut self, font_rid: Rid, size: i64, glyph: i64, advance: Vector2 )

pub fn font_get_glyph_offset( &self, font_rid: Rid, size: Vector2i, glyph: i64 ) -> Vector2

pub fn font_set_glyph_offset( &mut self, font_rid: Rid, size: Vector2i, glyph: i64, offset: Vector2 )

pub fn font_get_glyph_size( &self, font_rid: Rid, size: Vector2i, glyph: i64 ) -> Vector2

pub fn font_set_glyph_size( &mut self, font_rid: Rid, size: Vector2i, glyph: i64, gl_size: Vector2 )

pub fn font_get_glyph_uv_rect( &self, font_rid: Rid, size: Vector2i, glyph: i64 ) -> Rect2

pub fn font_set_glyph_uv_rect( &mut self, font_rid: Rid, size: Vector2i, glyph: i64, uv_rect: Rect2 )

pub fn font_get_glyph_texture_idx( &self, font_rid: Rid, size: Vector2i, glyph: i64 ) -> i64

pub fn font_set_glyph_texture_idx( &mut self, font_rid: Rid, size: Vector2i, glyph: i64, texture_idx: i64 )

pub fn font_get_glyph_texture_rid( &self, font_rid: Rid, size: Vector2i, glyph: i64 ) -> Rid

pub fn font_get_glyph_texture_size( &self, font_rid: Rid, size: Vector2i, glyph: i64 ) -> Vector2

pub fn font_get_glyph_contours( &self, font: Rid, size: i64, index: i64 ) -> Dictionary

pub fn font_get_kerning_list(&self, font_rid: Rid, size: i64) -> Array<Vector2i>

pub fn font_clear_kerning_map(&mut self, font_rid: Rid, size: i64)

pub fn font_remove_kerning( &mut self, font_rid: Rid, size: i64, glyph_pair: Vector2i )

pub fn font_set_kerning( &mut self, font_rid: Rid, size: i64, glyph_pair: Vector2i, kerning: Vector2 )

pub fn font_get_kerning( &self, font_rid: Rid, size: i64, glyph_pair: Vector2i ) -> Vector2

pub fn font_get_glyph_index( &self, font_rid: Rid, size: i64, char: i64, variation_selector: i64 ) -> i64

pub fn font_has_char(&self, font_rid: Rid, char: i64) -> bool

pub fn font_get_supported_chars(&self, font_rid: Rid) -> GodotString

pub fn font_render_range( &mut self, font_rid: Rid, size: Vector2i, start: i64, end: i64 )

pub fn font_render_glyph(&mut self, font_rid: Rid, size: Vector2i, index: i64)

pub fn font_draw_glyph( &self, font_rid: Rid, canvas: Rid, size: i64, pos: Vector2, index: i64, color: Color )

pub fn font_draw_glyph_outline( &self, font_rid: Rid, canvas: Rid, size: i64, outline_size: i64, pos: Vector2, index: i64, color: Color )

pub fn font_is_language_supported( &self, font_rid: Rid, language: GodotString ) -> bool

pub fn font_set_language_support_override( &mut self, font_rid: Rid, language: GodotString, supported: bool )

pub fn font_get_language_support_override( &mut self, font_rid: Rid, language: GodotString ) -> bool

pub fn font_remove_language_support_override( &mut self, font_rid: Rid, language: GodotString )

pub fn font_get_language_support_overrides( &mut self, font_rid: Rid ) -> PackedStringArray

pub fn font_is_script_supported(&self, font_rid: Rid, script: GodotString) -> bool

pub fn font_set_script_support_override( &mut self, font_rid: Rid, script: GodotString, supported: bool )

pub fn font_get_script_support_override( &mut self, font_rid: Rid, script: GodotString ) -> bool

pub fn font_remove_script_support_override( &mut self, font_rid: Rid, script: GodotString )

pub fn font_get_script_support_overrides( &mut self, font_rid: Rid ) -> PackedStringArray

pub fn font_set_opentype_feature_overrides( &mut self, font_rid: Rid, overrides: Dictionary )

pub fn font_get_opentype_feature_overrides(&self, font_rid: Rid) -> Dictionary

pub fn font_supported_feature_list(&self, font_rid: Rid) -> Dictionary

pub fn font_supported_variation_list(&self, font_rid: Rid) -> Dictionary

pub fn font_get_global_oversampling(&self) -> f64

pub fn font_set_global_oversampling(&mut self, oversampling: f64)

pub fn get_hex_code_box_size(&self, size: i64, index: i64) -> Vector2

pub fn draw_hex_code_box( &self, canvas: Rid, size: i64, pos: Vector2, index: i64, color: Color )

pub fn create_shaped_text( &mut self, direction: Direction, orientation: Orientation ) -> Rid

pub fn shaped_text_clear(&mut self, rid: Rid)

pub fn shaped_text_set_direction(&mut self, shaped: Rid, direction: Direction)

pub fn shaped_text_get_direction(&self, shaped: Rid) -> Direction

pub fn shaped_text_get_inferred_direction(&self, shaped: Rid) -> Direction

pub fn shaped_text_set_bidi_override( &mut self, shaped: Rid, override_: Array<Variant> )

pub fn shaped_text_set_custom_punctuation( &mut self, shaped: Rid, punct: GodotString )

pub fn shaped_text_get_custom_punctuation(&self, shaped: Rid) -> GodotString

pub fn shaped_text_set_orientation( &mut self, shaped: Rid, orientation: Orientation )

pub fn shaped_text_get_orientation(&self, shaped: Rid) -> Orientation

pub fn shaped_text_set_preserve_invalid(&mut self, shaped: Rid, enabled: bool)

pub fn shaped_text_get_preserve_invalid(&self, shaped: Rid) -> bool

pub fn shaped_text_set_preserve_control(&mut self, shaped: Rid, enabled: bool)

pub fn shaped_text_get_preserve_control(&self, shaped: Rid) -> bool

pub fn shaped_text_set_spacing( &mut self, shaped: Rid, spacing: SpacingType, value: i64 )

pub fn shaped_text_get_spacing(&self, shaped: Rid, spacing: SpacingType) -> i64

pub fn shaped_text_add_string( &mut self, shaped: Rid, text: GodotString, fonts: Array<Rid>, size: i64, opentype_features: Dictionary, language: GodotString, meta: Variant ) -> bool

pub fn shaped_text_add_object( &mut self, shaped: Rid, key: Variant, size: Vector2, inline_align: InlineAlignment, length: i64, baseline: f64 ) -> bool

pub fn shaped_text_resize_object( &mut self, shaped: Rid, key: Variant, size: Vector2, inline_align: InlineAlignment, baseline: f64 ) -> bool

pub fn shaped_get_span_count(&self, shaped: Rid) -> i64

pub fn shaped_get_span_meta(&self, shaped: Rid, index: i64) -> Variant

pub fn shaped_set_span_update_font( &mut self, shaped: Rid, index: i64, fonts: Array<Rid>, size: i64, opentype_features: Dictionary )

pub fn shaped_text_substr(&self, shaped: Rid, start: i64, length: i64) -> Rid

pub fn shaped_text_get_parent(&self, shaped: Rid) -> Rid

pub fn shaped_text_fit_to_width( &mut self, shaped: Rid, width: f64, jst_flags: JustificationFlag ) -> f64

pub fn shaped_text_tab_align( &mut self, shaped: Rid, tab_stops: PackedFloat32Array ) -> f64

pub fn shaped_text_shape(&mut self, shaped: Rid) -> bool

pub fn shaped_text_is_ready(&self, shaped: Rid) -> bool

pub fn shaped_text_get_glyphs(&self, shaped: Rid) -> Array<Dictionary>

pub fn shaped_text_sort_logical(&mut self, shaped: Rid) -> Array<Dictionary>

pub fn shaped_text_get_glyph_count(&self, shaped: Rid) -> i64

pub fn shaped_text_get_range(&self, shaped: Rid) -> Vector2i

pub fn shaped_text_get_line_breaks_adv( &self, shaped: Rid, width: PackedFloat32Array, start: i64, once: bool, break_flags: LineBreakFlag ) -> PackedInt32Array

pub fn shaped_text_get_line_breaks( &self, shaped: Rid, width: f64, start: i64, break_flags: LineBreakFlag ) -> PackedInt32Array

pub fn shaped_text_get_word_breaks( &self, shaped: Rid, grapheme_flags: GraphemeFlag ) -> PackedInt32Array

pub fn shaped_text_get_trim_pos(&self, shaped: Rid) -> i64

pub fn shaped_text_get_ellipsis_pos(&self, shaped: Rid) -> i64

pub fn shaped_text_get_ellipsis_glyphs(&self, shaped: Rid) -> Array<Dictionary>

pub fn shaped_text_get_ellipsis_glyph_count(&self, shaped: Rid) -> i64

pub fn shaped_text_overrun_trim_to_width( &mut self, shaped: Rid, width: f64, overrun_trim_flags: TextOverrunFlag )

pub fn shaped_text_get_objects(&self, shaped: Rid) -> Array<Variant>

pub fn shaped_text_get_object_rect(&self, shaped: Rid, key: Variant) -> Rect2

pub fn shaped_text_get_size(&self, shaped: Rid) -> Vector2

pub fn shaped_text_get_ascent(&self, shaped: Rid) -> f64

pub fn shaped_text_get_descent(&self, shaped: Rid) -> f64

pub fn shaped_text_get_width(&self, shaped: Rid) -> f64

pub fn shaped_text_get_underline_position(&self, shaped: Rid) -> f64

pub fn shaped_text_get_underline_thickness(&self, shaped: Rid) -> f64

pub fn shaped_text_get_carets(&self, shaped: Rid, position: i64) -> Dictionary

pub fn shaped_text_get_selection( &self, shaped: Rid, start: i64, end: i64 ) -> PackedVector2Array

pub fn shaped_text_hit_test_grapheme(&self, shaped: Rid, coords: f64) -> i64

pub fn shaped_text_hit_test_position(&self, shaped: Rid, coords: f64) -> i64

pub fn shaped_text_get_grapheme_bounds(&self, shaped: Rid, pos: i64) -> Vector2

pub fn shaped_text_next_grapheme_pos(&self, shaped: Rid, pos: i64) -> i64

pub fn shaped_text_prev_grapheme_pos(&self, shaped: Rid, pos: i64) -> i64

pub fn shaped_text_draw( &self, shaped: Rid, canvas: Rid, pos: Vector2, clip_l: f64, clip_r: f64, color: Color )

pub fn shaped_text_draw_outline( &self, shaped: Rid, canvas: Rid, pos: Vector2, clip_l: f64, clip_r: f64, outline_size: i64, color: Color )

pub fn shaped_text_get_dominant_direction_in_range( &self, shaped: Rid, start: i64, end: i64 ) -> Direction

pub fn format_number( &self, number: GodotString, language: GodotString ) -> GodotString

pub fn parse_number( &self, number: GodotString, language: GodotString ) -> GodotString

pub fn percent_sign(&self, language: GodotString) -> GodotString

pub fn string_get_word_breaks( &self, string: GodotString, language: GodotString, chars_per_line: i64 ) -> PackedInt32Array

pub fn is_confusable(&self, string: GodotString, dict: PackedStringArray) -> i64

pub fn spoof_check(&self, string: GodotString) -> bool

pub fn strip_diacritics(&self, string: GodotString) -> GodotString

pub fn is_valid_identifier(&self, string: GodotString) -> bool

pub fn string_to_upper( &self, string: GodotString, language: GodotString ) -> GodotString

pub fn string_to_lower( &self, string: GodotString, language: GodotString ) -> GodotString

pub fn parse_structured_text( &self, parser_type: StructuredTextParser, args: Array<Variant>, text: GodotString ) -> Array<Vector3i>