Struct godot::engine::TileSet

pub struct TileSet { /* private fields */ }

Godot class TileSet.

Inherits Resource.

Related symbols:

tile_set: sidecar module with related enum/flag types

TileSetVirtual: virtual methods

See also Godot docs for TileSet.

Implementations

impl TileSet

pub fn new() -> Gd<TileSet>

pub fn notify(&mut self, what: ObjectNotification)

⚠️ Sends a Godot notification to all classes inherited by the object.

Triggers calls to on_notification(), and depending on the notification, also to Godot’s lifecycle callbacks such as ready().

Starts from the highest ancestor (the Object class) and goes down the hierarchy. See also Godot docs for Object::notification().

Panics

If you call this method on a user-defined object while holding a GdRef or GdMut guard on the instance, you will encounter a panic. The reason is that the receiving virtual method on_notification() acquires a GdMut lock dynamically, which must be exclusive.

pub fn notify_reversed(&mut self, what: ObjectNotification)

⚠️ Like Self::notify(), but starts at the most-derived class and goes up the hierarchy.

See docs of that method, including the panics.

pub fn get_next_source_id(&self) -> i64

pub fn add_source( &mut self, source: Gd<TileSetSource>, atlas_source_id_override: i64 ) -> i64

pub fn remove_source(&mut self, source_id: i64)

pub fn set_source_id(&mut self, source_id: i64, new_source_id: i64)

pub fn get_source_count(&self) -> i64

pub fn get_source_id(&self, index: i64) -> i64

pub fn has_source(&self, source_id: i64) -> bool

pub fn get_source(&self, source_id: i64) -> Option<Gd<TileSetSource>>

pub fn set_tile_shape(&mut self, shape: TileShape)

pub fn get_tile_shape(&self) -> TileShape

pub fn set_tile_layout(&mut self, layout: TileLayout)

pub fn get_tile_layout(&self) -> TileLayout

pub fn set_tile_offset_axis(&mut self, alignment: TileOffsetAxis)

pub fn get_tile_offset_axis(&self) -> TileOffsetAxis

pub fn set_tile_size(&mut self, size: Vector2i)

pub fn get_tile_size(&self) -> Vector2i

pub fn set_uv_clipping(&mut self, uv_clipping: bool)

pub fn is_uv_clipping(&self) -> bool

pub fn get_occlusion_layers_count(&self) -> i64

pub fn add_occlusion_layer(&mut self, to_position: i64)

pub fn move_occlusion_layer(&mut self, layer_index: i64, to_position: i64)

pub fn remove_occlusion_layer(&mut self, layer_index: i64)

pub fn set_occlusion_layer_light_mask( &mut self, layer_index: i64, light_mask: i64 )

pub fn get_occlusion_layer_light_mask(&self, layer_index: i64) -> i64

pub fn set_occlusion_layer_sdf_collision( &mut self, layer_index: i64, sdf_collision: bool )

pub fn get_occlusion_layer_sdf_collision(&self, layer_index: i64) -> bool

pub fn get_physics_layers_count(&self) -> i64

pub fn add_physics_layer(&mut self, to_position: i64)

pub fn move_physics_layer(&mut self, layer_index: i64, to_position: i64)

pub fn remove_physics_layer(&mut self, layer_index: i64)

pub fn set_physics_layer_collision_layer(&mut self, layer_index: i64, layer: i64)

pub fn get_physics_layer_collision_layer(&self, layer_index: i64) -> i64

pub fn set_physics_layer_collision_mask(&mut self, layer_index: i64, mask: i64)

pub fn get_physics_layer_collision_mask(&self, layer_index: i64) -> i64

pub fn set_physics_layer_physics_material( &mut self, layer_index: i64, physics_material: Gd<PhysicsMaterial> )

pub fn get_physics_layer_physics_material( &self, layer_index: i64 ) -> Option<Gd<PhysicsMaterial>>

pub fn get_terrain_sets_count(&self) -> i64

pub fn add_terrain_set(&mut self, to_position: i64)

pub fn move_terrain_set(&mut self, terrain_set: i64, to_position: i64)

pub fn remove_terrain_set(&mut self, terrain_set: i64)

pub fn set_terrain_set_mode(&mut self, terrain_set: i64, mode: TerrainMode)

pub fn get_terrain_set_mode(&self, terrain_set: i64) -> TerrainMode

pub fn get_terrains_count(&self, terrain_set: i64) -> i64

pub fn add_terrain(&mut self, terrain_set: i64, to_position: i64)

pub fn move_terrain( &mut self, terrain_set: i64, terrain_index: i64, to_position: i64 )

pub fn remove_terrain(&mut self, terrain_set: i64, terrain_index: i64)

pub fn set_terrain_name( &mut self, terrain_set: i64, terrain_index: i64, name: GodotString )

pub fn get_terrain_name( &self, terrain_set: i64, terrain_index: i64 ) -> GodotString

pub fn set_terrain_color( &mut self, terrain_set: i64, terrain_index: i64, color: Color )

pub fn get_terrain_color(&self, terrain_set: i64, terrain_index: i64) -> Color

pub fn get_navigation_layers_count(&self) -> i64

pub fn add_navigation_layer(&mut self, to_position: i64)

pub fn move_navigation_layer(&mut self, layer_index: i64, to_position: i64)

pub fn remove_navigation_layer(&mut self, layer_index: i64)

pub fn set_navigation_layer_layers(&mut self, layer_index: i64, layers: i64)

pub fn get_navigation_layer_layers(&self, layer_index: i64) -> i64

pub fn set_navigation_layer_layer_value( &mut self, layer_index: i64, layer_number: i64, value: bool )

pub fn get_navigation_layer_layer_value( &self, layer_index: i64, layer_number: i64 ) -> bool

pub fn get_custom_data_layers_count(&self) -> i64

pub fn add_custom_data_layer(&mut self, to_position: i64)

pub fn move_custom_data_layer(&mut self, layer_index: i64, to_position: i64)

pub fn remove_custom_data_layer(&mut self, layer_index: i64)

pub fn get_custom_data_layer_by_name(&self, layer_name: GodotString) -> i64

pub fn set_custom_data_layer_name( &mut self, layer_index: i64, layer_name: GodotString )

pub fn get_custom_data_layer_name(&self, layer_index: i64) -> GodotString

pub fn set_custom_data_layer_type( &mut self, layer_index: i64, layer_type: VariantType )

pub fn get_custom_data_layer_type(&self, layer_index: i64) -> VariantType

pub fn set_source_level_tile_proxy(&mut self, source_from: i64, source_to: i64)

pub fn get_source_level_tile_proxy(&mut self, source_from: i64) -> i64

pub fn has_source_level_tile_proxy(&mut self, source_from: i64) -> bool

pub fn remove_source_level_tile_proxy(&mut self, source_from: i64)

pub fn set_coords_level_tile_proxy( &mut self, p_source_from: i64, coords_from: Vector2i, source_to: i64, coords_to: Vector2i )

pub fn get_coords_level_tile_proxy( &mut self, source_from: i64, coords_from: Vector2i ) -> Array<Variant>

pub fn has_coords_level_tile_proxy( &mut self, source_from: i64, coords_from: Vector2i ) -> bool

pub fn remove_coords_level_tile_proxy( &mut self, source_from: i64, coords_from: Vector2i )

pub fn set_alternative_level_tile_proxy( &mut self, source_from: i64, coords_from: Vector2i, alternative_from: i64, source_to: i64, coords_to: Vector2i, alternative_to: i64 )

pub fn get_alternative_level_tile_proxy( &mut self, source_from: i64, coords_from: Vector2i, alternative_from: i64 ) -> Array<Variant>

pub fn has_alternative_level_tile_proxy( &mut self, source_from: i64, coords_from: Vector2i, alternative_from: i64 ) -> bool

pub fn remove_alternative_level_tile_proxy( &mut self, source_from: i64, coords_from: Vector2i, alternative_from: i64 )

pub fn map_tile_proxy( &self, source_from: i64, coords_from: Vector2i, alternative_from: i64 ) -> Array<Variant>

pub fn cleanup_invalid_tile_proxies(&mut self)

pub fn clear_tile_proxies(&mut self)

pub fn add_pattern(&mut self, pattern: Gd<TileMapPattern>, index: i64) -> i64

pub fn get_pattern(&mut self, index: i64) -> Option<Gd<TileMapPattern>>

pub fn remove_pattern(&mut self, index: i64)

pub fn get_patterns_count(&mut self) -> i64