Struct godot::engine::TreeItem

pub struct TreeItem { /* private fields */ }

Godot class TreeItem.

Inherits Object.

Related symbols:

tree_item: sidecar module with related enum/flag types

TreeItemVirtual: virtual methods

See also Godot docs for TreeItem.

Implementations

impl TreeItem

pub fn notify(&mut self, what: ObjectNotification)

⚠️ Sends a Godot notification to all classes inherited by the object.

Triggers calls to on_notification(), and depending on the notification, also to Godot’s lifecycle callbacks such as ready().

Starts from the highest ancestor (the Object class) and goes down the hierarchy. See also Godot docs for Object::notification().

Panics

If you call this method on a user-defined object while holding a GdRef or GdMut guard on the instance, you will encounter a panic. The reason is that the receiving virtual method on_notification() acquires a GdMut lock dynamically, which must be exclusive.

pub fn notify_reversed(&mut self, what: ObjectNotification)

⚠️ Like Self::notify(), but starts at the most-derived class and goes up the hierarchy.

See docs of that method, including the panics.

pub fn set_cell_mode(&mut self, column: i64, mode: TreeCellMode)

pub fn get_cell_mode(&self, column: i64) -> TreeCellMode

pub fn set_checked(&mut self, column: i64, checked: bool)

pub fn set_indeterminate(&mut self, column: i64, indeterminate: bool)

pub fn is_checked(&self, column: i64) -> bool

pub fn is_indeterminate(&self, column: i64) -> bool

pub fn propagate_check(&mut self, column: i64, emit_signal: bool)

pub fn set_text(&mut self, column: i64, text: GodotString)

pub fn get_text(&self, column: i64) -> GodotString

pub fn set_text_direction(&mut self, column: i64, direction: TextDirection)

pub fn get_text_direction(&self, column: i64) -> TextDirection

pub fn set_structured_text_bidi_override( &mut self, column: i64, parser: StructuredTextParser )

pub fn get_structured_text_bidi_override( &self, column: i64 ) -> StructuredTextParser

pub fn set_structured_text_bidi_override_options( &mut self, column: i64, args: Array<Variant> )

pub fn get_structured_text_bidi_override_options( &self, column: i64 ) -> Array<Variant>

pub fn set_language(&mut self, column: i64, language: GodotString)

pub fn get_language(&self, column: i64) -> GodotString

pub fn set_suffix(&mut self, column: i64, text: GodotString)

pub fn get_suffix(&self, column: i64) -> GodotString

pub fn set_icon(&mut self, column: i64, texture: Gd<Texture2D>)

pub fn get_icon(&self, column: i64) -> Option<Gd<Texture2D>>

pub fn set_icon_region(&mut self, column: i64, region: Rect2)

pub fn get_icon_region(&self, column: i64) -> Rect2

pub fn set_icon_max_width(&mut self, column: i64, width: i64)

pub fn get_icon_max_width(&self, column: i64) -> i64

pub fn set_icon_modulate(&mut self, column: i64, modulate: Color)

pub fn get_icon_modulate(&self, column: i64) -> Color

pub fn set_range(&mut self, column: i64, value: f64)

pub fn get_range(&self, column: i64) -> f64

pub fn set_range_config( &mut self, column: i64, min: f64, max: f64, step: f64, expr: bool )

pub fn get_range_config(&mut self, column: i64) -> Dictionary

pub fn set_metadata(&mut self, column: i64, meta: Variant)

pub fn get_metadata(&self, column: i64) -> Variant

pub fn set_custom_draw( &mut self, column: i64, object: Gd<Object>, callback: StringName )

pub fn set_collapsed(&mut self, enable: bool)

pub fn is_collapsed(&mut self) -> bool

pub fn set_collapsed_recursive(&mut self, enable: bool)

pub fn is_any_collapsed(&mut self, only_visible: bool) -> bool

pub fn set_visible(&mut self, enable: bool)

pub fn is_visible(&mut self) -> bool

pub fn uncollapse_tree(&mut self)

pub fn set_custom_minimum_height(&mut self, height: i64)

pub fn get_custom_minimum_height(&self) -> i64

pub fn set_selectable(&mut self, column: i64, selectable: bool)

pub fn is_selectable(&self, column: i64) -> bool

pub fn is_selected(&mut self, column: i64) -> bool

pub fn select(&mut self, column: i64)

pub fn deselect(&mut self, column: i64)

pub fn set_editable(&mut self, column: i64, enabled: bool)

pub fn is_editable(&mut self, column: i64) -> bool

pub fn set_custom_color(&mut self, column: i64, color: Color)

pub fn get_custom_color(&self, column: i64) -> Color

pub fn clear_custom_color(&mut self, column: i64)

pub fn set_custom_font(&mut self, column: i64, font: Gd<Font>)

pub fn get_custom_font(&self, column: i64) -> Option<Gd<Font>>

pub fn set_custom_font_size(&mut self, column: i64, font_size: i64)

pub fn get_custom_font_size(&self, column: i64) -> i64

pub fn set_custom_bg_color( &mut self, column: i64, color: Color, just_outline: bool )

pub fn clear_custom_bg_color(&mut self, column: i64)

pub fn get_custom_bg_color(&self, column: i64) -> Color

pub fn set_custom_as_button(&mut self, column: i64, enable: bool)

pub fn is_custom_set_as_button(&self, column: i64) -> bool

pub fn add_button( &mut self, column: i64, button: Gd<Texture2D>, id: i64, disabled: bool, tooltip_text: GodotString )

pub fn get_button_count(&self, column: i64) -> i64

pub fn get_button_tooltip_text( &self, column: i64, button_index: i64 ) -> GodotString

pub fn get_button_id(&self, column: i64, button_index: i64) -> i64

pub fn get_button_by_id(&self, column: i64, id: i64) -> i64

pub fn get_button(&self, column: i64, button_index: i64) -> Option<Gd<Texture2D>>

pub fn set_button(&mut self, column: i64, button_index: i64, button: Gd<Texture2D>)

pub fn erase_button(&mut self, column: i64, button_index: i64)

pub fn set_button_disabled( &mut self, column: i64, button_index: i64, disabled: bool )

pub fn set_button_color(&mut self, column: i64, button_index: i64, color: Color)

pub fn is_button_disabled(&self, column: i64, button_index: i64) -> bool

pub fn set_tooltip_text(&mut self, column: i64, tooltip: GodotString)

pub fn get_tooltip_text(&self, column: i64) -> GodotString

pub fn set_text_alignment( &mut self, column: i64, text_alignment: HorizontalAlignment )

pub fn get_text_alignment(&self, column: i64) -> HorizontalAlignment

pub fn set_expand_right(&mut self, column: i64, enable: bool)

pub fn get_expand_right(&self, column: i64) -> bool

pub fn set_disable_folding(&mut self, disable: bool)

pub fn is_folding_disabled(&self) -> bool

pub fn create_child(&mut self, index: i64) -> Option<Gd<TreeItem>>

pub fn get_tree(&self) -> Option<Gd<Tree>>

pub fn get_next(&self) -> Option<Gd<TreeItem>>

pub fn get_prev(&mut self) -> Option<Gd<TreeItem>>

pub fn get_parent(&self) -> Option<Gd<TreeItem>>

pub fn get_first_child(&self) -> Option<Gd<TreeItem>>

pub fn get_next_visible(&mut self, wrap: bool) -> Option<Gd<TreeItem>>

pub fn get_prev_visible(&mut self, wrap: bool) -> Option<Gd<TreeItem>>

pub fn get_child(&mut self, index: i64) -> Option<Gd<TreeItem>>

pub fn get_child_count(&mut self) -> i64

pub fn get_children(&mut self) -> Array<Gd<TreeItem>>

pub fn get_index(&mut self) -> i64

pub fn move_before(&mut self, item: Gd<TreeItem>)

pub fn move_after(&mut self, item: Gd<TreeItem>)

pub fn remove_child(&mut self, child: Gd<TreeItem>)

pub fn call_recursive(&mut self, method: StringName, varargs: &[Variant])