Trait godot::engine::EditorSceneFormatImporterGltfVirtual

pub trait EditorSceneFormatImporterGltfVirtual: GodotClass + You_forgot_the_attribute__godot_api {
    // Provided methods
    fn init(base: Base<Self::Base>) -> Self { ... }
    fn to_string(&self) -> GodotString { ... }
    fn on_notification(&mut self, what: ObjectNotification) { ... }
    fn get_import_flags(&self) -> i64 { ... }
    fn get_extensions(&self) -> PackedStringArray { ... }
    fn import_scene(
        &mut self,
        path: GodotString,
        flags: i64,
        options: Dictionary
    ) -> Option<Gd<Object>> { ... }
    fn get_import_options(&mut self, path: GodotString) { ... }
    fn get_option_visibility(
        &self,
        path: GodotString,
        for_animation: bool,
        option: GodotString
    ) -> Variant { ... }
}

Virtual methods for class EditorSceneFormatImporterGltf.

These methods represent constructors (init) or callbacks invoked by the engine.

See also Godot docs for EditorSceneFormatImporterGLTF methods.

Provided Methods

fn init(base: Base<Self::Base>) -> Self

Godot constructor, accepting an injected base object.

base refers to the base instance of the class, which can either be stored in a #[base] field or discarded. This method returns a fully-constructed instance, which will then be moved into a Gd<T> pointer.

If the class has a #[class(init)] attribute, this method will be auto-generated and must not be overridden.

fn to_string(&self) -> GodotString

String representation of the Godot instance.

Override this method to define how the instance is represented as a string. Used by impl Display for Gd<T>, as well as str() and print() in GDScript.

fn on_notification(&mut self, what: ObjectNotification)

Called when the object receives a Godot notification.

The type of notification can be identified through what. The enum is designed to hold all possible NOTIFICATION_* constants that the current class can handle. However, this is not validated in Godot, so an enum variant Unknown exists to represent integers out of known constants (mistakes or future additions).

This method is named _notification in Godot, but on_notification in Rust. To send notifications, use the Object::notify method.

See also in Godot docs:

Object::_notification.

Notifications tutorial.

fn get_import_flags(&self) -> i64

fn get_extensions(&self) -> PackedStringArray

fn import_scene( &mut self, path: GodotString, flags: i64, options: Dictionary ) -> Option<Gd<Object>>

fn get_import_options(&mut self, path: GodotString)

fn get_option_visibility( &self, path: GodotString, for_animation: bool, option: GodotString ) -> Variant