Trait godot::engine::IPhysicsDirectBodyState2DExtension

pub trait IPhysicsDirectBodyState2DExtension: GodotClass + You_forgot_the_attribute__godot_api {

    // Provided methods
    fn init(base: Base<Self::Base>) -> Self { ... }
    fn to_string(&self) -> GString { ... }
    fn on_notification(&mut self, what: ObjectNotification) { ... }
    fn get_property(&self, property: StringName) -> Option<Variant> { ... }
    fn set_property(&mut self, property: StringName, value: Variant) -> bool { ... }
    fn get_total_gravity(&self) -> Vector2 { ... }
    fn get_total_linear_damp(&self) -> f32 { ... }
    fn get_total_angular_damp(&self) -> f32 { ... }
    fn get_center_of_mass(&self) -> Vector2 { ... }
    fn get_center_of_mass_local(&self) -> Vector2 { ... }
    fn get_inverse_mass(&self) -> f32 { ... }
    fn get_inverse_inertia(&self) -> f32 { ... }
    fn set_linear_velocity(&mut self, velocity: Vector2) { ... }
    fn get_linear_velocity(&self) -> Vector2 { ... }
    fn set_angular_velocity(&mut self, velocity: f32) { ... }
    fn get_angular_velocity(&self) -> f32 { ... }
    fn set_transform(&mut self, transform: Transform2D) { ... }
    fn get_transform(&self) -> Transform2D { ... }
    fn get_velocity_at_local_position(&self, local_position: Vector2) -> Vector2 { ... }
    fn apply_central_impulse(&mut self, impulse: Vector2) { ... }
    fn apply_impulse(&mut self, impulse: Vector2, position: Vector2) { ... }
    fn apply_torque_impulse(&mut self, impulse: f32) { ... }
    fn apply_central_force(&mut self, force: Vector2) { ... }
    fn apply_force(&mut self, force: Vector2, position: Vector2) { ... }
    fn apply_torque(&mut self, torque: f32) { ... }
    fn add_constant_central_force(&mut self, force: Vector2) { ... }
    fn add_constant_force(&mut self, force: Vector2, position: Vector2) { ... }
    fn add_constant_torque(&mut self, torque: f32) { ... }
    fn set_constant_force(&mut self, force: Vector2) { ... }
    fn get_constant_force(&self) -> Vector2 { ... }
    fn set_constant_torque(&mut self, torque: f32) { ... }
    fn get_constant_torque(&self) -> f32 { ... }
    fn set_sleep_state(&mut self, enabled: bool) { ... }
    fn is_sleeping(&self) -> bool { ... }
    fn get_contact_count(&self) -> i32 { ... }
    fn get_contact_local_position(&self, contact_idx: i32) -> Vector2 { ... }
    fn get_contact_local_normal(&self, contact_idx: i32) -> Vector2 { ... }
    fn get_contact_local_shape(&self, contact_idx: i32) -> i32 { ... }
    fn get_contact_local_velocity_at_position(
        &self,
        contact_idx: i32
    ) -> Vector2 { ... }
    fn get_contact_collider(&self, contact_idx: i32) -> Rid { ... }
    fn get_contact_collider_position(&self, contact_idx: i32) -> Vector2 { ... }
    fn get_contact_collider_id(&self, contact_idx: i32) -> u64 { ... }
    fn get_contact_collider_object(
        &self,
        contact_idx: i32
    ) -> Option<Gd<Object>> { ... }
    fn get_contact_collider_shape(&self, contact_idx: i32) -> i32 { ... }
    fn get_contact_collider_velocity_at_position(
        &self,
        contact_idx: i32
    ) -> Vector2 { ... }
    fn get_contact_impulse(&self, contact_idx: i32) -> Vector2 { ... }
    fn get_step(&self) -> f32 { ... }
    fn integrate_forces(&mut self) { ... }
    fn get_space_state(&mut self) -> Option<Gd<PhysicsDirectSpaceState2D>> { ... }

}

Virtual methods for class PhysicsDirectBodyState2DExtension.

These methods represent constructors (init) or callbacks invoked by the engine.

See also Godot docs for PhysicsDirectBodyState2DExtension methods.

Provided Methods

fn init(base: Base<Self::Base>) -> Self

Godot constructor, accepting an injected base object.

base refers to the base instance of the class, which can either be stored in a Base<T> field or discarded. This method returns a fully-constructed instance, which will then be moved into a Gd<T> pointer.

If the class has a #[class(init)] attribute, this method will be auto-generated and must not be overridden.

fn to_string(&self) -> GString

String representation of the Godot instance.

Override this method to define how the instance is represented as a string. Used by impl Display for Gd<T>, as well as str() and print() in GDScript.

fn on_notification(&mut self, what: ObjectNotification)

Called when the object receives a Godot notification.

The type of notification can be identified through what. The enum is designed to hold all possible NOTIFICATION_* constants that the current class can handle. However, this is not validated in Godot, so an enum variant Unknown exists to represent integers out of known constants (mistakes or future additions).

This method is named _notification in Godot, but on_notification in Rust. To send notifications, use the Object::notify method.

See also in Godot docs:

Object::_notification.

Notifications tutorial.

fn get_property(&self, property: StringName) -> Option<Variant>

Called whenever get() is called or Godot gets the value of a property.

Should return the given property’s value as Some(value), or None if the property should be handled normally.

See also in Godot docs:

Object::_get.

fn set_property(&mut self, property: StringName, value: Variant) -> bool

Called whenever Godot set() is called or Godot sets the value of a property.

Should set property to the given value and return true, or return false to indicate the property should be handled normally.

See also in Godot docs:

Object::_set.

fn get_total_gravity(&self) -> Vector2

fn get_total_linear_damp(&self) -> f32

fn get_total_angular_damp(&self) -> f32

fn get_center_of_mass(&self) -> Vector2

fn get_center_of_mass_local(&self) -> Vector2

fn get_inverse_mass(&self) -> f32

fn get_inverse_inertia(&self) -> f32

fn set_linear_velocity(&mut self, velocity: Vector2)

fn get_linear_velocity(&self) -> Vector2

fn set_angular_velocity(&mut self, velocity: f32)

fn get_angular_velocity(&self) -> f32

fn set_transform(&mut self, transform: Transform2D)

fn get_transform(&self) -> Transform2D

fn get_velocity_at_local_position(&self, local_position: Vector2) -> Vector2

fn apply_central_impulse(&mut self, impulse: Vector2)

fn apply_impulse(&mut self, impulse: Vector2, position: Vector2)

fn apply_torque_impulse(&mut self, impulse: f32)

fn apply_central_force(&mut self, force: Vector2)

fn apply_force(&mut self, force: Vector2, position: Vector2)

fn apply_torque(&mut self, torque: f32)

fn add_constant_central_force(&mut self, force: Vector2)

fn add_constant_force(&mut self, force: Vector2, position: Vector2)

fn add_constant_torque(&mut self, torque: f32)

fn set_constant_force(&mut self, force: Vector2)

fn get_constant_force(&self) -> Vector2

fn set_constant_torque(&mut self, torque: f32)

fn get_constant_torque(&self) -> f32

fn set_sleep_state(&mut self, enabled: bool)

fn is_sleeping(&self) -> bool

fn get_contact_count(&self) -> i32

fn get_contact_local_position(&self, contact_idx: i32) -> Vector2

fn get_contact_local_normal(&self, contact_idx: i32) -> Vector2

fn get_contact_local_shape(&self, contact_idx: i32) -> i32

fn get_contact_local_velocity_at_position(&self, contact_idx: i32) -> Vector2

fn get_contact_collider(&self, contact_idx: i32) -> Rid

fn get_contact_collider_position(&self, contact_idx: i32) -> Vector2

fn get_contact_collider_id(&self, contact_idx: i32) -> u64

fn get_contact_collider_object(&self, contact_idx: i32) -> Option<Gd<Object>>

fn get_contact_collider_shape(&self, contact_idx: i32) -> i32

fn get_contact_collider_velocity_at_position(&self, contact_idx: i32) -> Vector2

fn get_contact_impulse(&self, contact_idx: i32) -> Vector2

fn get_step(&self) -> f32

fn integrate_forces(&mut self)

fn get_space_state(&mut self) -> Option<Gd<PhysicsDirectSpaceState2D>>

Object Safety

This trait is not object safe.