Trait godot::engine::IPhysicsDirectSpaceState2DExtension

pub trait IPhysicsDirectSpaceState2DExtension: GodotClass + You_forgot_the_attribute__godot_api {
    // Provided methods
    fn init(base: Base<Self::Base>) -> Self { ... }
    fn to_string(&self) -> GString { ... }
    fn on_notification(&mut self, what: ObjectNotification) { ... }
    unsafe fn intersect_ray(
        &mut self,
        from: Vector2,
        to: Vector2,
        collision_mask: u32,
        collide_with_bodies: bool,
        collide_with_areas: bool,
        hit_from_inside: bool,
        result: *mut PhysicsServer2DExtensionRayResult
    ) -> bool { ... }
    unsafe fn intersect_point(
        &mut self,
        position: Vector2,
        canvas_instance_id: u64,
        collision_mask: u32,
        collide_with_bodies: bool,
        collide_with_areas: bool,
        results: *mut PhysicsServer2DExtensionShapeResult,
        max_results: i32
    ) -> i32 { ... }
    unsafe fn intersect_shape(
        &mut self,
        shape_rid: Rid,
        transform: Transform2D,
        motion: Vector2,
        margin: f32,
        collision_mask: u32,
        collide_with_bodies: bool,
        collide_with_areas: bool,
        result: *mut PhysicsServer2DExtensionShapeResult,
        max_results: i32
    ) -> i32 { ... }
    unsafe fn cast_motion(
        &mut self,
        shape_rid: Rid,
        transform: Transform2D,
        motion: Vector2,
        margin: f32,
        collision_mask: u32,
        collide_with_bodies: bool,
        collide_with_areas: bool,
        closest_safe: *mut f64,
        closest_unsafe: *mut f64
    ) -> bool { ... }
    unsafe fn collide_shape(
        &mut self,
        shape_rid: Rid,
        transform: Transform2D,
        motion: Vector2,
        margin: f32,
        collision_mask: u32,
        collide_with_bodies: bool,
        collide_with_areas: bool,
        results: *mut c_void,
        max_results: i32,
        result_count: *mut i32
    ) -> bool { ... }
    unsafe fn rest_info(
        &mut self,
        shape_rid: Rid,
        transform: Transform2D,
        motion: Vector2,
        margin: f32,
        collision_mask: u32,
        collide_with_bodies: bool,
        collide_with_areas: bool,
        rest_info: *mut PhysicsServer2DExtensionShapeRestInfo
    ) -> bool { ... }
}

Virtual methods for class PhysicsDirectSpaceState2DExtension.

These methods represent constructors (init) or callbacks invoked by the engine.

See also Godot docs for PhysicsDirectSpaceState2DExtension methods.

Provided Methods

fn init(base: Base<Self::Base>) -> Self

Godot constructor, accepting an injected base object.

base refers to the base instance of the class, which can either be stored in a #[base] field or discarded. This method returns a fully-constructed instance, which will then be moved into a Gd<T> pointer.

If the class has a #[class(init)] attribute, this method will be auto-generated and must not be overridden.

fn to_string(&self) -> GString

String representation of the Godot instance.

Override this method to define how the instance is represented as a string. Used by impl Display for Gd<T>, as well as str() and print() in GDScript.

fn on_notification(&mut self, what: ObjectNotification)

Called when the object receives a Godot notification.

The type of notification can be identified through what. The enum is designed to hold all possible NOTIFICATION_* constants that the current class can handle. However, this is not validated in Godot, so an enum variant Unknown exists to represent integers out of known constants (mistakes or future additions).

This method is named _notification in Godot, but on_notification in Rust. To send notifications, use the Object::notify method.

See also in Godot docs:

Object::_notification.

Notifications tutorial.

unsafe fn intersect_ray( &mut self, from: Vector2, to: Vector2, collision_mask: u32, collide_with_bodies: bool, collide_with_areas: bool, hit_from_inside: bool, result: *mut PhysicsServer2DExtensionRayResult ) -> bool

Safety

Godot currently does not document safety requirements on this method. Make sure you understand the underlying semantics.

unsafe fn intersect_point( &mut self, position: Vector2, canvas_instance_id: u64, collision_mask: u32, collide_with_bodies: bool, collide_with_areas: bool, results: *mut PhysicsServer2DExtensionShapeResult, max_results: i32 ) -> i32

Safety

Godot currently does not document safety requirements on this method. Make sure you understand the underlying semantics.

unsafe fn intersect_shape( &mut self, shape_rid: Rid, transform: Transform2D, motion: Vector2, margin: f32, collision_mask: u32, collide_with_bodies: bool, collide_with_areas: bool, result: *mut PhysicsServer2DExtensionShapeResult, max_results: i32 ) -> i32

Safety

Godot currently does not document safety requirements on this method. Make sure you understand the underlying semantics.

unsafe fn cast_motion( &mut self, shape_rid: Rid, transform: Transform2D, motion: Vector2, margin: f32, collision_mask: u32, collide_with_bodies: bool, collide_with_areas: bool, closest_safe: *mut f64, closest_unsafe: *mut f64 ) -> bool

Safety

Godot currently does not document safety requirements on this method. Make sure you understand the underlying semantics.

unsafe fn collide_shape( &mut self, shape_rid: Rid, transform: Transform2D, motion: Vector2, margin: f32, collision_mask: u32, collide_with_bodies: bool, collide_with_areas: bool, results: *mut c_void, max_results: i32, result_count: *mut i32 ) -> bool

Safety

Godot currently does not document safety requirements on this method. Make sure you understand the underlying semantics.

unsafe fn rest_info( &mut self, shape_rid: Rid, transform: Transform2D, motion: Vector2, margin: f32, collision_mask: u32, collide_with_bodies: bool, collide_with_areas: bool, rest_info: *mut PhysicsServer2DExtensionShapeRestInfo ) -> bool

Safety

Godot currently does not document safety requirements on this method. Make sure you understand the underlying semantics.

Object Safety

This trait is not object safe.