Trait godot::engine::ITextEdit

pub trait ITextEdit: GodotClass + You_forgot_the_attribute__godot_api {

    // Provided methods
    fn init(base: Base<Self::Base>) -> Self { ... }
    fn to_string(&self) -> GString { ... }
    fn on_notification(&mut self, what: ControlNotification) { ... }
    fn get_property(&self, property: StringName) -> Option<Variant> { ... }
    fn set_property(&mut self, property: StringName, value: Variant) -> bool { ... }
    fn handle_unicode_input(&mut self, unicode_char: i32, caret_index: i32) { ... }
    fn backspace(&mut self, caret_index: i32) { ... }
    fn cut(&mut self, caret_index: i32) { ... }
    fn copy(&mut self, caret_index: i32) { ... }
    fn paste(&mut self, caret_index: i32) { ... }
    fn paste_primary_clipboard(&mut self, caret_index: i32) { ... }
    fn has_point(&self, point: Vector2) -> bool { ... }
    fn structured_text_parser(
        &self,
        args: Array<Variant>,
        text: GString
    ) -> Array<Vector3i> { ... }
    fn get_minimum_size(&self) -> Vector2 { ... }
    fn get_tooltip(&self, at_position: Vector2) -> GString { ... }
    fn get_drag_data(&mut self, at_position: Vector2) -> Variant { ... }
    fn can_drop_data(&self, at_position: Vector2, data: Variant) -> bool { ... }
    fn drop_data(&mut self, at_position: Vector2, data: Variant) { ... }
    fn make_custom_tooltip(&self, for_text: GString) -> Option<Gd<Object>> { ... }
    fn gui_input(&mut self, event: Gd<InputEvent>) { ... }
    fn draw(&mut self) { ... }
    fn process(&mut self, delta: f64) { ... }
    fn physics_process(&mut self, delta: f64) { ... }
    fn enter_tree(&mut self) { ... }
    fn exit_tree(&mut self) { ... }
    fn ready(&mut self) { ... }
    fn get_configuration_warnings(&self) -> PackedStringArray { ... }
    fn input(&mut self, event: Gd<InputEvent>) { ... }
    fn shortcut_input(&mut self, event: Gd<InputEvent>) { ... }
    fn unhandled_input(&mut self, event: Gd<InputEvent>) { ... }
    fn unhandled_key_input(&mut self, event: Gd<InputEvent>) { ... }

}

Virtual methods for class TextEdit.

These methods represent constructors (init) or callbacks invoked by the engine.

See also Godot docs for TextEdit methods.

Provided Methods

fn init(base: Base<Self::Base>) -> Self

Godot constructor, accepting an injected base object.

base refers to the base instance of the class, which can either be stored in a Base<T> field or discarded. This method returns a fully-constructed instance, which will then be moved into a Gd<T> pointer.

If the class has a #[class(init)] attribute, this method will be auto-generated and must not be overridden.

fn to_string(&self) -> GString

String representation of the Godot instance.

Override this method to define how the instance is represented as a string. Used by impl Display for Gd<T>, as well as str() and print() in GDScript.

fn on_notification(&mut self, what: ControlNotification)

Called when the object receives a Godot notification.

The type of notification can be identified through what. The enum is designed to hold all possible NOTIFICATION_* constants that the current class can handle. However, this is not validated in Godot, so an enum variant Unknown exists to represent integers out of known constants (mistakes or future additions).

This method is named _notification in Godot, but on_notification in Rust. To send notifications, use the Object::notify method.

See also in Godot docs:

Object::_notification.

Notifications tutorial.

fn get_property(&self, property: StringName) -> Option<Variant>

Called whenever get() is called or Godot gets the value of a property.

Should return the given property’s value as Some(value), or None if the property should be handled normally.

See also in Godot docs:

Object::_get.

fn set_property(&mut self, property: StringName, value: Variant) -> bool

Called whenever Godot set() is called or Godot sets the value of a property.

Should set property to the given value and return true, or return false to indicate the property should be handled normally.

See also in Godot docs:

Object::_set.

fn handle_unicode_input(&mut self, unicode_char: i32, caret_index: i32)

fn backspace(&mut self, caret_index: i32)

fn cut(&mut self, caret_index: i32)

fn copy(&mut self, caret_index: i32)

fn paste(&mut self, caret_index: i32)

fn paste_primary_clipboard(&mut self, caret_index: i32)

fn has_point(&self, point: Vector2) -> bool

fn structured_text_parser( &self, args: Array<Variant>, text: GString ) -> Array<Vector3i>

fn get_minimum_size(&self) -> Vector2

fn get_tooltip(&self, at_position: Vector2) -> GString

fn get_drag_data(&mut self, at_position: Vector2) -> Variant

fn can_drop_data(&self, at_position: Vector2, data: Variant) -> bool

fn drop_data(&mut self, at_position: Vector2, data: Variant)

fn make_custom_tooltip(&self, for_text: GString) -> Option<Gd<Object>>

fn gui_input(&mut self, event: Gd<InputEvent>)

fn draw(&mut self)

fn process(&mut self, delta: f64)

fn physics_process(&mut self, delta: f64)

fn enter_tree(&mut self)

fn exit_tree(&mut self)

fn ready(&mut self)

fn get_configuration_warnings(&self) -> PackedStringArray

fn input(&mut self, event: Gd<InputEvent>)

fn shortcut_input(&mut self, event: Gd<InputEvent>)

fn unhandled_input(&mut self, event: Gd<InputEvent>)

fn unhandled_key_input(&mut self, event: Gd<InputEvent>)

Object Safety

This trait is not object safe.