Trait godot::engine::NodeExt

pub trait NodeExt {
    // Required method
    fn try_get_node_as<T>(&self, path: impl Into<NodePath>) -> Option<Gd<T>>
       where T: GodotClass + Inherits<Node>;

    // Provided method
    fn get_node_as<T>(&self, path: impl Into<NodePath>) -> Gd<T>
       where T: GodotClass + Inherits<Node> { ... }
}

Extension trait with convenience functions for the node tree.

Required Methods

fn try_get_node_as<T>(&self, path: impl Into<NodePath>) -> Option<Gd<T>>
where T: GodotClass + Inherits<Node>,

Retrieves the node at path path (fallible).

If the node is not found, or if it does not have type T or inherited, None will be returned.

Provided Methods

fn get_node_as<T>(&self, path: impl Into<NodePath>) -> Gd<T>
where T: GodotClass + Inherits<Node>,

Retrieves the node at path path, panicking if not found or bad type.

Panics

If the node is not found, or if it does not have type T or inherited.

Object Safety

This trait is not object safe.