Trait godot::engine::PhysicsServer2DExtensionVirtual

pub trait PhysicsServer2DExtensionVirtual: GodotClass + You_forgot_the_attribute__godot_api {

    // Provided methods
    fn init(base: Base<Self::Base>) -> Self { ... }
    fn to_string(&self) -> GodotString { ... }
    fn on_notification(&mut self, what: ObjectNotification) { ... }
    fn world_boundary_shape_create(&mut self) -> Rid { ... }
    fn separation_ray_shape_create(&mut self) -> Rid { ... }
    fn segment_shape_create(&mut self) -> Rid { ... }
    fn circle_shape_create(&mut self) -> Rid { ... }
    fn rectangle_shape_create(&mut self) -> Rid { ... }
    fn capsule_shape_create(&mut self) -> Rid { ... }
    fn convex_polygon_shape_create(&mut self) -> Rid { ... }
    fn concave_polygon_shape_create(&mut self) -> Rid { ... }
    fn shape_set_data(&mut self, shape: Rid, data: Variant) { ... }
    fn shape_set_custom_solver_bias(&mut self, shape: Rid, bias: f64) { ... }
    fn shape_get_type(&self, shape: Rid) -> ShapeType { ... }
    fn shape_get_data(&self, shape: Rid) -> Variant { ... }
    fn shape_get_custom_solver_bias(&self, shape: Rid) -> f64 { ... }
    fn space_create(&mut self) -> Rid { ... }
    fn space_set_active(&mut self, space: Rid, active: bool) { ... }
    fn space_is_active(&self, space: Rid) -> bool { ... }
    fn space_set_param(&mut self, space: Rid, param: SpaceParameter, value: f64) { ... }
    fn space_get_param(&self, space: Rid, param: SpaceParameter) -> f64 { ... }
    fn space_get_direct_state(
        &mut self,
        space: Rid
    ) -> Option<Gd<PhysicsDirectSpaceState2D>> { ... }
    fn space_set_debug_contacts(&mut self, space: Rid, max_contacts: i64) { ... }
    fn space_get_contacts(&self, space: Rid) -> PackedVector2Array { ... }
    fn space_get_contact_count(&self, space: Rid) -> i64 { ... }
    fn area_create(&mut self) -> Rid { ... }
    fn area_set_space(&mut self, area: Rid, space: Rid) { ... }
    fn area_get_space(&self, area: Rid) -> Rid { ... }
    fn area_add_shape(
        &mut self,
        area: Rid,
        shape: Rid,
        transform: Transform2D,
        disabled: bool
    ) { ... }
    fn area_set_shape(&mut self, area: Rid, shape_idx: i64, shape: Rid) { ... }
    fn area_set_shape_transform(
        &mut self,
        area: Rid,
        shape_idx: i64,
        transform: Transform2D
    ) { ... }
    fn area_set_shape_disabled(
        &mut self,
        area: Rid,
        shape_idx: i64,
        disabled: bool
    ) { ... }
    fn area_get_shape_count(&self, area: Rid) -> i64 { ... }
    fn area_get_shape(&self, area: Rid, shape_idx: i64) -> Rid { ... }
    fn area_get_shape_transform(&self, area: Rid, shape_idx: i64) -> Transform2D { ... }
    fn area_remove_shape(&mut self, area: Rid, shape_idx: i64) { ... }
    fn area_clear_shapes(&mut self, area: Rid) { ... }
    fn area_attach_object_instance_id(&mut self, area: Rid, id: i64) { ... }
    fn area_get_object_instance_id(&self, area: Rid) -> i64 { ... }
    fn area_attach_canvas_instance_id(&mut self, area: Rid, id: i64) { ... }
    fn area_get_canvas_instance_id(&self, area: Rid) -> i64 { ... }
    fn area_set_param(&mut self, area: Rid, param: AreaParameter, value: Variant) { ... }
    fn area_set_transform(&mut self, area: Rid, transform: Transform2D) { ... }
    fn area_get_param(&self, area: Rid, param: AreaParameter) -> Variant { ... }
    fn area_get_transform(&self, area: Rid) -> Transform2D { ... }
    fn area_set_collision_layer(&mut self, area: Rid, layer: i64) { ... }
    fn area_get_collision_layer(&self, area: Rid) -> i64 { ... }
    fn area_set_collision_mask(&mut self, area: Rid, mask: i64) { ... }
    fn area_get_collision_mask(&self, area: Rid) -> i64 { ... }
    fn area_set_monitorable(&mut self, area: Rid, monitorable: bool) { ... }
    fn area_set_pickable(&mut self, area: Rid, pickable: bool) { ... }
    fn area_set_monitor_callback(&mut self, area: Rid, callback: Callable) { ... }
    fn area_set_area_monitor_callback(&mut self, area: Rid, callback: Callable) { ... }
    fn body_create(&mut self) -> Rid { ... }
    fn body_set_space(&mut self, body: Rid, space: Rid) { ... }
    fn body_get_space(&self, body: Rid) -> Rid { ... }
    fn body_set_mode(&mut self, body: Rid, mode: BodyMode) { ... }
    fn body_get_mode(&self, body: Rid) -> BodyMode { ... }
    fn body_add_shape(
        &mut self,
        body: Rid,
        shape: Rid,
        transform: Transform2D,
        disabled: bool
    ) { ... }
    fn body_set_shape(&mut self, body: Rid, shape_idx: i64, shape: Rid) { ... }
    fn body_set_shape_transform(
        &mut self,
        body: Rid,
        shape_idx: i64,
        transform: Transform2D
    ) { ... }
    fn body_get_shape_count(&self, body: Rid) -> i64 { ... }
    fn body_get_shape(&self, body: Rid, shape_idx: i64) -> Rid { ... }
    fn body_get_shape_transform(&self, body: Rid, shape_idx: i64) -> Transform2D { ... }
    fn body_set_shape_disabled(
        &mut self,
        body: Rid,
        shape_idx: i64,
        disabled: bool
    ) { ... }
    fn body_set_shape_as_one_way_collision(
        &mut self,
        body: Rid,
        shape_idx: i64,
        enable: bool,
        margin: f64
    ) { ... }
    fn body_remove_shape(&mut self, body: Rid, shape_idx: i64) { ... }
    fn body_clear_shapes(&mut self, body: Rid) { ... }
    fn body_attach_object_instance_id(&mut self, body: Rid, id: i64) { ... }
    fn body_get_object_instance_id(&self, body: Rid) -> i64 { ... }
    fn body_attach_canvas_instance_id(&mut self, body: Rid, id: i64) { ... }
    fn body_get_canvas_instance_id(&self, body: Rid) -> i64 { ... }
    fn body_set_continuous_collision_detection_mode(
        &mut self,
        body: Rid,
        mode: CCDMode
    ) { ... }
    fn body_get_continuous_collision_detection_mode(&self, body: Rid) -> CCDMode { ... }
    fn body_set_collision_layer(&mut self, body: Rid, layer: i64) { ... }
    fn body_get_collision_layer(&self, body: Rid) -> i64 { ... }
    fn body_set_collision_mask(&mut self, body: Rid, mask: i64) { ... }
    fn body_get_collision_mask(&self, body: Rid) -> i64 { ... }
    fn body_set_collision_priority(&mut self, body: Rid, priority: f64) { ... }
    fn body_get_collision_priority(&self, body: Rid) -> f64 { ... }
    fn body_set_param(&mut self, body: Rid, param: BodyParameter, value: Variant) { ... }
    fn body_get_param(&self, body: Rid, param: BodyParameter) -> Variant { ... }
    fn body_reset_mass_properties(&mut self, body: Rid) { ... }
    fn body_set_state(&mut self, body: Rid, state: BodyState, value: Variant) { ... }
    fn body_get_state(&self, body: Rid, state: BodyState) -> Variant { ... }
    fn body_apply_central_impulse(&mut self, body: Rid, impulse: Vector2) { ... }
    fn body_apply_torque_impulse(&mut self, body: Rid, impulse: f64) { ... }
    fn body_apply_impulse(
        &mut self,
        body: Rid,
        impulse: Vector2,
        position: Vector2
    ) { ... }
    fn body_apply_central_force(&mut self, body: Rid, force: Vector2) { ... }
    fn body_apply_force(&mut self, body: Rid, force: Vector2, position: Vector2) { ... }
    fn body_apply_torque(&mut self, body: Rid, torque: f64) { ... }
    fn body_add_constant_central_force(&mut self, body: Rid, force: Vector2) { ... }
    fn body_add_constant_force(
        &mut self,
        body: Rid,
        force: Vector2,
        position: Vector2
    ) { ... }
    fn body_add_constant_torque(&mut self, body: Rid, torque: f64) { ... }
    fn body_set_constant_force(&mut self, body: Rid, force: Vector2) { ... }
    fn body_get_constant_force(&self, body: Rid) -> Vector2 { ... }
    fn body_set_constant_torque(&mut self, body: Rid, torque: f64) { ... }
    fn body_get_constant_torque(&self, body: Rid) -> f64 { ... }
    fn body_set_axis_velocity(&mut self, body: Rid, axis_velocity: Vector2) { ... }
    fn body_add_collision_exception(&mut self, body: Rid, excepted_body: Rid) { ... }
    fn body_remove_collision_exception(&mut self, body: Rid, excepted_body: Rid) { ... }
    fn body_get_collision_exceptions(&self, body: Rid) -> Array<Rid> { ... }
    fn body_set_max_contacts_reported(&mut self, body: Rid, amount: i64) { ... }
    fn body_get_max_contacts_reported(&self, body: Rid) -> i64 { ... }
    fn body_set_contacts_reported_depth_threshold(
        &mut self,
        body: Rid,
        threshold: f64
    ) { ... }
    fn body_get_contacts_reported_depth_threshold(&self, body: Rid) -> f64 { ... }
    fn body_set_omit_force_integration(&mut self, body: Rid, enable: bool) { ... }
    fn body_is_omitting_force_integration(&self, body: Rid) -> bool { ... }
    fn body_set_state_sync_callback(&mut self, body: Rid, callable: Callable) { ... }
    fn body_set_force_integration_callback(
        &mut self,
        body: Rid,
        callable: Callable,
        userdata: Variant
    ) { ... }
    fn body_set_pickable(&mut self, body: Rid, pickable: bool) { ... }
    fn body_get_direct_state(
        &mut self,
        body: Rid
    ) -> Option<Gd<PhysicsDirectBodyState2D>> { ... }
    fn joint_create(&mut self) -> Rid { ... }
    fn joint_clear(&mut self, joint: Rid) { ... }
    fn joint_set_param(&mut self, joint: Rid, param: JointParam, value: f64) { ... }
    fn joint_get_param(&self, joint: Rid, param: JointParam) -> f64 { ... }
    fn joint_disable_collisions_between_bodies(
        &mut self,
        joint: Rid,
        disable: bool
    ) { ... }
    fn joint_is_disabled_collisions_between_bodies(&self, joint: Rid) -> bool { ... }
    fn joint_make_pin(
        &mut self,
        joint: Rid,
        anchor: Vector2,
        body_a: Rid,
        body_b: Rid
    ) { ... }
    fn joint_make_groove(
        &mut self,
        joint: Rid,
        a_groove1: Vector2,
        a_groove2: Vector2,
        b_anchor: Vector2,
        body_a: Rid,
        body_b: Rid
    ) { ... }
    fn joint_make_damped_spring(
        &mut self,
        joint: Rid,
        anchor_a: Vector2,
        anchor_b: Vector2,
        body_a: Rid,
        body_b: Rid
    ) { ... }
    fn pin_joint_set_param(&mut self, joint: Rid, param: PinJointParam, value: f64) { ... }
    fn pin_joint_get_param(&self, joint: Rid, param: PinJointParam) -> f64 { ... }
    fn damped_spring_joint_set_param(
        &mut self,
        joint: Rid,
        param: DampedSpringParam,
        value: f64
    ) { ... }
    fn damped_spring_joint_get_param(
        &self,
        joint: Rid,
        param: DampedSpringParam
    ) -> f64 { ... }
    fn joint_get_type(&self, joint: Rid) -> JointType { ... }
    fn free_rid(&mut self, rid: Rid) { ... }
    fn set_active(&mut self, active: bool) { ... }
    fn init_ext(&mut self) { ... }
    fn step(&mut self, step: f64) { ... }
    fn sync(&mut self) { ... }
    fn flush_queries(&mut self) { ... }
    fn end_sync(&mut self) { ... }
    fn finish(&mut self) { ... }
    fn is_flushing_queries(&self) -> bool { ... }
    fn get_process_info(&mut self, process_info: ProcessInfo) -> i64 { ... }

}

Virtual methods for class PhysicsServer2DExtension.

These methods represent constructors (init) or callbacks invoked by the engine.

See also Godot docs for PhysicsServer2DExtension methods.

Provided Methods

fn init(base: Base<Self::Base>) -> Self

Godot constructor, accepting an injected base object.

base refers to the base instance of the class, which can either be stored in a #[base] field or discarded. This method returns a fully-constructed instance, which will then be moved into a Gd<T> pointer.

If the class has a #[class(init)] attribute, this method will be auto-generated and must not be overridden.

fn to_string(&self) -> GodotString

String representation of the Godot instance.

Override this method to define how the instance is represented as a string. Used by impl Display for Gd<T>, as well as str() and print() in GDScript.

fn on_notification(&mut self, what: ObjectNotification)

Called when the object receives a Godot notification.

The type of notification can be identified through what. The enum is designed to hold all possible NOTIFICATION_* constants that the current class can handle. However, this is not validated in Godot, so an enum variant Unknown exists to represent integers out of known constants (mistakes or future additions).

This method is named _notification in Godot, but on_notification in Rust. To send notifications, use the Object::notify method.

See also in Godot docs:

Object::_notification.

Notifications tutorial.

fn world_boundary_shape_create(&mut self) -> Rid

fn separation_ray_shape_create(&mut self) -> Rid

fn segment_shape_create(&mut self) -> Rid

fn circle_shape_create(&mut self) -> Rid

fn rectangle_shape_create(&mut self) -> Rid

fn capsule_shape_create(&mut self) -> Rid

fn convex_polygon_shape_create(&mut self) -> Rid

fn concave_polygon_shape_create(&mut self) -> Rid

fn shape_set_data(&mut self, shape: Rid, data: Variant)

fn shape_set_custom_solver_bias(&mut self, shape: Rid, bias: f64)

fn shape_get_type(&self, shape: Rid) -> ShapeType

fn shape_get_data(&self, shape: Rid) -> Variant

fn shape_get_custom_solver_bias(&self, shape: Rid) -> f64

fn space_create(&mut self) -> Rid

fn space_set_active(&mut self, space: Rid, active: bool)

fn space_is_active(&self, space: Rid) -> bool

fn space_set_param(&mut self, space: Rid, param: SpaceParameter, value: f64)

fn space_get_param(&self, space: Rid, param: SpaceParameter) -> f64

fn space_get_direct_state( &mut self, space: Rid ) -> Option<Gd<PhysicsDirectSpaceState2D>>

fn space_set_debug_contacts(&mut self, space: Rid, max_contacts: i64)

fn space_get_contacts(&self, space: Rid) -> PackedVector2Array

fn space_get_contact_count(&self, space: Rid) -> i64

fn area_create(&mut self) -> Rid

fn area_set_space(&mut self, area: Rid, space: Rid)

fn area_get_space(&self, area: Rid) -> Rid

fn area_add_shape( &mut self, area: Rid, shape: Rid, transform: Transform2D, disabled: bool )

fn area_set_shape(&mut self, area: Rid, shape_idx: i64, shape: Rid)

fn area_set_shape_transform( &mut self, area: Rid, shape_idx: i64, transform: Transform2D )

fn area_set_shape_disabled(&mut self, area: Rid, shape_idx: i64, disabled: bool)

fn area_get_shape_count(&self, area: Rid) -> i64

fn area_get_shape(&self, area: Rid, shape_idx: i64) -> Rid

fn area_get_shape_transform(&self, area: Rid, shape_idx: i64) -> Transform2D

fn area_remove_shape(&mut self, area: Rid, shape_idx: i64)

fn area_clear_shapes(&mut self, area: Rid)

fn area_attach_object_instance_id(&mut self, area: Rid, id: i64)

fn area_get_object_instance_id(&self, area: Rid) -> i64

fn area_attach_canvas_instance_id(&mut self, area: Rid, id: i64)

fn area_get_canvas_instance_id(&self, area: Rid) -> i64

fn area_set_param(&mut self, area: Rid, param: AreaParameter, value: Variant)

fn area_set_transform(&mut self, area: Rid, transform: Transform2D)

fn area_get_param(&self, area: Rid, param: AreaParameter) -> Variant

fn area_get_transform(&self, area: Rid) -> Transform2D

fn area_set_collision_layer(&mut self, area: Rid, layer: i64)

fn area_get_collision_layer(&self, area: Rid) -> i64

fn area_set_collision_mask(&mut self, area: Rid, mask: i64)

fn area_get_collision_mask(&self, area: Rid) -> i64

fn area_set_monitorable(&mut self, area: Rid, monitorable: bool)

fn area_set_pickable(&mut self, area: Rid, pickable: bool)

fn area_set_monitor_callback(&mut self, area: Rid, callback: Callable)

fn area_set_area_monitor_callback(&mut self, area: Rid, callback: Callable)

fn body_create(&mut self) -> Rid

fn body_set_space(&mut self, body: Rid, space: Rid)

fn body_get_space(&self, body: Rid) -> Rid

fn body_set_mode(&mut self, body: Rid, mode: BodyMode)

fn body_get_mode(&self, body: Rid) -> BodyMode

fn body_add_shape( &mut self, body: Rid, shape: Rid, transform: Transform2D, disabled: bool )

fn body_set_shape(&mut self, body: Rid, shape_idx: i64, shape: Rid)

fn body_set_shape_transform( &mut self, body: Rid, shape_idx: i64, transform: Transform2D )

fn body_get_shape_count(&self, body: Rid) -> i64

fn body_get_shape(&self, body: Rid, shape_idx: i64) -> Rid

fn body_get_shape_transform(&self, body: Rid, shape_idx: i64) -> Transform2D

fn body_set_shape_disabled(&mut self, body: Rid, shape_idx: i64, disabled: bool)

fn body_set_shape_as_one_way_collision( &mut self, body: Rid, shape_idx: i64, enable: bool, margin: f64 )

fn body_remove_shape(&mut self, body: Rid, shape_idx: i64)

fn body_clear_shapes(&mut self, body: Rid)

fn body_attach_object_instance_id(&mut self, body: Rid, id: i64)

fn body_get_object_instance_id(&self, body: Rid) -> i64

fn body_attach_canvas_instance_id(&mut self, body: Rid, id: i64)

fn body_get_canvas_instance_id(&self, body: Rid) -> i64

fn body_set_continuous_collision_detection_mode( &mut self, body: Rid, mode: CCDMode )

fn body_get_continuous_collision_detection_mode(&self, body: Rid) -> CCDMode

fn body_set_collision_layer(&mut self, body: Rid, layer: i64)

fn body_get_collision_layer(&self, body: Rid) -> i64

fn body_set_collision_mask(&mut self, body: Rid, mask: i64)

fn body_get_collision_mask(&self, body: Rid) -> i64

fn body_set_collision_priority(&mut self, body: Rid, priority: f64)

fn body_get_collision_priority(&self, body: Rid) -> f64

fn body_set_param(&mut self, body: Rid, param: BodyParameter, value: Variant)

fn body_get_param(&self, body: Rid, param: BodyParameter) -> Variant

fn body_reset_mass_properties(&mut self, body: Rid)

fn body_set_state(&mut self, body: Rid, state: BodyState, value: Variant)

fn body_get_state(&self, body: Rid, state: BodyState) -> Variant

fn body_apply_central_impulse(&mut self, body: Rid, impulse: Vector2)

fn body_apply_torque_impulse(&mut self, body: Rid, impulse: f64)

fn body_apply_impulse(&mut self, body: Rid, impulse: Vector2, position: Vector2)

fn body_apply_central_force(&mut self, body: Rid, force: Vector2)

fn body_apply_force(&mut self, body: Rid, force: Vector2, position: Vector2)

fn body_apply_torque(&mut self, body: Rid, torque: f64)

fn body_add_constant_central_force(&mut self, body: Rid, force: Vector2)

fn body_add_constant_force(&mut self, body: Rid, force: Vector2, position: Vector2)

fn body_add_constant_torque(&mut self, body: Rid, torque: f64)

fn body_set_constant_force(&mut self, body: Rid, force: Vector2)

fn body_get_constant_force(&self, body: Rid) -> Vector2

fn body_set_constant_torque(&mut self, body: Rid, torque: f64)

fn body_get_constant_torque(&self, body: Rid) -> f64

fn body_set_axis_velocity(&mut self, body: Rid, axis_velocity: Vector2)

fn body_add_collision_exception(&mut self, body: Rid, excepted_body: Rid)

fn body_remove_collision_exception(&mut self, body: Rid, excepted_body: Rid)

fn body_get_collision_exceptions(&self, body: Rid) -> Array<Rid>

fn body_set_max_contacts_reported(&mut self, body: Rid, amount: i64)

fn body_get_max_contacts_reported(&self, body: Rid) -> i64

fn body_set_contacts_reported_depth_threshold( &mut self, body: Rid, threshold: f64 )

fn body_get_contacts_reported_depth_threshold(&self, body: Rid) -> f64

fn body_set_omit_force_integration(&mut self, body: Rid, enable: bool)

fn body_is_omitting_force_integration(&self, body: Rid) -> bool

fn body_set_state_sync_callback(&mut self, body: Rid, callable: Callable)

fn body_set_force_integration_callback( &mut self, body: Rid, callable: Callable, userdata: Variant )

fn body_set_pickable(&mut self, body: Rid, pickable: bool)

fn body_get_direct_state( &mut self, body: Rid ) -> Option<Gd<PhysicsDirectBodyState2D>>

fn joint_create(&mut self) -> Rid

fn joint_clear(&mut self, joint: Rid)

fn joint_set_param(&mut self, joint: Rid, param: JointParam, value: f64)

fn joint_get_param(&self, joint: Rid, param: JointParam) -> f64

fn joint_disable_collisions_between_bodies(&mut self, joint: Rid, disable: bool)

fn joint_is_disabled_collisions_between_bodies(&self, joint: Rid) -> bool

fn joint_make_pin(&mut self, joint: Rid, anchor: Vector2, body_a: Rid, body_b: Rid)

fn joint_make_groove( &mut self, joint: Rid, a_groove1: Vector2, a_groove2: Vector2, b_anchor: Vector2, body_a: Rid, body_b: Rid )

fn joint_make_damped_spring( &mut self, joint: Rid, anchor_a: Vector2, anchor_b: Vector2, body_a: Rid, body_b: Rid )

fn pin_joint_set_param(&mut self, joint: Rid, param: PinJointParam, value: f64)

fn pin_joint_get_param(&self, joint: Rid, param: PinJointParam) -> f64

fn damped_spring_joint_set_param( &mut self, joint: Rid, param: DampedSpringParam, value: f64 )

fn damped_spring_joint_get_param( &self, joint: Rid, param: DampedSpringParam ) -> f64

fn joint_get_type(&self, joint: Rid) -> JointType

fn free_rid(&mut self, rid: Rid)

fn set_active(&mut self, active: bool)

fn init_ext(&mut self)

fn step(&mut self, step: f64)

fn sync(&mut self)

fn flush_queries(&mut self)

fn end_sync(&mut self)

fn finish(&mut self)

fn is_flushing_queries(&self) -> bool

fn get_process_info(&mut self, process_info: ProcessInfo) -> i64