Trait godot::engine::ScriptInstance

pub trait ScriptInstance {

    // Required methods
    fn class_name(&self) -> GString;
    fn set(&mut self, name: StringName, value: &Variant) -> bool;
    fn get(&self, name: StringName) -> Option<Variant>;
    fn get_property_list(&self) -> &[PropertyInfo];
    fn get_method_list(&self) -> &[MethodInfo];
    fn call(
        &mut self,
        method: StringName,
        args: &[&Variant]
    ) -> Result<Variant, u32>;
    fn is_placeholder(&self) -> bool;
    fn has_method(&self, method: StringName) -> bool;
    fn get_script(&self) -> &Gd<Script>;
    fn get_property_type(&self, name: StringName) -> VariantType;
    fn to_string(&self) -> GString;
    fn get_property_state(&self) -> Vec<(StringName, Variant)>;
    fn get_language(&self) -> Gd<ScriptLanguage>;
    fn on_refcount_decremented(&self) -> bool;
    fn on_refcount_incremented(&self);
    fn property_get_fallback(&self, name: StringName) -> Option<Variant>;
    fn property_set_fallback(
        &mut self,
        name: StringName,
        value: &Variant
    ) -> bool;

}

Interface for Godot’s GDExtensionScriptInstance, which cannot be directly constructed by an extension. Instead a type that implements this trait has to be passed to the create_script_instance function, which returns a [sys::GDExtensionScriptInstancePtr] pointer. This pointer can then be returned from IScriptExtension::instance_create.

Required Methods

fn class_name(&self) -> GString

Name of the new class the script implements.

fn set(&mut self, name: StringName, value: &Variant) -> bool

Property setter for Godots virtual dispatch system. The engine will call this function when it wants to change a property on the script.

fn get(&self, name: StringName) -> Option<Variant>

Property getter for Godots virtual dispatch system. The engine will call this function when it wants to read a property on the script.

fn get_property_list(&self) -> &[PropertyInfo]

A list of all the properties a script exposes to the engine.

fn get_method_list(&self) -> &[MethodInfo]

A list of all the methods a script exposes to the engine.

fn call( &mut self, method: StringName, args: &[&Variant] ) -> Result<Variant, u32>

Method invoker for Godots virtual dispatch system. The engine will call this function when it wants to call a method on the script.

All method calls are taking a mutable reference of the script instance, as the engine does not differentiate between immutable and mutable method calls like rust.

It’s important that the script does not cause a second call to this function while executing a method call. This would result in a panic.

fn is_placeholder(&self) -> bool

Identifies the script instance as a placeholder. If this function and IScriptExtension::is_placeholder_fallback_enabled return true, Godot will call Self::property_set_fallback instead of Self::set.

fn has_method(&self, method: StringName) -> bool

Validation function for the engine to verify if the script exposes a certain method.

fn get_script(&self) -> &Gd<Script>

Lets the engine get a reference to the script this instance was created for.

This function has to return a reference, because Scripts are reference counted in Godot and it has to be guaranteed that the object is not freed before the engine increased the reference count. (every time a Gd<T> which contains a reference counted object is dropped the reference count is decremented.)

fn get_property_type(&self, name: StringName) -> VariantType

Lets the engine fetch the type of a particular property.

fn to_string(&self) -> GString

String representation of the script instance.

fn get_property_state(&self) -> Vec<(StringName, Variant)>

A dump of all property names and values that are exposed to the engine.

fn get_language(&self) -> Gd<ScriptLanguage>

Lets the engine get a reference to the ScriptLanguage this instance belongs to.

fn on_refcount_decremented(&self) -> bool

Callback from the engine when the reference count of the base object has been decreased. When this method returns true the engine will not free the object the script is attached to.

fn on_refcount_incremented(&self)

Callback from the engine when the reference count of the base object has been increased.

fn property_get_fallback(&self, name: StringName) -> Option<Variant>

The engine may call this function if it failed to get a property value via ScriptInstance::get or the native types getter.

fn property_set_fallback(&mut self, name: StringName, value: &Variant) -> bool

The engine may call this function if ScriptLanguage::is_placeholder_fallback_enabled is enabled.