Trait godot::engine::ScriptLanguageExtensionVirtual

pub trait ScriptLanguageExtensionVirtual: GodotClass + You_forgot_the_attribute__godot_api {

    // Provided methods
    fn init(base: Base<Self::Base>) -> Self { ... }
    fn to_string(&self) -> GodotString { ... }
    fn on_notification(&mut self, what: ObjectNotification) { ... }
    fn get_name(&self) -> GodotString { ... }
    fn init_ext(&mut self) { ... }
    fn get_type(&self) -> GodotString { ... }
    fn get_extension(&self) -> GodotString { ... }
    fn finish(&mut self) { ... }
    fn get_reserved_words(&self) -> PackedStringArray { ... }
    fn is_control_flow_keyword(&self, keyword: GodotString) -> bool { ... }
    fn get_comment_delimiters(&self) -> PackedStringArray { ... }
    fn get_string_delimiters(&self) -> PackedStringArray { ... }
    fn make_template(
        &self,
        template: GodotString,
        class_name: GodotString,
        base_class_name: GodotString
    ) -> Option<Gd<Script>> { ... }
    fn get_built_in_templates(&self, object: StringName) -> Array<Dictionary> { ... }
    fn is_using_templates(&mut self) -> bool { ... }
    fn validate(
        &self,
        script: GodotString,
        path: GodotString,
        validate_functions: bool,
        validate_errors: bool,
        validate_warnings: bool,
        validate_safe_lines: bool
    ) -> Dictionary { ... }
    fn validate_path(&self, path: GodotString) -> GodotString { ... }
    fn create_script(&self) -> Option<Gd<Object>> { ... }
    fn has_named_classes(&self) -> bool { ... }
    fn supports_builtin_mode(&self) -> bool { ... }
    fn supports_documentation(&self) -> bool { ... }
    fn can_inherit_from_file(&self) -> bool { ... }
    fn find_function(
        &self,
        class_name: GodotString,
        function_name: GodotString
    ) -> i64 { ... }
    fn make_function(
        &self,
        class_name: GodotString,
        function_name: GodotString,
        function_args: PackedStringArray
    ) -> GodotString { ... }
    fn open_in_external_editor(
        &mut self,
        script: Gd<Script>,
        line: i64,
        column: i64
    ) -> Error { ... }
    fn overrides_external_editor(&mut self) -> bool { ... }
    fn complete_code(
        &self,
        code: GodotString,
        path: GodotString,
        owner: Gd<Object>
    ) -> Dictionary { ... }
    fn lookup_code(
        &self,
        code: GodotString,
        symbol: GodotString,
        path: GodotString,
        owner: Gd<Object>
    ) -> Dictionary { ... }
    fn auto_indent_code(
        &self,
        code: GodotString,
        from_line: i64,
        to_line: i64
    ) -> GodotString { ... }
    fn add_global_constant(&mut self, name: StringName, value: Variant) { ... }
    fn add_named_global_constant(&mut self, name: StringName, value: Variant) { ... }
    fn remove_named_global_constant(&mut self, name: StringName) { ... }
    fn thread_enter(&mut self) { ... }
    fn thread_exit(&mut self) { ... }
    fn debug_get_error(&self) -> GodotString { ... }
    fn debug_get_stack_level_count(&self) -> i64 { ... }
    fn debug_get_stack_level_line(&self, level: i64) -> i64 { ... }
    fn debug_get_stack_level_function(&self, level: i64) -> GodotString { ... }
    fn debug_get_stack_level_locals(
        &mut self,
        level: i64,
        max_subitems: i64,
        max_depth: i64
    ) -> Dictionary { ... }
    fn debug_get_stack_level_members(
        &mut self,
        level: i64,
        max_subitems: i64,
        max_depth: i64
    ) -> Dictionary { ... }
    fn debug_get_globals(
        &mut self,
        max_subitems: i64,
        max_depth: i64
    ) -> Dictionary { ... }
    fn debug_parse_stack_level_expression(
        &mut self,
        level: i64,
        expression: GodotString,
        max_subitems: i64,
        max_depth: i64
    ) -> GodotString { ... }
    fn debug_get_current_stack_info(&mut self) -> Array<Dictionary> { ... }
    fn reload_all_scripts(&mut self) { ... }
    fn reload_tool_script(&mut self, script: Gd<Script>, soft_reload: bool) { ... }
    fn get_recognized_extensions(&self) -> PackedStringArray { ... }
    fn get_public_functions(&self) -> Array<Dictionary> { ... }
    fn get_public_constants(&self) -> Dictionary { ... }
    fn get_public_annotations(&self) -> Array<Dictionary> { ... }
    fn profiling_start(&mut self) { ... }
    fn profiling_stop(&mut self) { ... }
    fn frame(&mut self) { ... }
    fn handles_global_class_type(&self, type_: GodotString) -> bool { ... }
    fn get_global_class_name(&self, path: GodotString) -> Dictionary { ... }

}

Virtual methods for class ScriptLanguageExtension.

These methods represent constructors (init) or callbacks invoked by the engine.

See also Godot docs for ScriptLanguageExtension methods.

Provided Methods

fn init(base: Base<Self::Base>) -> Self

Godot constructor, accepting an injected base object.

base refers to the base instance of the class, which can either be stored in a #[base] field or discarded. This method returns a fully-constructed instance, which will then be moved into a Gd<T> pointer.

If the class has a #[class(init)] attribute, this method will be auto-generated and must not be overridden.

fn to_string(&self) -> GodotString

String representation of the Godot instance.

Override this method to define how the instance is represented as a string. Used by impl Display for Gd<T>, as well as str() and print() in GDScript.

fn on_notification(&mut self, what: ObjectNotification)

Called when the object receives a Godot notification.

The type of notification can be identified through what. The enum is designed to hold all possible NOTIFICATION_* constants that the current class can handle. However, this is not validated in Godot, so an enum variant Unknown exists to represent integers out of known constants (mistakes or future additions).

This method is named _notification in Godot, but on_notification in Rust. To send notifications, use the Object::notify method.

See also in Godot docs:

Object::_notification.

Notifications tutorial.

fn get_name(&self) -> GodotString

fn init_ext(&mut self)

fn get_type(&self) -> GodotString

fn get_extension(&self) -> GodotString

fn finish(&mut self)

fn get_reserved_words(&self) -> PackedStringArray

fn is_control_flow_keyword(&self, keyword: GodotString) -> bool

fn get_comment_delimiters(&self) -> PackedStringArray

fn get_string_delimiters(&self) -> PackedStringArray

fn make_template( &self, template: GodotString, class_name: GodotString, base_class_name: GodotString ) -> Option<Gd<Script>>

fn get_built_in_templates(&self, object: StringName) -> Array<Dictionary>

fn is_using_templates(&mut self) -> bool

fn validate( &self, script: GodotString, path: GodotString, validate_functions: bool, validate_errors: bool, validate_warnings: bool, validate_safe_lines: bool ) -> Dictionary

fn validate_path(&self, path: GodotString) -> GodotString

fn create_script(&self) -> Option<Gd<Object>>

fn has_named_classes(&self) -> bool

fn supports_builtin_mode(&self) -> bool

fn supports_documentation(&self) -> bool

fn can_inherit_from_file(&self) -> bool

fn find_function( &self, class_name: GodotString, function_name: GodotString ) -> i64

fn make_function( &self, class_name: GodotString, function_name: GodotString, function_args: PackedStringArray ) -> GodotString

fn open_in_external_editor( &mut self, script: Gd<Script>, line: i64, column: i64 ) -> Error

fn overrides_external_editor(&mut self) -> bool

fn complete_code( &self, code: GodotString, path: GodotString, owner: Gd<Object> ) -> Dictionary

fn lookup_code( &self, code: GodotString, symbol: GodotString, path: GodotString, owner: Gd<Object> ) -> Dictionary

fn auto_indent_code( &self, code: GodotString, from_line: i64, to_line: i64 ) -> GodotString

fn add_global_constant(&mut self, name: StringName, value: Variant)

fn add_named_global_constant(&mut self, name: StringName, value: Variant)

fn remove_named_global_constant(&mut self, name: StringName)

fn thread_enter(&mut self)

fn thread_exit(&mut self)

fn debug_get_error(&self) -> GodotString

fn debug_get_stack_level_count(&self) -> i64

fn debug_get_stack_level_line(&self, level: i64) -> i64

fn debug_get_stack_level_function(&self, level: i64) -> GodotString

fn debug_get_stack_level_locals( &mut self, level: i64, max_subitems: i64, max_depth: i64 ) -> Dictionary

fn debug_get_stack_level_members( &mut self, level: i64, max_subitems: i64, max_depth: i64 ) -> Dictionary

fn debug_get_globals(&mut self, max_subitems: i64, max_depth: i64) -> Dictionary

fn debug_parse_stack_level_expression( &mut self, level: i64, expression: GodotString, max_subitems: i64, max_depth: i64 ) -> GodotString

fn debug_get_current_stack_info(&mut self) -> Array<Dictionary>

fn reload_all_scripts(&mut self)

fn reload_tool_script(&mut self, script: Gd<Script>, soft_reload: bool)

fn get_recognized_extensions(&self) -> PackedStringArray

fn get_public_functions(&self) -> Array<Dictionary>

fn get_public_constants(&self) -> Dictionary

fn get_public_annotations(&self) -> Array<Dictionary>

fn profiling_start(&mut self)

fn profiling_stop(&mut self)

fn frame(&mut self)

fn handles_global_class_type(&self, type_: GodotString) -> bool

fn get_global_class_name(&self, path: GodotString) -> Dictionary