Trait godot::engine::XrInterfaceExtensionVirtual

pub trait XrInterfaceExtensionVirtual: GodotClass + You_forgot_the_attribute__godot_api {

    // Provided methods
    fn init(base: Base<Self::Base>) -> Self { ... }
    fn to_string(&self) -> GodotString { ... }
    fn on_notification(&mut self, what: ObjectNotification) { ... }
    fn get_name(&self) -> StringName { ... }
    fn get_capabilities(&self) -> i64 { ... }
    fn is_initialized(&self) -> bool { ... }
    fn initialize(&mut self) -> bool { ... }
    fn uninitialize(&mut self) { ... }
    fn supports_play_area_mode(&self, mode: PlayAreaMode) -> bool { ... }
    fn get_play_area_mode(&self) -> PlayAreaMode { ... }
    fn set_play_area_mode(&self, mode: PlayAreaMode) -> bool { ... }
    fn get_play_area(&self) -> PackedVector3Array { ... }
    fn get_render_target_size(&mut self) -> Vector2 { ... }
    fn get_view_count(&mut self) -> i64 { ... }
    fn get_camera_transform(&mut self) -> Transform3D { ... }
    fn get_transform_for_view(
        &mut self,
        view: i64,
        cam_transform: Transform3D
    ) -> Transform3D { ... }
    fn get_projection_for_view(
        &mut self,
        view: i64,
        aspect: f64,
        z_near: f64,
        z_far: f64
    ) -> PackedFloat64Array { ... }
    fn get_vrs_texture(&mut self) -> Rid { ... }
    fn process(&mut self) { ... }
    fn pre_render(&mut self) { ... }
    fn pre_draw_viewport(&mut self, render_target: Rid) -> bool { ... }
    fn post_draw_viewport(&mut self, render_target: Rid, screen_rect: Rect2) { ... }
    fn end_frame(&mut self) { ... }
    fn get_suggested_tracker_names(&self) -> PackedStringArray { ... }
    fn get_suggested_pose_names(
        &self,
        tracker_name: StringName
    ) -> PackedStringArray { ... }
    fn get_tracking_status(&self) -> TrackingStatus { ... }
    fn trigger_haptic_pulse(
        &mut self,
        action_name: GodotString,
        tracker_name: StringName,
        frequency: f64,
        amplitude: f64,
        duration_sec: f64,
        delay_sec: f64
    ) { ... }
    fn get_anchor_detection_is_enabled(&self) -> bool { ... }
    fn set_anchor_detection_is_enabled(&mut self, enabled: bool) { ... }
    fn get_camera_feed_id(&self) -> i64 { ... }
    fn get_color_texture(&mut self) -> Rid { ... }
    fn get_depth_texture(&mut self) -> Rid { ... }
    fn get_velocity_texture(&mut self) -> Rid { ... }

}

Virtual methods for class XrInterfaceExtension.

These methods represent constructors (init) or callbacks invoked by the engine.

See also Godot docs for XRInterfaceExtension methods.

Provided Methods

fn init(base: Base<Self::Base>) -> Self

Godot constructor, accepting an injected base object.

base refers to the base instance of the class, which can either be stored in a #[base] field or discarded. This method returns a fully-constructed instance, which will then be moved into a Gd<T> pointer.

If the class has a #[class(init)] attribute, this method will be auto-generated and must not be overridden.

fn to_string(&self) -> GodotString

String representation of the Godot instance.

Override this method to define how the instance is represented as a string. Used by impl Display for Gd<T>, as well as str() and print() in GDScript.

fn on_notification(&mut self, what: ObjectNotification)

Called when the object receives a Godot notification.

The type of notification can be identified through what. The enum is designed to hold all possible NOTIFICATION_* constants that the current class can handle. However, this is not validated in Godot, so an enum variant Unknown exists to represent integers out of known constants (mistakes or future additions).

This method is named _notification in Godot, but on_notification in Rust. To send notifications, use the Object::notify method.

See also in Godot docs:

Object::_notification.

Notifications tutorial.

fn get_name(&self) -> StringName

fn get_capabilities(&self) -> i64

fn is_initialized(&self) -> bool

fn initialize(&mut self) -> bool

fn uninitialize(&mut self)

fn supports_play_area_mode(&self, mode: PlayAreaMode) -> bool

fn get_play_area_mode(&self) -> PlayAreaMode

fn set_play_area_mode(&self, mode: PlayAreaMode) -> bool

fn get_play_area(&self) -> PackedVector3Array

fn get_render_target_size(&mut self) -> Vector2

fn get_view_count(&mut self) -> i64

fn get_camera_transform(&mut self) -> Transform3D

fn get_transform_for_view( &mut self, view: i64, cam_transform: Transform3D ) -> Transform3D

fn get_projection_for_view( &mut self, view: i64, aspect: f64, z_near: f64, z_far: f64 ) -> PackedFloat64Array

fn get_vrs_texture(&mut self) -> Rid

fn process(&mut self)

fn pre_render(&mut self)

fn pre_draw_viewport(&mut self, render_target: Rid) -> bool

fn post_draw_viewport(&mut self, render_target: Rid, screen_rect: Rect2)

fn end_frame(&mut self)

fn get_suggested_tracker_names(&self) -> PackedStringArray

fn get_suggested_pose_names(&self, tracker_name: StringName) -> PackedStringArray

fn get_tracking_status(&self) -> TrackingStatus

fn trigger_haptic_pulse( &mut self, action_name: GodotString, tracker_name: StringName, frequency: f64, amplitude: f64, duration_sec: f64, delay_sec: f64 )

fn get_anchor_detection_is_enabled(&self) -> bool

fn set_anchor_detection_is_enabled(&mut self, enabled: bool)

fn get_camera_feed_id(&self) -> i64

fn get_color_texture(&mut self) -> Rid

fn get_depth_texture(&mut self) -> Rid

fn get_velocity_texture(&mut self) -> Rid