godot

Module init

Entry point and global init/shutdown of the library.

Structs

GdextBuild

Provides meta-information about the library and the Godot version in use.

Enums

EditorRunBehavior

Determines if and how an extension’s code is run in the editor.

Traits

ExtensionLibrary

Defines the entry point for a GDExtension Rust library.

Functions

is_main_threadNon-wasm_nothreads

Check if the current thread is the main thread.

main_thread_idNon-wasm_nothreads

Get the ThreadId of the main thread.

Type Aliases

InitLevel

Stage of the Godot initialization process.

Attribute Macros

gdextension

Proc-macro attribute to be used in combination with the ExtensionLibrary trait.