godot::init

Type Alias InitLevel

pub type InitLevel = InitLevel;

Stage of the Godot initialization process.

Godot’s initialization and deinitialization processes are split into multiple stages, like a stack. At each level, a different amount of engine functionality is available. Deinitialization happens in reverse order.

See also:

ExtensionLibrary::on_level_init()

ExtensionLibrary::on_level_deinit()

Aliased Type

enum InitLevel {
    Core,
    Servers,
    Scene,
    Editor,
}

Variants

Core

First level loaded by Godot. Builtin types are available, classes are not.

Servers

Second level loaded by Godot. Only server classes and builtins are available.

Scene

Third level loaded by Godot. Most classes are available.

Editor

Fourth level loaded by Godot, only in the editor. All classes are available.