godot::inner_builtins::central
Module global_enums

Global enums and constants, generated by Godot.

Structs

Error

HorizontalAlignment

InlineAlignment

JoyAxis

JoyButton

Key

KeyLocation

KeyModifierMask

MethodFlags

MidiMessage

Godot enum name: MIDIMessage.

MouseButton

MouseButtonMask

PropertyHint

PropertyUsageFlags

VerticalAlignment

Enums

ClockDirection

This enum is exhaustive; you should not expect future Godot versions to add new enumerators.

Orientation

This enum is exhaustive; you should not expect future Godot versions to add new enumerators.