godot::inner_builtins
Struct InnerArray

pub struct InnerArray<'a> { /* private fields */ }

Implementations

impl<'a> InnerArray<'a>

pub fn from_outer(outer: &Array<Variant>) -> InnerArray<'a>

pub fn from_outer_typed<T>(outer: &Array<T>) -> InnerArray<'a>
where T: ArrayElement,

pub fn size(&self) -> i64

pub fn is_empty(&self) -> bool

pub fn clear(&mut self)

pub fn hash(&self) -> i64

pub fn assign(&mut self, array: &Array<Variant>)

pub fn push_back(&mut self, value: &Variant)

pub fn push_front(&mut self, value: &Variant)

pub fn append(&mut self, value: &Variant)

pub fn append_array(&mut self, array: &Array<Variant>)

pub fn resize(&mut self, size: i64) -> i64

pub fn insert(&mut self, position: i64, value: &Variant) -> i64

pub fn remove_at(&mut self, position: i64)

pub fn fill(&mut self, value: &Variant)

pub fn erase(&mut self, value: &Variant)

pub fn front(&self) -> Variant

pub fn back(&self) -> Variant

pub fn pick_random(&self) -> Variant

pub fn find(&self, what: &Variant, from: i64) -> i64

pub fn rfind(&self, what: &Variant, from: i64) -> i64

pub fn count(&self, value: &Variant) -> i64

pub fn has(&self, value: &Variant) -> bool

pub fn pop_back(&mut self) -> Variant

pub fn pop_front(&mut self) -> Variant

pub fn pop_at(&mut self, position: i64) -> Variant

pub fn sort(&mut self)

pub fn sort_custom(&mut self, func: &Callable)

pub fn shuffle(&mut self)

pub fn bsearch(&self, value: &Variant, before: bool) -> i64

pub fn bsearch_custom( &self, value: &Variant, func: &Callable, before: bool, ) -> i64

pub fn reverse(&mut self)

pub unsafe fn duplicate(&self, deep: bool) -> Array<Variant>

Safety

You must ensure that the returned array fulfils the safety invariants of Array.

pub unsafe fn slice( &self, begin: i64, end: i64, step: i64, deep: bool, ) -> Array<Variant>

Safety

You must ensure that the returned array fulfils the safety invariants of Array.

pub unsafe fn filter(&self, method: &Callable) -> Array<Variant>

Safety

You must ensure that the returned array fulfils the safety invariants of Array.

pub unsafe fn map(&self, method: &Callable) -> Array<Variant>

Safety

You must ensure that the returned array fulfils the safety invariants of Array.

pub fn reduce(&self, method: &Callable, accum: &Variant) -> Variant

pub fn any(&self, method: &Callable) -> bool

pub fn all(&self, method: &Callable) -> bool

pub fn max(&self) -> Variant

pub fn min(&self) -> Variant

pub fn is_typed(&self) -> bool

pub fn is_same_typed(&self, array: &Array<Variant>) -> bool

pub fn get_typed_builtin(&self) -> i64

pub fn get_typed_class_name(&self) -> StringName

pub fn get_typed_script(&self) -> Variant

pub fn make_read_only(&mut self)

pub fn is_read_only(&self) -> bool