godot::meta

Module conv

Advanced conversion machinery.

Structs

RawPtr

Wrapper around a raw pointer, providing ToGodot/FromGodot for FFI passing.

Enums

ByObject

Pass arguments to Godot by object pointer (for objects only).

ByOption

Pass optional arguments by returning Option<&T::Via>, allowing delegation to underlying type’s strategy.

ByRef

Pass arguments to Godot by reference.

ByValue

Pass arguments to Godot by value.

ByVariant

Pass Variant arguments to Godot by reference.

Traits

ArgPassing

Determines whether arguments are passed by value or by reference to Godot.

AsDirectElement

Marker trait for types directly usable in the array! and dict! macros.

FfiRawPointer

Trait for raw pointers that can be passed over the Godot FFI boundary as i64 addresses.

InParamTuple

Represents a parameter list that is received from some external location (usually Godot).

ObjectToOwned

Obtain owned Gd from either &self or &Gd.

OutParamTuple

Represents a parameter list that is used to call some external code.

ParamTuple

Represents a parameter list as Rust tuple where each tuple element is one parameter.

UniformObjectDeref

Unifies dereferencing of user and engine classes, as &T/&mut T and Gd<T>.