godot::meta

Module error

Custom error types.

Modules

strat

Built-in ErrorToGodot strategies for mapping Result<T, E> return types of #[func] methods.

Macros

func_bail

Return early from a #[func], creating an error value from a format string (including string literals).

Structs

CallError

Error capable of representing failed function calls.

ConvertError

Represents errors that can occur when converting values from Godot.

IoError

Error that can occur while using godot-rust I/O utilities.

StringError

Error related to string encoding/decoding.

Enums

CallErrorType

Enum representing different errors during script instance calls.

CallOutcome

Outcome of mapping a Result<T, E> for a #[func] return value.

Traits

ErrorToGodot

Defines how Result<T, E> returned by #[func] is mapped to Godot.