godot::meta::error

Enum CallErrorType

#[non_exhaustive]

#[repr(C)]
pub enum CallErrorType {
    InvalidMethod = 1,
    InvalidArgument = 2,
    TooManyArguments = 3,
    TooFewArguments = 4,
    InstanceIsNull = 5,
    MethodNotConst = 6,
}

Enum representing different errors during script instance calls.

Provides a type-safe way to handle call errors when implementing ScriptInstance::call(). It maps to the underlying GDExtensionCallErrorType constants from Godotโs C API.

Note that the OK variant is not included here, as it represents a successful call. This type is meant to be used in Result<T, CallErrorType>, where Ok(T) indicates success.

See SiMut::base_mut() for an example.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

InvalidMethod = 1

The method is invalid or was not found.

InvalidArgument = 2

One or more arguments cannot be converted to the expected parameter types.

TooManyArguments = 3

Too many arguments were provided to the method.

TooFewArguments = 4

Too few arguments were provided to the method.

InstanceIsNull = 5

The instance is null.

MethodNotConst = 6

The method is not const, but was called on a const instance.