godot::meta
Struct MethodInfo

pub struct MethodInfo {
    pub id: i32,
    pub method_name: StringName,
    pub class_name: ClassName,
    pub return_type: PropertyInfo,
    pub arguments: Vec<PropertyInfo>,
    pub default_arguments: Vec<Variant>,
    pub flags: MethodFlags,
}

Describes a method in Godot.

Abstraction of the low-level sys::GDExtensionMethodInfo.

Fields
id: i32method_name: StringNameclass_name: ClassNamereturn_type: PropertyInfoarguments: Vec<PropertyInfo>default_arguments: Vec<Variant>flags: MethodFlags
Implementations

impl MethodInfo

pub fn into_owned_method_sys(self) -> GDExtensionMethodInfo

Consumes self and turns it into a sys::GDExtensionMethodInfo, should be used together with free_owned_method_sys.

This will leak memory unless used together with free_owned_method_sys.

pub unsafe fn free_owned_method_sys(info: GDExtensionMethodInfo)

Properly frees a sys::GDExtensionMethodInfo created by into_owned_method_sys.

Safety

Must only be used on a struct returned from a call to into_owned_method_sys, without modification.

Must not be called more than once on a sys::GDExtensionMethodInfo struct.