godot::meta
Struct RefArg

pub struct RefArg<'r, T> { /* private fields */ }

Simple reference wrapper, used when passing arguments by-ref to Godot APIs.

This type is often used as the result of ToGodot::to_godot(), if Self is not a Copy type.

Implementations

impl<'r, T> RefArg<'r, T>

pub fn new(shared_ref: &'r T) -> RefArg<'r, T>

Creates a new RefArg from a reference.

Unless you implement your own ToGodot impl, there is usually no reason to use this.

pub fn get_ref(&self) -> &T

Returns the stored reference.

Panics

If T is Option<Gd<...>>::None.

pub fn get_ref_or_none(&self) -> Option<&T>
where T: GodotNullableFfi,

Returns the stored reference.

Returns None if T is Option<Gd<...>>::None.

pub fn to_owned(&self) -> T
where T: Clone,

Returns the stored reference.

Panics

If T is Option<Gd<...>>::None.