godot::meta

Struct Signature

pub struct Signature<Params, Ret> { /* private fields */ }

Implementations

impl<Params, Ret> Signature<Params, Ret>
where Params: InParamList, Ret: ToGodot,

In-calls:

Calls going from the Godot engine to rust code.

pub unsafe fn in_varcall( instance_ptr: *mut __GdextClassInstance, call_ctx: &CallContext<'_>, args_ptr: *const *const __GdextVariant, arg_count: i64, ret: *mut __GdextVariant, err: *mut GDExtensionCallError, func: fn(_: *mut __GdextClassInstance, _: Params) -> Ret, ) -> Result<(), CallError>

pub unsafe fn in_ptrcall( instance_ptr: *mut __GdextClassInstance, call_ctx: &CallContext<'_>, args_ptr: *const *const __GdextType, ret: *mut __GdextType, func: fn(_: *mut __GdextClassInstance, _: Params) -> Ret, call_type: PtrcallType, )

impl<Params, Ret> Signature<Params, Ret>
where Params: OutParamList, Ret: FromGodot,

Out-calls:

Calls going from the rust code to the Godot engine.

pub unsafe fn out_class_varcall( method_bind: ClassMethodBind, class_name: &'static str, method_name: &'static str, object_ptr: *mut __GdextObject, maybe_instance_id: Option<InstanceId>, args: Params, varargs: &[Variant], ) -> Result<Ret, CallError>

pub unsafe fn out_script_virtual_call( class_name: &'static str, method_name: &'static str, method_sname_ptr: *const __GdextStringName, object_ptr: *mut __GdextObject, args: Params, ) -> Ret

Available on since_api="4.3" only.

pub unsafe fn out_utility_ptrcall_varargs( utility_fn: unsafe extern "C" fn(_: *mut __GdextType, _: *const *const __GdextType, _: i32), function_name: &'static str, args: Params, varargs: &[Variant], ) -> Ret

pub unsafe fn out_builtin_ptrcall_varargs( builtin_fn: unsafe extern "C" fn(_: *mut __GdextType, _: *const *const __GdextType, _: *mut __GdextType, _: i32), class_name: &'static str, method_name: &'static str, type_ptr: *mut __GdextType, args: Params, varargs: &[Variant], ) -> Ret

pub unsafe fn out_class_ptrcall( method_bind: ClassMethodBind, class_name: &'static str, method_name: &'static str, object_ptr: *mut __GdextObject, maybe_instance_id: Option<InstanceId>, args: Params, ) -> Ret

pub unsafe fn out_builtin_ptrcall( builtin_fn: unsafe extern "C" fn(_: *mut __GdextType, _: *const *const __GdextType, _: *mut __GdextType, _: i32), class_name: &'static str, method_name: &'static str, type_ptr: *mut __GdextType, args: Params, ) -> Ret

pub unsafe fn out_utility_ptrcall( utility_fn: unsafe extern "C" fn(_: *mut __GdextType, _: *const *const __GdextType, _: i32), function_name: &'static str, args: Params, ) -> Ret