godot::meta
Trait ApiParam

pub trait ApiParam: GodotType + Sized {
    type Arg<'v>: AsArg<Self>
       where Self: 'v;

    // Required method
    fn value_to_arg<'v>(self) -> Self::Arg<'v>;
}

Implemented for all parameter types T that are allowed to receive impl AsArg<T>.

Required Associated Types

type Arg<'v>: AsArg<Self> where Self: 'v

Canonical argument passing type, either T or an internally-used CoW type.

The general rule is that Copy types are passed by value, while the rest is passed by reference.

This associated type is closely related to ToGodot::ToVia<'v> and may be reorganized.

Required Methods

fn value_to_arg<'v>(self) -> Self::Arg<'v>

Converts an owned value to the canonical argument type, which can be passed to impl AsArg<T>.

Useful in generic contexts where only a value is available, and one doesn’t want to dispatch between value/reference.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.