godot::meta
Trait ArgTarget

pub trait ArgTarget: Sized {
    type Type<'v>: AsArg<Self>
       where Self: 'v;
}

Implemented for all types that can be stored in AsArg<T>.

Required Associated Types

type Type<'v>: AsArg<Self> where Self: 'v

Target type, either T or CowArg<'v, T>.

The general rule is that Copy types are passed by value, while the rest is passed by reference.

This associated type is closely related to [ToGodot::ToVia<'v>] and may be reorganized.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.