godot::meta

Trait EngineToGodot

pub trait EngineToGodot: Sized + GodotConvert {
    type Pass: ArgPassing;

    // Required methods
    fn engine_to_godot(
        &self,
    ) -> <Self::Pass as ArgPassing>::Output<'_, Self::Via>;
    fn engine_to_variant(&self) -> Variant;

    // Provided method
    fn engine_to_godot_owned(&self) -> Self::Via
       where Self::Via: Clone { ... }
}

Engine-internal variant of ToGodot, used for engine APIs and virtual methods.

This trait exists to support types like u64 that work in engine contexts (backed by C++ uint64_t), but cannot be used in user-facing #[func] methods (as they don’t fit in GDScript/Variant, which can only store i64).

On the FFI level, u64 are passed as i64 (same bit pattern). The C++ side reinterprets the bits again as uint64_t in engine APIs and bitfields. User-defined GDScript code generally does not get into contact with i64 (except for bitfields).

For internal use only; see ToGodot for user-facing conversions.

Required Associated Types

type Pass: ArgPassing

Whether arguments of this type are passed by value or by reference.

Required Methods

fn engine_to_godot(&self) -> <Self::Pass as ArgPassing>::Output<'_, Self::Via>

Converts this type to Godot representation, optimizing for zero-copy when possible.

fn engine_to_variant(&self) -> Variant

Provided Methods

fn engine_to_godot_owned(&self) -> Self::Via
where Self::Via: Clone,

Converts this type to owned Godot representation.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.