godot::meta

Trait FfiRawPointer

pub trait FfiRawPointer: Copy + Sealed { }

Trait for raw pointers that can be passed over the Godot FFI boundary as i64 addresses.

This trait is implemented for *const T and *mut T, and used as a bound inside RawPtr.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.