godot::meta

Trait GodotConvert

pub trait GodotConvert {
    type Via: GodotType;
}

Indicates that a type can be passed to/from Godot, either directly or through an intermediate “via” type.

The associated type Via specifies how this type is passed across the FFI boundary to/from Godot. Generally ToGodot needs to be implemented to pass a type to Godot, and FromGodot to receive this type from Godot.

GodotType is a stronger bound than GodotConvert, since it expresses that a type is directly representable in Godot (without intermediate “via”). Every GodotType also implements GodotConvert with Via = Self.

Please read the godot::meta module docs for further information about conversions.

Required Associated Types

type Via: GodotType

The type through which Self is represented in Godot.