godot::meta

Trait InParamTuple

pub trait InParamTuple: ParamTuple {
    // Required method
    fn from_variant_array(array: &[&Variant]) -> Self;
}

Represents a parameter list that is received from some external location (usually Godot).

As an example, this would be used for user-defined functions that will be called from Godot, however this is not used when calling a Godot function from Rust code.

Required Methods

fn from_variant_array(array: &[&Variant]) -> Self

Converts array to Self by calling from_variant on each argument.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.