godot::meta

Trait OutParamTuple

pub trait OutParamTuple: ParamTuple {
    // Required methods
    fn with_variants<F, R>(self, f: F) -> R
       where F: FnOnce(&[Variant]) -> R;
    fn with_variant_pointers<F, R>(self, f: F) -> R
       where F: FnOnce(&[*const __GdextVariant]) -> R;
    fn with_type_pointers<F, R>(self, f: F) -> R
       where F: FnOnce(&[*const __GdextType]) -> R;
    fn to_variant_array(&self) -> Vec<Variant>;
}

Represents a parameter list that is used to call some external code.

As an example, this would be used to call Godot functions through FFI, however this is not used when Godot calls a user-defined function.

Required Methods

fn with_variants<F, R>(self, f: F) -> R
where F: FnOnce(&[Variant]) -> R,

Call f on the tuple self by first converting self to an array of Variants.

fn with_variant_pointers<F, R>(self, f: F) -> R
where F: FnOnce(&[*const __GdextVariant]) -> R,

Call f on the tuple self by first converting self to an array of Variant pointers.

fn with_type_pointers<F, R>(self, f: F) -> R
where F: FnOnce(&[*const __GdextType]) -> R,

Call f on the tuple self by first converting self to an array of Godot type pointers.

fn to_variant_array(&self) -> Vec<Variant>

Converts array to Self by calling to_variant on each argument.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.