godot::meta

Trait ParamTuple

pub trait ParamTuple: Sized {
    const LEN: usize;

    // Required method
    fn format_args(&self) -> String;

    // Provided method
    fn property_info(index: usize, param_name: &str) -> Option<PropertyInfo> { ... }
}

Represents a parameter list as Rust tuple where each tuple element is one parameter.

This trait only contains metadata for the parameter list, the actual functionality is contained in InParamTuple and OutParamTuple.

Required Associated Constants

const LEN: usize

The number of elements in this parameter list.

Required Methods

fn format_args(&self) -> String

Return a string representing the arguments.

Provided Methods

fn property_info(index: usize, param_name: &str) -> Option<PropertyInfo>

The property info of the parameter at index index.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.