godot::meta

Trait ThreadSafeArg

pub trait ThreadSafeArg: ThreadSafeArgContext + GodotConvert
where
    Self::Via: Send + Sync,
{ }

Maker trait to implement ThreadSafeArgContext for thread-safe types.

Due to type system constraints ThreadSafeArgContext can not be implemented for all Sync + Send types. This marker trait allows to implement the sealed trait for types that are thread-safe.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.