godot::obj

Enum OnEditor

pub enum OnEditor<T> {
    Null,
    Uninitialized(T),
    Initialized(T),
}

Represents exported property which must not be null and must be set via the editor – or associated code – before use. Allows to use Gd<T> – which by itself never holds null objects – as an #[export] which should not be null during the runtime.

Panics during access if value hasn’t been set. Checks if value has been set before the ready is being run and panics if any of OnEditor fields is not properly initialized. OnEditor<T> should always be used as a property, preferably in tandem with an #[export] or #[var]. It should be used as it would be a value itself and lack thereof treated as a logical error.

#[init] can be used to provide default values. One can create new instance and set its required properties after the init, albeit Option<Gd<T>> and OnReady<Gd<T>> should be preferred instead for late initialization.

Using OnEditor<T> with Gd<T> and DynGd<T, D>

Primarily use of OnEditor<Gd<T>> is exposing properties to the Godot editor.

One can provide default values for OnEditor using #[init].

Example - auto-generated init

 use godot::prelude::*;

#[derive(GodotClass)]
#[class(init, base = Node)]
struct MyClass {
    #[export]
    editor_field: OnEditor<Gd<Resource>>,
    #[export]
    #[init(val = OnEditor::new(Node::new_alloc()))]
    required_with_default: OnEditor<Gd<Node>>,
    // Does **NOT** require base field to work.
    base: Base<Node>,
}

#[godot_api]
impl INode for MyClass {
    fn ready(&mut self) {
        // Field `required_with_default` can be either default value - specified in `#[init]` or value set via the Godot Editor.
       assert_eq!(self.required_with_default.get_class(), GString::from("Node"));
        // Will always be valid and must be set via editor – an additional check is being run before ready to make sure that given value can't be null.
        let some_variant = self.editor_field.get_meta("SomeName");
    }
}

Example - user-generated init

 use godot::prelude::*;

#[derive(GodotClass)]
#[class(base = Node)]
struct MyClass {
    #[export]
    required_node: OnEditor<Gd<Node>>,
    base: Base<Node>
}

#[godot_api]
impl INode for MyClass {
    fn init(base: Base<Node>) -> Self {
       Self {
           base,
           required_node: OnEditor::new(Node::new_alloc()),
       }
    }
}

Example - Late init

 use godot::prelude::*;

#[derive(GodotClass)]
#[class(init, base = Node)]
struct SomeClassThatCanBeInstantiatedInCode {
    #[export]
    required_node: OnEditor<Gd<Node>>,
}

fn foo(mut this: Gd<Node>) {
    let mut my_node_to_add = SomeClassThatCanBeInstantiatedInCode::new_alloc();
    // Would cause the panic:
    // this.add_child(&my_node_to_add);
    // Note: Remember that nodes are manually managed. They will leak memory if not added to tree and/or pruned.
    my_node_to_add.bind_mut().required_node = OnEditor::new(Node::new_alloc());
    // Will not cause the panic.
    this.add_child(&my_node_to_add);
}

Using OnEditor<T> with other GodotTypes

OnEditor<T> can be used with other built-ins to provide an extra validation logic and making sure that given properties has been set. Example usage might be checking if entities has been granted proper generated ids.

In such cases user must specify value which will be deemed invalid. Accessing uninitialized value will cause the panic. To initialize given value simply replace it with Onready::new(…).

Example - using OnEditor with primitives

 use godot::prelude::*;

#[derive(GodotClass)]
#[class(init, base = Node)]
struct SomeClassThatCanBeInstantiatedInCode {
    #[export]
    #[init(val = OnEditor::uninit(42))]
    some_primitive: OnEditor<i64>,
}

fn foo(mut this: Gd<Node>) {
    let mut my_node_to_add = SomeClassThatCanBeInstantiatedInCode::new_alloc();
    // Would cause the panic:
    // this.add_child(&my_node_to_add);
    my_node_to_add.bind_mut().some_primitive = OnEditor::new(45);
    // Will not cause the panic.
    this.add_child(&my_node_to_add);
}

Variants

Null

Uninitialized(T)

Initialized(T)

Implementations

impl<T> OnEditor<T>
where T: GodotConvert + Var + FromGodot + PartialEq,

pub fn new(val: T) -> OnEditor<T>

pub fn uninit(val: T) -> OnEditor<T>