godot::obj

Module rpc

Structs

RpcBuilder

Represents an RPC, and the object that it can be called on, and is usually obtained through the type-safe RPC API. See the relevant section in the book for more information about type-safe RPC calls.

Enums

UserRpcObject

Represents an object that RPCs can be called on.

Traits

RpcCollection

Represents a collection of RPCs that can be constructed with a UserRpcObject.