godot::obj::script
Struct ScriptBaseMut

pub struct ScriptBaseMut<'a, T>
where
    T: ScriptInstance,
{ /* private fields */ }

Mutable/exclusive reference guard for a Base pointer.

This can be used to call methods on the base object of a Rust object, which takes &self or &mut self as the receiver.

See SiMut::base_mut() for usage.