godot::obj
Struct DynGd

pub struct DynGd<T, D>
where
    T: GodotClass,
    D: ?Sized,
{ /* private fields */ }

Smart pointer integrating Rust traits via dyn dispatch.

DynGd<T, D> extends a Godot object Gd<T> with functionality for Rustโs trait dynamic dispatch.
In this context, the type parameters have the following meaning:

T is the Godot class.

D is a trait object dyn Trait, where T: Trait.

To register the T -> D relation with godot-rust, T must implement AsDyn<D>. This can be automated with the #[godot_dyn] attribute macro.

Construction and API

You can convert between Gd and DynGd using Gd::into_dyn() and DynGd::into_gd(). The former sometimes needs an explicit ::<dyn Trait> type argument, but can often be inferred.

The DynGd API is very close to Gd. In fact, both Deref and DerefMut are implemented for DynGd -> Gd, so you can access all the underlying Gd methods as well as Godot class APIs directly.

The main new parts are two methods dyn_bind() and dyn_bind_mut(). These are very similar to Gdโs bind() and bind_mut(), but return a reference guard to the trait object D instead of the Godot class T.

Example

use godot::obj::{Gd, DynGd,NewGd};
use godot::register::{godot_dyn, GodotClass};
use godot::classes::RefCounted;

#[derive(GodotClass)]
#[class(init)]
struct Monster {
   #[init(val = 100)]
    hitpoints: u16,
}

trait Health {
    fn is_alive(&self) -> bool;
    fn deal_damage(&mut self, damage: u16);
}

// The #[godot_dyn] attribute macro registers the dynamic relation in godot-rust.
// Traits are implemented as usual.
#[godot_dyn]
impl Health for Monster {
    fn is_alive(&self) -> bool {
        self.hitpoints > 0
    }

    fn deal_damage(&mut self, damage: u16) {
        self.hitpoints = self.hitpoints.saturating_sub(damage);
    }
}

// Create a Gd<Monster> and convert it -> DynGd<Monster, dyn Health>.
let monster = Monster::new_gd();
let dyn_monster = monster.into_dyn::<dyn Health>();

// Now upcast it to its base class -> type is DynGd<RefCounted, dyn Health>.
let mut dyn_monster = dyn_monster.upcast::<RefCounted>();

// Due to RefCounted abstraction, you can no longer access concrete Monster properties.
// However, the trait Health is still accessible through dyn_bind().
assert!(dyn_monster.dyn_bind().is_alive());

// To mutate the object, call dyn_bind_mut(). Rust borrow rules apply.
let mut guard = dyn_monster.dyn_bind_mut();
guard.deal_damage(120);
assert!(!guard.is_alive());

Implementations

impl<T, D> DynGd<T, D>
where T: GodotClass, D: ?Sized,

pub fn dyn_bind(&self) -> DynGdRef<'_, D>

Acquires a shared reference guard to the trait object D.

The resulting guard implements Deref<Target = D>, allowing shared access to the traitโs methods.

See Gd::bind() for borrow checking semantics and panics.

pub fn dyn_bind_mut(&mut self) -> DynGdMut<'_, D>

Acquires an exclusive reference guard to the trait object D.

The resulting guard implements DerefMut<Target = D>, allowing exclusive mutable access to the traitโs methods.

See Gd::bind_mut() for borrow checking semantics and panics.

pub fn upcast<Base>(self) -> DynGd<Base, D>
where Base: GodotClass, T: Inherits<Base>,

Upcast to a Godot base, while retaining the D trait object.

This is useful when you want to gather multiple objects under a common Godot base (e.g. Node), but still enable common functionality. The common functionality is still accessible through D even when upcasting.

See also Gd::upcast().

pub fn into_gd(self) -> Gd<T>

Downgrades to a Gd<T> pointer, abandoning the D abstraction.

impl<T, D> DynGd<T, D>
where T: GodotClass<Memory = MemManual> + Bounds, D: ?Sized,

pub fn free(self)