godot::obj

Struct OnEditor

pub struct OnEditor<T> { /* private fields */ }

Represents exported Gd<T> property which must not be null and must be set via the editor โ or associated code โ before use.

Panics during access if value hasnโt been set. Checks if value has been set before the ready is being run and panics if OnEditor fields are not properly initialized. Should always be used as a property, preferably in tandem with an #[export].

The underlying type is de facto a wrapper for an Option<Gd<T> which dereferences to underlying value. It should be used as it would be a value itself and lack thereof treated as a logical error.

#[init] can be used to provide default values โ for example default Resources supposed to be filled by user. One can create new instance and set its required properties after the init, albeit Option<Gd<T>> and OnReady<Gd<T>> should be preferred instead for late initialization.

Example - auto-generated init

 use godot::prelude::*;

#[derive(GodotClass)]
#[class(init, base = Node)]
struct MyClass {
    #[export]
    editor_field: OnEditor<Gd<Resource>>,
    #[export]
    #[init(val = OnEditor::new(Node::new_alloc()))]
    required_with_default: OnEditor<Gd<Node>>,
    // Does **NOT** require base field to work.
    base: Base<Node>,
}

#[godot_api]
impl INode for MyClass {
    fn ready(&mut self) {
        // Field `required_with_default` can be either default value - specified in `#[init]` or value set via the Godot Editor.
       assert_eq!(self.required_with_default.get_class(), GString::from("Node"));
        // Will always be valid and must be set via editor โ an additional check is being run before ready to make sure that given value can't be null.
        let some_variant = self.editor_field.get_meta("SomeName");
    }
}

Example - user-generated init

 use godot::prelude::*;

#[derive(GodotClass)]
#[class(base = Node)]
struct MyClass {
    #[export]
    required_node: OnEditor<Gd<Node>>,
    base: Base<Node>
}

#[godot_api]
impl INode for MyClass {
    fn init(base: Base<Node>) -> Self {
       Self {
           base,
           required_node: OnEditor::new(Node::new_alloc()),
       }
    }
}

Example - Late init

 use godot::prelude::*;

#[derive(GodotClass)]
#[class(init, base = Node)]
struct SomeClassThatCanBeInstantiatedInCode {
    #[export]
    required_node: OnEditor<Gd<Node>>,
    base: Base<Node>,
}

fn foo(mut this: Gd<Node>) {
    let mut my_node_to_add = SomeClassThatCanBeInstantiatedInCode::new_alloc();
    // Would cause the panic:
    // this.add_child(&my_node_to_add);
    my_node_to_add.bind_mut().required_node = OnEditor::new(Node::new_alloc());
    // Will not cause the panic.
    this.add_child(&my_node_to_add);
}

Implementations

impl<T> OnEditor<T>
where T: GodotConvert,

pub fn new(val: T) -> OnEditor<T>