Struct godot::obj::RawGd

pub struct RawGd<T>where
    T: GodotClass,{ /* private fields */ }

A raw pointer to an instance of a class.

This is not intended for general use, you usually want to use Gd instead of using this directly.

RawGd is not guaranteed to be live, nor non-null. But many of the methods are only safe to call on null or live instances.

Implementations

impl<T> RawGd<T>where T: GodotClass,

pub fn new_null() -> RawGd<T>

Construct a new null object. This does not require any ffi.

pub fn try_from_instance_id(instance_id: InstanceId) -> Option<RawGd<T>>

Looks up the given instance ID and returns the associated object, if possible.

If no such instance ID is registered, or if the dynamic type of the object behind that instance ID is not compatible with T, then None is returned.

pub fn instance_id_or_none_unchecked(&self) -> Option<InstanceId>

Return the cached instance id without checking if the object is still live or not.

pub fn instance_id_or_none(&self) -> Option<InstanceId>

Return the instance id of this instance, or None if the object is null or has been freed.

pub fn is_instance_valid(&self) -> bool

Checks if this pointer points to a live object.

Many unsafe methods of RawGd require the object to be live before they can be safely called, this is the main way to check that.

pub fn is_null(&self) -> bool

Checks if this is a null-pointer, this does not check if the pointer points to a live object, for that you should use [is_instance_valid()].

pub unsafe fn free(self)

Safety

self must not be a dead pointer.

self must not be used again after this, this includes dropping a Gd which inherits from RefCounted .

impl<T> RawGd<T>where T: GodotClass<Declarer = UserDomain>,

pub fn new(user_object: T) -> RawGd<T>

Moves a user-created object into this smart pointer, submitting ownership to the Godot engine.

This is only useful for types T which do not store their base objects (if they have a base, you cannot construct them standalone).

pub fn new_default() -> RawGd<T>where T: GodotInit,

Creates a default-constructed instance of T inside a smart pointer.

This is equivalent to the GDScript expression T.new().

pub fn with_base<F>(init: F) -> RawGd<T>where F: FnOnce(Base<<T as GodotClass>::Base>) -> T,

Creates a RawGd<T> using a function that constructs a T from a provided base.

See also [Gd::with_base].

pub fn bind(&self) -> GdRef<'_, T>

Hands out a guard for a shared borrow, through which the user instance can be read.

The pattern is very similar to interior mutability with standard RefCell. You can either have multiple GdRef shared guards, or a single GdMut exclusive guard to a Rust GodotClass instance, independently of how many smart pointers point to it. There are runtime checks to ensure that Rust safety rules (e.g. no & and &mut coexistence) are upheld.

Panics

If another smart pointer pointing to the same Rust instance has a live GdMut guard bound.

If there is an ongoing function call from GDScript to Rust, which currently holds a &mut T reference to the user instance. This can happen through re-entrancy (Rust -> GDScript -> Rust call).

pub fn bind_mut(&mut self) -> GdMut<'_, T>

Hands out a guard for an exclusive borrow, through which the user instance can be read and written.

The pattern is very similar to interior mutability with standard RefCell. You can either have multiple GdRef shared guards, or a single GdMut exclusive guard to a Rust GodotClass instance, independently of how many smart pointers point to it. There are runtime checks to ensure that Rust safety rules (e.g. no &mut aliasing) are upheld.

Panics

If another smart pointer pointing to the same Rust instance has a live GdRef or GdMut guard bound.

If there is an ongoing function call from GDScript to Rust, which currently holds a &T or &mut T reference to the user instance. This can happen through re-entrancy (Rust -> GDScript -> Rust call).

impl<T> RawGd<T>where T: GodotClass<Declarer = EngineDomain>,

pub unsafe fn as_inner(&self) -> &T

Get a reference to the underlying Godot object.

Safety

self must not have been previously freed.

pub unsafe fn as_inner_mut(&mut self) -> &mut T

Get a mutable reference to the underlying Godot object.

Safety

self must not have been previously freed.