Trait godot::obj::AsArg

pub trait AsArg<T>
where
    T: GodotClass<Declarer = DeclEngine> + Bounds,
{
    // Required method
    fn as_arg(&self) -> RawGd<T>;
}

Objects that can be passed as arguments to Godot engine functions.

This trait is implemented for the following types:

Gd<T> and &Gd<T>, to pass objects. Subclasses of T are explicitly supported.

Option<Gd<T>> and Option<&Gd<T>>, to pass optional objects. None is mapped to a null argument.

The GDExtension API does not provide information about nullability of its function parameters. It is up to you to verify that the arguments you pass are only null when this is allowed. Doing this wrong _should_ be safe, but can lead to the function call failing.

Required Methods

fn as_arg(&self) -> RawGd<T>