Trait godot::obj::ExportableObject

pub trait ExportableObject: GodotClass { }

Trait implemented for all objects that inherit from Resource or Node.

Those are the only objects you can export to the editor.

Object Safety

This trait is not object safe.