Trait godot::obj::GodotSingleton

pub unsafe trait GodotSingleton: GodotClass<Declarer = EngineDomain> { }

Trait to indicate what classes are thread-safe global singletons.

Safety

You must guarantee that the type implementing this trait fulfils the safety requirements as described by https://docs.godotengine.org/en/latest/tutorials/performance/thread_safe_apis.html#global-scope.

Singletons must live for the entire duration of a gdextension library.