Trait godot::obj::Share

pub trait Share {
    // Required method
    fn share(&self) -> Self;
}

Trait to create more references from a smart pointer or collection.

Required Methods

fn share(&self) -> Self

Creates a new reference that points to the same object.

If the referred-to object is reference-counted, this will increment the count.