godot::obj

Trait WithDeferredCall

pub trait WithDeferredCall<T>
where
    T: GodotClass,
{
    // Required method
    fn apply_deferred<F>(&mut self, rust_function: F)
       where F: FnMut(&mut T) + 'static;
}

Available on since_api="4.2" only.

Trait that is automatically implemented for engine classes and user classes containing a Base<T> field.

This trait enables type safe deferred method calls.

Usage

fn some_fn(mut shape: Gd<CollisionShape2D>)
{
    shape.apply_deferred(|shape_mut| shape_mut.rotate(PI))
}

Required Methods

fn apply_deferred<F>(&mut self, rust_function: F)
where F: FnMut(&mut T) + 'static,

Runs the given Closure deferred.

This can be a type-safe alternative to crate::classes::Object::call_deferred. This method must be used on the main thread.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.