godot::obj

Trait WithSignals

pub trait WithSignals: GodotClass {
    type SignalCollection<'c>;
}

Implemented for all classes with registered signals, both engine- and user-declared.

This trait enables the Gd::signals() method.

User-defined classes with #[signal] additionally implement WithUserSignals.

Required Associated Types

type SignalCollection<'c>

The associated struct listing all signals of this class.

'c denotes the lifetime during which the class instance is borrowed and its signals can be modified.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.