godot::prelude

Function get_autoload

pub fn get_autoload<T>() -> Gd<T>
where
    T: Inherits<Node>,

Retrieves an autoload by its type, using the class name as the autoload name.

See Godot docs for an explanation of the autoload concept. Godot sometimes uses the term “autoload” interchangeably with “singleton”; we strictly refer to the former to separate from Singleton objects.

This is a convenience function that calls try_get_named_autoload() with the class name derived from T::class_id().

Panics

If the autoload is not found or cannot be cast to type T.

Example

use godot::prelude::*;

#[derive(GodotClass)]
#[class(init, base=Node)]
struct GlobalStats {
    base: Base<Node>,
}

// Assuming "GlobalStats" is registered as an autoload in `project.godot`,
// this returns the one instance of type Gd<MyAutoload>.
let stats = get_autoload::<GlobalStats>();