godot::prelude

Macro match_class

macro_rules! match_class {
    ($subject:expr, { $($arms:tt)* }) => { ... };
    (@munch $evt:ident;
        $var:ident @ $($class:ident)::+ => $body:expr,
        $($rest:tt)*
    ) => { ... };
    (@munch $evt:ident;
        $fallback_var:ident @ $pat:tt => $fallback:expr
        $(,)?
    ) => { ... };
    (@munch $evt:ident;
        _ => $fallback:expr
        $(,)?
    ) => { ... };
}

Dispatches a class to different subclasses.

Similar to a match statement, but with downcasts. Earlier matches dominate, so keep more-derived classes first. The current implementation checks Gd::try_cast() linearly with the number of branches. This may change in the future.

Requires a fallback branch, even if all direct known classes are handled. The reason for this is that there may be other subclasses which are not statically known by godot-rust (e.g. from a script or GDExtension). The fallback branch can either be _ (discard object), or variable @ _ to access the original object inside the fallback arm.

Example

// Basic syntax.
let event: Gd<InputEvent> = some_input();

let simple_dispatch: i32 = match_class!(event, {
   button @ InputEventMouseButton => 1,
   motion @ InputEventMouseMotion => 2,
   action @ InputEventAction => 3,
   _ => 0, // Fallback.
});

// More diverse dispatch patterns are also supported.
let fancy_dispatch: i32 = match_class!(some_input(), {
    button @ InputEventMouseButton => 1,

    // Block syntax for multiple statements:
    motion @ InputEventMouseMotion => {
        godot_print!("motion");
        2
    },

    // Qualified types supported:
    action @ godot::classes::InputEventAction => 3,

    // Fallback with variable -- retrieves original Gd<InputEvent>.
    // Equivalent to pattern `original @ InputEvent`.
    original @ _ => 0,
});

// event_type is now 0, 1, 2, or 3

Limitations

The expression block is currently wrapped by a closure, so you cannot use control-flow statements like ?, return, continue, break.