godot::prelude

Macro varray

macro_rules! varray {
    ($($elements:expr),* $(,)?) => { ... };
}

Constructs VariantArray literals, similar to Rust’s standard vec! macro.

The type of the array elements is always Variant.

Example

let arr: VariantArray = varray![42_i64, "hello", true];

See also

To create a typed Array with a single element type, see the array! macro.

For dictionaries, a similar macro dict! exists.