Module godot::prelude::math

Math-related functions and traits like ApproxEq.

Macros

assert_eq_approx

Asserts that two values are approximately equal

assert_ne_approx

Asserts that two values are not approximately equal, using the provided func for equality checking.

Constants

CMP_EPSILON

Traits

ApproxEq

Approximate equality-comparison of geometric types.

Functions

bezier_derivative

bezier_interpolate

cubic_interpolate

cubic_interpolate_in_time

fposmod

is_angle_equal_approx

Check if two angles are approximately equal, by comparing the distance between the points on the unit circle with 0 using is_equal_approx.

is_equal_approx

Approximate equality of floating-point types.

is_zero_approx

lerp

lerp_angle

Linearly interpolates between two angles (in radians) by a weight value between 0.0 and 1.0.

sign

snapped