Struct godot::prelude::Array

#[repr(C)]
pub struct Array<T>where
    T: GodotType,{ /* private fields */ }

Godot’s Array type.

Unlike GDScript, all indices and sizes are unsigned, so negative indices are not supported.

Typed arrays

Godot’s Array can be either typed or untyped.

An untyped array can contain any kind of Variant, even different types in the same array. We represent this in Rust as VariantArray, which is just a type alias for Array<Variant>.

Godot also supports typed arrays, which are also just Variant arrays under the hood, but with runtime checks that no values of the wrong type are put into the array. We represent this as Array<T>, where the type T implements GodotType.

Reference semantics

Like in GDScript, Array acts as a reference type: multiple Array instances may refer to the same underlying array, and changes to one are visible in the other.

To create a copy that shares data with the original array, use Clone::clone(). If you want to create a copy of the data, use duplicate_shallow() or duplicate_deep().

Thread safety

Usage is safe if the Array is used on a single thread only. Concurrent reads on different threads are also safe, but any writes must be externally synchronized. The Rust compiler will enforce this as long as you use only Rust threads, but it cannot protect against concurrent modification on other threads (e.g. created through GDScript).

Implementations

impl<T> Array<T>where T: GodotType,

pub fn len(&self) -> usize

Returns the number of elements in the array. Equivalent of size() in Godot.

Retrieving the size incurs an FFI call. If you know the size hasn’t changed, you may consider storing it in a variable. For loops, prefer iterators.

pub fn is_empty(&self) -> bool

Returns true if the array is empty.

Checking for emptiness incurs an FFI call. If you know the size hasn’t changed, you may consider storing it in a variable. For loops, prefer iterators.

pub fn hash(&self) -> u32

Returns a 32-bit integer hash value representing the array and its contents.

Note: Arrays with equal content will always produce identical hash values. However, the reverse is not true. Returning identical hash values does not imply the arrays are equal, because different arrays can have identical hash values due to hash collisions.

pub fn clear(&mut self)

Clears the array, removing all elements.

pub fn resize(&mut self, size: usize)

Resizes the array to contain a different number of elements. If the new size is smaller, elements are removed from the end. If the new size is larger, new elements are set to Variant::nil().

pub fn reverse(&mut self)

Reverses the order of the elements in the array.

pub fn sort_unstable(&mut self)

Sorts the array.

Note: The sorting algorithm used is not stable. This means that values considered equal may have their order changed when using sort_unstable.

pub fn shuffle(&mut self)

Shuffles the array such that the items will have a random order. This method uses the global random number generator common to methods such as randi. Call randomize to ensure that a new seed will be used each time if you want non-reproducible shuffling.

impl<T> Array<T>where T: GodotType,

pub fn new() -> Array<T>

Constructs an empty Array.

pub fn duplicate_shallow(&self) -> Array<T>

Returns a shallow copy of the array. All array elements are copied, but any reference types (such as Array, Dictionary and Object) will still refer to the same value.

To create a deep copy, use duplicate_deep() instead. To create a new reference to the same array data, use clone().

pub fn duplicate_deep(&self) -> Array<T>

Returns a deep copy of the array. All nested arrays and dictionaries are duplicated and will not be shared with the original array. Note that any Object-derived elements will still be shallow copied.

To create a shallow copy, use duplicate_shallow() instead. To create a new reference to the same array data, use clone().

pub fn subarray_shallow( &self, begin: usize, end: usize, step: Option<isize> ) -> Array<T>

Returns a sub-range begin..end, as a new array.

The values of begin (inclusive) and end (exclusive) will be clamped to the array size.

If specified, step is the relative index between source elements. It can be negative, in which case begin must be higher than end. For example, TypedArray::from(&[0, 1, 2, 3, 4, 5]).slice(5, 1, -2) returns [5, 3].

Array elements are copied to the slice, but any reference types (such as Array, Dictionary and Object) will still refer to the same value. To create a deep copy, use subarray_deep() instead.

pub fn subarray_deep( &self, begin: usize, end: usize, step: Option<isize> ) -> Array<T>

Returns a sub-range begin..end, as a new Array.

The values of begin (inclusive) and end (exclusive) will be clamped to the array size.

If specified, step is the relative index between source elements. It can be negative, in which case begin must be higher than end. For example, TypedArray::from(&[0, 1, 2, 3, 4, 5]).slice(5, 1, -2) returns [5, 3].

All nested arrays and dictionaries are duplicated and will not be shared with the original array. Note that any Object-derived elements will still be shallow copied. To create a shallow copy, use subarray_shallow() instead.

pub fn extend_array(&mut self, other: Array<T>)

Appends another array at the end of this array. Equivalent of append_array in GDScript.

impl<T> Array<T>where T: GodotType + FromGodot,

pub fn iter_shared(&self) -> Iter<'_, T>

Returns an iterator over the elements of the Array. Note that this takes the array by reference but returns its elements by value, since they are internally converted from Variant.

Notice that it’s possible to modify the Array through another reference while iterating over it. This will not result in unsoundness or crashes, but will cause the iterator to behave in an unspecified way.

pub fn get(&self, index: usize) -> T

Returns the value at the specified index.

Panics

If index is out of bounds.

pub fn first(&self) -> Option<T>

Returns the first element in the array, or None if the array is empty. Equivalent of front() in GDScript.

pub fn last(&self) -> Option<T>

Returns the last element in the array, or None if the array is empty. Equivalent of back() in GDScript.

pub fn min(&self) -> Option<T>

Returns the minimum value contained in the array if all elements are of comparable types. If the elements can’t be compared or the array is empty, None is returned.

pub fn max(&self) -> Option<T>

Returns the maximum value contained in the array if all elements are of comparable types. If the elements can’t be compared or the array is empty, None is returned.

pub fn pick_random(&self) -> Option<T>

Returns a random element from the array, or None if it is empty.

pub fn pop(&mut self) -> Option<T>

Removes and returns the last element of the array. Returns None if the array is empty. Equivalent of pop_back in GDScript.

pub fn pop_front(&mut self) -> Option<T>

Removes and returns the first element of the array. Returns None if the array is empty.

Note: On large arrays, this method is much slower than pop as it will move all the array’s elements. The larger the array, the slower pop_front will be.

pub fn remove(&mut self, index: usize) -> T

Removes and returns the element at the specified index. Equivalent of pop_at in GDScript.

On large arrays, this method is much slower than pop_back as it will move all the array’s elements after the removed element. The larger the array, the slower remove will be.

Panics

If index is out of bounds.

impl<T> Array<T>where T: GodotType + ToGodot,

pub fn binary_search(&self, value: &T) -> usize

Finds the index of an existing value in a sorted array using binary search. Equivalent of bsearch in GDScript.

If the value is not present in the array, returns the insertion index that would maintain sorting order.

Calling binary_search on an unsorted array results in unspecified behavior.

pub fn count(&self, value: &T) -> usize

Returns the number of times a value is in the array.

pub fn contains(&self, value: &T) -> bool

Returns true if the array contains the given value. Equivalent of has in GDScript.

pub fn find(&self, value: &T, from: Option<usize>) -> Option<usize>

Searches the array for the first occurrence of a value and returns its index, or None if not found. Starts searching at index from; pass None to search the entire array.

pub fn rfind(&self, value: &T, from: Option<usize>) -> Option<usize>

Searches the array backwards for the last occurrence of a value and returns its index, or None if not found. Starts searching at index from; pass None to search the entire array.

pub fn set(&mut self, index: usize, value: T)

Sets the value at the specified index.

Panics

If index is out of bounds.

pub fn push(&mut self, value: T)

Appends an element to the end of the array. Equivalent of append and push_back in GDScript.

pub fn push_front(&mut self, value: T)

Adds an element at the beginning of the array. See also push.

Note: On large arrays, this method is much slower than push as it will move all the array’s elements. The larger the array, the slower push_front will be.

pub fn insert(&mut self, index: usize, value: T)

Inserts a new element at a given index in the array. The index must be valid, or at the end of the array (index == len()).

Note: On large arrays, this method is much slower than push as it will move all the array’s elements after the inserted element. The larger the array, the slower insert will be.

pub fn erase(&mut self, value: &T)

Removes the first occurrence of a value from the array. If the value does not exist in the array, nothing happens. To remove an element by index, use remove instead.

On large arrays, this method is much slower than pop_back as it will move all the array’s elements after the removed element. The larger the array, the slower remove will be.

pub fn fill(&mut self, value: &T)

Assigns the given value to all elements in the array. This can be used together with resize to create an array with a given size and initialized elements.