Struct godot::prelude::Callable

pub struct Callable { /* private fields */ }

A Callable represents a function in Godot.

Usually a callable is a reference to an Object and a method name, this is a standard callable. But can also be a custom callable, which is usually created from bind, unbind, or a GDScript lambda. See Callable::is_custom.

Currently it is impossible to use bind and unbind in GDExtension, see godot-cpp#802.

Implementations

impl Callable

pub fn from_object_method<T, S>(object: Gd<T>, method_name: S) -> Callablewhere T: GodotClass, S: Into<StringName>,

Create a callable for the method object::method_name.

Godot equivalent: Callable(Object object, StringName method)

pub fn callv(&self, arguments: Array<Variant>) -> Variant

Calls the method represented by this callable.

Arguments passed should match the method’s signature.

If called with more arguments than expected by the method, the extra arguments will be ignored and the call continues as normal.

If called with fewer arguments than expected it will crash Godot, without triggering UB.

If called with arguments of the wrong type then an error will be printed and the call will return NIL.

If called on an invalid Callable then no error is printed, and NIL is returned.

Godot equivalent: callv

pub fn method_name(&self) -> Option<StringName>

Returns the name of the method represented by this callable. If the callable is a lambda function, returns the function’s name.

Known Bugs

Getting the name of a lambda errors instead of returning its name, see godot#73052.

Godot equivalent: get_method

pub fn object(&self) -> Option<Gd<Object>>

Returns the object on which this callable is called.

Returns None when this callable doesn’t have any target object to call a method on, regardless of if the method exists for that target or not.

Godot equivalent: get_object

pub fn object_id(&self) -> Option<InstanceId>

Returns the ID of this callable’s object, see also Gd::instance_id.

Returns None when this callable doesn’t have any target to call a method on.

Godot equivalent: get_object_id

pub fn hash(&self) -> u32

Returns the 32-bit hash value of this callable’s object.

Godot equivalent: hash

pub fn is_custom(&self) -> bool

Returns true if this callable is a custom callable.

Custom callables are mainly created from bind or unbind. In GDScript, lambda functions are also custom callables.

If a callable is not a custom callable, then it is considered a standard callable, this function is the opposite of Callable.is_standard.

Godot equivalent: is_custom

pub fn is_null(&self) -> bool

Returns true if this callable has no target to call the method on.

This is not the negated form of [is_valid], as is_valid will return false if the callable has a target but the method does not exist.

Godot equivalent: is_null

pub fn is_valid(&self) -> bool

Returns true if the callable’s object exists and has a valid method name assigned, or is a custom callable.

Godot equivalent: is_valid