Struct godot::prelude::GFile

pub struct GFile { /* private fields */ }

Implementations

impl GFile

pub const DEFAULT_BUFFER_SIZE: usize = 4_096usize

Default GFile buffer size. Relevant for its BufRead implementation

pub fn set_buffer_size(&mut self, buffer_size: usize)

Sets the internal buffer size. Relevant for its BufRead implementation

pub fn get_buffer_size(&self) -> usize

Gets current internal buffer size. Relevant for its BufRead implementation

pub fn open(path: GodotString, flags: ModeFlags) -> Result<GFile, Error>

Opens a file located at path. For ModeFlags description check the GFile documentation

pub fn open_compressed( path: GodotString, flags: ModeFlags, compression_mode: CompressionMode ) -> Result<GFile, Error>

Opens a compressed file located at path. Can read only files compressed by Godot compression formats. For ModeFlags description check the GFile documentation

pub fn open_encrypted( path: GodotString, flags: ModeFlags, key: PackedByteArray ) -> Result<GFile, Error>

Opens a file encrypted by 32-byte long PackedByteArray located at path. For ModeFlags description check the GFile documentation

pub fn open_encrypted_with_pass( path: GodotString, flags: ModeFlags, pass: GodotString ) -> Result<GFile, Error>

Opens a file encrypted with a pass located at path. For ModeFlags description check the GFile documentation

pub fn close(self)

Closes the inner Gd<FileAccess> and destroys the GFile instance

pub fn into_inner(self) -> Gd<FileAccess>

Deconstructs the GFile intance and retrieves the inner Gd<FileAccess>

pub fn get_8(&self) -> Result<u8, Error>

Reads the next 8 bits from the file as u8. Exposed Godot method.

pub fn get_16(&self) -> Result<u16, Error>

Reads the next 16 bits from the file as u16. For details check the Godot documentation

pub fn get_32(&self) -> Result<u32, Error>

Reads the next 32 bits from the file as u32. Exposed Godot method. For details check the Godot documentation

pub fn get_64(&self) -> Result<u64, Error>

Reads the next 64 bits from the file as u64. Exposed Godot method. For details check the Godot documentation

pub fn get_pascal_string(&mut self) -> Result<GodotString, Error>

Reads the next 64 bits from the file as u64. Exposed Godot method. For details check the Godot documentation

pub fn get_as_text(&self, skip_cr: bool) -> Result<GodotString, Error>

Reads the whole file as GodotString. Text is interpeted as UTF-8 encoded. If skip_cr is set to true, carriage returns will be ingored when parsing and only line feed characters will represents a new line For details check the Godot documentation

pub fn get_csv_line( &self, delim: GodotString ) -> Result<PackedStringArray, Error>

Reads the next line of the file in delimiter-separated file. For traditional CSV format, provide comma (",") as delim For details check the Godot documentation

pub fn get_float(&self) -> Result<f32, Error>

Reads the next 32 bits from file as f32. POSSIBLE DUPLICATE / For details check the Godot documentation

pub fn get_real(&self) -> Result<f32, Error>

Reads the next 32 bits from file as f32. POSSIBLE DUPLICATE /
For details check the Godot documentation

pub fn get_double(&self) -> Result<f64, Error>

Reads the next 64 bits from file as f64. For details check the Godot documentation

pub fn get_line(&self) -> Result<GodotString, Error>

Reads the next line of the file as GodotString For details check the Godot documentation

pub fn get_var(&self, allow_objects: bool) -> Result<Variant, Error>

Reads the next Variant value from file. If allow_objects is set to true, objects will be decoded For details check the Godot documentation

pub fn store_8(&mut self, value: u8) -> Result<(), Error>

pub fn store_16(&mut self, value: u16) -> Result<(), Error>

pub fn store_32(&mut self, value: u32) -> Result<(), Error>

pub fn store_64(&mut self, value: u64) -> Result<(), Error>

pub fn store_double(&mut self, value: f64) -> Result<(), Error>

pub fn store_float(&mut self, value: f32) -> Result<(), Error>

pub fn store_real(&mut self, value: f32) -> Result<(), Error>

pub fn store_string(&mut self, value: GodotString) -> Result<(), Error>

pub fn store_pascal_string(&mut self, value: GodotString) -> Result<(), Error>

pub fn store_line(&mut self, value: GodotString) -> Result<(), Error>

pub fn store_csv_line( &mut self, values: PackedStringArray, delim: GodotString ) -> Result<(), Error>

pub fn store_variant( &mut self, value: Variant, full_objects: bool ) -> Result<(), Error>

pub fn set_big_endian(&mut self, value: bool)

pub fn is_big_endian(&self) -> bool

pub fn get_path(&self) -> GodotString

pub fn get_path_absolute(&self) -> GodotString

pub fn get_position(&self) -> u64

Returns the current file position