Struct godot::prelude::Projection

pub struct Projection {
    pub cols: [Vector4; 4],
}

A 4x4 matrix used for 3D projective transformations. It can represent transformations such as translation, rotation, scaling, shearing, and perspective division. It consists of four Vector4 columns.

For purely linear transformations (translation, rotation, and scale), it is recommended to use Transform3D, as it is more performant and has a lower memory footprint.

Used internally as Camera3D’s projection matrix.

Note: The current implementation largely makes calls to godot for its methods and as such are not as performant as other types.

Fields
cols: [Vector4; 4]

The columns of the projection matrix.

Implementations

impl Projection

pub const IDENTITY: Projection = Self::from_diagonal(1.0, 1.0, 1.0, 1.0)

A Projection with no transformation defined. When applied to other data structures, no transformation is performed.

Godot equivalent: Projection.IDENTITY

pub const ZERO: Projection = Self::from_diagonal(0.0, 0.0, 0.0, 0.0)

A Projection with all values initialized to 0. When applied to other data structures, they will be zeroed.

Godot equivalent: Projection.ZERO

pub const fn new(cols: [Vector4; 4]) -> Projection

Create a new projection from a list of column vectors.

pub const fn from_diagonal(x: f32, y: f32, z: f32, w: f32) -> Projection

Create a diagonal matrix from the given values.

pub const fn from_cols( x: Vector4, y: Vector4, z: Vector4, w: Vector4 ) -> Projection

Create a matrix from four column vectors.

Godot equivalent: Projection(Vector4 x_axis, Vector4 y_axis, Vector4 z_axis, Vector4 w_axis)

pub fn create_depth_correction(flip_y: bool) -> Projection

Creates a new Projection that projects positions from a depth range of -1 to 1 to one that ranges from 0 to 1, and flips the projected positions vertically, according to flip_y.

Godot equivalent: Projection.create_depth_correction()

pub fn create_for_hmd( eye: ProjectionEye, aspect: f32, intraocular_dist: f32, display_width: f32, display_to_lens: f32, oversample: f32, near: f32, far: f32 ) -> Projection

Creates a new Projection for projecting positions onto a head-mounted display with the given X:Y aspect ratio, distance between eyes, display width, distance to lens, oversampling factor, and depth clipping planes.

Godot equivalent: Projection.create_for_hmd()

pub fn create_frustum( left: f32, right: f32, bottom: f32, top: f32, near: f32, far: f32 ) -> Projection

Creates a new Projection that projects positions in a frustum with the given clipping planes.

Godot equivalent: Projection.create_frustum()

pub fn create_frustum_aspect( size: f32, aspect: f32, offset: Vector2, near: f32, far: f32, flip_fov: bool ) -> Projection

Creates a new Projection that projects positions in a frustum with the given size, X:Y aspect ratio, offset, and clipping planes.

flip_fov determines whether the projection’s field of view is flipped over its diagonal.

Godot equivalent: Projection.create_frustum_aspect()

pub fn create_orthogonal( left: f32, right: f32, bottom: f32, top: f32, near: f32, far: f32 ) -> Projection

Creates a new Projection that projects positions using an orthogonal projection with the given clipping planes.

Godot equivalent: Projection.create_orthogonal()

pub fn create_orthogonal_aspect( size: f32, aspect: f32, near: f32, far: f32, flip_fov: bool ) -> Projection

Creates a new Projection that projects positions using an orthogonal projection with the given size, X:Y aspect ratio, and clipping planes.

flip_fov determines whether the projection’s field of view is flipped over its diagonal.

Godot equivalent: Projection.create_orthogonal_aspect()

pub fn create_perspective( fov_y: f32, aspect: f32, near: f32, far: f32, flip_fov: bool ) -> Projection

Creates a new Projection that projects positions using a perspective projection with the given Y-axis field of view (in degrees), X:Y aspect ratio, and clipping planes

flip_fov determines whether the projection’s field of view is flipped over its diagonal.

Godot equivalent: Projection.create_perspective()

pub fn create_perspective_hmd( fov_y: f32, aspect: f32, near: f32, far: f32, flip_fov: bool, eye: ProjectionEye, intraocular_dist: f32, convergence_dist: f32 ) -> Projection

Creates a new Projection that projects positions using a perspective projection with the given Y-axis field of view (in degrees), X:Y aspect ratio, and clipping distances. The projection is adjusted for a head-mounted display with the given distance between eyes and distance to a point that can be focused on.

flip_fov determines whether the projection’s field of view is flipped over its diagonal.

Godot equivalent: Projection.create_perspective_hmd()

pub fn determinant(&self) -> f32

Return the determinant of the matrix.

Godot equivalent: Projection.determinant()

pub fn flipped_y(self) -> Projection

Returns a copy of this Projection with the signs of the values of the Y column flipped.

Godot equivalent: Projection.flipped_y()

pub fn aspect(&self) -> f32

Returns the X:Y aspect ratio of this Projection’s viewport.

Godot equivalent: Projection.get_aspect()

pub fn far_plane_half_extents(&self) -> Vector2

Returns the dimensions of the far clipping plane of the projection, divided by two.

Godot equivalent: Projection.get_far_plane_half_extents()

pub fn fov(&self) -> f32

Returns the horizontal field of view of the projection (in degrees).

Godot equivalent: Projection.get_fov()

pub fn fovy_of(fov_x: f32, aspect: f32) -> f32

Returns the vertical field of view of a projection (in degrees) which has the given horizontal field of view (in degrees) and aspect ratio.

Godot equivalent: Projection.get_fovy()

pub fn lod_multiplier(&self) -> f32

Returns the factor by which the visible level of detail is scaled by this Projection.

Godot equivalent: Projection.get_lod_multiplier()

pub fn pixels_per_meter(&self, pixel_width: i64) -> i64

Returns the number of pixels with the given pixel width displayed per meter, after this Projection is applied.

Godot equivalent: Projection.get_pixels_per_meter()

pub fn projection_plane(&self, plane: ProjectionPlane) -> Plane

Returns the clipping plane of this Projection whose index is given by plane.

Godot equivalent: Projection.get_projection_plane()

pub fn viewport_half_extents(&self) -> Vector2

Returns the dimensions of the viewport plane that this Projection projects positions onto, divided by two.

Godot equivalent: Projection.get_viewport_half_extents()

pub fn z_far(&self) -> f32

Returns the distance for this Projection beyond which positions are clipped.

Godot equivalent: Projection.get_z_far()

pub fn z_near(&self) -> f32

Returns the distance for this Projection before which positions are clipped.

Godot equivalent: Projection.get_z_near()

pub fn inverse(self) -> Projection

Returns a Projection that performs the inverse of this Projection’s projective transformation.

Godot equivalent: Projection.inverse()

pub fn is_orthogonal(&self) -> bool

Returns true if this Projection performs an orthogonal projection.

Godot equivalent: Projection.is_orthogonal()

pub fn jitter_offset(&self, offset: Vector2) -> Projection

Returns a Projection with the X and Y values from the given Vector2 added to the first and second values of the final column respectively.

Godot equivalent: Projection.jitter_offseted()

pub fn perspective_znear_adjusted(&self, new_znear: f32) -> Projection

Returns a Projection with the near clipping distance adjusted to be new_znear.

Note: The original Projection must be a perspective projection.

Godot equivalent: Projection.perspective_znear_adjusted()