Struct godot::prelude::RefCounted

pub struct RefCounted { /* private fields */ }

Godot class RefCounted.

Inherits Object.

Related symbols:

RefCountedVirtual: virtual methods

See also Godot docs for RefCounted.

Implementations

impl RefCounted

pub fn new() -> Gd<RefCounted>

pub fn notify(&mut self, what: ObjectNotification)

Sends a Godot notification to all classes inherited by the object.

Triggers calls to on_notification(), and depending on the notification, also to Godots lifecycle callbacks such as ready().

Starts from the highest ancestor (the Object class) and goes down the hierarchy. See also Godot docs for Object::notification().

Panics

If you call this method on a user-defined object while holding a GdRef or GdMut guard on the instance, you will encounter a panic. The reason is that the receiving virtual method on_notification() acquires a GdMut lock dynamically, which must be exclusive.

pub fn notify_reversed(&mut self, what: ObjectNotification)

Like Self::notify(), but starts at the most-derived class and goes up the hierarchy.

See docs of that method, including the panics.

pub fn get_reference_count(&self) -> i64