godot::prelude

Struct Resource

pub struct Resource { /* private fields */ }

Godot class Resource.

Inherits RefCounted.

Related symbols:

resource: sidecar module with related enum/flag types

IResource: virtual methods

SignalsOfResource: signal collection

See also Godot docs for Resource.

Construction

This class is reference-counted. You can create a new instance using Resource::new_gd().

Godot docs

Resource is the base class for all Godot-specific resource types, serving primarily as data containers. Since they inherit from RefCounted, resources are reference-counted and freed when no longer in use. They can also be nested within other resources, and saved on disk. PackedScene, one of the most common Objects in a Godot project, is also a resource, uniquely capable of storing and instantiating the Nodes it contains as many times as desired.

In GDScript, resources can loaded from disk by their [member resource_path] using load or [method @GDScript.preload].

The engine keeps a global cache of all loaded resources, referenced by paths (see has_cached). A resource will be cached when loaded for the first time and removed from cache once all references are released. When a resource is cached, subsequent loads using its path will return the cached reference.

Note: In C#, resources will not be freed instantly after they are no longer in use. Instead, garbage collection will run periodically and will free resources that are no longer in use. This means that unused resources will remain in memory for a while before being removed.

Implementations

impl Resource

pub fn set_path(&mut self, path: impl AsArg<GString>)

pub fn take_over_path(&mut self, path: impl AsArg<GString>)

Sets the [member resource_path] to path, potentially overriding an existing cache entry for this path. Further attempts to load an overridden resource by path will instead return this resource.

pub fn get_path(&self) -> GString

pub fn set_path_cache(&mut self, path: impl AsArg<GString>)

Sets the resource’s path to path without involving the resource cache. Useful for handling [enum ResourceFormatLoader.CacheMode] values when implementing a custom resource format by extending ResourceFormatLoader and ResourceFormatSaver.

pub fn set_name(&mut self, name: impl AsArg<GString>)

pub fn get_name(&self) -> GString

pub fn get_rid(&self) -> Rid

Returns the RID of this resource (or an empty RID). Many resources (such as Texture2D, Mesh, and so on) are high-level abstractions of resources stored in a specialized server (DisplayServer, RenderingServer, etc.), so this function will return the original RID.

pub fn set_local_to_scene(&mut self, enable: bool)

pub fn is_local_to_scene(&self) -> bool

pub fn get_local_scene(&self) -> Option<Gd<Node>>

If [member resource_local_to_scene] is set to true and the resource has been loaded from a PackedScene instantiation, returns the root Node of the scene where this resource is used. Otherwise, returns null.

pub fn setup_local_to_scene(&mut self)

Calls setup_local_to_scene. If [member resource_local_to_scene] is set to true, this method is automatically called from instantiate by the newly duplicated resource within the scene instance.

pub fn reset_state(&mut self)

Makes the resource clear its non-exported properties. See also reset_state. Useful when implementing a custom resource format by extending ResourceFormatLoader and ResourceFormatSaver.

pub fn set_id_for_path( &mut self, path: impl AsArg<GString>, id: impl AsArg<GString>, )

In the internal cache for scene-unique IDs, sets the ID of this resource to id for the scene at path. If id is empty, the cache entry for path is cleared. Useful to keep scene-unique IDs the same when implementing a VCS-friendly custom resource format by extending ResourceFormatLoader and ResourceFormatSaver.

Note: This method is only implemented when running in an editor context.

pub fn get_id_for_path(&self, path: impl AsArg<GString>) -> GString

From the internal cache for scene-unique IDs, returns the ID of this resource for the scene at path. If there is no entry, an empty string is returned. Useful to keep scene-unique IDs the same when implementing a VCS-friendly custom resource format by extending ResourceFormatLoader and ResourceFormatSaver.

Note: This method is only implemented when running in an editor context. At runtime, it returns an empty string.

pub fn is_built_in(&self) -> bool

Returns true if the resource is saved on disk as a part of another resource’s file.

pub fn generate_scene_unique_id() -> GString

Generates a unique identifier for a resource to be contained inside a PackedScene, based on the current date, time, and a random value. The returned string is only composed of letters (a to y) and numbers (0 to 8). See also [member resource_scene_unique_id].

pub fn set_scene_unique_id(&mut self, id: impl AsArg<GString>)

pub fn get_scene_unique_id(&self) -> GString

pub fn emit_changed(&mut self)

Emits the changed signal. This method is called automatically for some built-in resources.

Note: For custom resources, it’s recommended to call this method whenever a meaningful change occurs, such as a modified property. This ensures that custom Objects depending on the resource are properly updated.

var damage:
	set(new_value):
		if damage != new_value:
			damage = new_value
			emit_changed()

pub fn duplicate(&self) -> Option<Gd<Resource>>

👎Deprecated:
Use Gd::duplicate_resource() or Gd::duplicate_resource_ex().

To set the default parameters, use duplicate_ex and its builder methods. See the book for detailed usage instructions. Duplicates this resource, returning a new resource with its exported or PropertyUsageFlags::STORAGE properties copied from the original.

If deep is false, a shallow copy is returned: nested Array, Dictionary, and Resource properties are not duplicated and are shared with the original resource.

If deep is true, a deep copy is returned: all nested arrays, dictionaries, and packed arrays are also duplicated (recursively). Any Resource found inside will only be duplicated if it’s local, like DeepDuplicateMode::INTERNAL used with duplicate_deep.

The following exceptions apply:

Subresource properties with the PropertyUsageFlags::ALWAYS_DUPLICATE flag are always duplicated (recursively or not, depending on deep).

Subresource properties with the PropertyUsageFlags::NEVER_DUPLICATE flag are never duplicated.

Note: For custom resources, this method will fail if init has been defined with required parameters.

Note: When duplicating with deep set to true, each resource found, including the one on which this method is called, will be only duplicated once and referenced as many times as needed in the duplicate. For instance, if you are duplicating resource A that happens to have resource B referenced twice, you’ll get a new resource A’ referencing a new resource B’ twice.

pub fn duplicate_ex<'ex>(&'ex self) -> ExDuplicate<'ex>

👎Deprecated:
Use Gd::duplicate_resource() or Gd::duplicate_resource_ex().

Duplicates this resource, returning a new resource with its exported or PropertyUsageFlags::STORAGE properties copied from the original.

If deep is false, a shallow copy is returned: nested Array, Dictionary, and Resource properties are not duplicated and are shared with the original resource.

If deep is true, a deep copy is returned: all nested arrays, dictionaries, and packed arrays are also duplicated (recursively). Any Resource found inside will only be duplicated if it’s local, like DeepDuplicateMode::INTERNAL used with duplicate_deep.

The following exceptions apply:

Subresource properties with the PropertyUsageFlags::ALWAYS_DUPLICATE flag are always duplicated (recursively or not, depending on deep).

Subresource properties with the PropertyUsageFlags::NEVER_DUPLICATE flag are never duplicated.

Note: For custom resources, this method will fail if init has been defined with required parameters.

Note: When duplicating with deep set to true, each resource found, including the one on which this method is called, will be only duplicated once and referenced as many times as needed in the duplicate. For instance, if you are duplicating resource A that happens to have resource B referenced twice, you’ll get a new resource A’ referencing a new resource B’ twice.

pub fn duplicate_deep(&self) -> Option<Gd<Resource>>

👎Deprecated:
Use Gd::duplicate_resource() or Gd::duplicate_resource_ex().

To set the default parameters, use duplicate_deep_ex and its builder methods. See the book for detailed usage instructions. Duplicates this resource, deeply, like duplicate when passing true, with extra control over how subresources are handled.

pub fn duplicate_deep_ex<'ex>(&'ex self) -> ExDuplicateDeep<'ex>

👎Deprecated:
Use Gd::duplicate_resource() or Gd::duplicate_resource_ex().

Duplicates this resource, deeply, like duplicate when passing true, with extra control over how subresources are handled.