godot::prelude
Struct Vector2

#[repr(C)]
pub struct Vector2 {
    pub x: f32,
    pub y: f32,
}

Vector used for 2D math using floating point coordinates.

2-element structure that can be used to represent continuous positions or directions in 2D space, as well as any other pair of numeric values.

It uses floating-point coordinates of 32-bit precision, unlike the engine’s float type which is always 64-bit. The engine can be compiled with the option precision=double to use 64-bit vectors; use the gdext library with the double-precision feature in that case.

Conversions are provided via various from_* and to_* functions, not via the From trait. This encourages new() as the main way to construct vectors, is explicit about the conversion taking place, needs no type inference, and works in const contexts.

Navigation to impl blocks within this page

Constants

Constructors and general vector functions

Specialized Vector2 functions

Float-specific functions

2D functions

2D and 3D functions

Trait impls + operators

All vector types

Dimension	Floating-point	Integer

2D	Vector2	Vector2i

3D	Vector3	Vector3i

4D	Vector4	Vector4i

You can convert to Vector2i using cast_int().

Godot docs

Vector2 (stable)

Fields
x: f32

The vector’s X component.

y: f32

The vector’s Y component.

Implementations

impl Vector2

Constants

pub const ZERO: Vector2

Zero vector, a vector with all components set to 0.

pub const ONE: Vector2

One vector, a vector with all components set to 1.

pub const INF: Vector2

Infinity vector, a vector with all components set to real::INFINITY.

pub const LEFT: Vector2

Left unit vector. Represents the direction of left.

pub const RIGHT: Vector2

Right unit vector. Represents the direction of right.

pub const UP: Vector2

Up unit vector. Y is down in 2D, so this vector points -Y.

pub const DOWN: Vector2

Down unit vector. Y is down in 2D, so this vector points +Y.

impl Vector2

Constructors and general vector functions

The following associated functions and methods are available on all vectors (2D, 3D, 4D; float and int).

pub const fn new(x: f32, y: f32) -> Vector2

Creates a vector with the given components.

pub const fn splat(v: f32) -> Vector2

Creates a vector with all components set to v.

pub const fn from_tuple(tuple: (f32, f32)) -> Vector2

Creates a vector from the given tuple.

pub const fn from_array(array: [f32; 2]) -> Vector2

Creates a vector from the given array.

pub const fn to_tuple(&self) -> (f32, f32)

Returns a tuple with the components of the vector.

pub const fn to_array(&self) -> [f32; 2]

Returns an array with the components of the vector.

pub fn abs(self) -> Vector2

Returns a new vector with all components in absolute values (i.e. positive or zero).

pub fn clamp(self, min: Vector2, max: Vector2) -> Vector2

Returns a new vector with all components clamped between the components of min and max.

Panics

If min > max, min is NaN, or max is NaN.

pub fn length(self) -> f32

Returns the length (magnitude) of this vector.

pub fn length_squared(self) -> f32

Squared length (squared magnitude) of this vector.

Runs faster than length(), so prefer it if you need to compare vectors or need the squared distance for some formula.

pub fn coord_min(self, other: Vector2) -> Vector2

Returns a new vector containing the minimum of the two vectors, component-wise.

You may consider using the fully-qualified syntax Vector2::coord_min(a, b) for symmetry.

pub fn coord_max(self, other: Vector2) -> Vector2

Returns a new vector containing the maximum of the two vectors, component-wise.

You may consider using the fully-qualified syntax Vector2::coord_max(a, b) for symmetry.

pub fn sign(self) -> Vector2

Returns a new vector with each component set to 1 if the component is positive, -1 if negative, and 0 if zero.

impl Vector2

Specialized Vector2 functions

pub const fn from_vector2i(v: Vector2i) -> Vector2

👎Deprecated: Moved to Vector2i::cast_float()

pub fn from_angle(angle: f32) -> Vector2

Creates a unit Vector2 rotated to the given angle in radians. This is equivalent to doing Vector2::new(angle.cos(), angle.sin()) or Vector2::RIGHT.rotated(angle).

use godot::prelude::*;

let a = Vector2::from_angle(0.0);                       // (1.0, 0.0)
let b = Vector2::new(1.0, 0.0).angle();                 // 0.0
let c = Vector2::from_angle(real_consts::PI / 2.0);     // (0.0, 1.0)

pub fn angle(self) -> f32

Returns this vector’s angle with respect to the positive X axis, or (1.0, 0.0) vector, in radians.

For example, Vector2::RIGHT.angle() will return zero, Vector2::DOWN.angle() will return PI / 2 (a quarter turn, or 90 degrees), and Vector2::new(1.0, -1.0).angle() will return -PI / 4 (a negative eighth turn, or -45 degrees).

Illustration of the returned angle.

Equivalent to the result of y.atan2(x).

pub fn angle_to(self, to: Vector2) -> f32

Returns the signed angle between self and the given vector, as radians in [-π, +π].

Note that behavior is different from 3D [Vector3::angle_to()] which returns the unsigned angle.

pub fn angle_to_point(self, to: Vector2) -> f32

Returns the angle to the given vector, in radians.

Illustration of the returned angle.

pub fn cross(self, with: Vector2) -> f32

Returns the 2D analog of the cross product for this vector and with.

This is the signed area of the parallelogram formed by the two vectors. If the second vector is clockwise from the first vector, then the cross product is the positive area. If counter-clockwise, the cross product is the negative area. If the two vectors are parallel this returns zero, making it useful for testing if two vectors are parallel.

Note: Cross product is not defined in 2D mathematically. This method embeds the 2D vectors in the XY plane of 3D space and uses their cross product’s Z component as the analog.

pub fn orthogonal(self) -> Vector2

Returns a perpendicular vector rotated 90 degrees counter-clockwise compared to the original, with the same length.

pub fn rotated(self, angle: f32) -> Vector2

Returns the result of rotating this vector by angle (in radians).

pub fn slerp(self, to: Vector2, weight: f32) -> Vector2

Returns the result of spherical linear interpolation between this vector and to, by amount weight. weight is on the range of 0.0 to 1.0, representing the amount of interpolation.

This method also handles interpolating the lengths if the input vectors have different lengths. For the special case of one or both input vectors having zero length, this method behaves like Vector2::lerp.

impl Vector2

Float-specific functions

The following methods are only available on floating-point vectors.

pub const fn cast_int(self) -> Vector2i

Converts to a vector with integer components, using as casts.

pub fn floor(self) -> Vector2

Returns a new vector with all components rounded down (towards negative infinity).

pub fn ceil(self) -> Vector2

Returns a new vector with all components rounded up (towards positive infinity).

pub fn cubic_interpolate( self, b: Vector2, pre_a: Vector2, post_b: Vector2, weight: f32, ) -> Vector2

Cubic interpolation between self and b using pre_a and post_b as handles, and returns the result at position weight.

weight is on the range of 0.0 to 1.0, representing the amount of interpolation.

pub fn cubic_interpolate_in_time( self, b: Vector2, pre_a: Vector2, post_b: Vector2, weight: f32, b_t: f32, pre_a_t: f32, post_b_t: f32, ) -> Vector2

Cubic interpolation between self and b using pre_a and post_b as handles, and returns the result at position weight.

weight is on the range of 0.0 to 1.0, representing the amount of interpolation. It can perform smoother interpolation than cubic_interpolate() by the time values.

pub fn try_direction_to(self, to: Vector2) -> Option<Vector2>

Returns the normalized vector pointing from this vector to to or None, if self and to are equal.

This is equivalent to using (b - a).try_normalized(). See also direction_to().

pub fn direction_to(self, to: Vector2) -> Vector2

⚠️ Returns the normalized vector pointing from this vector to to.

This is equivalent to using (b - a).normalized(). See also try_direction_to().

Panics

If self and to are equal.

pub fn distance_squared_to(self, to: Vector2) -> f32

Returns the squared distance between this vector and to.

This method runs faster than Self::distance_to, so prefer it if you need to compare vectors or need the squared distance for some formula.

pub fn distance_to(self, to: Vector2) -> f32

Returns the distance between this vector and to.

pub fn dot(self, with: Vector2) -> f32

Returns the dot product of this vector and with.

pub fn is_finite(self) -> bool

Returns true if each component of this vector is finite.

pub fn is_normalized(self) -> bool

Returns true if the vector is normalized, i.e. its length is approximately equal to 1.

pub fn is_zero_approx(self) -> bool

Returns true if this vector’s values are approximately zero.

This method is faster than using approx_eq() with one value as a zero vector.

pub fn lerp(self, other: Vector2, weight: f32) -> Vector2

Returns the result of the linear interpolation between this vector and to by amount weight.

weight is on the range of 0.0 to 1.0, representing the amount of interpolation.

pub fn try_normalized(self) -> Option<Vector2>

Returns the vector scaled to unit length or None, if called on a zero vector.

Computes self / self.length(). See also normalized() and is_normalized().

pub fn normalized(self) -> Vector2

⚠️ Returns the vector scaled to unit length.

Computes self / self.length(). See also try_normalized() and is_normalized().

Panics

If called on a zero vector.

pub fn normalized_or_zero(self) -> Vector2

Returns the vector scaled to unit length or Self::ZERO, if called on a zero vector.

Computes self / self.length(). See also try_normalized() and is_normalized().

pub fn posmod(self, pmod: f32) -> Vector2

Returns a vector composed of the FloatExt::fposmod() of this vector’s components and pmod.

pub fn posmodv(self, modv: Vector2) -> Vector2

Returns a vector composed of the FloatExt::fposmod() of this vector’s components and modv’s components.

pub fn round(self) -> Vector2

Returns a new vector with all components rounded to the nearest integer, with halfway cases rounded away from zero.

pub fn snapped(self, step: Vector2) -> Vector2

A new vector with each component snapped to the closest multiple of the corresponding component in step.

impl Vector2

2D functions

The following methods are only available on 2D vectors (for both float and int).

pub fn aspect(self) -> f32

Returns the aspect ratio of this vector, the ratio of Self::x to Self::y.

pub fn max_axis(self) -> Option<Vector2Axis>

Returns the axis of the vector’s highest value. See Vector2Axis enum. If all components are equal, this method returns None.

To mimic Godot’s behavior, unwrap this function’s result with unwrap_or(Vector2Axis::X).

Godot equivalent: Vector2.max_axis_index

pub fn min_axis(self) -> Option<Vector2Axis>

Returns the axis of the vector’s lowest value. See Vector2Axis enum. If all components are equal, this method returns None.

To mimic Godot’s behavior, unwrap this function’s result with unwrap_or(Vector2Axis::Y).

Godot equivalent: Vector2.min_axis_index

impl Vector2

2D and 3D functions

The following methods are available on both 2D and 3D float vectors.

pub fn bezier_derivative( self, control_1: Vector2, control_2: Vector2, end: Vector2, t: f32, ) -> Vector2

Returns the derivative at the given t on the Bézier curve defined by this vector and the given control_1, control_2, and end points.

pub fn bezier_interpolate( self, control_1: Vector2, control_2: Vector2, end: Vector2, t: f32, ) -> Vector2

Returns the point at the given t on the Bézier curve defined by this vector and the given control_1, control_2, and end points.

pub fn bounce(self, n: Vector2) -> Vector2

Returns a new vector “bounced off” from a plane defined by the given normal.

Panics

If n is not normalized.

pub fn limit_length(self, length: Option<f32>) -> Vector2

Returns the vector with a maximum length by limiting its length to length.

pub fn move_toward(self, to: Vector2, delta: f32) -> Vector2

Returns a new vector moved toward to by the fixed delta amount. Will not go past the final value.

pub fn project(self, b: Vector2) -> Vector2

Returns the result of projecting the vector onto the given vector b.

pub fn reflect(self, n: Vector2) -> Vector2

Returns the result of reflecting the vector defined by the given direction vector n.

Panics

If n is not normalized.

pub fn slide(self, n: Vector2) -> Vector2

Returns a new vector slid along a plane defined by the given normal.

Panics

If n is not normalized.