Struct godot::prelude::Vector3i

#[repr(C)]
pub struct Vector3i {
    pub x: i32,
    pub y: i32,
    pub z: i32,
}

Vector used for 3D math using integer coordinates.

3-element structure that can be used to represent positions in 3D space or any other triple of numeric values.

It uses integer coordinates and is therefore preferable to Vector3 when exact precision is required. Note that the values are limited to 32 bits, and unlike Vector3 this cannot be configured with an engine build option. Use i64 or PackedInt64Array if 64-bit values are needed.

Fields
x: i32

The vector’s X component.

y: i32

The vector’s Y component.

z: i32

The vector’s Z component.

Implementations

impl Vector3i

pub const ZERO: Vector3i = _

Zero vector, a vector with all components set to 0.

pub const ONE: Vector3i = _

One vector, a vector with all components set to 1.

impl Vector3i

pub const MIN: Vector3i = _

Min vector, a vector with all components equal to i32::MIN. Can be used as a negative integer equivalent of real::INF.

pub const MAX: Vector3i = _

Max vector, a vector with all components equal to i32::MAX. Can be used as an integer equivalent of real::INF.

impl Vector3i

pub const LEFT: Vector3i = _

Unit vector in -X direction. Can be interpreted as left in an untransformed 3D world.

pub const RIGHT: Vector3i = _

Unit vector in +X direction. Can be interpreted as right in an untransformed 3D world.

pub const UP: Vector3i = _

Unit vector in +Y direction. Typically interpreted as up in a 3D world.

pub const DOWN: Vector3i = _

Unit vector in -Y direction. Typically interpreted as down in a 3D world.

pub const FORWARD: Vector3i = _

Unit vector in -Z direction. Can be interpreted as “into the screen” in an untransformed 3D world.

pub const BACK: Vector3i = _

Unit vector in +Z direction. Can be interpreted as “out of the screen” in an untransformed 3D world.

impl Vector3i

pub const fn new(x: i32, y: i32, z: i32) -> Vector3i

Returns a vector with the given components.

pub const fn splat(v: i32) -> Vector3i

Returns a new vector with all components set to v.

pub fn abs(self) -> Vector3i

Returns a new vector with all components in absolute values (i.e. positive or zero).

pub fn clamp(self, min: Vector3i, max: Vector3i) -> Vector3i

Returns a new vector with all components clamped between the components of min and max.

Panics

If min > max, min is NaN, or max is NaN.

pub fn length(self) -> f32

Returns the length (magnitude) of this vector.

pub fn length_squared(self) -> i32

Squared length (squared magnitude) of this vector.

Runs faster than Self::length, so prefer it if you need to compare vectors or need the squared distance for some formula.

pub fn coord_min(self, other: Vector3i) -> Vector3i

Returns a new vector containing the minimum of the two vectors, component-wise.

pub fn coord_max(self, other: Vector3i) -> Vector3i

Returns a new vector containing the maximum of the two vectors, component-wise.

pub fn sign(self) -> Vector3i

Returns a new vector with each component set to 1 if it’s positive, -1 if it’s negative, and 0 if it’s zero.

impl Vector3i

pub fn max_axis(self) -> Option<Vector3Axis>

Returns the axis of the vector’s highest value. See Vector3Axis enum. If all components are equal, this method returns None.

To mimic Godot’s behavior, unwrap this function’s result with unwrap_or(Vector3Axis::X).

pub fn min_axis(self) -> Option<Vector3Axis>

Returns the axis of the vector’s lowest value. See Vector3Axis enum. If all components are equal, this method returns None.

To mimic Godot’s behavior, unwrap this function’s result with unwrap_or(Vector3Axis::Z).

impl Vector3i

pub fn snapped(self, step: Vector3i) -> Vector3i

A new vector with each component snapped to the closest multiple of the corresponding component in step.

Panics

On under- or overflow:

If any component of self is i32::MIN while the same component on step is -1.

If any component of self plus half of the same component of step is not in range on i32.

pub const fn from_vector3(v: Vector3) -> Vector3i

Constructs a new Vector3i from a Vector3. The floating point coordinates will be truncated.