Trait godot::prelude::Export

pub trait Export {
    // Required method
    fn export(&self) -> Self;
}

Trait implemented for types that can be used as #[export] fields. This creates a copy of the value, for some type-specific definition of “copy”. For example, Array and Gd are returned via Share::share() instead of copying the actual data.

Required Methods

fn export(&self) -> Self

Creates a copy to be returned from a getter.