Trait godot::prelude::FromGodot

pub trait FromGodot: Sized + GodotConvert {
    // Required method
    fn try_from_godot(via: Self::Via) -> Result<Self, ConvertError>;

    // Provided methods
    fn from_godot(via: Self::Via) -> Self { ... }
    fn try_from_variant(variant: &Variant) -> Result<Self, ConvertError> { ... }
    fn from_variant(variant: &Variant) -> Self { ... }
}

Defines the canonical conversion from Godot for a type.

It is assumed that all the methods return equal values given equal inputs. Additionally it is assumed that if ToGodot is implemented, converting to Godot and back again will return a value equal to the starting value.

Violating these assumptions is safe but will give unexpected results.

Required Methods

fn try_from_godot(via: Self::Via) -> Result<Self, ConvertError>

Performs the conversion.

Provided Methods

fn from_godot(via: Self::Via) -> Self

⚠️ Performs the conversion.

Panics

If the conversion fails.

fn try_from_variant(variant: &Variant) -> Result<Self, ConvertError>

Performs the conversion from a Variant.

fn from_variant(variant: &Variant) -> Self

⚠️ Performs the conversion from a Variant.

Panics

If the conversion fails.

Object Safety

This trait is not object safe.