Trait godot::prelude::ToVariant

pub trait ToVariant {
    // Required method
    fn to_variant(&self) -> Variant;
}

Trait to enable conversions of types to the Variant type.

Required Methods

fn to_variant(&self) -> Variant

Infallible conversion from Self type to Variant.

This method must not panic. If your conversion is fallible, this trait should not be used.