Derive Macro godot::register::Var

#[derive(Var)]

Derive macro for Var on enums.

Currently has some tight requirements which are expected to be softened as implementation expands:

Only works for enums, structs aren’t supported by this derive macro at the moment.

The enum must have an explicit #[repr(u*/i*)] type.

This will likely stay this way, since isize, the default repr type, is not a concept in Godot.

The enum variants must not have any fields - currently only unit variants are supported.

The enum variants must have explicit discriminants, that is, e.g. A = 2, not just A

Example

#[derive(Var)]
#[repr(i32)]
enum MyEnum {
    A = 0,
    B = 1,
}

#[derive(GodotClass)]
struct MyClass {
    #[var]
    foo: MyEnum,
}

fn main() {
    let mut class = MyClass { foo: MyEnum::B };
    assert_eq!(class.get_foo(), MyEnum::B as i32);

    class.set_foo(MyEnum::A as i32);
    assert_eq!(class.foo, MyEnum::A);
}