godot::register::property

Enum GodotElementShape

#[non_exhaustive]
pub enum GodotElementShape {
    Variant,
    Builtin {
        variant_type: VariantType,
    },
    Class {
        class_id: ClassId,
        heritage: ClassHeritage,
    },
    Enum {
        variant_type: VariantType,
        enumerators: Cow<'static, [Enumerator]>,
        godot_name: Option<Cow<'static, str>>,
        is_bitfield: bool,
    },
    Custom {
        variant_type: VariantType,
        var_hint: PropertyHintInfo,
        export_hint: PropertyHintInfo,
        class_name: Option<Cow<'static, str>>,
        usage_flags: PropertyUsageFlags,
    },
}

Same as GodotShape, but for element types nested in typed arrays/dictionaries.

Matches the same layout as GodotShape, exists to avoid recursive definition (and also Box allocations). Also constrains the possible shapes (elements cannot be typed arrays/dictionaries themselves).

In contrast to ElementType, this is a static type description for Godot registration purposes.

Use into_outer() to convert into a full GodotShape.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Variant

Builtin

Fields

variant_type: VariantType

Class

Fields

class_id: ClassId

heritage: ClassHeritage

Enum

Fields

variant_type: VariantType

enumerators: Cow<'static, [Enumerator]>

godot_name: Option<Cow<'static, str>>

is_bitfield: bool

Custom

Fields

variant_type: VariantType

var_hint: PropertyHintInfo

export_hint: PropertyHintInfo

class_name: Option<Cow<'static, str>>

usage_flags: PropertyUsageFlags

Implementations

impl GodotElementShape

pub fn into_outer(self) -> GodotShape

Converts this nested shape into a full GodotShape. Infallible.