godot::register::property

Trait BuiltinExport

pub trait BuiltinExport { }

Marker trait to identify GodotTypes that can be directly used with an #[export].

Implemented pretty much for all the GodotTypes that are not GodotClass. Provides a few blanket implementations and, by itself, has no implications for the Var or Export traits.

Types which don’t implement the BuiltinExport trait can’t be used directly as an #[export] and must be handled using associated algebraic types, such as:

Option<T>, which represents optional value that can be null when used.

OnEditor<T>, which represents value that must not be null when used.