godot::register::property

Trait Export

pub trait Export: Var {
    // Provided method
    fn export_hint() -> PropertyHintInfo { ... }
}

Trait implemented for types that can be used as #[export] fields.

To export objects, see the Exporting section of Gd<T>.

For enums, this trait can be derived using the #[derive(Export)] macro.

Provided Methods

fn export_hint() -> PropertyHintInfo

The export info to use for an exported field of this type, if no other export info is specified.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.