godot::register::property

Trait VarPub

pub trait VarPub: Var {
    type Type;

    // Required methods
    fn var_pub_get(property: &Self) -> Self::Type;
    fn var_pub_set(property: &mut Self, value: Self::Type);
}

Customizes how #[var(pub)] generates public getter/setter accessors for a property.

#[var] properties only need Var. The pub modifier additionally requires a way to expose the field through generated methods without giving direct field access.

This trait has a blanket implementation for all T: Clone, using clone() in the getter, so in most cases you don’t need to do anything. The generated getters are not designed to provide customization beyond that; Clone will always be the behavior if available. If more special types are required, define getters/setters manually.

Associated type

Type is the type used by the generated public accessors. It may differ from the stored type (Self) and from the property conversion type (<Self as GodotConvert>::Via). As an example, OnEditor<T> uses Option<T> for Type.

Required Associated Types

type Type

Type of the value returned by the getter, and accepted by the setter.

Required Methods

fn var_pub_get(property: &Self) -> Self::Type

Implementation of the getter generated by #[var(pub)].

fn var_pub_set(property: &mut Self, value: Self::Type)

Implementation of the setter generated by #[var(pub)].

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.