godot::register

Struct IndirectSignalReceiver

pub struct IndirectSignalReceiver<'view, I, Ps, F>
where
    Ps: InParamTuple,
    F: SignalReceiver<I, Ps> + 'static,
{ /* private fields */ }

A special “identity struct” which allows to use type inference while specifying various closures for connections.

rustc can’t infer types in closures when dealing with Fn/FnMut/FnOnce traits abstracted behind another trait (in our case SignalReceiver). To make type inference work in such cases, we need to specify concrete type – such as FnMut(...) -> R or <&mut F as Into<IndirectSignalReceiver<'_, Instance, Params, Func>>>.

In other words, IndirectSignalReceiver allows us to “smuggle” in a Fn* trait, forcing rustc to deal with type inference while resolving its inner type.

Example usage

IndirectSignalReceiver::from(&mut function)
    .function()
    .call((), args);

Further reading

rustc issue #63702

identity function trick

rustc comments around closure type-check

Implementations

impl<'view, I, Ps, F> IndirectSignalReceiver<'view, I, Ps, F>
where Ps: InParamTuple, F: SignalReceiver<I, Ps> + 'static,

pub fn function(&'view mut self) -> &'view mut F

Retrieves inner &mut F function ready to be used as SignalReceiver.