Crate godot::sys

Low level bindings to the provided C core API

Modules

paste

github crates-io docs-rs

types

Macros

ffi_methods

Provides “sys” style methods for FFI and ptrcall integration with Godot. The generated implementations follow one of three patterns:

static_assert

Verifies a condition at compile time.

static_assert_eq_size

Verifies at compile time that two types T and U have the same size.

Structs

GDExtensionCallError

GDExtensionClassCreationInfo

GDExtensionClassMethodInfo

GDExtensionInitialization

GDExtensionInstanceBindingCallbacks

GDExtensionInterface

GDExtensionMethodInfo

GDExtensionPropertyInfo

GDExtensionScriptInstanceInfo

GlobalMethodTable

__GdextClassInstance

__GdextClassLibrary

__GdextMethodBind

__GdextObject

__GdextRef

__GdextScriptInstance

__GdextScriptInstanceData

__GdextScriptLanguage

__GdextString

__GdextStringName

__GdextType

__GdextVariant

__fsid_t

max_align_t

Enums

PtrcallType

An indication of what type of pointer call is being made.

VariantOperator

VariantType

Constants

GDEXTENSION_CALL_ERROR_INSTANCE_IS_NULL

GDEXTENSION_CALL_ERROR_INVALID_ARGUMENT

GDEXTENSION_CALL_ERROR_INVALID_METHOD

GDEXTENSION_CALL_ERROR_METHOD_NOT_CONST

GDEXTENSION_CALL_ERROR_TOO_FEW_ARGUMENTS

GDEXTENSION_CALL_ERROR_TOO_MANY_ARGUMENTS

GDEXTENSION_CALL_OK

GDEXTENSION_INITIALIZATION_CORE

GDEXTENSION_INITIALIZATION_EDITOR

GDEXTENSION_INITIALIZATION_SCENE

GDEXTENSION_INITIALIZATION_SERVERS

GDEXTENSION_MAX_INITIALIZATION_LEVEL

GDEXTENSION_METHOD_ARGUMENT_METADATA_INT_IS_INT8

GDEXTENSION_METHOD_ARGUMENT_METADATA_INT_IS_INT16

GDEXTENSION_METHOD_ARGUMENT_METADATA_INT_IS_INT32

GDEXTENSION_METHOD_ARGUMENT_METADATA_INT_IS_INT64

GDEXTENSION_METHOD_ARGUMENT_METADATA_INT_IS_UINT8

GDEXTENSION_METHOD_ARGUMENT_METADATA_INT_IS_UINT16

GDEXTENSION_METHOD_ARGUMENT_METADATA_INT_IS_UINT32

GDEXTENSION_METHOD_ARGUMENT_METADATA_INT_IS_UINT64

GDEXTENSION_METHOD_ARGUMENT_METADATA_NONE

GDEXTENSION_METHOD_ARGUMENT_METADATA_REAL_IS_DOUBLE

GDEXTENSION_METHOD_ARGUMENT_METADATA_REAL_IS_FLOAT

GDEXTENSION_METHOD_FLAGS_DEFAULT

GDEXTENSION_METHOD_FLAG_CONST

GDEXTENSION_METHOD_FLAG_EDITOR

GDEXTENSION_METHOD_FLAG_NORMAL

GDEXTENSION_METHOD_FLAG_STATIC

GDEXTENSION_METHOD_FLAG_VARARG

GDEXTENSION_METHOD_FLAG_VIRTUAL

GDEXTENSION_VARIANT_OP_ADD

GDEXTENSION_VARIANT_OP_AND

GDEXTENSION_VARIANT_OP_BIT_AND

GDEXTENSION_VARIANT_OP_BIT_NEGATE

GDEXTENSION_VARIANT_OP_BIT_OR

GDEXTENSION_VARIANT_OP_BIT_XOR

GDEXTENSION_VARIANT_OP_DIVIDE

GDEXTENSION_VARIANT_OP_EQUAL

GDEXTENSION_VARIANT_OP_GREATER

GDEXTENSION_VARIANT_OP_GREATER_EQUAL

GDEXTENSION_VARIANT_OP_IN

GDEXTENSION_VARIANT_OP_LESS

GDEXTENSION_VARIANT_OP_LESS_EQUAL

GDEXTENSION_VARIANT_OP_MAX

GDEXTENSION_VARIANT_OP_MODULE

GDEXTENSION_VARIANT_OP_MULTIPLY

GDEXTENSION_VARIANT_OP_NEGATE

GDEXTENSION_VARIANT_OP_NOT

GDEXTENSION_VARIANT_OP_NOT_EQUAL

GDEXTENSION_VARIANT_OP_OR

GDEXTENSION_VARIANT_OP_POSITIVE

GDEXTENSION_VARIANT_OP_POWER

GDEXTENSION_VARIANT_OP_SHIFT_LEFT

GDEXTENSION_VARIANT_OP_SHIFT_RIGHT

GDEXTENSION_VARIANT_OP_SUBTRACT

GDEXTENSION_VARIANT_OP_XOR

GDEXTENSION_VARIANT_TYPE_AABB

GDEXTENSION_VARIANT_TYPE_ARRAY

GDEXTENSION_VARIANT_TYPE_BASIS

GDEXTENSION_VARIANT_TYPE_BOOL

GDEXTENSION_VARIANT_TYPE_CALLABLE

GDEXTENSION_VARIANT_TYPE_COLOR

GDEXTENSION_VARIANT_TYPE_DICTIONARY

GDEXTENSION_VARIANT_TYPE_FLOAT

GDEXTENSION_VARIANT_TYPE_INT

GDEXTENSION_VARIANT_TYPE_NIL

GDEXTENSION_VARIANT_TYPE_NODE_PATH

GDEXTENSION_VARIANT_TYPE_OBJECT

GDEXTENSION_VARIANT_TYPE_PACKED_BYTE_ARRAY

GDEXTENSION_VARIANT_TYPE_PACKED_COLOR_ARRAY

GDEXTENSION_VARIANT_TYPE_PACKED_FLOAT32_ARRAY

GDEXTENSION_VARIANT_TYPE_PACKED_FLOAT64_ARRAY

GDEXTENSION_VARIANT_TYPE_PACKED_INT32_ARRAY

GDEXTENSION_VARIANT_TYPE_PACKED_INT64_ARRAY

GDEXTENSION_VARIANT_TYPE_PACKED_STRING_ARRAY

GDEXTENSION_VARIANT_TYPE_PACKED_VECTOR2_ARRAY

GDEXTENSION_VARIANT_TYPE_PACKED_VECTOR3_ARRAY

GDEXTENSION_VARIANT_TYPE_PLANE

GDEXTENSION_VARIANT_TYPE_PROJECTION

GDEXTENSION_VARIANT_TYPE_QUATERNION

GDEXTENSION_VARIANT_TYPE_RECT2

GDEXTENSION_VARIANT_TYPE_RECT2I

GDEXTENSION_VARIANT_TYPE_RID

GDEXTENSION_VARIANT_TYPE_SIGNAL

GDEXTENSION_VARIANT_TYPE_STRING

GDEXTENSION_VARIANT_TYPE_STRING_NAME

GDEXTENSION_VARIANT_TYPE_TRANSFORM2D

GDEXTENSION_VARIANT_TYPE_TRANSFORM3D

GDEXTENSION_VARIANT_TYPE_VARIANT_MAX

GDEXTENSION_VARIANT_TYPE_VECTOR2

GDEXTENSION_VARIANT_TYPE_VECTOR2I

GDEXTENSION_VARIANT_TYPE_VECTOR3

GDEXTENSION_VARIANT_TYPE_VECTOR3I

GDEXTENSION_VARIANT_TYPE_VECTOR4

GDEXTENSION_VARIANT_TYPE_VECTOR4I

INT8_MAX

INT8_MIN

INT16_MAX

INT16_MIN

INT32_MAX

INT32_MIN

INTPTR_MAX

INTPTR_MIN

INT_FAST8_MAX

INT_FAST8_MIN

INT_FAST16_MAX

INT_FAST16_MIN

INT_FAST32_MAX

INT_FAST32_MIN

INT_LEAST8_MAX

INT_LEAST8_MIN

INT_LEAST16_MAX

INT_LEAST16_MIN

INT_LEAST32_MAX

INT_LEAST32_MIN

PTRDIFF_MAX

PTRDIFF_MIN

SIG_ATOMIC_MAX

SIG_ATOMIC_MIN

SIZE_MAX

UINT8_MAX

UINT16_MAX

UINT32_MAX

UINTPTR_MAX

UINT_FAST8_MAX

UINT_FAST16_MAX

UINT_FAST32_MAX

UINT_LEAST8_MAX

UINT_LEAST16_MAX

UINT_LEAST32_MAX

WINT_MAX

WINT_MIN

_ATFILE_SOURCE

_BITS_STDINT_INTN_H

_BITS_STDINT_UINTN_H

_BITS_TIME64_H

_BITS_TYPESIZES_H

_BITS_TYPES_H

_BITS_WCHAR_H

_DEFAULT_SOURCE

_FEATURES_H

_POSIX_C_SOURCE

_POSIX_SOURCE

_STDC_PREDEF_H

_STDINT_H

_SYS_CDEFS_H

__FD_SETSIZE

__GLIBC_MINOR__

__GLIBC_USE_DEPRECATED_GETS

__GLIBC_USE_DEPRECATED_SCANF

__GLIBC_USE_IEC_60559_BFP_EXT

__GLIBC_USE_IEC_60559_BFP_EXT_C2X

__GLIBC_USE_IEC_60559_FUNCS_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT_C2X

__GLIBC_USE_IEC_60559_TYPES_EXT

__GLIBC_USE_ISOC2X

__GLIBC_USE_LIB_EXT2

__GLIBC__

__GNU_LIBRARY__

__HAVE_GENERIC_SELECTION

__INO_T_MATCHES_INO64_T

__LONG_DOUBLE_USES_FLOAT128

__OFF_T_MATCHES_OFF64_T

__RLIM_T_MATCHES_RLIM64_T

__STATFS_MATCHES_STATFS64

__STDC_IEC_559_COMPLEX__

__STDC_IEC_559__

__STDC_ISO_10646__

__SYSCALL_WORDSIZE

__TIMESIZE

__USE_ATFILE

__USE_FORTIFY_LEVEL

__USE_ISOC11

__USE_ISOC95

__USE_ISOC99

__USE_MISC

__USE_POSIX

__USE_POSIX2

__USE_POSIX199309

__USE_POSIX199506

__USE_POSIX_IMPLICITLY

__USE_XOPEN2K

__USE_XOPEN2K8

__WORDSIZE

__WORDSIZE_TIME64_COMPAT32

__glibc_c99_flexarr_available

Traits

GodotFfi

Adds methods to convert from and to Godot FFI pointers. See crate::ffi_methods for ergonomic implementation.

GodotFuncMarshal

Trait implemented for all types that can be passed to and from user-defined #[func] methods through Godot’s ptrcall calling convention.

Functions

force_mut_ptr

Explicitly cast away const from a pointer, similar to C++ const_cast.

get_interface⚠

Safety

get_library⚠

Safety

initialize⚠

Safety

method_table⚠

Safety

ptr_then

If ptr is not null, returns Some(mapper(ptr)); otherwise None.

to_const_ptr

Add const to a mut ptr.

unbox

Extract value from box before into_inner() is stable

Type Definitions

GDExtensionBool

GDExtensionCallErrorType

GDExtensionClassCallVirtual

GDExtensionClassConstructor

GDExtensionClassCreateInstance

GDExtensionClassFreeInstance

GDExtensionClassFreePropertyList

GDExtensionClassGet

GDExtensionClassGetPropertyList

GDExtensionClassGetRID

GDExtensionClassGetVirtual

GDExtensionClassInstancePtr

GDExtensionClassLibraryPtr

GDExtensionClassMethodArgumentMetadata

GDExtensionClassMethodCall

GDExtensionClassMethodFlags

GDExtensionClassMethodPtrCall

GDExtensionClassNotification

GDExtensionClassPropertyCanRevert

GDExtensionClassPropertyGetRevert

GDExtensionClassReference

GDExtensionClassSet

GDExtensionClassToString

GDExtensionClassUnreference

GDExtensionConstObjectPtr

GDExtensionConstRefPtr

GDExtensionConstStringNamePtr

GDExtensionConstStringPtr

GDExtensionConstTypePtr

GDExtensionConstVariantPtr

GDExtensionInitializationFunction

GDExtensionInitializationLevel

GDExtensionInstanceBindingCreateCallback

GDExtensionInstanceBindingFreeCallback

GDExtensionInstanceBindingReferenceCallback

GDExtensionInt

GDExtensionMethodBindPtr

GDExtensionObjectPtr

GDExtensionPtrBuiltInMethod

GDExtensionPtrConstructor

GDExtensionPtrDestructor

GDExtensionPtrGetter

GDExtensionPtrIndexedGetter

GDExtensionPtrIndexedSetter

GDExtensionPtrKeyedChecker

GDExtensionPtrKeyedGetter

GDExtensionPtrKeyedSetter

GDExtensionPtrOperatorEvaluator

GDExtensionPtrSetter

GDExtensionPtrUtilityFunction

GDExtensionRefPtr

GDExtensionScriptInstanceCall

GDExtensionScriptInstanceDataPtr

GDExtensionScriptInstanceFree

GDExtensionScriptInstanceFreeMethodList

GDExtensionScriptInstanceFreePropertyList

GDExtensionScriptInstanceGet

GDExtensionScriptInstanceGetLanguage

GDExtensionScriptInstanceGetMethodList

GDExtensionScriptInstanceGetOwner

GDExtensionScriptInstanceGetPropertyList

GDExtensionScriptInstanceGetPropertyState

GDExtensionScriptInstanceGetPropertyType

GDExtensionScriptInstanceGetScript

GDExtensionScriptInstanceHasMethod

GDExtensionScriptInstanceIsPlaceholder

GDExtensionScriptInstanceNotification

GDExtensionScriptInstancePropertyCanRevert

GDExtensionScriptInstancePropertyGetRevert

GDExtensionScriptInstancePropertyStateAdd

GDExtensionScriptInstancePtr

GDExtensionScriptInstanceRefCountDecremented

GDExtensionScriptInstanceRefCountIncremented

GDExtensionScriptInstanceSet

GDExtensionScriptInstanceToString

GDExtensionScriptLanguagePtr

GDExtensionStringNamePtr

GDExtensionStringPtr

GDExtensionTypeFromVariantConstructorFunc

GDExtensionTypePtr

GDExtensionVariantFromTypeConstructorFunc

GDExtensionVariantOperator

GDExtensionVariantPtr

GDExtensionVariantType

GDObjectInstanceID

__blkcnt64_t

__blkcnt_t

__blksize_t

__caddr_t

__clock_t

__clockid_t

__daddr_t

__dev_t

__fsblkcnt64_t

__fsblkcnt_t

__fsfilcnt64_t

__fsfilcnt_t

__fsword_t

__gid_t

__id_t

__ino64_t

__ino_t

__int8_t

__int16_t

__int32_t

__int64_t

__int_least8_t

__int_least16_t

__int_least32_t

__int_least64_t

__intmax_t

__intptr_t

__key_t

__loff_t

__mode_t

__nlink_t

__off64_t

__off_t

__pid_t

__quad_t

__rlim64_t

__rlim_t

__sig_atomic_t

__socklen_t

__ssize_t

__suseconds_t

__syscall_slong_t

__syscall_ulong_t

__time_t

__timer_t

__u_char

__u_int

__u_long

__u_quad_t

__u_short

__uid_t

__uint8_t

__uint16_t

__uint32_t

__uint64_t

__uint_least8_t

__uint_least16_t

__uint_least32_t

__uint_least64_t

__uintmax_t

__useconds_t

char16_t

char32_t

int_fast8_t

int_fast16_t

int_fast32_t

int_fast64_t

int_least8_t

int_least16_t

int_least32_t

int_least64_t

intmax_t

uint_fast8_t

uint_fast16_t

uint_fast32_t

uint_fast64_t

uint_least8_t

uint_least16_t

uint_least32_t

uint_least64_t

uintmax_t

wchar_t