Macro godot::sys::ffi_methods

macro_rules! ffi_methods {
    ( // Sys pointer = address of opaque
        type $Ptr:ty = *mut Opaque;
        $( $rest:tt )*
    ) => { ... };
    ( // Sys pointer = value of opaque
        type $Ptr:ty = Opaque;
        $( $rest:tt )*
    ) => { ... };
    ( // Sys pointer = address of self
        type $Ptr:ty = *mut Self;
        $( $rest:tt )*
    ) => { ... };
}

Provides “sys” style methods for FFI and ptrcall integration with Godot. The generated implementations follow one of three patterns:

*mut Opaque
Implements FFI methods for a type with Opaque data that stores a value type (e.g. Vector2). The address of the Opaque field is used as the sys pointer. Expects a from_opaque() constructor and a opaque field.

Opaque
Implements FFI methods for a type with Opaque data. The sys pointer is directly reinterpreted from/to the Opaque and not its address. Expects a from_opaque() constructor and a opaque field.

*mut Self
Implements FFI methods for a type implemented with standard Rust fields (not opaque). The address of Self is directly reinterpreted as the sys pointer. The size of the corresponding sys type (the N in Opaque*<N>) must not be bigger than size_of::<Self>(). This cannot be checked easily, because Self cannot be used in size_of(). There would of course be workarounds.

Using this macro as a complete implementation for GodotFfi is sound only when:

Using *mut Opaque

Turning pointer call arguments into a value is simply calling from_opaque on the dereferenced argument pointer. Returning a value from a pointer call is simply calling std::ptr::swap on the return pointer and the address to the opaque field.

Using Opaque

Turning pointer call arguments into a value is simply calling from_opaque on the argument pointer. Returning a value from a pointer call is simply calling std::ptr::swap on the return pointer and the opaque field transmuted into a pointer.

Using *mut Self

Turning pointer call arguments into a value is a dereference. Returning a value from a pointer call is *ret_ptr = value.