Trait godot::sys::GodotFuncMarshal

pub trait GodotFuncMarshal: Sized {
    type Via: GodotFfi;
    type FromViaError: Error;
    type IntoViaError: Error;

    // Required methods
    fn try_from_via(via: Self::Via) -> Result<Self, Self::FromViaError>;
    fn try_into_via(self) -> Result<Self::Via, Self::IntoViaError>;

    // Provided methods
    unsafe fn try_from_arg(
        ptr: *mut __GdextType,
        call_type: PtrcallType
    ) -> Result<Self, Self::FromViaError> { ... }
    unsafe fn try_return(
        self,
        dst: *mut __GdextType,
        call_type: PtrcallType
    ) -> Result<(), Self::IntoViaError> { ... }
}

Trait implemented for all types that can be passed to and from Godot via function calls.

Required Associated Types

type Via: GodotFfi

The type used when passing a value to/from Godot.

type FromViaError: Error

Error type returned when a value of type Via fails to be converted into Self.

type IntoViaError: Error

Error type returned when a value of type Self fails to be converted into Via.

Required Methods

fn try_from_via(via: Self::Via) -> Result<Self, Self::FromViaError>

Try to convert the value from Self::Via to Self.

fn try_into_via(self) -> Result<Self::Via, Self::IntoViaError>

Try to convert the value from Self to Self::Via.

Provided Methods

unsafe fn try_from_arg( ptr: *mut __GdextType, call_type: PtrcallType ) -> Result<Self, Self::FromViaError>

Used for function arguments. On failure, the argument which can’t be converted to Self is returned.

Safety

The value behind ptr must be the C FFI type that corresponds to Self::Via. See also GodotFfi::from_arg_ptr.

unsafe fn try_return( self, dst: *mut __GdextType, call_type: PtrcallType ) -> Result<(), Self::IntoViaError>

Used for function return values. On failure, self which can’t be converted to Via is returned.

Safety

dst must point to an initialized value of type Self::Via. See also GodotFfi::move_return_ptr.