godot

Module task

Integrates async rust code with the engine.

This module contains:

Implementations of Future for Signal and TypedSignal.

A way to spawn new async tasks by using the engine as the async runtime.

Structs

FallibleSignalFuture

A future that tries to resolve as soon as the provided Godot signal was emitted.

FallibleSignalFutureError

Error that might be returned by the FallibleSignalFuture.

SignalFuture

The panicking counter part to the FallibleSignalFuture.

TaskHandle

Handle for an active background task.

Traits

DynamicSend

Runtime-checked Send capability.

IntoDynamicSend

Convert a value into a type that is Send at compile-time while the value might not be.

Functions

spawn

Create a new async background task.