godot::task

Trait IntoDynamicSend

pub trait IntoDynamicSend: Sealed + 'static {
    type Target: DynamicSend<Inner = Self>;

    // Required method
    fn into_dynamic_send(self) -> Self::Target;
}

Convert a value into a type that is Send at compile-time while the value might not be.

This allows to turn any implementor into a type that is Send, but requires to also implement DynamicSend as well. The later trait will verify if a value can actually be sent between threads at runtime.

Required Associated Types

type Target: DynamicSend<Inner = Self>

Required Methods

fn into_dynamic_send(self) -> Self::Target