Module godot::builtin

Built-in types like Vector2, GodotString and Variant.

Background on the design of vector algebra types

The basic vector algebra types like Vector2, Matrix4 and Quaternion are re-implemented here, with an API similar to that in the Godot engine itself. There are other approaches, but they all have their disadvantages:

We could invoke API methods from the engine. The implementations could be generated, but it is slower and prevents inlining.

We could re-export types from an existing vector algebra crate, like glam. This removes the duplication, but it would create a strong dependency on a volatile API outside our control. The gdnative crate started out this way, using types from euclid, but found it impractical. Moreover, the API would not match Godot’s own, which would make porting from GDScript (slightly) harder.

We could opaquely wrap types from an existing vector algebra crate. This protects users of gdextension from changes in the wrapped crate. However, direct field access using .x, .y, .z is no longer possible. Instead of v.y += a; you would have to write v.set_y(v.get_y() + a);. (A union could be used to add these fields in the public API, but would make every field access unsafe, which is also not great.)

We could re-export types from the mint crate, which was explicitly designed to solve this problem. However, it falls short because operator overloading would become impossible.

Modules

array

Specialized types related to arrays.

dictionary

Specialized types related to dictionaries.

meta

Meta-information about variant types, properties and class names.

real_consts

Basic mathematical constants.

Macros

array

Constructs Array literals, similar to Rust’s standard vec! macro.

dict

Constructs Dictionary literals, close to Godot’s own syntax.

real

A macro to coerce float-literals into the real type. Mainly used where you’d normally use a suffix to specity the type, such as 115.0f32.

varray

Constructs VariantArray literals, similar to Rust’s standard vec! macro.

Structs

Aabb

Axis-aligned bounding box in 3D space.

Array

Godot’s Array type.

Basis

A 3x3 matrix, typically used as an orthogonal basis for Transform3D.

Callable

A Callable represents a function in Godot.

Color

Color built-in type, in floating-point RGBA format.

Dictionary

Godot’s Dictionary type.

GodotString

Godot’s reference counted string type.

NodePath

A pre-parsed scene tree path.

PackedByteArray

Implements Godot’s $PackedArray type, which is an efficient array of $Elements.

PackedColorArray

Implements Godot’s $PackedArray type, which is an efficient array of $Elements.

PackedFloat32Array

Implements Godot’s $PackedArray type, which is an efficient array of $Elements.

PackedFloat64Array

Implements Godot’s $PackedArray type, which is an efficient array of $Elements.

PackedInt32Array

Implements Godot’s $PackedArray type, which is an efficient array of $Elements.

PackedInt64Array

Implements Godot’s $PackedArray type, which is an efficient array of $Elements.

PackedStringArray

Implements Godot’s $PackedArray type, which is an efficient array of $Elements.

PackedVector2Array

Implements Godot’s $PackedArray type, which is an efficient array of $Elements.

PackedVector3Array

Implements Godot’s $PackedArray type, which is an efficient array of $Elements.

Plane

3D plane in Hessian normal form.

Projection

A 4x4 matrix used for 3D projective transformations. It can represent transformations such as translation, rotation, scaling, shearing, and perspective division. It consists of four Vector4 columns.

Quaternion

Rect2

2D axis-aligned bounding box.

Rect2i

2D axis-aligned integer bounding box.

Signal

StringName

A string optimized for unique names.

Transform2D

Affine 2D transform (2x3 matrix).

Transform3D

Affine 3D transform (3x4 matrix).

Variant

VariantConversionError

Vector2

Vector used for 2D math using floating point coordinates.

Vector2i

Vector used for 2D math using integer coordinates.

Vector3

Vector used for 3D math using floating point coordinates.

Vector3i

Vector used for 3D math using integer coordinates.

Vector4

Vector used for 4D math using floating point coordinates.

Vector4i

Vector used for 4D math using integer coordinates.

Enums

ColorChannelOrder

EulerOrder

The ordering used to interpret a set of euler angles as extrinsic rotations.

ProjectionEye

The eye to create a projection for, when creating a projection adjusted for head-mounted displays.

ProjectionPlane

A projections clipping plane.

RectSide

The side of a Rect2 or Rect2i.

Rid

A RID (“resource ID”) is an opaque handle that refers to a Godot Resource.

VariantOperator

VariantType

Vector2Axis

Enumerates the axes in a Vector2.

Vector2iAxis

Enumerates the axes in a Vector2i.

Vector3Axis

Enumerates the axes in a Vector3.

Vector3iAxis

Enumerates the axes in a Vector3i.

Vector4Axis

Enumerates the axes in a Vector4.

Vector4iAxis

Enumerates the axes in a Vector4i.

Constants

CMP_EPSILON

Traits

FromVariant

Trait to enable conversions of types from the Variant type.

RealConv

Clippy often complains if you do f as f64 when f is already an f64. This trait exists to make it easy to convert between the different reals and floats without a lot of allowing clippy lints for your code.

ToVariant

Trait to enable conversions of types to the Variant type.

Functions

bezier_derivative

bezier_interpolate

cubic_interpolate

cubic_interpolate_in_time

fposmod

is_angle_equal_approx

Check if two angles are approximately equal, by comparing the distance between the points on the unit circle with 0 using is_equal_approx.

is_equal_approx

is_zero_approx

lerp

lerp_angle

Linearly interpolates between two angles (in radians) by a weight value between 0.0 and 1.0.

sign

snapped

Type Definitions

VariantArray

A Godot Array without an assigned type.

real

Floating point type used for many structs and functions in Godot.