Struct godot::builtin::NodePath

pub struct NodePath { /* private fields */ }

A pre-parsed scene tree path.

Null bytes

Note that Godot ignores any bytes after a null-byte. This means that for instance "hello, world!" and "hello, world!\0 ignored by Godot" will be treated as the same string if converted to a NodePath.

All string types

Intended use case	String type

General purpose	GString

Interned names	StringName

Scene-node paths	NodePath

Implementations

impl NodePath

pub fn is_empty(&self) -> bool

pub fn hash(&self) -> u32

Returns a 32-bit integer hash value representing the string.