Struct godot::builtin::Rect2

pub struct Rect2 {
    pub position: Vector2,
    pub size: Vector2,
}

2D axis-aligned bounding box.

Rect2 consists of a position, a size, and several utility functions. It is typically used for fast overlap tests.

Currently most methods are only available through InnerRect2.

The 3D counterpart to Rect2 is Aabb.

Fields
position: Vector2size: Vector2
Implementations

impl Rect2

pub const fn new(position: Vector2, size: Vector2) -> Rect2

Create a new Rect2 from a position and a size.

Godot equivalent: Rect2(Vector2 position, Vector2 size)

pub const fn from_components(x: f32, y: f32, width: f32, height: f32) -> Rect2

Create a new Rect2 from four reals representing position (x,y) and size (width,height).

Godot equivalent: Rect2(float x, float y, float width, float height)

pub const fn from_rect2i(rect: Rect2i) -> Rect2

Create a new Rect2 from a Rect2i, using as for i32 to real conversions.

Godot equivalent: Rect2(Rect2i from)

pub fn from_corners(position: Vector2, end: Vector2) -> Rect2

Create a new Rect2 with the first corner at position and the opposite corner at end.

pub fn end(&self) -> Vector2

The end of the Rect2 calculated as position + size.

Godot equivalent: Rect2.size property

pub fn set_end(&mut self, end: Vector2)

Set size based on desired end-point.

Godot equivalent: Rect2.size property

pub fn is_equal_approx(&self, other: &Rect2) -> bool

Returns true if the two Rect2s are approximately equal, by calling is_equal_approx on position and size.

Godot equivalent: Rect2.is_equal_approx()