Struct godot::builtin::Transform3D

pub struct Transform3D {
    pub basis: Basis,
    pub origin: Vector3,
}

Affine 3D transform (3x4 matrix).

Used for 3D linear transformations. Uses a basis + origin representation.

Expressed as a 3x4 matrix, this transform consists of 3 basis (column) vectors a, b, c as well as an origin o:

[ a.x  b.x  c.x  o.x ]
[ a.y  b.y  c.y  o.y ]
[ a.z  b.z  c.z  o.z ]

Fields
basis: Basis

The basis is a matrix containing 3 vectors as its columns. They can be interpreted as the basis vectors of the transformed coordinate system.

origin: Vector3

The new origin of the transformed coordinate system.

Implementations

impl Transform3D

pub const IDENTITY: Transform3D = Self::new(Basis::IDENTITY, Vector3::ZERO)

The identity transform, with no translation, rotation or scaling applied. When applied to other data structures, IDENTITY performs no transformation.

Godot equivalent: Transform3D.IDENTITY

pub const FLIP_X: Transform3D = Self::new(Basis::FLIP_X, Vector3::ZERO)

Transform3D with mirroring applied perpendicular to the YZ plane.

Godot equivalent: Transform3D.FLIP_X

pub const FLIP_Y: Transform3D = Self::new(Basis::FLIP_Y, Vector3::ZERO)

Transform3D with mirroring applied perpendicular to the XZ plane.

Godot equivalent: Transform3D.FLIP_Y

pub const FLIP_Z: Transform3D = Self::new(Basis::FLIP_Z, Vector3::ZERO)

Transform3D with mirroring applied perpendicular to the XY plane.

Godot equivalent: Transform3D.FLIP_Z

pub const fn new(basis: Basis, origin: Vector3) -> Transform3D

Create a new transform from a Basis and a Vector3.

Godot equivalent: Transform3D(Basis basis, Vector3 origin)

pub const fn from_cols( a: Vector3, b: Vector3, c: Vector3, origin: Vector3 ) -> Transform3D

Create a new transform from 4 matrix-columns.

Godot equivalent: Transform3D(Vector3 x_axis, Vector3 y_axis, Vector3 z_axis, Vector3 origin)

pub fn from_projection(proj: Projection) -> Transform3D

Constructs a Transform3d from a Projection by trimming the last row of the projection matrix.

Godot equivalent: Transform3D(Projection from)

pub fn affine_inverse(self) -> Transform3D

Returns the inverse of the transform, under the assumption that the transformation is composed of rotation, scaling and translation.

Godot equivalent: Transform3D.affine_inverse()

pub fn interpolate_with(self, other: Transform3D, weight: f32) -> Transform3D

Returns a transform interpolated between this transform and another by a given weight (on the range of 0.0 to 1.0).

Godot equivalent: Transform3D.interpolate_with()

pub fn is_equal_approx(&self, other: &Transform3D) -> bool

Returns `true if this transform and transform are approximately equal, by calling is_equal_approx each basis and origin.

Godot equivalent: Transform3D.is_equal_approx()

pub fn is_finite(&self) -> bool

Returns true if this transform is finite by calling is_finite` on the basis and origin.

Godot equivalent: Transform3D.is_finite()

pub fn looking_at(self, target: Vector3, up: Vector3) -> Transform3D

Returns a copy of the transform rotated such that the forward axis (-Z) points towards the target position.

See Basis::new_looking_at() for more information.

Godot equivalent: Transform3D.looking_at()

pub fn orthonormalized(self) -> Transform3D

Returns the transform with the basis orthogonal (90 degrees), and normalized axis vectors (scale of 1 or -1).

Godot equivalent: Transform3D.orthonormalized()

pub fn rotated(self, axis: Vector3, angle: f32) -> Transform3D

Returns a copy of the transform rotated by the given angle (in radians). This method is an optimized version of multiplying the given transform X with a corresponding rotation transform R from the left, i.e., R * X. This can be seen as transforming with respect to the global/parent frame.

Godot equivalent: Transform2D.rotated()

pub fn rotated_local(self, axis: Vector3, angle: f32) -> Transform3D

Returns a copy of the transform rotated by the given angle (in radians). This method is an optimized version of multiplying the given transform X with a corresponding rotation transform R from the right, i.e., X * R. This can be seen as transforming with respect to the local frame.

Godot equivalent: Transform2D.rotated_local()

pub fn scaled(self, scale: Vector3) -> Transform3D

Returns a copy of the transform scaled by the given scale factor. This method is an optimized version of multiplying the given transform X with a corresponding scaling transform S from the left, i.e., S * X. This can be seen as transforming with respect to the global/parent frame.

Godot equivalent: Transform2D.scaled()

pub fn scaled_local(self, scale: Vector3) -> Transform3D

Returns a copy of the transform scaled by the given scale factor. This method is an optimized version of multiplying the given transform X with a corresponding scaling transform S from the right, i.e., X * S. This can be seen as transforming with respect to the local frame.

Godot equivalent: Transform2D.scaled_local()

pub fn translated(self, offset: Vector3) -> Transform3D

Returns a copy of the transform translated by the given offset. This method is an optimized version of multiplying the given transform X with a corresponding translation transform T from the left, i.e., T * X. This can be seen as transforming with respect to the global/parent frame.

Godot equivalent: Transform2D.translated()

pub fn translated_local(self, offset: Vector3) -> Transform3D

Returns a copy of the transform translated by the given offset. This method is an optimized version of multiplying the given transform X with a corresponding translation transform T from the right, i.e., X * T. This can be seen as transforming with respect to the local frame.

Godot equivalent: Transform2D.translated()