Struct godot::builtin::Vector3

pub struct Vector3 {
    pub x: f32,
    pub y: f32,
    pub z: f32,
}

Vector used for 3D math using floating point coordinates.

3-element structure that can be used to represent positions in 3D space or any other triple of numeric values.

It uses floating-point coordinates of 32-bit precision, unlike the engine’s float type which is always 64-bit. The engine can be compiled with the option precision=double to use 64-bit vectors; use the gdext library with the double-precision feature in that case.

See Vector3i for its integer counterpart.

Fields
x: f32

The vector’s X component.

y: f32

The vector’s Y component.

z: f32

The vector’s Z component.

Implementations

impl Vector3

pub const ZERO: Vector3 = Self::splat(0.0)

Vector with all components set to 0.0.

pub const ONE: Vector3 = Self::splat(1.0)

Vector with all components set to 1.0.

pub const LEFT: Vector3 = Self::new(-1.0, 0.0, 0.0)

Unit vector in -X direction. Can be interpreted as left in an untransformed 3D world.

pub const RIGHT: Vector3 = Self::new(1.0, 0.0, 0.0)

Unit vector in +X direction. Can be interpreted as right in an untransformed 3D world.

pub const UP: Vector3 = Self::new(0.0, 1.0, 0.0)

Unit vector in +Y direction. Typically interpreted as up in a 3D world.

pub const DOWN: Vector3 = Self::new(0.0, -1.0, 0.0)

Unit vector in -Y direction. Typically interpreted as down in a 3D world.

pub const FORWARD: Vector3 = Self::new(0.0, 0.0, -1.0)

Unit vector in -Z direction. Can be interpreted as “into the screen” in an untransformed 3D world.

pub const BACK: Vector3 = Self::new(0.0, 0.0, 1.0)

Unit vector in +Z direction. Can be interpreted as “out of the screen” in an untransformed 3D world.

pub const fn new(x: f32, y: f32, z: f32) -> Vector3

Returns a Vector3 with the given components.

pub const fn splat(v: f32) -> Vector3

Returns a new Vector3 with all components set to v.

pub const fn from_vector3i(v: Vector3i) -> Vector3

Constructs a new Vector3 from a Vector3i.

pub fn angle_to(self, to: Vector3) -> f32

pub fn bezier_derivative( self, control_1: Vector3, control_2: Vector3, end: Vector3, t: f32 ) -> Vector3

pub fn bezier_interpolate( self, control_1: Vector3, control_2: Vector3, end: Vector3, t: f32 ) -> Vector3

pub fn bounce(self, normal: Vector3) -> Vector3

pub fn ceil(self) -> Vector3

pub fn clamp(self, min: Vector3, max: Vector3) -> Vector3

pub fn cross(self, with: Vector3) -> Vector3

pub fn cubic_interpolate( self, b: Vector3, pre_a: Vector3, post_b: Vector3, weight: f32 ) -> Vector3

pub fn cubic_interpolate_in_time( self, b: Vector3, pre_a: Vector3, post_b: Vector3, weight: f32, b_t: f32, pre_a_t: f32, post_b_t: f32 ) -> Vector3

pub fn direction_to(self, to: Vector3) -> Vector3

pub fn distance_squared_to(self, to: Vector3) -> f32

pub fn distance_to(self, to: Vector3) -> f32

pub fn dot(self, with: Vector3) -> f32

pub fn floor(self) -> Vector3

pub fn inverse(self) -> Vector3

pub fn is_equal_approx(self, to: Vector3) -> bool

pub fn is_finite(self) -> bool

pub fn is_normalized(self) -> bool

pub fn is_zero_approx(self) -> bool

pub fn length_squared(self) -> f32

pub fn lerp(self, to: Vector3, weight: f32) -> Vector3

pub fn limit_length(self, length: Option<f32>) -> Vector3

pub fn max_axis_index(self) -> Vector3Axis

pub fn min_axis_index(self) -> Vector3Axis

pub fn move_toward(self, to: Vector3, delta: f32) -> Vector3

pub fn posmod(self, pmod: f32) -> Vector3

pub fn posmodv(self, modv: Vector3) -> Vector3

pub fn project(self, b: Vector3) -> Vector3

pub fn reflect(self, normal: Vector3) -> Vector3

pub fn round(self) -> Vector3

pub fn sign(self) -> Vector3

pub fn signed_angle_to(self, to: Vector3, axis: Vector3) -> f32

pub fn slide(self, normal: Vector3) -> Vector3

pub fn snapped(self, step: Vector3) -> Vector3

pub fn rotated(self, axis: Vector3, angle: f32) -> Vector3

Returns this vector rotated around axis by angle radians. axis must be normalized.

Panics

If axis is not normalized.

impl Vector3

pub fn abs(self) -> Vector3

Returns a new vector with all components in absolute values (i.e. positive or zero).

pub fn coord_min(self, other: Vector3) -> Vector3

Returns a new vector containing the minimum of the two vectors, component-wise.

pub fn coord_max(self, other: Vector3) -> Vector3

Returns a new vector containing the maximum of the two vectors, component-wise.

impl Vector3

pub fn length(self) -> f32

Returns the length (magnitude) of this vector.

pub fn normalized(self) -> Vector3

Returns the vector scaled to unit length. Equivalent to self / self.length(). See also is_normalized().

If the vector is zero, the result is also zero.