godot::classes

Trait IEditorExportPlugin

pub trait IEditorExportPlugin: GodotClass<Base = EditorExportPlugin> + You_forgot_the_attribute__godot_api {

    // Required methods
    fn customize_resource(
        &mut self,
        resource: Gd<Resource>,
        path: GString,
    ) -> Option<Gd<Resource>>;
    fn customize_scene(
        &mut self,
        scene: Gd<Node>,
        path: GString,
    ) -> Option<Gd<Node>>;
    fn get_customization_configuration_hash(&self) -> u64;
    fn get_name(&self) -> GString;

    // Provided methods
    fn init(base: Base<Self::Base>) -> Self { ... }
    fn to_string(&self) -> GString { ... }
    fn on_notification(&mut self, what: ObjectNotification) { ... }
    fn get_property(&self, property: StringName) -> Option<Variant> { ... }
    fn set_property(&mut self, property: StringName, value: Variant) -> bool { ... }
    fn get_property_list(&mut self) -> Vec<PropertyInfo> { ... }
    fn property_get_revert(&self, property: StringName) -> Option<Variant> { ... }
    fn export_file(
        &mut self,
        path: GString,
        type_: GString,
        features: PackedStringArray,
    ) { ... }
    fn export_begin(
        &mut self,
        features: PackedStringArray,
        is_debug: bool,
        path: GString,
        flags: u32,
    ) { ... }
    fn export_end(&mut self) { ... }
    fn begin_customize_resources(
        &self,
        platform: Option<Gd<EditorExportPlatform>>,
        features: PackedStringArray,
    ) -> bool { ... }
    fn begin_customize_scenes(
        &self,
        platform: Option<Gd<EditorExportPlatform>>,
        features: PackedStringArray,
    ) -> bool { ... }
    fn end_customize_scenes(&mut self) { ... }
    fn end_customize_resources(&mut self) { ... }
    fn get_export_options(
        &self,
        platform: Option<Gd<EditorExportPlatform>>,
    ) -> Array<Dictionary> { ... }
    fn get_export_options_overrides(
        &self,
        platform: Option<Gd<EditorExportPlatform>>,
    ) -> Dictionary { ... }
    fn should_update_export_options(
        &self,
        platform: Option<Gd<EditorExportPlatform>>,
    ) -> bool { ... }
    fn get_export_option_warning(
        &self,
        platform: Option<Gd<EditorExportPlatform>>,
        option: GString,
    ) -> GString { ... }
    fn get_export_features(
        &self,
        platform: Option<Gd<EditorExportPlatform>>,
        debug: bool,
    ) -> PackedStringArray { ... }
    fn supports_platform(
        &self,
        platform: Option<Gd<EditorExportPlatform>>,
    ) -> bool { ... }
    fn get_android_dependencies(
        &self,
        platform: Option<Gd<EditorExportPlatform>>,
        debug: bool,
    ) -> PackedStringArray { ... }
    fn get_android_dependencies_maven_repos(
        &self,
        platform: Option<Gd<EditorExportPlatform>>,
        debug: bool,
    ) -> PackedStringArray { ... }
    fn get_android_libraries(
        &self,
        platform: Option<Gd<EditorExportPlatform>>,
        debug: bool,
    ) -> PackedStringArray { ... }
    fn get_android_manifest_activity_element_contents(
        &self,
        platform: Option<Gd<EditorExportPlatform>>,
        debug: bool,
    ) -> GString { ... }
    fn get_android_manifest_application_element_contents(
        &self,
        platform: Option<Gd<EditorExportPlatform>>,
        debug: bool,
    ) -> GString { ... }
    fn get_android_manifest_element_contents(
        &self,
        platform: Option<Gd<EditorExportPlatform>>,
        debug: bool,
    ) -> GString { ... }

}

Virtual methods for class EditorExportPlugin.

These methods represent constructors (init) or callbacks invoked by the engine.

See also Godot docs for EditorExportPlugin methods.

Required Methods

fn customize_resource( &mut self, resource: Gd<Resource>, path: GString, ) -> Option<Gd<Resource>>

fn customize_scene( &mut self, scene: Gd<Node>, path: GString, ) -> Option<Gd<Node>>

fn get_customization_configuration_hash(&self) -> u64

fn get_name(&self) -> GString

Provided Methods

fn init(base: Base<Self::Base>) -> Self

Godot constructor, accepting an injected base object.

base refers to the base instance of the class, which can either be stored in a Base<T> field or discarded. This method returns a fully-constructed instance, which will then be moved into a Gd<T> pointer.

If the class has a #[class(init)] attribute, this method will be auto-generated and must not be overridden.

fn to_string(&self) -> GString

String representation of the Godot instance.

Override this method to define how the instance is represented as a string. Used by impl Display for Gd<T>, as well as str() and print() in GDScript.

fn on_notification(&mut self, what: ObjectNotification)

Called when the object receives a Godot notification.

The type of notification can be identified through what. The enum is designed to hold all possible NOTIFICATION_* constants that the current class can handle. However, this is not validated in Godot, so an enum variant Unknown exists to represent integers out of known constants (mistakes or future additions).

This method is named _notification in Godot, but on_notification in Rust. To send notifications, use the Object::notify method.

See also in Godot docs:

Object::_notification.

Notifications tutorial.

fn get_property(&self, property: StringName) -> Option<Variant>

Called whenever get() is called or Godot gets the value of a property.

Should return the given property’s value as Some(value), or None if the property should be handled normally.

See also in Godot docs:

Object::_get.

fn set_property(&mut self, property: StringName, value: Variant) -> bool

Called whenever Godot set() is called or Godot sets the value of a property.

Should set property to the given value and return true, or return false to indicate the property should be handled normally.

See also in Godot docs:

Object::_set.

fn get_property_list(&mut self) -> Vec<PropertyInfo>

Available on since_api="4.3" only.

Called whenever Godot get_property_list() is called, the returned vector here is appended to the existing list of properties.

This should mainly be used for advanced purposes, such as dynamically updating the property list in the editor.

See also in Godot docs:

Object::_get_property_list

fn property_get_revert(&self, property: StringName) -> Option<Variant>

Called by Godot to tell if a property has a custom revert or not.

Return None for no custom revert, and return Some(value) to specify the custom revert.

This is a combination of Godot’s Object::_property_get_revert and Object::_property_can_revert. This means that this function will usually be called twice by Godot to find the revert.

Note that this should be a pure function. That is, it should always return the same value for a property as long as self remains unchanged. Otherwise, this may lead to unexpected (safe) behavior.

fn export_file( &mut self, path: GString, type_: GString, features: PackedStringArray, )

fn export_begin( &mut self, features: PackedStringArray, is_debug: bool, path: GString, flags: u32, )

fn export_end(&mut self)

fn begin_customize_resources( &self, platform: Option<Gd<EditorExportPlatform>>, features: PackedStringArray, ) -> bool

fn begin_customize_scenes( &self, platform: Option<Gd<EditorExportPlatform>>, features: PackedStringArray, ) -> bool

fn end_customize_scenes(&mut self)

fn end_customize_resources(&mut self)

fn get_export_options( &self, platform: Option<Gd<EditorExportPlatform>>, ) -> Array<Dictionary>

fn get_export_options_overrides( &self, platform: Option<Gd<EditorExportPlatform>>, ) -> Dictionary

fn should_update_export_options( &self, platform: Option<Gd<EditorExportPlatform>>, ) -> bool

fn get_export_option_warning( &self, platform: Option<Gd<EditorExportPlatform>>, option: GString, ) -> GString

fn get_export_features( &self, platform: Option<Gd<EditorExportPlatform>>, debug: bool, ) -> PackedStringArray

fn supports_platform(&self, platform: Option<Gd<EditorExportPlatform>>) -> bool

fn get_android_dependencies( &self, platform: Option<Gd<EditorExportPlatform>>, debug: bool, ) -> PackedStringArray

fn get_android_dependencies_maven_repos( &self, platform: Option<Gd<EditorExportPlatform>>, debug: bool, ) -> PackedStringArray

fn get_android_libraries( &self, platform: Option<Gd<EditorExportPlatform>>, debug: bool, ) -> PackedStringArray

fn get_android_manifest_activity_element_contents( &self, platform: Option<Gd<EditorExportPlatform>>, debug: bool, ) -> GString

fn get_android_manifest_application_element_contents( &self, platform: Option<Gd<EditorExportPlatform>>, debug: bool, ) -> GString

fn get_android_manifest_element_contents( &self, platform: Option<Gd<EditorExportPlatform>>, debug: bool, ) -> GString

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.