godot::classes
Trait IOpenXrExtensionWrapperExtension

pub trait IOpenXrExtensionWrapperExtension: GodotClass<Base = OpenXrExtensionWrapperExtension> + You_forgot_the_attribute__godot_api {

    // Provided methods
    fn init(base: Base<Self::Base>) -> Self { ... }
    fn to_string(&self) -> GString { ... }
    fn on_notification(&mut self, what: ObjectNotification) { ... }
    fn get_property(&self, property: StringName) -> Option<Variant> { ... }
    fn set_property(&mut self, property: StringName, value: Variant) -> bool { ... }
    fn get_property_list(&mut self) -> Vec<PropertyInfo> { ... }
    fn property_get_revert(&self, property: StringName) -> Option<Variant> { ... }
    fn get_requested_extensions(&mut self) -> Dictionary { ... }
    unsafe fn set_system_properties_and_get_next_pointer(
        &mut self,
        next_pointer: *mut c_void,
    ) -> u64 { ... }
    unsafe fn set_instance_create_info_and_get_next_pointer(
        &mut self,
        next_pointer: *mut c_void,
    ) -> u64 { ... }
    unsafe fn set_session_create_and_get_next_pointer(
        &mut self,
        next_pointer: *mut c_void,
    ) -> u64 { ... }
    unsafe fn set_swapchain_create_info_and_get_next_pointer(
        &mut self,
        next_pointer: *mut c_void,
    ) -> u64 { ... }
    unsafe fn set_hand_joint_locations_and_get_next_pointer(
        &mut self,
        hand_index: i32,
        next_pointer: *mut c_void,
    ) -> u64 { ... }
    fn get_composition_layer_count(&mut self) -> i32 { ... }
    fn get_composition_layer(&mut self, index: i32) -> u64 { ... }
    fn get_composition_layer_order(&mut self, index: i32) -> i32 { ... }
    fn get_suggested_tracker_names(&mut self) -> PackedStringArray { ... }
    fn on_register_metadata(&mut self) { ... }
    fn on_before_instance_created(&mut self) { ... }
    fn on_instance_created(&mut self, instance: u64) { ... }
    fn on_instance_destroyed(&mut self) { ... }
    fn on_session_created(&mut self, session: u64) { ... }
    fn on_process(&mut self) { ... }
    fn on_pre_render(&mut self) { ... }
    fn on_main_swapchains_created(&mut self) { ... }
    fn on_session_destroyed(&mut self) { ... }
    fn on_state_idle(&mut self) { ... }
    fn on_state_ready(&mut self) { ... }
    fn on_state_synchronized(&mut self) { ... }
    fn on_state_visible(&mut self) { ... }
    fn on_state_focused(&mut self) { ... }
    fn on_state_stopping(&mut self) { ... }
    fn on_state_loss_pending(&mut self) { ... }
    fn on_state_exiting(&mut self) { ... }
    unsafe fn on_event_polled(&mut self, event: *const c_void) -> bool { ... }
    unsafe fn set_viewport_composition_layer_and_get_next_pointer(
        &mut self,
        layer: *const c_void,
        property_values: Dictionary,
        next_pointer: *mut c_void,
    ) -> u64 { ... }
    fn get_viewport_composition_layer_extension_properties(
        &mut self,
    ) -> Array<Dictionary> { ... }
    fn get_viewport_composition_layer_extension_property_defaults(
        &mut self,
    ) -> Dictionary { ... }
    unsafe fn on_viewport_composition_layer_destroyed(
        &mut self,
        layer: *const c_void,
    ) { ... }

}

Virtual methods for class OpenXrExtensionWrapperExtension.

These methods represent constructors (init) or callbacks invoked by the engine.

See also Godot docs for OpenXRExtensionWrapperExtension methods.

Provided Methods

fn init(base: Base<Self::Base>) -> Self

Godot constructor, accepting an injected base object.

base refers to the base instance of the class, which can either be stored in a Base<T> field or discarded. This method returns a fully-constructed instance, which will then be moved into a Gd<T> pointer.

If the class has a #[class(init)] attribute, this method will be auto-generated and must not be overridden.

fn to_string(&self) -> GString

String representation of the Godot instance.

Override this method to define how the instance is represented as a string. Used by impl Display for Gd<T>, as well as str() and print() in GDScript.

fn on_notification(&mut self, what: ObjectNotification)

Called when the object receives a Godot notification.

The type of notification can be identified through what. The enum is designed to hold all possible NOTIFICATION_* constants that the current class can handle. However, this is not validated in Godot, so an enum variant Unknown exists to represent integers out of known constants (mistakes or future additions).

This method is named _notification in Godot, but on_notification in Rust. To send notifications, use the Object::notify method.

See also in Godot docs:

Object::_notification.

Notifications tutorial.

fn get_property(&self, property: StringName) -> Option<Variant>

Called whenever get() is called or Godot gets the value of a property.

Should return the given property’s value as Some(value), or None if the property should be handled normally.

See also in Godot docs:

Object::_get.

fn set_property(&mut self, property: StringName, value: Variant) -> bool

Called whenever Godot set() is called or Godot sets the value of a property.

Should set property to the given value and return true, or return false to indicate the property should be handled normally.

See also in Godot docs:

Object::_set.

fn get_property_list(&mut self) -> Vec<PropertyInfo>

Available on since_api="4.3" only.

Called whenever Godot get_property_list() is called, the returned vector here is appended to the existing list of properties.

This should mainly be used for advanced purposes, such as dynamically updating the property list in the editor.

See also in Godot docs:

Object::_get_property_list

fn property_get_revert(&self, property: StringName) -> Option<Variant>

Called by Godot to tell if a property has a custom revert or not.

Return None for no custom revert, and return Some(value) to specify the custom revert.

This is a combination of Godot’s Object::_property_get_revert and Object::_property_can_revert. This means that this function will usually be called twice by Godot to find the revert.

Note that this should be a pure function. That is, it should always return the same value for a property as long as self remains unchanged. Otherwise, this may lead to unexpected (safe) behavior.

fn get_requested_extensions(&mut self) -> Dictionary

unsafe fn set_system_properties_and_get_next_pointer( &mut self, next_pointer: *mut c_void, ) -> u64

Safety

This method has automatically been marked unsafe because it accepts raw pointers as parameters. If Godot does not document any safety requirements, make sure you understand the underlying semantics.

unsafe fn set_instance_create_info_and_get_next_pointer( &mut self, next_pointer: *mut c_void, ) -> u64

Safety

This method has automatically been marked unsafe because it accepts raw pointers as parameters. If Godot does not document any safety requirements, make sure you understand the underlying semantics.

unsafe fn set_session_create_and_get_next_pointer( &mut self, next_pointer: *mut c_void, ) -> u64

Safety

This method has automatically been marked unsafe because it accepts raw pointers as parameters. If Godot does not document any safety requirements, make sure you understand the underlying semantics.

unsafe fn set_swapchain_create_info_and_get_next_pointer( &mut self, next_pointer: *mut c_void, ) -> u64

Safety

This method has automatically been marked unsafe because it accepts raw pointers as parameters. If Godot does not document any safety requirements, make sure you understand the underlying semantics.

unsafe fn set_hand_joint_locations_and_get_next_pointer( &mut self, hand_index: i32, next_pointer: *mut c_void, ) -> u64

Safety

This method has automatically been marked unsafe because it accepts raw pointers as parameters. If Godot does not document any safety requirements, make sure you understand the underlying semantics.

fn get_composition_layer_count(&mut self) -> i32

fn get_composition_layer(&mut self, index: i32) -> u64

fn get_composition_layer_order(&mut self, index: i32) -> i32

fn get_suggested_tracker_names(&mut self) -> PackedStringArray

fn on_register_metadata(&mut self)

fn on_before_instance_created(&mut self)

fn on_instance_created(&mut self, instance: u64)

fn on_instance_destroyed(&mut self)

fn on_session_created(&mut self, session: u64)

fn on_process(&mut self)

fn on_pre_render(&mut self)

fn on_main_swapchains_created(&mut self)

fn on_session_destroyed(&mut self)

fn on_state_idle(&mut self)

fn on_state_ready(&mut self)

fn on_state_synchronized(&mut self)

fn on_state_visible(&mut self)

fn on_state_focused(&mut self)

fn on_state_stopping(&mut self)

fn on_state_loss_pending(&mut self)

fn on_state_exiting(&mut self)

unsafe fn on_event_polled(&mut self, event: *const c_void) -> bool

Safety

This method has automatically been marked unsafe because it accepts raw pointers as parameters. If Godot does not document any safety requirements, make sure you understand the underlying semantics.

unsafe fn set_viewport_composition_layer_and_get_next_pointer( &mut self, layer: *const c_void, property_values: Dictionary, next_pointer: *mut c_void, ) -> u64

Safety

This method has automatically been marked unsafe because it accepts raw pointers as parameters. If Godot does not document any safety requirements, make sure you understand the underlying semantics.

fn get_viewport_composition_layer_extension_properties( &mut self, ) -> Array<Dictionary>

fn get_viewport_composition_layer_extension_property_defaults( &mut self, ) -> Dictionary

unsafe fn on_viewport_composition_layer_destroyed( &mut self, layer: *const c_void, )

Safety

This method has automatically been marked unsafe because it accepts raw pointers as parameters. If Godot does not document any safety requirements, make sure you understand the underlying semantics.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.