Struct godot::engine::RenderingServer

pub struct RenderingServer { /* private fields */ }

Godot class RenderingServer.

Inherits Object.

Related symbols:

rendering_server: sidecar module with related enum/flag types

RenderingServerVirtual: virtual methods

See also Godot docs for RenderingServer.

Implementations

impl RenderingServer

pub fn singleton() -> Gd<RenderingServer>

pub fn notify(&mut self, what: ObjectNotification)

⚠️ Sends a Godot notification to all classes inherited by the object.

Triggers calls to on_notification(), and depending on the notification, also to Godot’s lifecycle callbacks such as ready().

Starts from the highest ancestor (the Object class) and goes down the hierarchy. See also Godot docs for Object::notification().

Panics

If you call this method on a user-defined object while holding a GdRef or GdMut guard on the instance, you will encounter a panic. The reason is that the receiving virtual method on_notification() acquires a GdMut lock dynamically, which must be exclusive.

pub fn notify_reversed(&mut self, what: ObjectNotification)

⚠️ Like Self::notify(), but starts at the most-derived class and goes up the hierarchy.

See docs of that method, including the panics.

pub fn texture_2d_create(&mut self, image: Gd<Image>) -> Rid

pub fn texture_2d_layered_create( &mut self, layers: Array<Gd<Image>>, layered_type: TextureLayeredType ) -> Rid

pub fn texture_3d_create( &mut self, format: Format, width: i64, height: i64, depth: i64, mipmaps: bool, data: Array<Gd<Image>> ) -> Rid

pub fn texture_proxy_create(&mut self, base: Rid) -> Rid

pub fn texture_2d_update(&mut self, texture: Rid, image: Gd<Image>, layer: i64)

pub fn texture_3d_update(&mut self, texture: Rid, data: Array<Gd<Image>>)

pub fn texture_proxy_update(&mut self, texture: Rid, proxy_to: Rid)

pub fn texture_2d_placeholder_create(&mut self) -> Rid

pub fn texture_2d_layered_placeholder_create( &mut self, layered_type: TextureLayeredType ) -> Rid

pub fn texture_3d_placeholder_create(&mut self) -> Rid

pub fn texture_2d_get(&self, texture: Rid) -> Option<Gd<Image>>

pub fn texture_2d_layer_get(&self, texture: Rid, layer: i64) -> Option<Gd<Image>>

pub fn texture_3d_get(&self, texture: Rid) -> Array<Gd<Image>>

pub fn texture_replace(&mut self, texture: Rid, by_texture: Rid)

pub fn texture_set_size_override(&mut self, texture: Rid, width: i64, height: i64)

pub fn texture_set_path(&mut self, texture: Rid, path: GodotString)

pub fn texture_get_path(&self, texture: Rid) -> GodotString

pub fn texture_set_force_redraw_if_visible(&mut self, texture: Rid, enable: bool)

pub fn texture_get_rd_texture(&self, texture: Rid, srgb: bool) -> Rid

pub fn shader_create(&mut self) -> Rid

pub fn shader_set_code(&mut self, shader: Rid, code: GodotString)

pub fn shader_set_path_hint(&mut self, shader: Rid, path: GodotString)

pub fn shader_get_code(&self, shader: Rid) -> GodotString

pub fn get_shader_parameter_list(&self, shader: Rid) -> Array<Dictionary>

pub fn shader_get_parameter_default( &self, shader: Rid, name: StringName ) -> Variant

pub fn shader_set_default_texture_parameter( &mut self, shader: Rid, name: StringName, texture: Rid, index: i64 )

pub fn shader_get_default_texture_parameter( &self, shader: Rid, name: StringName, index: i64 ) -> Rid

pub fn material_create(&mut self) -> Rid

pub fn material_set_shader(&mut self, shader_material: Rid, shader: Rid)

pub fn material_set_param( &mut self, material: Rid, parameter: StringName, value: Variant )

pub fn material_get_param(&self, material: Rid, parameter: StringName) -> Variant

pub fn material_set_render_priority(&mut self, material: Rid, priority: i64)

pub fn material_set_next_pass(&mut self, material: Rid, next_material: Rid)

pub fn mesh_create_from_surfaces( &mut self, surfaces: Array<Dictionary>, blend_shape_count: i64 ) -> Rid

pub fn mesh_create(&mut self) -> Rid

pub fn mesh_surface_get_format_offset( &self, format: ArrayFormat, vertex_count: i64, array_index: i64 ) -> i64

pub fn mesh_surface_get_format_vertex_stride( &self, format: ArrayFormat, vertex_count: i64 ) -> i64

pub fn mesh_surface_get_format_attribute_stride( &self, format: ArrayFormat, vertex_count: i64 ) -> i64

pub fn mesh_surface_get_format_skin_stride( &self, format: ArrayFormat, vertex_count: i64 ) -> i64

pub fn mesh_add_surface(&mut self, mesh: Rid, surface: Dictionary)

pub fn mesh_add_surface_from_arrays( &mut self, mesh: Rid, primitive: PrimitiveType, arrays: Array<Variant>, blend_shapes: Array<Variant>, lods: Dictionary, compress_format: ArrayFormat )

pub fn mesh_get_blend_shape_count(&self, mesh: Rid) -> i64

pub fn mesh_set_blend_shape_mode(&mut self, mesh: Rid, mode: BlendShapeMode)

pub fn mesh_get_blend_shape_mode(&self, mesh: Rid) -> BlendShapeMode

pub fn mesh_surface_set_material(&mut self, mesh: Rid, surface: i64, material: Rid)

pub fn mesh_surface_get_material(&self, mesh: Rid, surface: i64) -> Rid

pub fn mesh_get_surface(&mut self, mesh: Rid, surface: i64) -> Dictionary

pub fn mesh_surface_get_arrays(&self, mesh: Rid, surface: i64) -> Array<Variant>

pub fn mesh_surface_get_blend_shape_arrays( &self, mesh: Rid, surface: i64 ) -> Array<Array<Variant>>

pub fn mesh_get_surface_count(&self, mesh: Rid) -> i64

pub fn mesh_set_custom_aabb(&mut self, mesh: Rid, aabb: Aabb)

pub fn mesh_get_custom_aabb(&self, mesh: Rid) -> Aabb

pub fn mesh_clear(&mut self, mesh: Rid)

pub fn mesh_surface_update_vertex_region( &mut self, mesh: Rid, surface: i64, offset: i64, data: PackedByteArray )

pub fn mesh_surface_update_attribute_region( &mut self, mesh: Rid, surface: i64, offset: i64, data: PackedByteArray )

pub fn mesh_surface_update_skin_region( &mut self, mesh: Rid, surface: i64, offset: i64, data: PackedByteArray )

pub fn mesh_set_shadow_mesh(&mut self, mesh: Rid, shadow_mesh: Rid)

pub fn multimesh_create(&mut self) -> Rid

pub fn multimesh_allocate_data( &mut self, multimesh: Rid, instances: i64, transform_format: MultimeshTransformFormat, color_format: bool, custom_data_format: bool )

pub fn multimesh_get_instance_count(&self, multimesh: Rid) -> i64

pub fn multimesh_set_mesh(&mut self, multimesh: Rid, mesh: Rid)

pub fn multimesh_instance_set_transform( &mut self, multimesh: Rid, index: i64, transform: Transform3D )

pub fn multimesh_instance_set_transform_2d( &mut self, multimesh: Rid, index: i64, transform: Transform2D )

pub fn multimesh_instance_set_color( &mut self, multimesh: Rid, index: i64, color: Color )

pub fn multimesh_instance_set_custom_data( &mut self, multimesh: Rid, index: i64, custom_data: Color )

pub fn multimesh_get_mesh(&self, multimesh: Rid) -> Rid

pub fn multimesh_get_aabb(&self, multimesh: Rid) -> Aabb

pub fn multimesh_instance_get_transform( &self, multimesh: Rid, index: i64 ) -> Transform3D

pub fn multimesh_instance_get_transform_2d( &self, multimesh: Rid, index: i64 ) -> Transform2D

pub fn multimesh_instance_get_color(&self, multimesh: Rid, index: i64) -> Color

pub fn multimesh_instance_get_custom_data( &self, multimesh: Rid, index: i64 ) -> Color

pub fn multimesh_set_visible_instances(&mut self, multimesh: Rid, visible: i64)

pub fn multimesh_get_visible_instances(&self, multimesh: Rid) -> i64

pub fn multimesh_set_buffer(&mut self, multimesh: Rid, buffer: PackedFloat32Array)

pub fn multimesh_get_buffer(&self, multimesh: Rid) -> PackedFloat32Array

pub fn skeleton_create(&mut self) -> Rid

pub fn skeleton_allocate_data( &mut self, skeleton: Rid, bones: i64, is_2d_skeleton: bool )

pub fn skeleton_get_bone_count(&self, skeleton: Rid) -> i64

pub fn skeleton_bone_set_transform( &mut self, skeleton: Rid, bone: i64, transform: Transform3D )

pub fn skeleton_bone_get_transform(&self, skeleton: Rid, bone: i64) -> Transform3D

pub fn skeleton_bone_set_transform_2d( &mut self, skeleton: Rid, bone: i64, transform: Transform2D )

pub fn skeleton_bone_get_transform_2d( &self, skeleton: Rid, bone: i64 ) -> Transform2D

pub fn skeleton_set_base_transform_2d( &mut self, skeleton: Rid, base_transform: Transform2D )

pub fn directional_light_create(&mut self) -> Rid

pub fn omni_light_create(&mut self) -> Rid

pub fn spot_light_create(&mut self) -> Rid

pub fn light_set_color(&mut self, light: Rid, color: Color)

pub fn light_set_param(&mut self, light: Rid, param: LightParam, value: f64)

pub fn light_set_shadow(&mut self, light: Rid, enabled: bool)

pub fn light_set_projector(&mut self, light: Rid, texture: Rid)

pub fn light_set_negative(&mut self, light: Rid, enable: bool)

pub fn light_set_cull_mask(&mut self, light: Rid, mask: i64)

pub fn light_set_distance_fade( &mut self, decal: Rid, enabled: bool, begin: f64, shadow: f64, length: f64 )

pub fn light_set_reverse_cull_face_mode(&mut self, light: Rid, enabled: bool)

pub fn light_set_bake_mode(&mut self, light: Rid, bake_mode: LightBakeMode)

pub fn light_set_max_sdfgi_cascade(&mut self, light: Rid, cascade: i64)

pub fn light_omni_set_shadow_mode( &mut self, light: Rid, mode: LightOmniShadowMode )

pub fn light_directional_set_shadow_mode( &mut self, light: Rid, mode: LightDirectionalShadowMode )

pub fn light_directional_set_blend_splits(&mut self, light: Rid, enable: bool)

pub fn light_directional_set_sky_mode( &mut self, light: Rid, mode: LightDirectionalSkyMode )

pub fn light_projectors_set_filter(&mut self, filter: LightProjectorFilter)

pub fn positional_soft_shadow_filter_set_quality( &mut self, quality: ShadowQuality )

pub fn directional_soft_shadow_filter_set_quality( &mut self, quality: ShadowQuality )

pub fn directional_shadow_atlas_set_size(&mut self, size: i64, is_16bits: bool)

pub fn reflection_probe_create(&mut self) -> Rid

pub fn reflection_probe_set_update_mode( &mut self, probe: Rid, mode: ReflectionProbeUpdateMode )

pub fn reflection_probe_set_intensity(&mut self, probe: Rid, intensity: f64)

pub fn reflection_probe_set_ambient_mode( &mut self, probe: Rid, mode: ReflectionProbeAmbientMode )

pub fn reflection_probe_set_ambient_color(&mut self, probe: Rid, color: Color)

pub fn reflection_probe_set_ambient_energy(&mut self, probe: Rid, energy: f64)

pub fn reflection_probe_set_max_distance(&mut self, probe: Rid, distance: f64)

pub fn reflection_probe_set_size(&mut self, probe: Rid, size: Vector3)

pub fn reflection_probe_set_origin_offset(&mut self, probe: Rid, offset: Vector3)

pub fn reflection_probe_set_as_interior(&mut self, probe: Rid, enable: bool)

pub fn reflection_probe_set_enable_box_projection( &mut self, probe: Rid, enable: bool )

pub fn reflection_probe_set_enable_shadows(&mut self, probe: Rid, enable: bool)

pub fn reflection_probe_set_cull_mask(&mut self, probe: Rid, layers: i64)

pub fn reflection_probe_set_resolution(&mut self, probe: Rid, resolution: i64)

pub fn reflection_probe_set_mesh_lod_threshold(&mut self, probe: Rid, pixels: f64)

pub fn decal_create(&mut self) -> Rid

pub fn decal_set_size(&mut self, decal: Rid, size: Vector3)

pub fn decal_set_texture(&mut self, decal: Rid, type_: DecalTexture, texture: Rid)

pub fn decal_set_emission_energy(&mut self, decal: Rid, energy: f64)

pub fn decal_set_albedo_mix(&mut self, decal: Rid, albedo_mix: f64)

pub fn decal_set_modulate(&mut self, decal: Rid, color: Color)

pub fn decal_set_cull_mask(&mut self, decal: Rid, mask: i64)

pub fn decal_set_distance_fade( &mut self, decal: Rid, enabled: bool, begin: f64, length: f64 )

pub fn decal_set_fade(&mut self, decal: Rid, above: f64, below: f64)

pub fn decal_set_normal_fade(&mut self, decal: Rid, fade: f64)

pub fn decals_set_filter(&mut self, filter: DecalFilter)

pub fn gi_set_use_half_resolution(&mut self, half_resolution: bool)

pub fn voxel_gi_create(&mut self) -> Rid

pub fn voxel_gi_allocate_data( &mut self, voxel_gi: Rid, to_cell_xform: Transform3D, aabb: Aabb, octree_size: Vector3i, octree_cells: PackedByteArray, data_cells: PackedByteArray, distance_field: PackedByteArray, level_counts: PackedInt32Array )

pub fn voxel_gi_get_octree_size(&self, voxel_gi: Rid) -> Vector3i

pub fn voxel_gi_get_octree_cells(&self, voxel_gi: Rid) -> PackedByteArray

pub fn voxel_gi_get_data_cells(&self, voxel_gi: Rid) -> PackedByteArray

pub fn voxel_gi_get_distance_field(&self, voxel_gi: Rid) -> PackedByteArray

pub fn voxel_gi_get_level_counts(&self, voxel_gi: Rid) -> PackedInt32Array

pub fn voxel_gi_get_to_cell_xform(&self, voxel_gi: Rid) -> Transform3D

pub fn voxel_gi_set_dynamic_range(&mut self, voxel_gi: Rid, range: f64)

pub fn voxel_gi_set_propagation(&mut self, voxel_gi: Rid, amount: f64)

pub fn voxel_gi_set_energy(&mut self, voxel_gi: Rid, energy: f64)

pub fn voxel_gi_set_baked_exposure_normalization( &mut self, voxel_gi: Rid, baked_exposure: f64 )

pub fn voxel_gi_set_bias(&mut self, voxel_gi: Rid, bias: f64)

pub fn voxel_gi_set_normal_bias(&mut self, voxel_gi: Rid, bias: f64)

pub fn voxel_gi_set_interior(&mut self, voxel_gi: Rid, enable: bool)

pub fn voxel_gi_set_use_two_bounces(&mut self, voxel_gi: Rid, enable: bool)

pub fn voxel_gi_set_quality(&mut self, quality: VoxelGIQuality)

pub fn lightmap_create(&mut self) -> Rid

pub fn lightmap_set_textures(&mut self, lightmap: Rid, light: Rid, uses_sh: bool)

pub fn lightmap_set_probe_bounds(&mut self, lightmap: Rid, bounds: Aabb)

pub fn lightmap_set_probe_interior(&mut self, lightmap: Rid, interior: bool)

pub fn lightmap_set_probe_capture_data( &mut self, lightmap: Rid, points: PackedVector3Array, point_sh: PackedColorArray, tetrahedra: PackedInt32Array, bsp_tree: PackedInt32Array )

pub fn lightmap_get_probe_capture_points( &self, lightmap: Rid ) -> PackedVector3Array

pub fn lightmap_get_probe_capture_sh(&self, lightmap: Rid) -> PackedColorArray

pub fn lightmap_get_probe_capture_tetrahedra( &self, lightmap: Rid ) -> PackedInt32Array

pub fn lightmap_get_probe_capture_bsp_tree( &self, lightmap: Rid ) -> PackedInt32Array

pub fn lightmap_set_baked_exposure_normalization( &mut self, lightmap: Rid, baked_exposure: f64 )

pub fn lightmap_set_probe_capture_update_speed(&mut self, speed: f64)

pub fn particles_create(&mut self) -> Rid

pub fn particles_set_mode(&mut self, particles: Rid, mode: ParticlesMode)

pub fn particles_set_emitting(&mut self, particles: Rid, emitting: bool)

pub fn particles_get_emitting(&mut self, particles: Rid) -> bool

pub fn particles_set_amount(&mut self, particles: Rid, amount: i64)

pub fn particles_set_lifetime(&mut self, particles: Rid, lifetime: f64)

pub fn particles_set_one_shot(&mut self, particles: Rid, one_shot: bool)

pub fn particles_set_pre_process_time(&mut self, particles: Rid, time: f64)

pub fn particles_set_explosiveness_ratio(&mut self, particles: Rid, ratio: f64)

pub fn particles_set_randomness_ratio(&mut self, particles: Rid, ratio: f64)

pub fn particles_set_custom_aabb(&mut self, particles: Rid, aabb: Aabb)

pub fn particles_set_speed_scale(&mut self, particles: Rid, scale: f64)

pub fn particles_set_use_local_coordinates( &mut self, particles: Rid, enable: bool )

pub fn particles_set_process_material(&mut self, particles: Rid, material: Rid)

pub fn particles_set_fixed_fps(&mut self, particles: Rid, fps: i64)

pub fn particles_set_interpolate(&mut self, particles: Rid, enable: bool)

pub fn particles_set_fractional_delta(&mut self, particles: Rid, enable: bool)

pub fn particles_set_collision_base_size(&mut self, particles: Rid, size: f64)

pub fn particles_set_transform_align( &mut self, particles: Rid, align: ParticlesTransformAlign )

pub fn particles_set_trails( &mut self, particles: Rid, enable: bool, length_sec: f64 )

pub fn particles_set_trail_bind_poses( &mut self, particles: Rid, bind_poses: Array<Transform3D> )

pub fn particles_is_inactive(&mut self, particles: Rid) -> bool

pub fn particles_request_process(&mut self, particles: Rid)

pub fn particles_restart(&mut self, particles: Rid)

pub fn particles_set_subemitter( &mut self, particles: Rid, subemitter_particles: Rid )

pub fn particles_emit( &mut self, particles: Rid, transform: Transform3D, velocity: Vector3, color: Color, custom: Color, emit_flags: i64 )

pub fn particles_set_draw_order( &mut self, particles: Rid, order: ParticlesDrawOrder )

pub fn particles_set_draw_passes(&mut self, particles: Rid, count: i64)

pub fn particles_set_draw_pass_mesh( &mut self, particles: Rid, pass: i64, mesh: Rid )

pub fn particles_get_current_aabb(&mut self, particles: Rid) -> Aabb

pub fn particles_set_emission_transform( &mut self, particles: Rid, transform: Transform3D )

pub fn particles_collision_create(&mut self) -> Rid

pub fn particles_collision_set_collision_type( &mut self, particles_collision: Rid, type_: ParticlesCollisionType )

pub fn particles_collision_set_cull_mask( &mut self, particles_collision: Rid, mask: i64 )

pub fn particles_collision_set_sphere_radius( &mut self, particles_collision: Rid, radius: f64 )

pub fn particles_collision_set_box_extents( &mut self, particles_collision: Rid, extents: Vector3 )

pub fn particles_collision_set_attractor_strength( &mut self, particles_collision: Rid, setrngth: f64 )

pub fn particles_collision_set_attractor_directionality( &mut self, particles_collision: Rid, amount: f64 )

pub fn particles_collision_set_attractor_attenuation( &mut self, particles_collision: Rid, curve: f64 )

pub fn particles_collision_set_field_texture( &mut self, particles_collision: Rid, texture: Rid )

pub fn particles_collision_height_field_update( &mut self, particles_collision: Rid )

pub fn particles_collision_set_height_field_resolution( &mut self, particles_collision: Rid, resolution: ParticlesCollisionHeightfieldResolution )

pub fn fog_volume_create(&mut self) -> Rid

pub fn fog_volume_set_shape(&mut self, fog_volume: Rid, shape: FogVolumeShape)

pub fn fog_volume_set_size(&mut self, fog_volume: Rid, size: Vector3)

pub fn fog_volume_set_material(&mut self, fog_volume: Rid, material: Rid)

pub fn visibility_notifier_create(&mut self) -> Rid

pub fn visibility_notifier_set_aabb(&mut self, notifier: Rid, aabb: Aabb)

pub fn visibility_notifier_set_callbacks( &mut self, notifier: Rid, enter_callable: Callable, exit_callable: Callable )

pub fn occluder_create(&mut self) -> Rid

pub fn occluder_set_mesh( &mut self, occluder: Rid, vertices: PackedVector3Array, indices: PackedInt32Array )

pub fn camera_create(&mut self) -> Rid

pub fn camera_set_perspective( &mut self, camera: Rid, fovy_degrees: f64, z_near: f64, z_far: f64 )

pub fn camera_set_orthogonal( &mut self, camera: Rid, size: f64, z_near: f64, z_far: f64 )

pub fn camera_set_frustum( &mut self, camera: Rid, size: f64, offset: Vector2, z_near: f64, z_far: f64 )

pub fn camera_set_transform(&mut self, camera: Rid, transform: Transform3D)

pub fn camera_set_cull_mask(&mut self, camera: Rid, layers: i64)

pub fn camera_set_environment(&mut self, camera: Rid, env: Rid)

pub fn camera_set_camera_attributes(&mut self, camera: Rid, effects: Rid)

pub fn camera_set_use_vertical_aspect(&mut self, camera: Rid, enable: bool)

pub fn viewport_create(&mut self) -> Rid

pub fn viewport_set_use_xr(&mut self, viewport: Rid, use_xr: bool)

pub fn viewport_set_size(&mut self, viewport: Rid, width: i64, height: i64)

pub fn viewport_set_active(&mut self, viewport: Rid, active: bool)

pub fn viewport_set_parent_viewport( &mut self, viewport: Rid, parent_viewport: Rid )

pub fn viewport_attach_to_screen( &mut self, viewport: Rid, rect: Rect2, screen: i64 )

pub fn viewport_set_render_direct_to_screen( &mut self, viewport: Rid, enabled: bool )

pub fn viewport_set_canvas_cull_mask( &mut self, viewport: Rid, canvas_cull_mask: i64 )

pub fn viewport_set_scaling_3d_mode( &mut self, viewport: Rid, scaling_3d_mode: ViewportScaling3DMode )

pub fn viewport_set_scaling_3d_scale(&mut self, viewport: Rid, scale: f64)

pub fn viewport_set_fsr_sharpness(&mut self, viewport: Rid, sharpness: f64)

pub fn viewport_set_texture_mipmap_bias( &mut self, viewport: Rid, mipmap_bias: f64 )

pub fn viewport_set_update_mode( &mut self, viewport: Rid, update_mode: ViewportUpdateMode )

pub fn viewport_set_clear_mode( &mut self, viewport: Rid, clear_mode: ViewportClearMode )

pub fn viewport_get_texture(&self, viewport: Rid) -> Rid

pub fn viewport_set_disable_3d(&mut self, viewport: Rid, disable: bool)

pub fn viewport_set_disable_2d(&mut self, viewport: Rid, disable: bool)

pub fn viewport_set_environment_mode( &mut self, viewport: Rid, mode: ViewportEnvironmentMode )

pub fn viewport_attach_camera(&mut self, viewport: Rid, camera: Rid)

pub fn viewport_set_scenario(&mut self, viewport: Rid, scenario: Rid)

pub fn viewport_attach_canvas(&mut self, viewport: Rid, canvas: Rid)

pub fn viewport_remove_canvas(&mut self, viewport: Rid, canvas: Rid)

pub fn viewport_set_snap_2d_transforms_to_pixel( &mut self, viewport: Rid, enabled: bool )

pub fn viewport_set_snap_2d_vertices_to_pixel( &mut self, viewport: Rid, enabled: bool )

pub fn viewport_set_default_canvas_item_texture_filter( &mut self, viewport: Rid, filter: CanvasItemTextureFilter )

pub fn viewport_set_default_canvas_item_texture_repeat( &mut self, viewport: Rid, repeat: CanvasItemTextureRepeat )

pub fn viewport_set_canvas_transform( &mut self, viewport: Rid, canvas: Rid, offset: Transform2D )

pub fn viewport_set_canvas_stacking( &mut self, viewport: Rid, canvas: Rid, layer: i64, sublayer: i64 )

pub fn viewport_set_transparent_background( &mut self, viewport: Rid, enabled: bool )

pub fn viewport_set_global_canvas_transform( &mut self, viewport: Rid, transform: Transform2D )

pub fn viewport_set_sdf_oversize_and_scale( &mut self, viewport: Rid, oversize: ViewportSDFOversize, scale: ViewportSDFScale )

pub fn viewport_set_positional_shadow_atlas_size( &mut self, viewport: Rid, size: i64, use_16_bits: bool )

pub fn viewport_set_positional_shadow_atlas_quadrant_subdivision( &mut self, viewport: Rid, quadrant: i64, subdivision: i64 )

pub fn viewport_set_msaa_3d(&mut self, viewport: Rid, msaa: ViewportMSAA)

pub fn viewport_set_msaa_2d(&mut self, viewport: Rid, msaa: ViewportMSAA)

pub fn viewport_set_screen_space_aa( &mut self, viewport: Rid, mode: ViewportScreenSpaceAA )

pub fn viewport_set_use_taa(&mut self, viewport: Rid, enable: bool)

pub fn viewport_set_use_debanding(&mut self, viewport: Rid, enable: bool)

pub fn viewport_set_use_occlusion_culling(&mut self, viewport: Rid, enable: bool)

pub fn viewport_set_occlusion_rays_per_thread(&mut self, rays_per_thread: i64)

pub fn viewport_set_occlusion_culling_build_quality( &mut self, quality: ViewportOcclusionCullingBuildQuality )

pub fn viewport_get_render_info( &mut self, viewport: Rid, type_: ViewportRenderInfoType, info: ViewportRenderInfo ) -> i64

pub fn viewport_set_debug_draw(&mut self, viewport: Rid, draw: ViewportDebugDraw)

pub fn viewport_set_measure_render_time(&mut self, viewport: Rid, enable: bool)

pub fn viewport_get_measured_render_time_cpu(&self, viewport: Rid) -> f64

pub fn viewport_get_measured_render_time_gpu(&self, viewport: Rid) -> f64

pub fn viewport_set_vrs_mode(&mut self, viewport: Rid, mode: ViewportVRSMode)

pub fn viewport_set_vrs_texture(&mut self, viewport: Rid, texture: Rid)

pub fn sky_create(&mut self) -> Rid

pub fn sky_set_radiance_size(&mut self, sky: Rid, radiance_size: i64)

pub fn sky_set_mode(&mut self, sky: Rid, mode: SkyMode)

pub fn sky_set_material(&mut self, sky: Rid, material: Rid)

pub fn sky_bake_panorama( &mut self, sky: Rid, energy: f64, bake_irradiance: bool, size: Vector2i ) -> Option<Gd<Image>>

pub fn environment_create(&mut self) -> Rid

pub fn environment_set_background(&mut self, env: Rid, bg: EnvironmentBG)

pub fn environment_set_sky(&mut self, env: Rid, sky: Rid)

pub fn environment_set_sky_custom_fov(&mut self, env: Rid, scale: f64)

pub fn environment_set_sky_orientation(&mut self, env: Rid, orientation: Basis)

pub fn environment_set_bg_color(&mut self, env: Rid, color: Color)

pub fn environment_set_bg_energy( &mut self, env: Rid, multiplier: f64, exposure_value: f64 )

pub fn environment_set_canvas_max_layer(&mut self, env: Rid, max_layer: i64)

pub fn environment_set_ambient_light( &mut self, env: Rid, color: Color, ambient: EnvironmentAmbientSource, energy: f64, sky_contibution: f64, reflection_source: EnvironmentReflectionSource )

pub fn environment_set_glow( &mut self, env: Rid, enable: bool, levels: PackedFloat32Array, intensity: f64, strength: f64, mix: f64, bloom_threshold: f64, blend_mode: EnvironmentGlowBlendMode, hdr_bleed_threshold: f64, hdr_bleed_scale: f64, hdr_luminance_cap: f64, glow_map_strength: f64, glow_map: Rid )

pub fn environment_set_tonemap( &mut self, env: Rid, tone_mapper: EnvironmentToneMapper, exposure: f64, white: f64 )

pub fn environment_set_adjustment( &mut self, env: Rid, enable: bool, brightness: f64, contrast: f64, saturation: f64, use_1d_color_correction: bool, color_correction: Rid )

pub fn environment_set_ssr( &mut self, env: Rid, enable: bool, max_steps: i64, fade_in: f64, fade_out: f64, depth_tolerance: f64 )

pub fn environment_set_ssao( &mut self, env: Rid, enable: bool, radius: f64, intensity: f64, power: f64, detail: f64, horizon: f64, sharpness: f64, light_affect: f64, ao_channel_affect: f64 )

pub fn environment_set_fog( &mut self, env: Rid, enable: bool, light_color: Color, light_energy: f64, sun_scatter: f64, density: f64, height: f64, height_density: f64, aerial_perspective: f64, sky_affect: f64 )

pub fn environment_set_sdfgi( &mut self, env: Rid, enable: bool, cascades: i64, min_cell_size: f64, y_scale: EnvironmentSDFGIYScale, use_occlusion: bool, bounce_feedback: f64, read_sky: bool, energy: f64, normal_bias: f64, probe_bias: f64 )

pub fn environment_set_volumetric_fog( &mut self, env: Rid, enable: bool, density: f64, albedo: Color, emission: Color, emission_energy: f64, anisotropy: f64, length: f64, p_detail_spread: f64, gi_inject: f64, temporal_reprojection: bool, temporal_reprojection_amount: f64, ambient_inject: f64, sky_affect: f64 )

pub fn environment_glow_set_use_bicubic_upscale(&mut self, enable: bool)

pub fn environment_set_ssr_roughness_quality( &mut self, quality: EnvironmentSSRRoughnessQuality )

pub fn environment_set_ssao_quality( &mut self, quality: EnvironmentSSAOQuality, half_size: bool, adaptive_target: f64, blur_passes: i64, fadeout_from: f64, fadeout_to: f64 )

pub fn environment_set_ssil_quality( &mut self, quality: EnvironmentSSILQuality, half_size: bool, adaptive_target: f64, blur_passes: i64, fadeout_from: f64, fadeout_to: f64 )

pub fn environment_set_sdfgi_ray_count( &mut self, ray_count: EnvironmentSDFGIRayCount )

pub fn environment_set_sdfgi_frames_to_converge( &mut self, frames: EnvironmentSDFGIFramesToConverge )

pub fn environment_set_sdfgi_frames_to_update_light( &mut self, frames: EnvironmentSDFGIFramesToUpdateLight )

pub fn environment_set_volumetric_fog_volume_size( &mut self, size: i64, depth: i64 )

pub fn environment_set_volumetric_fog_filter_active(&mut self, active: bool)

pub fn environment_bake_panorama( &mut self, environment: Rid, bake_irradiance: bool, size: Vector2i ) -> Option<Gd<Image>>

pub fn screen_space_roughness_limiter_set_active( &mut self, enable: bool, amount: f64, limit: f64 )

pub fn sub_surface_scattering_set_quality( &mut self, quality: SubSurfaceScatteringQuality )

pub fn sub_surface_scattering_set_scale(&mut self, scale: f64, depth_scale: f64)

pub fn camera_attributes_create(&mut self) -> Rid

pub fn camera_attributes_set_dof_blur_quality( &mut self, quality: DOFBlurQuality, use_jitter: bool )

pub fn camera_attributes_set_dof_blur_bokeh_shape( &mut self, shape: DOFBokehShape )

pub fn camera_attributes_set_dof_blur( &mut self, camera_attributes: Rid, far_enable: bool, far_distance: f64, far_transition: f64, near_enable: bool, near_distance: f64, near_transition: f64, amount: f64 )

pub fn camera_attributes_set_exposure( &mut self, camera_attributes: Rid, multiplier: f64, normalization: f64 )

pub fn camera_attributes_set_auto_exposure( &mut self, camera_attributes: Rid, enable: bool, min_sensitivity: f64, max_sensitivity: f64, speed: f64, scale: f64 )

pub fn scenario_create(&mut self) -> Rid

pub fn scenario_set_environment(&mut self, scenario: Rid, environment: Rid)

pub fn scenario_set_fallback_environment( &mut self, scenario: Rid, environment: Rid )

pub fn scenario_set_camera_attributes(&mut self, scenario: Rid, effects: Rid)

pub fn instance_create2(&mut self, base: Rid, scenario: Rid) -> Rid

pub fn instance_create(&mut self) -> Rid

pub fn instance_set_base(&mut self, instance: Rid, base: Rid)

pub fn instance_set_scenario(&mut self, instance: Rid, scenario: Rid)

pub fn instance_set_layer_mask(&mut self, instance: Rid, mask: i64)

pub fn instance_set_pivot_data( &mut self, instance: Rid, sorting_offset: f64, use_aabb_center: bool )

pub fn instance_set_transform(&mut self, instance: Rid, transform: Transform3D)

pub fn instance_attach_object_instance_id(&mut self, instance: Rid, id: i64)

pub fn instance_set_blend_shape_weight( &mut self, instance: Rid, shape: i64, weight: f64 )

pub fn instance_set_surface_override_material( &mut self, instance: Rid, surface: i64, material: Rid )

pub fn instance_set_visible(&mut self, instance: Rid, visible: bool)

pub fn instance_geometry_set_transparency( &mut self, instance: Rid, transparency: f64 )

pub fn instance_set_custom_aabb(&mut self, instance: Rid, aabb: Aabb)

pub fn instance_attach_skeleton(&mut self, instance: Rid, skeleton: Rid)

pub fn instance_set_extra_visibility_margin(&mut self, instance: Rid, margin: f64)

pub fn instance_set_visibility_parent(&mut self, instance: Rid, parent: Rid)

pub fn instance_set_ignore_culling(&mut self, instance: Rid, enabled: bool)

pub fn instance_geometry_set_flag( &mut self, instance: Rid, flag: InstanceFlags, enabled: bool )

pub fn instance_geometry_set_cast_shadows_setting( &mut self, instance: Rid, shadow_casting_setting: ShadowCastingSetting )

pub fn instance_geometry_set_material_override( &mut self, instance: Rid, material: Rid )

pub fn instance_geometry_set_material_overlay( &mut self, instance: Rid, material: Rid )

pub fn instance_geometry_set_visibility_range( &mut self, instance: Rid, min: f64, max: f64, min_margin: f64, max_margin: f64, fade_mode: VisibilityRangeFadeMode )

pub fn instance_geometry_set_lightmap( &mut self, instance: Rid, lightmap: Rid, lightmap_uv_scale: Rect2, lightmap_slice: i64 )

pub fn instance_geometry_set_lod_bias(&mut self, instance: Rid, lod_bias: f64)

pub fn instance_geometry_set_shader_parameter( &mut self, instance: Rid, parameter: StringName, value: Variant )

pub fn instance_geometry_get_shader_parameter( &self, instance: Rid, parameter: StringName ) -> Variant

pub fn instance_geometry_get_shader_parameter_default_value( &self, instance: Rid, parameter: StringName ) -> Variant

pub fn instance_geometry_get_shader_parameter_list( &self, instance: Rid ) -> Array<Dictionary>

pub fn instances_cull_aabb(&self, aabb: Aabb, scenario: Rid) -> PackedInt64Array

pub fn instances_cull_ray( &self, from: Vector3, to: Vector3, scenario: Rid ) -> PackedInt64Array

pub fn instances_cull_convex( &self, convex: Array<Plane>, scenario: Rid ) -> PackedInt64Array

pub fn bake_render_uv2( &mut self, base: Rid, material_overrides: Array<Rid>, image_size: Vector2i ) -> Array<Gd<Image>>

pub fn canvas_create(&mut self) -> Rid

pub fn canvas_set_item_mirroring( &mut self, canvas: Rid, item: Rid, mirroring: Vector2 )

pub fn canvas_set_modulate(&mut self, canvas: Rid, color: Color)

pub fn canvas_set_disable_scale(&mut self, disable: bool)

pub fn canvas_texture_create(&mut self) -> Rid

pub fn canvas_texture_set_channel( &mut self, canvas_texture: Rid, channel: CanvasTextureChannel, texture: Rid )

pub fn canvas_texture_set_shading_parameters( &mut self, canvas_texture: Rid, base_color: Color, shininess: f64 )

pub fn canvas_texture_set_texture_filter( &mut self, canvas_texture: Rid, filter: CanvasItemTextureFilter )

pub fn canvas_texture_set_texture_repeat( &mut self, canvas_texture: Rid, repeat: CanvasItemTextureRepeat )

pub fn canvas_item_create(&mut self) -> Rid

pub fn canvas_item_set_parent(&mut self, item: Rid, parent: Rid)

pub fn canvas_item_set_default_texture_filter( &mut self, item: Rid, filter: CanvasItemTextureFilter )

pub fn canvas_item_set_default_texture_repeat( &mut self, item: Rid, repeat: CanvasItemTextureRepeat )

pub fn canvas_item_set_visible(&mut self, item: Rid, visible: bool)

pub fn canvas_item_set_light_mask(&mut self, item: Rid, mask: i64)

pub fn canvas_item_set_visibility_layer( &mut self, item: Rid, visibility_layer: i64 )

pub fn canvas_item_set_transform(&mut self, item: Rid, transform: Transform2D)

pub fn canvas_item_set_clip(&mut self, item: Rid, clip: bool)

pub fn canvas_item_set_distance_field_mode(&mut self, item: Rid, enabled: bool)

pub fn canvas_item_set_custom_rect( &mut self, item: Rid, use_custom_rect: bool, rect: Rect2 )

pub fn canvas_item_set_modulate(&mut self, item: Rid, color: Color)

pub fn canvas_item_set_self_modulate(&mut self, item: Rid, color: Color)

pub fn canvas_item_set_draw_behind_parent(&mut self, item: Rid, enabled: bool)

pub fn canvas_item_add_line( &mut self, item: Rid, from: Vector2, to: Vector2, color: Color, width: f64, antialiased: bool )

pub fn canvas_item_add_polyline( &mut self, item: Rid, points: PackedVector2Array, colors: PackedColorArray, width: f64, antialiased: bool )

pub fn canvas_item_add_rect(&mut self, item: Rid, rect: Rect2, color: Color)

pub fn canvas_item_add_circle( &mut self, item: Rid, pos: Vector2, radius: f64, color: Color )

pub fn canvas_item_add_texture_rect( &mut self, item: Rid, rect: Rect2, texture: Rid, tile: bool, modulate: Color, transpose: bool )

pub fn canvas_item_add_msdf_texture_rect_region( &mut self, item: Rid, rect: Rect2, texture: Rid, src_rect: Rect2, modulate: Color, outline_size: i64, px_range: f64, scale: f64 )

pub fn canvas_item_add_lcd_texture_rect_region( &mut self, item: Rid, rect: Rect2, texture: Rid, src_rect: Rect2, modulate: Color )

pub fn canvas_item_add_texture_rect_region( &mut self, item: Rid, rect: Rect2, texture: Rid, src_rect: Rect2, modulate: Color, transpose: bool, clip_uv: bool )

pub fn canvas_item_add_nine_patch( &mut self, item: Rid, rect: Rect2, source: Rect2, texture: Rid, topleft: Vector2, bottomright: Vector2, x_axis_mode: NinePatchAxisMode, y_axis_mode: NinePatchAxisMode, draw_center: bool, modulate: Color )

pub fn canvas_item_add_primitive( &mut self, item: Rid, points: PackedVector2Array, colors: PackedColorArray, uvs: PackedVector2Array, texture: Rid )

pub fn canvas_item_add_polygon( &mut self, item: Rid, points: PackedVector2Array, colors: PackedColorArray, uvs: PackedVector2Array, texture: Rid )

pub fn canvas_item_add_triangle_array( &mut self, item: Rid, indices: PackedInt32Array, points: PackedVector2Array, colors: PackedColorArray, uvs: PackedVector2Array, bones: PackedInt32Array, weights: PackedFloat32Array, texture: Rid, count: i64 )

pub fn canvas_item_add_mesh( &mut self, item: Rid, mesh: Rid, transform: Transform2D, modulate: Color, texture: Rid )

pub fn canvas_item_add_multimesh(&mut self, item: Rid, mesh: Rid, texture: Rid)

pub fn canvas_item_add_particles( &mut self, item: Rid, particles: Rid, texture: Rid )

pub fn canvas_item_add_set_transform(&mut self, item: Rid, transform: Transform2D)

pub fn canvas_item_add_clip_ignore(&mut self, item: Rid, ignore: bool)

pub fn canvas_item_add_animation_slice( &mut self, item: Rid, animation_length: f64, slice_begin: f64, slice_end: f64, offset: f64 )

pub fn canvas_item_set_sort_children_by_y(&mut self, item: Rid, enabled: bool)

pub fn canvas_item_set_z_index(&mut self, item: Rid, z_index: i64)

pub fn canvas_item_set_z_as_relative_to_parent( &mut self, item: Rid, enabled: bool )

pub fn canvas_item_set_copy_to_backbuffer( &mut self, item: Rid, enabled: bool, rect: Rect2 )

pub fn canvas_item_clear(&mut self, item: Rid)

pub fn canvas_item_set_draw_index(&mut self, item: Rid, index: i64)

pub fn canvas_item_set_material(&mut self, item: Rid, material: Rid)

pub fn canvas_item_set_use_parent_material(&mut self, item: Rid, enabled: bool)

pub fn canvas_item_set_visibility_notifier( &mut self, item: Rid, enable: bool, area: Rect2, enter_callable: Callable, exit_callable: Callable )

pub fn canvas_item_set_canvas_group_mode( &mut self, item: Rid, mode: CanvasGroupMode, clear_margin: f64, fit_empty: bool, fit_margin: f64, blur_mipmaps: bool )

pub fn canvas_light_create(&mut self) -> Rid

pub fn canvas_light_attach_to_canvas(&mut self, light: Rid, canvas: Rid)

pub fn canvas_light_set_enabled(&mut self, light: Rid, enabled: bool)

pub fn canvas_light_set_texture_scale(&mut self, light: Rid, scale: f64)

pub fn canvas_light_set_transform(&mut self, light: Rid, transform: Transform2D)

pub fn canvas_light_set_texture(&mut self, light: Rid, texture: Rid)

pub fn canvas_light_set_texture_offset(&mut self, light: Rid, offset: Vector2)

pub fn canvas_light_set_color(&mut self, light: Rid, color: Color)

pub fn canvas_light_set_height(&mut self, light: Rid, height: f64)

pub fn canvas_light_set_energy(&mut self, light: Rid, energy: f64)

pub fn canvas_light_set_z_range(&mut self, light: Rid, min_z: i64, max_z: i64)

pub fn canvas_light_set_layer_range( &mut self, light: Rid, min_layer: i64, max_layer: i64 )

pub fn canvas_light_set_item_cull_mask(&mut self, light: Rid, mask: i64)

pub fn canvas_light_set_item_shadow_cull_mask(&mut self, light: Rid, mask: i64)

pub fn canvas_light_set_mode(&mut self, light: Rid, mode: CanvasLightMode)

pub fn canvas_light_set_shadow_enabled(&mut self, light: Rid, enabled: bool)

pub fn canvas_light_set_shadow_filter( &mut self, light: Rid, filter: CanvasLightShadowFilter )

pub fn canvas_light_set_shadow_color(&mut self, light: Rid, color: Color)

pub fn canvas_light_set_shadow_smooth(&mut self, light: Rid, smooth: f64)

pub fn canvas_light_set_blend_mode( &mut self, light: Rid, mode: CanvasLightBlendMode )

pub fn canvas_light_occluder_create(&mut self) -> Rid

pub fn canvas_light_occluder_attach_to_canvas( &mut self, occluder: Rid, canvas: Rid )

pub fn canvas_light_occluder_set_enabled(&mut self, occluder: Rid, enabled: bool)

pub fn canvas_light_occluder_set_polygon(&mut self, occluder: Rid, polygon: Rid)

pub fn canvas_light_occluder_set_as_sdf_collision( &mut self, occluder: Rid, enable: bool )

pub fn canvas_light_occluder_set_transform( &mut self, occluder: Rid, transform: Transform2D )

pub fn canvas_light_occluder_set_light_mask(&mut self, occluder: Rid, mask: i64)

pub fn canvas_occluder_polygon_create(&mut self) -> Rid

pub fn canvas_occluder_polygon_set_shape( &mut self, occluder_polygon: Rid, shape: PackedVector2Array, closed: bool )

pub fn canvas_occluder_polygon_set_cull_mode( &mut self, occluder_polygon: Rid, mode: CanvasOccluderPolygonCullMode )

pub fn canvas_set_shadow_texture_size(&mut self, size: i64)

pub fn global_shader_parameter_add( &mut self, name: StringName, type_: GlobalShaderParameterType, default_value: Variant )

pub fn global_shader_parameter_remove(&mut self, name: StringName)

pub fn global_shader_parameter_get_list(&self) -> PackedStringArray

pub fn global_shader_parameter_set(&mut self, name: StringName, value: Variant)

pub fn global_shader_parameter_set_override( &mut self, name: StringName, value: Variant )

pub fn global_shader_parameter_get(&self, name: StringName) -> Variant

pub fn global_shader_parameter_get_type( &self, name: StringName ) -> GlobalShaderParameterType

pub fn free_rid(&mut self, rid: Rid)

pub fn request_frame_drawn_callback(&mut self, callable: Callable)

pub fn has_changed(&self) -> bool

pub fn get_rendering_info(&mut self, info: RenderingInfo) -> i64

pub fn get_video_adapter_name(&self) -> GodotString

pub fn get_video_adapter_vendor(&self) -> GodotString

pub fn get_video_adapter_type(&self) -> DeviceType

pub fn get_video_adapter_api_version(&self) -> GodotString

pub fn make_sphere_mesh( &mut self, latitudes: i64, longitudes: i64, radius: f64 ) -> Rid

pub fn get_test_cube(&mut self) -> Rid

pub fn get_test_texture(&mut self) -> Rid

pub fn get_white_texture(&mut self) -> Rid

pub fn set_boot_image( &mut self, image: Gd<Image>, color: Color, scale: bool, use_filter: bool )

pub fn get_default_clear_color(&mut self) -> Color

pub fn set_default_clear_color(&mut self, color: Color)

pub fn has_feature(&self, feature: Features) -> bool

pub fn has_os_feature(&self, feature: GodotString) -> bool

pub fn set_debug_generate_wireframes(&mut self, generate: bool)

pub fn is_render_loop_enabled(&self) -> bool

pub fn set_render_loop_enabled(&mut self, enabled: bool)

pub fn get_frame_setup_time_cpu(&self) -> f64

pub fn force_sync(&mut self)

pub fn force_draw(&mut self, swap_buffers: bool, frame_step: f64)

pub fn get_rendering_device(&self) -> Option<Gd<RenderingDevice>>

pub fn create_local_rendering_device(&self) -> Option<Gd<RenderingDevice>>

pub const NO_INDEX_ARRAY: i32 = -1i32

pub const ARRAY_WEIGHTS_SIZE: i32 = 4i32

pub const CANVAS_ITEM_Z_MIN: i32 = -4_096i32

pub const CANVAS_ITEM_Z_MAX: i32 = 4_096i32

pub const MAX_GLOW_LEVELS: i32 = 7i32

pub const MAX_CURSORS: i32 = 8i32

pub const MAX_2D_DIRECTIONAL_LIGHTS: i32 = 8i32

pub const MATERIAL_RENDER_PRIORITY_MIN: i32 = -128i32

pub const MATERIAL_RENDER_PRIORITY_MAX: i32 = 127i32

pub const ARRAY_CUSTOM_COUNT: i32 = 4i32

pub const PARTICLES_EMIT_FLAG_POSITION: i32 = 1i32

pub const PARTICLES_EMIT_FLAG_ROTATION_SCALE: i32 = 2i32

pub const PARTICLES_EMIT_FLAG_VELOCITY: i32 = 4i32

pub const PARTICLES_EMIT_FLAG_COLOR: i32 = 8i32

pub const PARTICLES_EMIT_FLAG_CUSTOM: i32 = 16i32