Module godot::init

Entry point and global init/shutdown of the library.

Structs

GdextBuild

Provides meta-information about the library and the Godot version in use.

Enums

EditorRunBehavior

Determines if and how an extension’s code is run in the editor.

Traits

ExtensionLibrary

Defines the entry point for a GDExtension Rust library.

Type Aliases

InitLevel

Stage of the Godot initialization process.

Attribute Macros

gdextension

Proc-macro attribute to be used in combination with the ExtensionLibrary trait.