godot

Module meta

Meta-information about Godot types, their properties and conversions between them.

Modules

conv

Advanced conversion machinery.

error

Custom error types.

inspect

Runtime type information for Godot arrays and dictionaries.

shape

Static type descriptors for Rust types towards Godot.

Structs

ClassId

Globally unique ID of a class registered with Godot.

Traits

AsArg

Implicit conversions for arguments passed to Godot APIs.

Element

Marker trait to identify types that can be stored in Array<T> and Dictionary<K, V>.

FromGodot

Defines the canonical conversion from Godot for a type.

GodotConvert

Indicates that a type can be passed to/from Godot, either directly or through an intermediate “via” type.

GodotImmutable

Implemented for types that can be used as immutable default parameters in #[func] methods.

GodotType

Type that is directly representable in the engine.

PackedElement

Marker trait to identify types that can be stored in PackedArray<T>.

SignedRange

Trait supporting regular usize ranges, as well as negative indices.

ToGodot

Defines the canonical conversion to Godot for a type.

Functions

owned_into_arg

Generic abstraction over T owned values that should be passed as AsArg<T>.

ref_to_arg

Generic abstraction over &T references that should be passed as AsArg<T>.

wrapped

Accepts negative bounds, interpreted relative to the end of the collection.

Derive Macros

GodotConvert

Derive macro for GodotConvert on structs.