godot::meta

Trait InParamTuple

pub trait InParamTuple: ParamTuple {
    // Required methods
    unsafe fn from_varcall_args(
        args_ptr: *const *const __GdextVariant,
        call_ctx: &CallContext<'_>,
    ) -> Result<Self, CallError>;
    unsafe fn from_ptrcall_args(
        args_ptr: *const *const __GdextType,
        call_type: PtrcallType,
        call_ctx: &CallContext<'_>,
    ) -> Self;
    fn from_variant_array(array: &[&Variant]) -> Self;
}

Represents a parameter list that is received from some external location (usually Godot).

As an example, this would be used for user-defined functions that will be called from Godot, however this is not used when calling a Godot function from Rust code.

Required Methods

unsafe fn from_varcall_args( args_ptr: *const *const __GdextVariant, call_ctx: &CallContext<'_>, ) -> Result<Self, CallError>

Converts args_ptr to Self by first going through Variant.

Safety

args_ptr must be a pointer to an array of length Self::LEN

Each element of args_ptr must be reborrowable as a &Variant with a lifetime that lasts for the duration of the call.

unsafe fn from_ptrcall_args( args_ptr: *const *const __GdextType, call_type: PtrcallType, call_ctx: &CallContext<'_>, ) -> Self

Converts args_ptr to Self directly.

Safety

args_ptr must be a pointer to a valid array of length Self::LEN

each element of args_ptr must be of the same type as each element of Self

fn from_variant_array(array: &[&Variant]) -> Self

Converts array to Self by calling from_variant on each argument.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.