Module godot::obj

Types and traits related to objects.

The most important symbols in this module are:

GodotClass, which is implemented for every class that Godot can work with (either engine- or user-provided).

Gd, a smart pointer that manages instances of Godot classes.

Modules

cap

Capability traits, providing dedicated functionalities for Godot classes

dom

mem

Structs

Base

Restricted version of Gd, to hold the base instance inside a user’s GodotClass.

Gd

Smart pointer to objects owned by the Godot engine.

GdMut

Mutably/exclusively bound reference guard for a Gd smart pointer.

GdRef

Immutably/shared bound reference guard for a Gd smart pointer.

InstanceId

Represents a non-zero instance ID.

Traits

AsArg

EngineClass

Auto-implemented for all engine-provided classes

EngineEnum

Auto-implemented for all engine-provided enums

Export

Trait implemented for types that can be used as #[export] fields. This creates a copy of the value, for some type-specific definition of “copy”. For example, Array and Gd are returned via Share::share() instead of copying the actual data.

GodotClass

Makes T eligible to be managed by Godot and stored in Gd<T> pointers.

Inherits

Non-strict inheritance relationship in the Godot class hierarchy.

Share

Trait to create more references from a smart pointer or collection.