godot::obj
Enum NoBase

pub enum NoBase {}

Type representing the absence of a base class, at the root of the hierarchy.

NoBase is used as the base class for exactly one class: Object.

This is an enum without any variants, as we should never construct an instance of this class.