godot::obj
Module script

Functionality related to script instances (Rust code that can be attached as node scripts).

The features in this module are complemented by the ScriptExtension class and the IScriptExtension trait.

See ScriptInstance for usage.

Structs

ScriptBaseMut

Mutable/exclusive reference guard for a Base pointer.

ScriptBaseRef

Shared reference guard for a Base pointer.

SiMut

Mutable/exclusive reference guard for a T where T implements ScriptInstance.

Traits

ScriptInstance

Implement custom scripts that can be attached to objects in Godot.

Functions

create_script_instance⚠

Creates a new from a type that implements ScriptInstance.