godot::obj::script
Function script_instance_exists

pub fn script_instance_exists<O, S, L>(
    object: &Gd<O>,
    script: &Gd<S>,
    language: &Gd<L>,
) -> bool
where
    O: Inherits<Object>,
    S: Inherits<Script>,
    L: Inherits<ScriptLanguage>,

Available on since_api="4.2" only.

Checks if a script instance exists for a given object.

The engine does not expose the ScriptLanguage of a Script (as of 4.3) and it’s therefore necessary to pass the expected language as well.

Use this function to implement IScriptExtension::instance_has.