godot::obj::script

Function script_instance_exists

pub fn script_instance_exists<O, S>(object: &Gd<O>, script: &Gd<S>) -> bool
where
    O: Inherits<Object>,
    S: Inherits<Script, Declarer = DeclUser> + IScriptExtension + Bounds,

Available on since_api="4.2" only.

Checks if an instance of the script exists for a given object.

This function both checks if the passed script matches the one currently assigned to the passed object, as well as verifies that there is an instance for the script.

Use this function to implement IScriptExtension::instance_has.