Struct godot::obj::RawGd

#[repr(C)]
pub struct RawGd<T>where
    T: GodotClass,{ /* private fields */ }

Low-level bindings for object pointers in Godot.

This should not be used directly, you should either use Gd<T> or Option<Gd<T>> depending on whether you need a nullable object pointer or not.

Implementations

impl<T> RawGd<T>where T: GodotClass<Declarer = UserDomain>,

pub fn bind(&self) -> GdRef<'_, T>

Hands out a guard for a shared borrow, through which the user instance can be read.

See crate::obj::Gd::bind() for a more in depth explanation.

pub fn bind_mut(&mut self) -> GdMut<'_, T>

Hands out a guard for an exclusive borrow, through which the user instance can be read and written.

See crate::obj::Gd::bind_mut() for a more in depth explanation.

impl<T, M> RawGd<T>where T: GodotClass<Mem = M>, M: PossiblyManual + Memory,

This impl block contains no items.

The methods in this impl block are only available for objects T that are manually managed, i.e. anything that is not RefCounted or inherited from it.