Trait godot::obj::GodotClass

pub trait GodotClass: 'static + Sized {
    type Base: GodotClass;
    type Declarer: Domain;
    type Mem: Memory;

    const CLASS_NAME: &'static str;
}

Makes T eligible to be managed by Godot and stored in Gd<T> pointers.

The behavior of types implementing this trait is influenced by the associated types; check their documentation for information.

You wouldn’t usually implement this trait yourself; use the GodotClass derive macro instead.

Required Associated Types

type Base: GodotClass

The immediate superclass of T. This is always a Godot engine class.

type Declarer: Domain

Whether this class is a core Godot class provided by the engine, or declared by the user as a Rust struct.

type Mem: Memory

Defines the memory strategy.

Required Associated Constants

const CLASS_NAME: &'static str

The name of the class, under which it is registered in Godot.

This may deviate from the Rust struct name: HttpRequest::CLASS_NAME == "HTTPRequest".