godot::obj

Trait WithSignals

pub trait WithSignals: GodotClass + Inherits<Object> {
    type SignalCollection<'c, C: WithSignals>;
}

Implemented for all classes with registered signals, both engine- and user-declared.

This trait enables the Gd::signals() method.

User-defined classes with #[signal] additionally implement WithUserSignals.

Required Associated Types

type SignalCollection<'c, C: WithSignals>

The associated struct listing all signals of this class.

Parameters:

'c denotes the lifetime during which the class instance is borrowed and its signals can be modified.

C is the concrete class on which the signals are provided. This can be different than Self in case of derived classes (e.g. a user-defined node) connecting/emitting signals of a base class (e.g. Node).

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.