Module godot::prelude

Often-imported symbols.

Re-exports

pub use super::engine::NodeExt as _;

pub use super::obj::EngineEnum as _;

Modules

array

Specialized types related to arrays.

dictionary

Specialized types related to dictionaries.

meta

Meta-information about variant types, properties and class names.

real_consts

Basic mathematical constants.

utilities

Global utility functions.

Macros

array

Constructs Array literals, similar to Rust’s standard vec! macro.

dict

Constructs Dictionary literals, close to Godot’s own syntax.

godot_error

Pushes an error message to Godot’s built-in debugger and to the OS terminal.

godot_print

Prints to the Godot console.

godot_script_error

godot_warn

Pushes a warning message to Godot’s built-in debugger and to the OS terminal.

real

A macro to coerce float-literals into the real type. Mainly used where you’d normally use a suffix to specity the type, such as 115.0f32.

varray

Constructs VariantArray literals, similar to Rust’s standard vec! macro.

Structs

Aabb

Axis-aligned bounding box in 3D space.

Array

Godot’s Array type.

AudioStreamPlayer

Godot class AudioStreamPlayer.

Base

Restricted version of Gd, to hold the base instance inside a user’s GodotClass.

Basis

A 3x3 matrix, typically used as an orthogonal basis for Transform3D.

Callable

A Callable represents a function in Godot.

Camera2D

Godot class Camera2D.

Camera3D

Godot class Camera3D.

Color

Color built-in type, in floating-point RGBA format.

Dictionary

Godot’s Dictionary type.

Gd

Smart pointer to objects owned by the Godot engine.

GdMut

Mutably/exclusively bound reference guard for a Gd smart pointer.

GdRef

Immutably/shared bound reference guard for a Gd smart pointer.

GodotString

Godot’s reference counted string type.

InitHandle

Input

Godot class Input.

InstanceId

Represents a non-zero instance ID.

Node

Godot class Node.

Node2D

Godot class Node2D.

Node3D

Godot class Node3D.

NodePath

A pre-parsed scene tree path.

Object

Godot class Object.

PackedByteArray

Implements Godot’s $PackedArray type, which is an efficient array of $Elements.

PackedColorArray

Implements Godot’s $PackedArray type, which is an efficient array of $Elements.

PackedFloat32Array

Implements Godot’s $PackedArray type, which is an efficient array of $Elements.

PackedFloat64Array

Implements Godot’s $PackedArray type, which is an efficient array of $Elements.

PackedInt32Array

Implements Godot’s $PackedArray type, which is an efficient array of $Elements.

PackedInt64Array

Implements Godot’s $PackedArray type, which is an efficient array of $Elements.

PackedScene

Godot class PackedScene.

PackedStringArray

Implements Godot’s $PackedArray type, which is an efficient array of $Elements.

PackedVector2Array

Implements Godot’s $PackedArray type, which is an efficient array of $Elements.

PackedVector3Array

Implements Godot’s $PackedArray type, which is an efficient array of $Elements.

Plane

3D plane in Hessian normal form.

Projection

A 4x4 matrix used for 3D projective transformations. It can represent transformations such as translation, rotation, scaling, shearing, and perspective division. It consists of four Vector4 columns.

Quaternion

Rect2

2D axis-aligned bounding box.

Rect2i

2D axis-aligned integer bounding box.

RefCounted

Godot class RefCounted.

Resource

Godot class Resource.

SceneTree

Godot class SceneTree.

Signal

StringName

A string optimized for unique names.

Transform2D

Affine 2D transform (2x3 matrix).

Transform3D

Affine 3D transform (3x4 matrix).

Variant

VariantConversionError

Vector2

Vector used for 2D math using floating point coordinates.

Vector2i

Vector used for 2D math using integer coordinates.

Vector3

Vector used for 3D math using floating point coordinates.

Vector3i

Vector used for 3D math using integer coordinates.

Vector4

Vector used for 4D math using floating point coordinates.

Vector4i

Vector used for 4D math using integer coordinates.

Enums

ColorChannelOrder

EulerOrder

The ordering used to interpret a set of euler angles as extrinsic rotations.

InitLevel

ProjectionEye

The eye to create a projection for, when creating a projection adjusted for head-mounted displays.

ProjectionPlane

A projections clipping plane.

RectSide

The side of a Rect2 or Rect2i.

Rid

A RID (“resource ID”) is an opaque handle that refers to a Godot Resource.

VariantOperator

VariantType

Vector2Axis

Enumerates the axes in a Vector2.

Vector2iAxis

Enumerates the axes in a Vector2i.

Vector3Axis

Enumerates the axes in a Vector3.

Vector3iAxis

Enumerates the axes in a Vector3i.

Vector4Axis

Enumerates the axes in a Vector4.

Vector4iAxis

Enumerates the axes in a Vector4i.

Constants

CMP_EPSILON

Traits

AsAStar2D

Safety

AsAStar3D

Safety

AsAStarGrid2D

Safety

AsAcceptDialog

Safety

AsAesContext

Safety

AsAnimatableBody2D

Safety

AsAnimatableBody3D

Safety

AsAnimatedSprite2D

Safety

AsAnimatedSprite3D

Safety

AsAnimatedTexture

Safety

AsAnimation

Safety

AsAnimationLibrary

Safety

AsAnimationNode

Safety

AsAnimationNodeAdd2

Safety

AsAnimationNodeAdd3

Safety

AsAnimationNodeAnimation

Safety

AsAnimationNodeBlend2

Safety

AsAnimationNodeBlend3

Safety

AsAnimationNodeBlendSpace1D

Safety

AsAnimationNodeBlendSpace2D

Safety

AsAnimationNodeBlendTree

Safety

AsAnimationNodeOneShot

Safety

AsAnimationNodeOutput

Safety

AsAnimationNodeStateMachine

Safety

AsAnimationNodeStateMachinePlayback

Safety

AsAnimationNodeStateMachineTransition

Safety

AsAnimationNodeSync

Safety

AsAnimationNodeTimeScale

Safety

AsAnimationNodeTimeSeek

Safety

AsAnimationNodeTransition

Safety

AsAnimationPlayer

Safety

AsAnimationRootNode

Safety

AsAnimationTrackEditPlugin

Safety

AsAnimationTree

Safety

AsArea2D

Safety

AsArea3D

Safety

AsArrayMesh

Safety

AsArrayOccluder3D

Safety

AsAspectRatioContainer

Safety

AsAtlasTexture

Safety

AsAudioBusLayout

Safety

AsAudioEffect

Safety

AsAudioEffectAmplify

Safety

AsAudioEffectBandLimitFilter

Safety

AsAudioEffectBandPassFilter

Safety

AsAudioEffectCapture

Safety

AsAudioEffectChorus

Safety

AsAudioEffectCompressor

Safety

AsAudioEffectDelay

Safety

AsAudioEffectDistortion

Safety

AsAudioEffectEq

Safety

AsAudioEffectEq6

Safety

AsAudioEffectEq10

Safety

AsAudioEffectEq21

Safety

AsAudioEffectFilter

Safety

AsAudioEffectHighPassFilter

Safety

AsAudioEffectHighShelfFilter

Safety

AsAudioEffectInstance

Safety

AsAudioEffectLimiter

Safety

AsAudioEffectLowPassFilter

Safety

AsAudioEffectLowShelfFilter

Safety

AsAudioEffectNotchFilter

Safety

AsAudioEffectPanner

Safety

AsAudioEffectPhaser

Safety

AsAudioEffectPitchShift

Safety

AsAudioEffectRecord

Safety

AsAudioEffectReverb

Safety

AsAudioEffectSpectrumAnalyzer

Safety

AsAudioEffectSpectrumAnalyzerInstance

Safety

AsAudioEffectStereoEnhance

Safety

AsAudioListener2D

Safety

AsAudioListener3D

Safety

AsAudioServer

Safety

AsAudioStream

Safety

AsAudioStreamGenerator

Safety

AsAudioStreamGeneratorPlayback

Safety

AsAudioStreamMicrophone

Safety

AsAudioStreamMp3

Safety

AsAudioStreamOggVorbis

Safety

AsAudioStreamPlayback

Safety

AsAudioStreamPlaybackOggVorbis

Safety

AsAudioStreamPlaybackPolyphonic

Safety

AsAudioStreamPlaybackResampled

Safety

AsAudioStreamPlayer

Safety

AsAudioStreamPlayer2D

Safety

AsAudioStreamPlayer3D

Safety

AsAudioStreamPolyphonic

Safety

AsAudioStreamRandomizer

Safety

AsAudioStreamWav

Safety

AsBackBufferCopy

Safety

AsBaseButton

Safety

AsBaseMaterial3D

Safety

AsBitMap

Safety

AsBone2D

Safety

AsBoneAttachment3D

Safety

AsBoneMap

Safety

AsBoxContainer

Safety

AsBoxMesh

Safety

AsBoxOccluder3D

Safety

AsBoxShape3D

Safety

AsButton

Safety

AsButtonGroup

Safety

AsCallbackTweener

Safety

AsCamera2D

Safety

AsCamera3D

Safety

AsCameraAttributes

Safety

AsCameraAttributesPhysical

Safety

AsCameraAttributesPractical

Safety

AsCameraFeed

Safety

AsCameraServer

Safety

AsCameraTexture

Safety

AsCanvasGroup

Safety

AsCanvasItem

Safety

AsCanvasItemMaterial

Safety

AsCanvasLayer

Safety

AsCanvasModulate

Safety

AsCanvasTexture

Safety

AsCapsuleMesh

Safety

AsCapsuleShape2D

Safety

AsCapsuleShape3D

Safety

AsCenterContainer

Safety

AsCharFxTransform

Safety

AsCharacterBody2D

Safety

AsCharacterBody3D

Safety

AsCheckBox

Safety

AsCheckButton

Safety

AsCircleShape2D

Safety

AsClassDb

Safety

AsCodeEdit

Safety

AsCodeHighlighter

Safety

AsCollisionObject2D

Safety

AsCollisionObject3D

Safety

AsCollisionPolygon2D

Safety

AsCollisionPolygon3D

Safety

AsCollisionShape2D

Safety

AsCollisionShape3D

Safety

AsColorPicker

Safety

AsColorPickerButton

Safety

AsColorRect

Safety

AsCompressedCubemap

Safety

AsCompressedCubemapArray

Safety

AsCompressedTexture2D

Safety

AsCompressedTexture2DArray

Safety

AsCompressedTexture3D

Safety

AsCompressedTextureLayered

Safety

AsConcavePolygonShape2D

Safety

AsConcavePolygonShape3D

Safety

AsConeTwistJoint3D

Safety

AsConfigFile

Safety

AsConfirmationDialog

Safety

AsContainer

Safety

AsControl

Safety

AsConvexPolygonShape2D

Safety

AsConvexPolygonShape3D

Safety

AsCpuParticles2D

Safety

AsCpuParticles3D

Safety

AsCrypto

Safety

AsCryptoKey

Safety

AsCsgBox3D

Safety

AsCsgCombiner3D

Safety

AsCsgCylinder3D

Safety

AsCsgMesh3D

Safety

AsCsgPolygon3D

Safety

AsCsgPrimitive3D

Safety

AsCsgShape3D

Safety

AsCsgSphere3D

Safety

AsCsgTorus3D

Safety

AsCubemap

Safety

AsCubemapArray

Safety

AsCurve

Safety

AsCurve2D

Safety

AsCurve3D

Safety

AsCurveTexture

Safety

AsCurveXyzTexture

Safety

AsCylinderMesh

Safety

AsCylinderShape3D

Safety

AsDampedSpringJoint2D

Safety

AsDecal

Safety

AsDirAccess

Safety

AsDirectionalLight2D

Safety

AsDirectionalLight3D

Safety

AsDisplayServer

Safety

AsDtlsServer

Safety

AsENetConnection

Safety

AsENetMultiplayerPeer

Safety

AsENetPacketPeer

Safety

AsEditorCommandPalette

Safety

AsEditorDebuggerPlugin

Safety

AsEditorDebuggerSession

Safety

AsEditorExportPlatform

Safety

AsEditorExportPlugin

Safety

AsEditorFeatureProfile

Safety

AsEditorFileDialog

Safety

AsEditorFileSystem

Safety

AsEditorFileSystemDirectory

Safety

AsEditorFileSystemImportFormatSupportQuery

Safety

AsEditorImportPlugin

Safety

AsEditorInspector

Safety

AsEditorInspectorPlugin

Safety

AsEditorInterface

Safety

AsEditorNode3DGizmo

Safety

AsEditorNode3DGizmoPlugin

Safety

AsEditorPaths

Safety

AsEditorPlugin

Safety

AsEditorProperty

Safety

AsEditorResourceConversionPlugin

Safety

AsEditorResourcePicker

Safety

AsEditorResourcePreview

Safety

AsEditorResourcePreviewGenerator

Safety

AsEditorSceneFormatImporter

Safety

AsEditorSceneFormatImporterBlend

Safety

AsEditorSceneFormatImporterFbx

Safety

AsEditorSceneFormatImporterGltf

Safety

AsEditorScenePostImport

Safety

AsEditorScenePostImportPlugin

Safety

AsEditorScript

Safety

AsEditorScriptPicker

Safety

AsEditorSelection

Safety

AsEditorSettings

Safety

AsEditorSpinSlider

Safety

AsEditorSyntaxHighlighter

Safety

AsEditorTranslationParserPlugin

Safety

AsEditorUndoRedoManager

Safety

AsEditorVcsInterface

Safety

AsEncodedObjectAsId

Safety

AsEngine

Safety

AsEngineDebugger

Safety

AsEngineProfiler

Safety

AsEnvironment

Safety

AsExpression

Safety

AsFastNoiseLite

Safety

AsFileAccess

Safety

AsFileDialog

Safety

AsFileSystemDock

Safety

AsFlowContainer

Safety

AsFogMaterial

Safety

AsFogVolume

Safety

AsFont

Safety

AsFontFile

Safety

AsFontVariation

Safety

AsGDExtension

Safety

AsGDExtensionManager

Safety

AsGdScript

Safety

AsGdScriptEditorTranslationParserPlugin

Safety

AsGdScriptNativeClass

Safety

AsGeneric6DofJoint3D

Safety

AsGeometry2D

Safety

AsGeometry3D

Safety

AsGeometryInstance3D

Safety

AsGltfAccessor

Safety

AsGltfAnimation

Safety

AsGltfBufferView

Safety

AsGltfCamera

Safety

AsGltfDocument

Safety

AsGltfDocumentExtension

Safety

AsGltfDocumentExtensionConvertImporterMesh

Safety

AsGltfLight

Safety

AsGltfMesh

Safety

AsGltfNode

Safety

AsGltfSkeleton

Safety

AsGltfSkin

Safety

AsGltfSpecGloss

Safety

AsGltfState

Safety

AsGltfTexture

Safety

AsGltfTextureSampler

Safety

AsGodotPhysicsServer2D

Safety

AsGodotPhysicsServer3D

Safety

AsGpuParticles2D

Safety

AsGpuParticles3D

Safety

AsGpuParticlesAttractor3D

Safety

AsGpuParticlesAttractorBox3D

Safety

AsGpuParticlesAttractorSphere3D

Safety

AsGpuParticlesAttractorVectorField3D

Safety

AsGpuParticlesCollision3D

Safety

AsGpuParticlesCollisionBox3D

Safety

AsGpuParticlesCollisionHeightField3D

Safety

AsGpuParticlesCollisionSdf3d

Safety

AsGpuParticlesCollisionSphere3D

Safety

AsGradient

Safety

AsGradientTexture1D

Safety

AsGradientTexture2D

Safety

AsGraphEdit

Safety

AsGraphNode

Safety

AsGridContainer

Safety

AsGridMap

Safety

AsGrooveJoint2D

Safety

AsHBoxContainer

Safety

AsHFlowContainer

Safety

AsHScrollBar

Safety

AsHSeparator

Safety

AsHSlider

Safety

AsHSplitContainer

Safety

AsHashingContext

Safety

AsHeightMapShape3D

Safety

AsHingeJoint3D

Safety

AsHmacContext

Safety

AsHttpClient

Safety

AsHttpRequest

Safety

AsImage

Safety

AsImageFormatLoader

Safety

AsImageFormatLoaderExtension

Safety

AsImageTexture

Safety

AsImageTexture3D

Safety

AsImageTextureLayered

Safety

AsImmediateMesh

Safety

AsImporterMesh

Safety

AsImporterMeshInstance3D

Safety

AsInput

Safety

AsInputEvent

Safety

AsInputEventAction

Safety

AsInputEventFromWindow

Safety

AsInputEventGesture

Safety

AsInputEventJoypadButton

Safety

AsInputEventJoypadMotion

Safety

AsInputEventKey

Safety

AsInputEventMagnifyGesture

Safety

AsInputEventMidi

Safety

AsInputEventMouse

Safety

AsInputEventMouseButton

Safety

AsInputEventMouseMotion

Safety

AsInputEventPanGesture

Safety

AsInputEventScreenDrag

Safety

AsInputEventScreenTouch

Safety

AsInputEventShortcut

Safety

AsInputEventWithModifiers

Safety

AsInputMap

Safety

AsInstancePlaceholder

Safety

AsIntervalTweener

Safety

AsIp

Safety

AsIpUnix

Safety

AsItemList

Safety

AsJavaClass

Safety

AsJavaClassWrapper

Safety

AsJavaScriptBridge

Safety

AsJavaScriptObject

Safety

AsJniSingleton

Safety

AsJoint2D

Safety

AsJoint3D

Safety

AsJson

Safety

AsJsonRpc

Safety

AsKinematicCollision2D

Safety

AsKinematicCollision3D

Safety

AsLabel

Safety

AsLabel3D

Safety

AsLabelSettings

Safety

AsLight2D

Safety

AsLight3D

Safety

AsLightOccluder2D

Safety

AsLightmapGi

Safety

AsLightmapGiData

Safety

AsLightmapProbe

Safety

AsLightmapper

Safety

AsLightmapperRd

Safety

AsLine2D

Safety

AsLineEdit

Safety

AsLinkButton

Safety

AsMainLoop

Safety

AsMarginContainer

Safety

AsMarker2D

Safety

AsMarker3D

Safety

AsMarshalls

Safety

AsMaterial

Safety

AsMenuBar

Safety

AsMenuButton

Safety

AsMesh

Safety

AsMeshDataTool

Safety

AsMeshInstance2D

Safety

AsMeshInstance3D

Safety

AsMeshLibrary

Safety

AsMeshTexture

Safety

AsMethodTweener

Safety

AsMissingNode

Safety

AsMissingResource

Safety

AsMobileVrInterface

Safety

AsMovieWriter

Safety

AsMovieWriterMjpeg

Safety

AsMovieWriterPngwav

Safety

AsMultiMesh

Safety

AsMultiMeshInstance2D

Safety

AsMultiMeshInstance3D

Safety

AsMultiplayerApi

Safety

AsMultiplayerApiExtension

Safety

AsMultiplayerPeer

Safety

AsMultiplayerPeerExtension

Safety

AsMultiplayerSpawner

Safety

AsMultiplayerSynchronizer

Safety

AsNavigationAgent2D

Safety

AsNavigationAgent3D

Safety

AsNavigationLink2D

Safety

AsNavigationLink3D

Safety

AsNavigationMesh

Safety

AsNavigationMeshGenerator

Safety

AsNavigationObstacle2D

Safety

AsNavigationObstacle3D

Safety

AsNavigationPathQueryParameters2D

Safety

AsNavigationPathQueryParameters3D

Safety

AsNavigationPathQueryResult2D

Safety

AsNavigationPathQueryResult3D

Safety

AsNavigationPolygon

Safety

AsNavigationRegion2D

Safety

AsNavigationRegion3D

Safety

AsNavigationServer2D

Safety

AsNavigationServer3D

Safety

AsNinePatchRect

Safety

AsNode

Safety

AsNode2D

Safety

AsNode3D

Safety

AsNode3DGizmo

Safety

AsNoise

Safety

AsNoiseTexture2D

Safety

AsObject

Safety

AsOccluder3D

Safety

AsOccluderInstance3D

Safety

AsOccluderPolygon2D

Safety

AsOfflineMultiplayerPeer

Safety

AsOggPacketSequence

Safety

AsOggPacketSequencePlayback

Safety

AsOmniLight3D

Safety

AsOpenXrAction

Safety

AsOpenXrActionMap

Safety

AsOpenXrActionSet

Safety

AsOpenXrHand

Safety

AsOpenXrInteractionProfile

Safety

AsOpenXrInterface

Safety

AsOpenXripBinding

Safety

AsOptimizedTranslation

Safety

AsOptionButton

Safety

AsOrmMaterial3D

Safety

AsOs

Safety

AsPackedDataContainer

Safety

AsPackedDataContainerRef

Safety

AsPackedScene

Safety

AsPacketPeer

Safety

AsPacketPeerDtls

Safety

AsPacketPeerExtension

Safety

AsPacketPeerStream

Safety

AsPacketPeerUdp

Safety

AsPanel

Safety

AsPanelContainer

Safety

AsPanoramaSkyMaterial

Safety

AsParallaxBackground

Safety

AsParallaxLayer

Safety

AsParticleProcessMaterial

Safety

AsPath2D

Safety

AsPath3D

Safety

AsPathFollow2D

Safety

AsPathFollow3D

Safety

AsPckPacker

Safety

AsPerformance

Safety

AsPhysicalBone2D

Safety

AsPhysicalBone3D

Safety

AsPhysicalSkyMaterial

Safety

AsPhysicsBody2D

Safety

AsPhysicsBody3D

Safety

AsPhysicsDirectBodyState2D

Safety

AsPhysicsDirectBodyState2DExtension

Safety

AsPhysicsDirectBodyState3D

Safety

AsPhysicsDirectBodyState3DExtension

Safety

AsPhysicsDirectSpaceState2D

Safety

AsPhysicsDirectSpaceState2DExtension

Safety

AsPhysicsDirectSpaceState3D

Safety

AsPhysicsDirectSpaceState3DExtension

Safety

AsPhysicsMaterial

Safety

AsPhysicsPointQueryParameters2D

Safety

AsPhysicsPointQueryParameters3D

Safety

AsPhysicsRayQueryParameters2D

Safety

AsPhysicsRayQueryParameters3D

Safety

AsPhysicsServer2D

Safety

AsPhysicsServer2DExtension

Safety

AsPhysicsServer2DManager

Safety

AsPhysicsServer3D

Safety

AsPhysicsServer3DExtension

Safety

AsPhysicsServer3DManager

Safety

AsPhysicsServer3DRenderingServerHandler

Safety

AsPhysicsShapeQueryParameters2D

Safety

AsPhysicsShapeQueryParameters3D

Safety

AsPhysicsTestMotionParameters2D

Safety

AsPhysicsTestMotionParameters3D

Safety

AsPhysicsTestMotionResult2D

Safety

AsPhysicsTestMotionResult3D

Safety

AsPinJoint2D

Safety

AsPinJoint3D

Safety

AsPlaceholderCubemap

Safety

AsPlaceholderCubemapArray

Safety

AsPlaceholderMaterial

Safety

AsPlaceholderMesh

Safety

AsPlaceholderTexture2D

Safety

AsPlaceholderTexture2DArray

Safety

AsPlaceholderTexture3D

Safety

AsPlaceholderTextureLayered

Safety

AsPlaneMesh

Safety

AsPointLight2D

Safety

AsPointMesh

Safety

AsPolygon2D

Safety

AsPolygonOccluder3D

Safety

AsPolygonPathFinder

Safety

AsPopup

Safety

AsPopupMenu

Safety

AsPopupPanel

Safety

AsPortableCompressedTexture2D

Safety

AsPrimitiveMesh

Safety

AsPrismMesh

Safety

AsProceduralSkyMaterial

Safety

AsProgressBar

Safety

AsProjectSettings

Safety

AsPropertyTweener

Safety

AsQuadMesh

Safety

AsQuadOccluder3D

Safety

AsRandomNumberGenerator

Safety

AsRange

Safety

AsRayCast2D

Safety

AsRayCast3D

Safety

AsRdAttachmentFormat

Safety

AsRdFramebufferPass

Safety

AsRdPipelineColorBlendState

Safety

AsRdPipelineColorBlendStateAttachment

Safety

AsRdPipelineDepthStencilState

Safety

AsRdPipelineMultisampleState

Safety

AsRdPipelineRasterizationState

Safety

AsRdPipelineSpecializationConstant

Safety

AsRdSamplerState

Safety

AsRdShaderFile

Safety

AsRdShaderSource

Safety

AsRdShaderSpirv

Safety

AsRdTextureFormat

Safety

AsRdTextureView

Safety

AsRdUniform

Safety

AsRdVertexAttribute

Safety

AsRectangleShape2D

Safety

AsRefCounted

Safety

AsReferenceRect

Safety

AsReflectionProbe

Safety

AsRegEx

Safety

AsRegExMatch

Safety

AsRemoteTransform2D

Safety

AsRemoteTransform3D

Safety

AsRenderingDevice

Safety

AsRenderingServer

Safety

AsResource

Safety

AsResourceFormatImporterSaver

Safety

AsResourceFormatLoader

Safety

AsResourceFormatSaver

Safety

AsResourceImporter

Safety

AsResourceLoader

Safety

AsResourcePreloader

Safety

AsResourceSaver

Safety

AsResourceUid

Safety

AsRibbonTrailMesh

Safety

AsRichTextEffect

Safety

AsRichTextLabel

Safety

AsRigidBody2D

Safety

AsRigidBody3D

Safety

AsRootMotionView

Safety

AsSceneMultiplayer

Safety

AsSceneReplicationConfig

Safety

AsSceneState

Safety

AsSceneTree

Safety

AsSceneTreeTimer

Safety

AsScript

Safety

AsScriptCreateDialog

Safety

AsScriptEditor

Safety

AsScriptEditorBase

Safety

AsScriptExtension

Safety

AsScriptLanguage

Safety

AsScriptLanguageExtension

Safety

AsScrollBar

Safety

AsScrollContainer

Safety

AsSegmentShape2D

Safety

AsSeparationRayShape2D

Safety

AsSeparationRayShape3D

Safety

AsSeparator

Safety

AsShader

Safety

AsShaderGlobalsOverride

Safety

AsShaderInclude

Safety

AsShaderMaterial

Safety

AsShape2D

Safety

AsShape3D

Safety

AsShapeCast2D

Safety

AsShapeCast3D

Safety

AsShortcut

Safety

AsSkeleton2D

Safety

AsSkeleton3D

Safety

AsSkeletonIk3d

Safety

AsSkeletonModification2D

Safety

AsSkeletonModification2DJiggle

Safety

AsSkeletonModification2DLookAt

Safety

AsSkeletonModification2DPhysicalBones

Safety

AsSkeletonModification2DStackHolder

Safety

AsSkeletonModification2DTwoBoneIk

Safety

AsSkeletonModification2Dccdik

Safety

AsSkeletonModification2Dfabrik

Safety

AsSkeletonModificationStack2D

Safety

AsSkeletonProfile

Safety

AsSkeletonProfileHumanoid

Safety

AsSkin

Safety

AsSkinReference

Safety

AsSky

Safety

AsSlider

Safety

AsSliderJoint3D

Safety

AsSoftBody3D

Safety

AsSphereMesh

Safety

AsSphereOccluder3D

Safety

AsSphereShape3D

Safety

AsSpinBox

Safety

AsSplitContainer

Safety

AsSpotLight3D

Safety

AsSpringArm3D

Safety

AsSprite2D

Safety

AsSprite3D

Safety

AsSpriteBase3D

Safety

AsSpriteFrames

Safety

AsStandardMaterial3D

Safety

AsStaticBody2D

Safety

AsStaticBody3D

Safety

AsStreamPeer

Safety

AsStreamPeerBuffer

Safety

AsStreamPeerExtension

Safety

AsStreamPeerGzip

Safety

AsStreamPeerTcp

Safety

AsStreamPeerTls

Safety

AsStyleBox

Safety

AsStyleBoxEmpty

Safety

AsStyleBoxFlat

Safety

AsStyleBoxLine

Safety

AsStyleBoxTexture

Safety

AsSubViewport

Safety

AsSubViewportContainer

Safety

AsSurfaceTool

Safety

AsSyntaxHighlighter

Safety

AsSystemFont

Safety

AsTabBar

Safety

AsTabContainer

Safety

AsTcpServer

Safety

AsTextEdit

Safety

AsTextLine

Safety

AsTextMesh

Safety

AsTextParagraph

Safety

AsTextServer

Safety

AsTextServerAdvanced

Safety

AsTextServerDummy

Safety

AsTextServerExtension

Safety

AsTextServerManager

Safety

AsTexture

Safety

AsTexture2D

Safety

AsTexture2DArray

Safety

AsTexture3D

Safety

AsTextureButton

Safety

AsTextureLayered

Safety

AsTextureProgressBar

Safety

AsTextureRect

Safety

AsTheme

Safety

AsThemeDb

Safety

AsTileData

Safety

AsTileMap

Safety

AsTileMapPattern

Safety

AsTileSet

Safety

AsTileSetAtlasSource

Safety

AsTileSetScenesCollectionSource

Safety

AsTileSetSource

Safety

AsTime

Safety

AsTimer

Safety

AsTlsOptions

Safety

AsTorusMesh

Safety

AsTouchScreenButton

Safety

AsTranslation

Safety

AsTranslationServer

Safety

AsTree

Safety

AsTreeItem

Safety

AsTriangleMesh

Safety

AsTubeTrailMesh

Safety

AsTween

Safety

AsTweener

Safety

AsUdpServer

Safety

AsUndoRedo

Safety

AsUpnp

Safety

AsUpnpDevice

Safety

AsVBoxContainer

Safety

AsVFlowContainer

Safety

AsVScrollBar

Safety

AsVSeparator

Safety

AsVSlider

Safety

AsVSplitContainer

Safety

AsVehicleBody3D

Safety

AsVehicleWheel3D

Safety

AsVideoStream

Safety

AsVideoStreamPlayback

Safety

AsVideoStreamPlayer

Safety

AsVideoStreamTheora

Safety

AsViewport

Safety

AsViewportTexture

Safety

AsVisibleOnScreenEnabler2D

Safety

AsVisibleOnScreenEnabler3D

Safety

AsVisibleOnScreenNotifier2D

Safety

AsVisibleOnScreenNotifier3D

Safety

AsVisualInstance3D

Safety

AsVisualShader

Safety

AsVisualShaderNode

Safety

AsVisualShaderNodeBillboard

Safety

AsVisualShaderNodeBooleanConstant

Safety

AsVisualShaderNodeBooleanParameter

Safety

AsVisualShaderNodeClamp

Safety

AsVisualShaderNodeColorConstant

Safety

AsVisualShaderNodeColorFunc

Safety

AsVisualShaderNodeColorOp

Safety

AsVisualShaderNodeColorParameter

Safety

AsVisualShaderNodeComment

Safety

AsVisualShaderNodeCompare

Safety

AsVisualShaderNodeConstant

Safety

AsVisualShaderNodeCubemap

Safety

AsVisualShaderNodeCubemapParameter

Safety

AsVisualShaderNodeCurveTexture

Safety

AsVisualShaderNodeCurveXyzTexture

Safety

AsVisualShaderNodeCustom

Safety

AsVisualShaderNodeDerivativeFunc

Safety

AsVisualShaderNodeDeterminant

Safety

AsVisualShaderNodeDistanceFade

Safety

AsVisualShaderNodeDotProduct

Safety

AsVisualShaderNodeExpression

Safety

AsVisualShaderNodeFaceForward

Safety

AsVisualShaderNodeFloatConstant

Safety

AsVisualShaderNodeFloatFunc

Safety

AsVisualShaderNodeFloatOp

Safety

AsVisualShaderNodeFloatParameter

Safety

AsVisualShaderNodeFresnel

Safety

AsVisualShaderNodeGlobalExpression

Safety

AsVisualShaderNodeGroupBase

Safety

AsVisualShaderNodeIf

Safety

AsVisualShaderNodeInput

Safety

AsVisualShaderNodeIntConstant

Safety

AsVisualShaderNodeIntFunc

Safety

AsVisualShaderNodeIntOp

Safety

AsVisualShaderNodeIntParameter

Safety

AsVisualShaderNodeIs

Safety

AsVisualShaderNodeLinearSceneDepth

Safety

AsVisualShaderNodeMix

Safety

AsVisualShaderNodeMultiplyAdd

Safety

AsVisualShaderNodeOuterProduct

Safety

AsVisualShaderNodeOutput

Safety

AsVisualShaderNodeParameter

Safety

AsVisualShaderNodeParameterRef

Safety

AsVisualShaderNodeParticleAccelerator

Safety

AsVisualShaderNodeParticleBoxEmitter

Safety

AsVisualShaderNodeParticleConeVelocity

Safety

AsVisualShaderNodeParticleEmit

Safety

AsVisualShaderNodeParticleEmitter

Safety

AsVisualShaderNodeParticleMeshEmitter

Safety

AsVisualShaderNodeParticleMultiplyByAxisAngle

Safety

AsVisualShaderNodeParticleOutput

Safety

AsVisualShaderNodeParticleRandomness

Safety

AsVisualShaderNodeParticleRingEmitter

Safety

AsVisualShaderNodeParticleSphereEmitter

Safety

AsVisualShaderNodeProximityFade

Safety

AsVisualShaderNodeRandomRange

Safety

AsVisualShaderNodeRemap

Safety

AsVisualShaderNodeResizableBase

Safety

AsVisualShaderNodeSample3D

Safety

AsVisualShaderNodeScreenUvToSdf

Safety

AsVisualShaderNodeSdfRaymarch

Safety

AsVisualShaderNodeSdfToScreenUv

Safety

AsVisualShaderNodeSmoothStep

Safety

AsVisualShaderNodeStep

Safety

AsVisualShaderNodeSwitch

Safety

AsVisualShaderNodeTexture

Safety

AsVisualShaderNodeTexture2DArray

Safety

AsVisualShaderNodeTexture2DArrayParameter

Safety

AsVisualShaderNodeTexture2DParameter

Safety

AsVisualShaderNodeTexture3D

Safety

AsVisualShaderNodeTexture3DParameter

Safety

AsVisualShaderNodeTextureParameter

Safety

AsVisualShaderNodeTextureParameterTriplanar

Safety

AsVisualShaderNodeTextureSdf

Safety

AsVisualShaderNodeTextureSdfNormal

Safety

AsVisualShaderNodeTransformCompose

Safety

AsVisualShaderNodeTransformConstant

Safety

AsVisualShaderNodeTransformDecompose

Safety

AsVisualShaderNodeTransformFunc

Safety

AsVisualShaderNodeTransformOp

Safety

AsVisualShaderNodeTransformParameter

Safety

AsVisualShaderNodeTransformVecMult

Safety

AsVisualShaderNodeUIntConstant

Safety

AsVisualShaderNodeUIntFunc

Safety

AsVisualShaderNodeUIntOp

Safety

AsVisualShaderNodeUIntParameter

Safety

AsVisualShaderNodeUvFunc

Safety

AsVisualShaderNodeUvPolarCoord

Safety

AsVisualShaderNodeVarying

Safety

AsVisualShaderNodeVaryingGetter

Safety

AsVisualShaderNodeVaryingSetter

Safety

AsVisualShaderNodeVec2Constant

Safety

AsVisualShaderNodeVec2Parameter

Safety

AsVisualShaderNodeVec3Constant

Safety

AsVisualShaderNodeVec3Parameter

Safety

AsVisualShaderNodeVec4Constant

Safety

AsVisualShaderNodeVec4Parameter

Safety

AsVisualShaderNodeVectorBase

Safety

AsVisualShaderNodeVectorCompose

Safety

AsVisualShaderNodeVectorDecompose

Safety

AsVisualShaderNodeVectorDistance

Safety

AsVisualShaderNodeVectorFunc

Safety

AsVisualShaderNodeVectorLen

Safety

AsVisualShaderNodeVectorOp

Safety

AsVisualShaderNodeVectorRefract

Safety

AsVoxelGi

Safety

AsVoxelGiData

Safety

AsWeakRef

Safety

AsWebRtcDataChannel

Safety

AsWebRtcDataChannelExtension

Safety

AsWebRtcMultiplayerPeer

Safety

AsWebRtcPeerConnection

Safety

AsWebRtcPeerConnectionExtension

Safety

AsWebSocketMultiplayerPeer

Safety

AsWebSocketPeer

Safety

AsWebXrInterface

Safety

AsWindow

Safety

AsWorkerThreadPool

Safety

AsWorld2D

Safety

AsWorld3D

Safety

AsWorldBoundaryShape2D

Safety

AsWorldBoundaryShape3D

Safety

AsWorldEnvironment

Safety

AsX509Certificate

Safety

AsXmlParser

Safety

AsXrAnchor3D

Safety

AsXrCamera3D

Safety

AsXrController3D

Safety

AsXrInterface

Safety

AsXrInterfaceExtension

Safety

AsXrNode3D

Safety

AsXrOrigin3D

Safety

AsXrPose

Safety

AsXrPositionalTracker

Safety

AsXrServer

Safety

AsZipPacker

Safety

AsZipReader

Safety

AudioStreamPlayerVirtual

Virtual methods for class AudioStreamPlayer.

Camera2DVirtual

Virtual methods for class Camera2D.

Camera3DVirtual

Virtual methods for class Camera3D.

Export

Trait implemented for types that can be used as #[export] fields. This creates a copy of the value, for some type-specific definition of “copy”. For example, Array and Gd are returned via Share::share() instead of copying the actual data.

ExtensionLayer

ExtensionLibrary

Defines the entry point for a GDExtension Rust library.

FromVariant

Trait to enable conversions of types from the Variant type.

GodotClass

Makes T eligible to be managed by Godot and stored in Gd<T> pointers.

Inherits

Non-strict inheritance relationship in the Godot class hierarchy.

Node2DVirtual

Virtual methods for class Node2D.

Node3DVirtual

Virtual methods for class Node3D.

NodeVirtual

Virtual methods for class Node.

ObjectVirtual

Virtual methods for class Object.

PackedSceneVirtual

Virtual methods for class PackedScene.

RealConv

Clippy often complains if you do f as f64 when f is already an f64. This trait exists to make it easy to convert between the different reals and floats without a lot of allowing clippy lints for your code.

RefCountedVirtual

Virtual methods for class RefCounted.

ResourceVirtual

Virtual methods for class Resource.

SceneTreeVirtual

Virtual methods for class SceneTree.

Share

Trait to create more references from a smart pointer or collection.

ToVariant

Trait to enable conversions of types to the Variant type.

TypeStringHint

Trait for types that can be represented as a type string for use with PropertyHint::PROPERTY_HINT_TYPE_STRING.

Functions

bezier_derivative

bezier_interpolate

cubic_interpolate

cubic_interpolate_in_time

fposmod

is_angle_equal_approx

Check if two angles are approximately equal, by comparing the distance between the points on the unit circle with 0 using is_equal_approx.

is_equal_approx

is_zero_approx

lerp

lerp_angle

Linearly interpolates between two angles (in radians) by a weight value between 0.0 and 1.0.

load

Loads a resource from the filesystem located at path, panicking on error.

print

Prints to the Godot console, used by the godot_print! macro.

sign

snapped

try_load

Loads a resource from the filesystem located at path.

Type Definitions

VariantArray

A Godot Array without an assigned type.

real

Floating point type used for many structs and functions in Godot.

Attribute Macros

gdextension

Proc-macro attribute to be used in combination with the ExtensionLibrary trait.

godot_api

Derive Macros

GodotClass

Derive macro for GodotClass on structs. You should normally use this macro, rather than implement GodotClass manually for your type.