Module godot::prelude

Often-imported symbols.

Re-exports

pub use super::engine::NodeExt as _;

pub use super::obj::EngineEnum as _;

Modules

array

Specialized types related to arrays.

dictionary

Specialized types related to dictionaries.

meta

Meta-information about variant types, properties and class names.

real_consts

Basic mathematical constants.

utilities

Global utility functions.

Macros

array

Constructs Array literals, similar to Rust’s standard vec! macro.

dict

Constructs Dictionary literals, close to Godot’s own syntax.

godot_error

Pushes an error message to Godot’s built-in debugger and to the OS terminal.

godot_print

Prints to the Godot console.

godot_script_error

godot_warn

Pushes a warning message to Godot’s built-in debugger and to the OS terminal.

real

A macro to coerce float-literals into the real type. Mainly used where you’d normally use a suffix to specity the type, such as 115.0f32.

varray

Constructs VariantArray literals, similar to Rust’s standard vec! macro.

Structs

Aabb

Axis-aligned bounding box in 3D space.

Array

Godot’s Array type.

AudioStreamPlayer

Godot class AudioStreamPlayer.

Base

Restricted version of Gd, to hold the base instance inside a user’s GodotClass.

Basis

A 3x3 matrix, typically used as an orthogonal basis for Transform3D.

Callable

A Callable represents a function in Godot.

Camera2D

Godot class Camera2D.

Camera3D

Godot class Camera3D.

Color

Color built-in type, in floating-point RGBA format.

Dictionary

Godot’s Dictionary type.

Gd

Smart pointer to objects owned by the Godot engine.

GdMut

Mutably/exclusively bound reference guard for a Gd smart pointer.

GdRef

Immutably/shared bound reference guard for a Gd smart pointer.

GodotString

Godot’s reference counted string type.

InitHandle

Input

Godot class Input.

InstanceId

Represents a non-zero instance ID.

Node

Godot class Node.

Node2D

Godot class Node2D.

Node3D

Godot class Node3D.

NodePath

A pre-parsed scene tree path.

Object

Godot class Object.

PackedByteArray

Implements Godot’s $PackedArray type, which is an efficient array of $Elements.

PackedColorArray

Implements Godot’s $PackedArray type, which is an efficient array of $Elements.

PackedFloat32Array

Implements Godot’s $PackedArray type, which is an efficient array of $Elements.

PackedFloat64Array

Implements Godot’s $PackedArray type, which is an efficient array of $Elements.

PackedInt32Array

Implements Godot’s $PackedArray type, which is an efficient array of $Elements.

PackedInt64Array

Implements Godot’s $PackedArray type, which is an efficient array of $Elements.

PackedScene

Godot class PackedScene.

PackedStringArray

Implements Godot’s $PackedArray type, which is an efficient array of $Elements.

PackedVector2Array

Implements Godot’s $PackedArray type, which is an efficient array of $Elements.

PackedVector3Array

Implements Godot’s $PackedArray type, which is an efficient array of $Elements.

Plane

3D plane in Hessian normal form.

Projection

A 4x4 matrix used for 3D projective transformations. It can represent transformations such as translation, rotation, scaling, shearing, and perspective division. It consists of four Vector4 columns.

Quaternion

Rect2

2D axis-aligned bounding box.

Rect2i

2D axis-aligned integer bounding box.

RefCounted

Godot class RefCounted.

Resource

Godot class Resource.

SceneTree

Godot class SceneTree.

Signal

StringName

A string optimized for unique names.

Transform2D

Affine 2D transform (2x3 matrix).

Transform3D

Affine 3D transform (3x4 matrix).

Variant

VariantConversionError

Vector2

Vector used for 2D math using floating point coordinates.

Vector2i

Vector used for 2D math using integer coordinates.

Vector3

Vector used for 3D math using floating point coordinates.

Vector3i

Vector used for 3D math using integer coordinates.

Vector4

Vector used for 4D math using floating point coordinates.

Vector4i

Vector used for 4D math using integer coordinates.

Enums

ColorChannelOrder

EulerOrder

The ordering used to interpret a set of euler angles as extrinsic rotations.

InitLevel

ProjectionEye

The eye to create a projection for, when creating a projection adjusted for head-mounted displays.

ProjectionPlane

A projections clipping plane.

RectSide

The side of a Rect2 or Rect2i.

Rid

A RID (“resource ID”) is an opaque handle that refers to a Godot Resource.

VariantOperator

VariantType

Vector2Axis

Enumerates the axes in a Vector2.

Vector2iAxis

Enumerates the axes in a Vector2i.

Vector3Axis

Enumerates the axes in a Vector3.

Vector3iAxis

Enumerates the axes in a Vector3i.

Vector4Axis

Enumerates the axes in a Vector4.

Vector4iAxis

Enumerates the axes in a Vector4i.

Constants

CMP_EPSILON

Traits

AudioStreamPlayerVirtual

Virtual methods for class AudioStreamPlayer.

Camera2DVirtual

Virtual methods for class Camera2D.

Camera3DVirtual

Virtual methods for class Camera3D.

Export

Trait implemented for types that can be used as #[export] fields. This creates a copy of the value, for some type-specific definition of “copy”. For example, Array and Gd are returned via Share::share() instead of copying the actual data.

ExtensionLayer

ExtensionLibrary

Defines the entry point for a GDExtension Rust library.

FromVariant

Trait to enable conversions of types from the Variant type.

GodotClass

Makes T eligible to be managed by Godot and stored in Gd<T> pointers.

Inherits

Non-strict inheritance relationship in the Godot class hierarchy.

Node2DVirtual

Virtual methods for class Node2D.

Node3DVirtual

Virtual methods for class Node3D.

NodeVirtual

Virtual methods for class Node.

ObjectVirtual

Virtual methods for class Object.

PackedSceneVirtual

Virtual methods for class PackedScene.

RealConv

Clippy often complains if you do f as f64 when f is already an f64. This trait exists to make it easy to convert between the different reals and floats without a lot of allowing clippy lints for your code.

RefCountedVirtual

Virtual methods for class RefCounted.

ResourceVirtual

Virtual methods for class Resource.

SceneTreeVirtual

Virtual methods for class SceneTree.

Share

Trait to create more references from a smart pointer or collection.

ToVariant

Trait to enable conversions of types to the Variant type.

Functions

bezier_derivative

bezier_interpolate

cubic_interpolate

cubic_interpolate_in_time

fposmod

is_angle_equal_approx

Check if two angles are approximately equal, by comparing the distance between the points on the unit circle with 0 using is_equal_approx.

is_equal_approx

is_zero_approx

lerp

lerp_angle

Linearly interpolates between two angles (in radians) by a weight value between 0.0 and 1.0.

load

Loads a resource from the filesystem located at path, panicking on error.

print

Prints to the Godot console, used by the godot_print! macro.

sign

snapped

try_load

Loads a resource from the filesystem located at path.

Type Definitions

VariantArray

A Godot Array without an assigned type.

real

Floating point type used for many structs and functions in Godot.

Attribute Macros

gdextension

Proc-macro attribute to be used in combination with the ExtensionLibrary trait.

godot_api

Derive Macros

GodotClass

Derive macro for GodotClass on structs. You should normally use this macro, rather than implement GodotClass manually for your type.