Attribute Macro godot::prelude::godot_api

#[godot_api]

Proc-macro attribute to be used with impl blocks of #[derive(GodotClass)] structs.

Can be used in two ways:

#[derive(GodotClass)]
#[class(base=Node)]
struct MyClass {}

// 1) inherent impl block: user-defined, custom API
#[godot_api]
impl MyClass { /* ... */ }

// 2) trait impl block: implement Godot-specific APIs
#[godot_api]
impl INode for MyClass { /* ... */ }

The second case works by implementing the corresponding trait I<Base> for the base class of your class (for example IRefCounted or INode3D). Then, you can add functionality such as:

init constructors

lifecycle methods like ready or process

on_notification method

to_string method

Neither #[godot_api] attribute is required. For small data bundles inheriting RefCounted, you may be fine with accessing properties directly from GDScript.

Examples

RefCounted as a base, overridden init

#[derive(GodotClass)]
struct MyStruct;

#[godot_api]
impl MyStruct {
    #[func]
    pub fn hello_world(&mut self) {
        godot_print!("Hello World!")
    }
}

#[godot_api]
impl IRefCounted for MyStruct {
    fn init(_base: Base<RefCounted>) -> Self {
        MyStruct
    }
}

Note that init can be either provided by overriding it, or generated with a #[class(init)] attribute on the struct. Classes without init cannot be instantiated from GDScript.

Node as a base, generated init

#[derive(GodotClass)]
#[class(init, base=Node)]
pub struct MyNode {
    #[base]
    base: Base<Node>,
}

#[godot_api]
impl INode for MyNode {
    fn ready(&mut self) {
        godot_print!("Hello World!");
    }
}