godot::prelude
Struct Callable

pub struct Callable { /* private fields */ }

A Callable represents a function in Godot.

Callables can be created in many ways:

From an Object and a (non-static) method name. This is a standard callable.

From a GDScript class name and a static function name. (This typically works because classes are instances of GDScript).

From a GDScript lambda function.

By modifying an existing Callable with bind() or unbind().

By creating a custom callable from Rust.

Godot docs

Callable (stable)

Implementations

impl Callable

pub fn from_object_method<T, S>(object: &Gd<T>, method_name: S) -> Callable
where T: GodotClass, S: AsArg<StringName>,

Create a callable for the non-static method object.method_name.

See also Gd::callable().

Godot equivalent: Callable(Object object, StringName method)

pub fn from_local_static( class_name: impl AsArg<StringName>, function_name: impl AsArg<StringName>, ) -> Callable

Create a callable for the static method class_name::function (single-threaded).

Allows you to call static functions through Callable.

Note that due to varying support across different engine versions, the resulting Callable has unspecified behavior for methods such as method_name(), object(), object_id() or [get_argument_count()][Self::arg_len] among others. It is recommended to only use this for calling the function.

Compatibility

Up until and including Godot 4.3, this method has some limitations:

is_valid() will return false, even though the call itself succeeds.

You cannot use statics to connect signals to such callables. Use the new typed signal API instead.

pub fn from_local_fn<F, S>(name: S, rust_function: F) -> Callable
where F: 'static + FnMut(&[&Variant]) -> Result<Variant, ()>, S: AsArg<GString>,

Available on since_api="4.2" only.

Create callable from single-threaded Rust function or closure.

name is used for the string representation of the closure, which helps debugging.

This constructor only allows the callable to be invoked from the same thread as creating it. If you need to invoke it from any thread, use from_sync_fn instead (requires crate feature experimental-threads; only enable if really needed).

pub fn from_sync_fn<F, S>(name: S, rust_function: F) -> Callable
where F: 'static + Send + Sync + FnMut(&[&Variant]) -> Result<Variant, ()>, S: AsArg<GString>,

Available on since_api="4.2" only.

Create callable from thread-safe Rust function or closure.

name is used for the string representation of the closure, which helps debugging.

This constructor requires Send + Sync bound and allows the callable to be invoked from any thread. If you guarantee that you invoke it from the same thread as creating it, use from_local_fn instead.

Callables created through multiple from_local_fn or from_sync_fn() calls are never equal, even if they refer to the same function. If you want to use equality, either clone an existing Callable instance, or define your own PartialEq impl with Callable::from_custom.

Example

let callable = Callable::from_sync_fn("sum", |args: &[&Variant]| {
    let sum: i32 = args.iter().map(|arg| arg.to::<i32>()).sum();
    Ok(sum.to_variant())
});

pub fn from_fn<F, S>(name: S, rust_function: F) -> Callable
where F: 'static + Send + Sync + FnMut(&[&Variant]) -> Result<Variant, ()>, S: Into<GString>,

👎Deprecated: Now split into from_local_fn (single-threaded) and from_sync_fn (multi-threaded).

Available on since_api="4.2" only.

pub fn from_custom<C>(callable: C) -> Callable
where C: RustCallable,

Available on since_api="4.2" only.

Create a highly configurable callable from Rust.

See RustCallable for requirements on the type.

pub fn invalid() -> Callable

Creates an invalid/empty object that cannot be called.

Godot equivalent: Callable()

pub fn callv(&self, arguments: &Array<Variant>) -> Variant

Calls the method represented by this callable.

Arguments passed should match the method’s signature.

If called with more arguments than expected by the method, the extra arguments will be ignored and the call continues as normal.

If called with fewer arguments than expected it will crash Godot, without triggering UB.

If called with arguments of the wrong type then an error will be printed and the call will return NIL.

If called on an invalid Callable then no error is printed, and NIL is returned.

Godot equivalent: callv

pub fn bindv(&self, arguments: &Array<Variant>) -> Callable

Returns a copy of this Callable with one or more arguments bound, reading them from an array.

Godot equivalent: bindv

pub fn method_name(&self) -> Option<StringName>

Returns the name of the method represented by this callable. If the callable is a lambda function, returns the surrounding function’s name.

Known Bugs

Getting the name of a lambda errors instead of returning its name, see godot#73052.

Godot equivalent: get_method

pub fn object(&self) -> Option<Gd<Object>>

Returns the object on which this callable is called.

Returns None when this callable doesn’t have any target object to call a method on (regardless of whether the method exists for that target or not). Also returns None if the object is dead. You can differentiate these two cases using object_id().

Godot equivalent: get_object

pub fn object_id(&self) -> Option<InstanceId>

Returns the ID of this callable’s object, see also Gd::instance_id.

Returns None when this callable doesn’t have any target to call a method on.

If the pointed-to object is dead, the ID will still be returned. Use object() to check for liveness.

Godot equivalent: get_object_id

pub fn hash(&self) -> u32

Returns the 32-bit hash value of this callable’s object.

Godot equivalent: hash

pub fn is_custom(&self) -> bool

Returns true if this callable is a custom callable.

Custom callables are mainly created from bind or unbind. In GDScript, lambda functions are also custom callables.

If a callable is not a custom callable, then it is considered a standard callable, this function is the opposite of Callable.is_standard.

Godot equivalent: is_custom

pub fn is_null(&self) -> bool

Returns true if this callable has no target to call the method on.

This is not the negated form of is_valid, as is_valid will return false if the callable has a target but the method does not exist.

Godot equivalent: is_null

pub fn is_valid(&self) -> bool

Returns true if the callable’s object exists and has a valid method name assigned, or is a custom callable.

Godot equivalent: is_valid

pub fn unbind(&self, args: usize) -> Callable

Returns a copy of the callable, ignoring args user arguments.

Despite its name, this does not directly undo previous bind() calls. See Godot docs for up-to-date semantics.

pub fn get_argument_count(&self) -> usize

Available on since_api="4.3" only.

pub fn get_bound_arguments_count(&self) -> usize

Get number of bound arguments.

Note: for Godot < 4.4, this function returns incorrect results when applied on a callable that used unbind(). See #98713 for details.

impl Callable

pub fn get_bound_arguments(&self) -> Array<Variant>

pub fn call(&self, varargs: &[Variant]) -> Variant

pub fn call_deferred(&self, varargs: &[Variant])

pub fn rpc(&self, varargs: &[Variant])

pub fn rpc_id(&self, peer_id: i64, varargs: &[Variant])

pub fn bind(&self, varargs: &[Variant]) -> Callable