Struct godot::prelude::GFile

pub struct GFile { /* private fields */ }

Open file for reading or writing.

It’s a convienience wrapper around Gd<FileAccess>, providing multiple both safety and quality of life upgrades over the inner type.

Exposes reading and writing capabilities of FileAccess in a safe way, returning Result<T> instead of the T itself.

Makes the FileAccess exclusive to its instance, not allowing parallel reads and writes, which could introduce hard to track bugs by unpredicted cursor movement. Exclusivity also ensures that when the GFile moves out of scope, inner FileAccess does also, automatically closing the file. Alternatively, you can drop() the GFile to close the file manually.

Implements useful Rust traits allowing better file processing, namely Read, BufRead, Write, Seek and making it better integrated with various tools in Rust ecosystem (eg. serialization tools).

Files by default are always opened with little-endian, as most files are saved as such. To switch into big-endian, use GFile.set_big_endian().

ModeFlags

Every constructor (open_* associated function) intakes a flags argument, opening the file for different types of operations. Regardless of provided flags, cursor is always positioned at the beginning of the file upon opening. To adjust its position use Seek-provided methods.

ModeFlags::READ opens the file for read operations.

ModeFlags::WRITE opens the file for write operations. If the file doesn’t exist at provided path, it is created. If it exists, it is truncated after the file is closed.

ModeFlags::READ_WRITE opens the file for read and write operations. The file is not truncated after closing.

ModeFlags::WRITE_READ opens the file for read and write operations. If the file doesn’t exist at provided path, it is created. If it exists, it is truncated after the file is closed.

Examples

use godot::engine::GFile;
use godot::engine::file_access::ModeFlags;
use godot::builtin::GodotString;

// Open file in write mode
let mut my_file = GFile::open("user://save_game.sav", ModeFlags::WRITE)
    .expect("couldn't open file for writing");

// Write some lines into it
my_file.write_gstring_line("This is my saved game")
    .expect("couldn't write first line");
my_file.write_gstring_line("I played for 5 minutes")
    .expect("couldn't write second line");

// Consume object and close the file
drop(my_file);

// Open file in read mode
let mut my_file = GFile::open("user://save_game.sav", ModeFlags::READ)
    .expect("couldn't open file for reading");

// Read lines
let first_line = my_file.read_gstring_line()
    .expect("couldn't read first line");
assert_eq!(first_line, GodotString::from("This is my saved game"));
let second_line = my_file.read_gstring_line()
    .expect("couldn't read second line");
assert_eq!(second_line, GodotString::from("I played for 5 minutes"));

// Consume object and close the file
drop(my_file);

References

FileAccess class Godot documentation

Implementations

impl GFile

pub fn open( path: impl Into<GodotString>, flags: ModeFlags ) -> Result<GFile, Error>

Open a file

Opens a file located at path, creating new GFile object. For ModeFlags description check the GFile documentation.

pub fn open_compressed( path: impl Into<GodotString>, flags: ModeFlags, compression_mode: CompressionMode ) -> Result<GFile, Error>

Open a compressed file

Opens a compressed file located at path, creating new GFile object. Can read only files compressed by Godot compression formats. For ModeFlags description check the GFile documentation.

pub fn open_encrypted( path: impl Into<GodotString>, flags: ModeFlags, key: PackedByteArray ) -> Result<GFile, Error>

Open a file encrypted by byte key

Opens a file encrypted by 32-byte long PackedByteArray located at path, creating new GFile object. For ModeFlags description check the GFile documentation.

pub fn open_encrypted_with_pass( path: impl Into<GodotString>, flags: ModeFlags, pass: GodotString ) -> Result<GFile, Error>

Open a file encrypted by password

Opens a file encrypted by a pass located at path, creating new GFile object. For ModeFlags description check the GFile documentation.

pub fn from_unique(file_access: Gd<FileAccess>) -> Result<GFile, NotUniqueError>

Creates new GFile from provided Gd<FileAccess>.

For this method to work, the provided file_access must be unique - no other symbol can reference it. Its state is retained: ModeFlags with which it was open and current internal cursor position.

pub fn into_inner(self) -> Gd<FileAccess>

Retrieve inner Gd<FileAccess>

Deconstructs the GFile intance and retrieves the inner Gd<FileAccess>. Useful when the Gd<FileAccess> is needed in some other function or method, and you would like to preserve its state or just keep the file open while transferring.

pub fn modified_time(path: impl Into<GodotString>) -> Result<u64, Error>

Get last modified time

Get last modified time as an unix timestamp

pub fn md5(path: impl Into<GodotString>) -> Result<GodotString, Error>

Get file md5 representation

Returns a md5 GodotString representing the file at the given path

pub fn sha256(path: impl Into<GodotString>) -> Result<GodotString, Error>

Get file sha256 representation

Returns a sha-256 GodotString representing the file at the given path

pub fn read_u8(&mut self) -> Result<u8, Error>

Reads the next 8 bits from the file as u8.

For more information about underlying method see Godot documentation.

pub fn read_u16(&mut self) -> Result<u16, Error>

Reads the next 16 bits from the file as u16.

For more information about underlying method see Godot documentation.

pub fn read_u32(&mut self) -> Result<u32, Error>

Reads the next 32 bits from the file as u32.

For more information about underlying method see Godot documentation.

pub fn read_u64(&mut self) -> Result<u64, Error>

Reads the next 64 bits from the file as u64.

For more information about underlying method see Godot documentation.

pub fn read_pascal_string(&mut self) -> Result<GodotString, Error>

Reads the Pascal String from file as GodotString.

Pascal String is useful for writing and retrieving verying-length string data from binary files. It is saved with length-prefix, instead of null terminator as in C strings. Text is interpreted as being UTF-8 encoded.

See also:

Wikipedia article

Godot documentation

pub fn read_as_gstring_entire( &mut self, skip_cr: bool ) -> Result<GodotString, Error>

Reads the whole file as GodotString.

If skip_cr is set to true, carriage returns will be ingored when parsing and only line feed characters will represents a new line. Text is interpreted as being UTF-8 encoded.

To retrieve the file as String instead, use GFile::read_to_string() Read trait method.

For more information about underlying method see Godot documentation.

pub fn read_csv_line( &mut self, delim: impl Into<GodotString> ) -> Result<PackedStringArray, Error>

Reads the next line of the file in delimiter-separated file.

For reading traditional CSV format, provide comma (',') as delim.

For more information about underlying method see Godot documentation.

pub fn read_f32(&mut self) -> Result<f32, Error>

Reads the next 32 bits from file as f32.

For more information about underlying method see Godot documentation.

pub fn read_f64(&mut self) -> Result<f64, Error>

Reads the next 64 bits from file as f64.

For more information about underlying method see Godot documentation.

pub fn read_gstring_line(&mut self) -> Result<GodotString, Error>

Reads the next line of the file as GodotString.

To retrieve the file as String instead, use GFile::read_line() Read trait method.

For more information about underlying method see Godot documentation.

pub fn read_variant(&mut self, allow_objects: bool) -> Result<Variant, Error>

Reads the next Variant value from file.

If allow_objects is set to true, objects will be decoded. For more information about underlying method see Godot documentation.

pub fn write_u8(&mut self, value: u8) -> Result<(), Error>

Writes u8 as the 8 bits in the file.

For more information about underlying method see Godot documentation.

pub fn write_u16(&mut self, value: u16) -> Result<(), Error>

Writes u16 as the 16 bits in the file.

For more information about underlying method see Godot documentation.

pub fn write_u32(&mut self, value: u32) -> Result<(), Error>

Writes u32 as the 32 bits in the file.

For more information about underlying method see Godot documentation.

pub fn write_u64(&mut self, value: u64) -> Result<(), Error>

Writes u64 as the 64 bits in the file.

For more information about underlying method see Godot documentation.

pub fn write_f32(&mut self, value: f32) -> Result<(), Error>

Writes f32 as the 32 bits in the file.

For more information about underlying method see Godot documentation.

pub fn write_f64(&mut self, value: f64) -> Result<(), Error>

Writes f64 as the 64 bits in the file.

For more information about underlying method see Godot documentation.

pub fn write_gstring( &mut self, value: impl Into<GodotString> ) -> Result<(), Error>

Writes string to the file.

This function is meant to be used in text files. To store a string in a binary file, use store_pascal_string()

For more information about underlying method see Godot documentation.

pub fn write_pascal_string( &mut self, value: impl Into<GodotString> ) -> Result<(), Error>

Writes string to the file as Pascal String.

This function is meant to be used in binary files. To store a string in a text file, use store_string()

Pascal String is useful for writing and retrieving verying-length string data from binary files. It is saved with length-prefix, instead of null terminator as in C strings.

See also:

Wikipedia article

Godot documentation

pub fn write_gstring_line( &mut self, value: impl Into<GodotString> ) -> Result<(), Error>

Write string to the file as a line.

For more information about underlying method see Godot documentation.

pub fn write_csv_line( &mut self, values: PackedStringArray, delim: impl Into<GodotString> ) -> Result<(), Error>

Write PackedStringArray to the file as delimited line.

For writing traditional CSV format, provide comma (',') as delim.

For more information about underlying method see Godot documentation.

pub fn write_variant( &mut self, value: Variant, full_objects: bool ) -> Result<(), Error>

Write Variant to the file.

If full_objects is set to true, encoding objects is allowed (and can potentially include GDScript code). Not all properties of the Variant are included. Only properties that are exported (have #[export] derive attribute) will be serialized.

For more information about underlying method see Godot documentation.

pub fn set_big_endian(&mut self, value: bool)

Set true to use big-endian, false to use little-endian.

endianness can be set mid-file, not only at the start position. It makes it possible to write different sections of binary file with different endianness, though it is not recommended - can lead to confusion and mistakes during consequent read operations.

pub fn is_big_endian(&self) -> bool

Check endianness of file connection.

pub fn path(&self) -> GodotString

Get path of the opened file.

pub fn path_absolute(&self) -> GodotString

Get absolute path of the opened file.

pub fn position(&self) -> u64

Returns the current cursor position.

pub fn length(&self) -> u64

Get file length in bytes

pub fn eof_reached(&self) -> bool

Checks if the file cursor has read past the end of the file.