Struct godot::prelude::Node

pub struct Node { /* private fields */ }

Godot class Node.

Inherits Object.

Related symbols:

node: sidecar module with related enum/flag types

NodeVirtual: virtual methods

NodeNotification: notification type

See also Godot docs for Node.

Implementations

impl Node

pub fn new_alloc() -> Gd<Node>

pub fn notify(&mut self, what: NodeNotification)

⚠️ Sends a Godot notification to all classes inherited by the object.

Triggers calls to on_notification(), and depending on the notification, also to Godot’s lifecycle callbacks such as ready().

Starts from the highest ancestor (the Object class) and goes down the hierarchy. See also Godot docs for Object::notification().

Panics

If you call this method on a user-defined object while holding a GdRef or GdMut guard on the instance, you will encounter a panic. The reason is that the receiving virtual method on_notification() acquires a GdMut lock dynamically, which must be exclusive.

pub fn notify_reversed(&mut self, what: NodeNotification)

⚠️ Like Self::notify(), but starts at the most-derived class and goes up the hierarchy.

See docs of that method, including the panics.

pub fn print_orphan_nodes()

pub fn add_sibling(&mut self, sibling: Gd<Node>, force_readable_name: bool)

pub fn set_name(&mut self, name: GodotString)

pub fn get_name(&self) -> StringName

pub fn add_child( &mut self, node: Gd<Node>, force_readable_name: bool, internal: InternalMode )

pub fn remove_child(&mut self, node: Gd<Node>)

pub fn reparent(&mut self, new_parent: Gd<Node>, keep_global_transform: bool)

pub fn get_child_count(&self, include_internal: bool) -> i64

pub fn get_children(&self, include_internal: bool) -> Array<Gd<Node>>

pub fn get_child(&self, idx: i64, include_internal: bool) -> Option<Gd<Node>>

pub fn has_node(&self, path: NodePath) -> bool

pub fn get_node(&self, path: NodePath) -> Option<Gd<Node>>

pub fn get_node_or_null(&self, path: NodePath) -> Option<Gd<Node>>

pub fn get_parent(&self) -> Option<Gd<Node>>

pub fn find_child( &self, pattern: GodotString, recursive: bool, owned: bool ) -> Option<Gd<Node>>

pub fn find_children( &self, pattern: GodotString, type_: GodotString, recursive: bool, owned: bool ) -> Array<Gd<Node>>

pub fn find_parent(&self, pattern: GodotString) -> Option<Gd<Node>>

pub fn has_node_and_resource(&self, path: NodePath) -> bool

pub fn get_node_and_resource(&mut self, path: NodePath) -> Array<Variant>

pub fn is_inside_tree(&self) -> bool

pub fn is_ancestor_of(&self, node: Gd<Node>) -> bool

pub fn is_greater_than(&self, node: Gd<Node>) -> bool

pub fn get_path(&self) -> NodePath

pub fn get_path_to(&self, node: Gd<Node>, use_unique_path: bool) -> NodePath

pub fn add_to_group(&mut self, group: StringName, persistent: bool)

pub fn remove_from_group(&mut self, group: StringName)

pub fn is_in_group(&self, group: StringName) -> bool

pub fn move_child(&mut self, child_node: Gd<Node>, to_index: i64)

pub fn get_groups(&self) -> Array<StringName>

pub fn set_owner(&mut self, owner: Gd<Node>)

pub fn get_owner(&self) -> Option<Gd<Node>>

pub fn get_index(&self, include_internal: bool) -> i64

pub fn print_tree(&mut self)

pub fn print_tree_pretty(&mut self)

pub fn set_scene_file_path(&mut self, scene_file_path: GodotString)

pub fn get_scene_file_path(&self) -> GodotString

pub fn propagate_notification(&mut self, what: i64)

pub fn propagate_call( &mut self, method: StringName, args: Array<Variant>, parent_first: bool )

pub fn set_physics_process(&mut self, enable: bool)

pub fn get_physics_process_delta_time(&self) -> f64

pub fn is_physics_processing(&self) -> bool

pub fn get_process_delta_time(&self) -> f64

pub fn set_process(&mut self, enable: bool)

pub fn set_process_priority(&mut self, priority: i64)

pub fn get_process_priority(&self) -> i64

pub fn is_processing(&self) -> bool

pub fn set_process_input(&mut self, enable: bool)

pub fn is_processing_input(&self) -> bool

pub fn set_process_shortcut_input(&mut self, enable: bool)

pub fn is_processing_shortcut_input(&self) -> bool

pub fn set_process_unhandled_input(&mut self, enable: bool)

pub fn is_processing_unhandled_input(&self) -> bool

pub fn set_process_unhandled_key_input(&mut self, enable: bool)

pub fn is_processing_unhandled_key_input(&self) -> bool

pub fn set_process_mode(&mut self, mode: ProcessMode)

pub fn get_process_mode(&self) -> ProcessMode

pub fn can_process(&self) -> bool

pub fn set_display_folded(&mut self, fold: bool)

pub fn is_displayed_folded(&self) -> bool

pub fn set_process_internal(&mut self, enable: bool)

pub fn is_processing_internal(&self) -> bool

pub fn set_physics_process_internal(&mut self, enable: bool)

pub fn is_physics_processing_internal(&self) -> bool

pub fn get_window(&self) -> Option<Gd<Window>>

pub fn get_tree(&self) -> Option<Gd<SceneTree>>

pub fn create_tween(&mut self) -> Option<Gd<Tween>>

pub fn duplicate(&self, flags: i64) -> Option<Gd<Node>>

pub fn replace_by(&mut self, node: Gd<Node>, keep_groups: bool)

pub fn set_scene_instance_load_placeholder(&mut self, load_placeholder: bool)

pub fn get_scene_instance_load_placeholder(&self) -> bool

pub fn set_editable_instance(&mut self, node: Gd<Node>, is_editable: bool)

pub fn is_editable_instance(&self, node: Gd<Node>) -> bool

pub fn get_viewport(&self) -> Option<Gd<Viewport>>

pub fn queue_free(&mut self)

pub fn request_ready(&mut self)

pub fn set_multiplayer_authority(&mut self, id: i64, recursive: bool)

pub fn get_multiplayer_authority(&self) -> i64

pub fn is_multiplayer_authority(&self) -> bool

pub fn get_multiplayer(&self) -> Option<Gd<MultiplayerApi>>

pub fn rpc_config(&mut self, method: StringName, config: Variant)

pub fn set_editor_description(&mut self, editor_description: GodotString)

pub fn get_editor_description(&self) -> GodotString

pub fn set_unique_name_in_owner(&mut self, enable: bool)

pub fn is_unique_name_in_owner(&self) -> bool

pub fn rpc(&mut self, method: StringName, varargs: &[Variant]) -> Error

pub fn rpc_id( &mut self, peer_id: i64, method: StringName, varargs: &[Variant] ) -> Error

pub fn update_configuration_warnings(&mut self)