Struct godot::prelude::Resource

pub struct Resource { /* private fields */ }

Godot class Resource.

Inherits RefCounted.

Related symbols:

ResourceVirtual: virtual methods

See also Godot docs for Resource.

Implementations

impl Resource

pub fn new() -> Gd<Resource>

pub fn notify(&mut self, what: ObjectNotification)

⚠️ Sends a Godot notification to all classes inherited by the object.

Triggers calls to on_notification(), and depending on the notification, also to Godot’s lifecycle callbacks such as ready().

Starts from the highest ancestor (the Object class) and goes down the hierarchy. See also Godot docs for Object::notification().

Panics

If you call this method on a user-defined object while holding a GdRef or GdMut guard on the instance, you will encounter a panic. The reason is that the receiving virtual method on_notification() acquires a GdMut lock dynamically, which must be exclusive.

pub fn notify_reversed(&mut self, what: ObjectNotification)

⚠️ Like Self::notify(), but starts at the most-derived class and goes up the hierarchy.

See docs of that method, including the panics.

pub fn set_path(&mut self, path: GodotString)

pub fn take_over_path(&mut self, path: GodotString)

pub fn get_path(&self) -> GodotString

pub fn set_name(&mut self, name: GodotString)

pub fn get_name(&self) -> GodotString

pub fn get_rid(&self) -> Rid

pub fn set_local_to_scene(&mut self, enable: bool)

pub fn is_local_to_scene(&self) -> bool

pub fn get_local_scene(&self) -> Option<Gd<Node>>

pub fn setup_local_to_scene(&mut self)

pub fn emit_changed(&mut self)

pub fn duplicate(&self, subresources: bool) -> Option<Gd<Resource>>