Struct godot::prelude::Transform2D

pub struct Transform2D {
    pub a: Vector2,
    pub b: Vector2,
    pub origin: Vector2,
}

Affine 2D transform (2x3 matrix).

Represents transformations such as translation, rotation, or scaling.

Expressed as a 2x3 matrix, this transform consists of a two column vectors a and b representing the basis of the transform, as well as the origin:

[ a.x  b.x  origin.x ]
[ a.y  b.y  origin.y ]

For methods that don’t take translation into account, see [Basis2D].

Fields
a: Vector2

The first basis vector.

This is equivalent to the x field from godot.

b: Vector2

The second basis vector.

This is equivalent to the y field from godot.

origin: Vector2

The origin of the transform. The coordinate space defined by this transform starts at this point.

Godot equivalent: Transform2D.origin

Implementations

impl Transform2D

pub const IDENTITY: Transform2D = Self::from_basis_origin(Basis2D::IDENTITY, Vector2::ZERO)

The identity transform, with no translation, rotation or scaling applied. When applied to other data structures, IDENTITY performs no transformation.

Godot equivalent: Transform2D.IDENTITY

pub const FLIP_X: Transform2D = Self::from_basis_origin(Basis2D::FLIP_X, Vector2::ZERO)

The Transform2D that will flip something along its x axis.

Godot equivalent: Transform2D.FLIP_X

pub const FLIP_Y: Transform2D = Self::from_basis_origin(Basis2D::FLIP_Y, Vector2::ZERO)

The Transform2D that will flip something along its y axis.

Godot equivalent: Transform2D.FLIP_Y

pub const fn from_cols(a: Vector2, b: Vector2, origin: Vector2) -> Transform2D

Create a new Transform2D with the given column vectors.

Godot equivalent: `Transform2D(Vector2 x_axis, Vector2 y_axis, Vector2 origin)

pub fn from_angle(angle: f32) -> Transform2D

Create a new Transform2D which will rotate by the given angle.

pub fn from_angle_origin(angle: f32, origin: Vector2) -> Transform2D

Create a new Transform2D which will rotate by angle and translate by origin.

Godot equivalent: Transform2D(float rotation, Vector2 position)

pub fn from_angle_scale_skew_origin( angle: f32, scale: Vector2, skew: f32, origin: Vector2 ) -> Transform2D

Create a new Transform2D which will rotate by angle, scale by scale, skew by skew and translate by origin.

Godot equivalent: Transform2D(float rotation, Vector2 scale, float skew, Vector2 position)

pub fn affine_inverse(self) -> Transform2D

Returns the inverse of the transform, under the assumption that the transformation is composed of rotation, scaling and translation.

Godot equivalent: Transform2D.affine_inverse()

pub fn rotation(&self) -> f32

Returns the transform’s rotation (in radians).

Godot equivalent: Transform2D.get_rotation()

pub fn scale(&self) -> Vector2

Returns the transform’s scale.

Godot equivalent: Transform2D.get_scale()

pub fn skew(&self) -> f32

Returns the transform’s skew (in radians).

Godot equivalent: Transform2D.get_skew()

pub fn interpolate_with(self, other: Transform2D, weight: f32) -> Transform2D

Returns a transform interpolated between this transform and another by a given weight (on the range of 0.0 to 1.0).

Godot equivalent: Transform2D.interpolate_with()

pub fn is_equal_approx(&self, other: &Transform2D) -> bool

Returns true if this transform and transform are approximately equal, by calling is_equal_approx on each component.

Godot equivalent: Transform2D.is_equal_approx()

pub fn is_finite(&self) -> bool

Returns true if this transform is finite, by calling Vector2::is_finite() on each component.

Godot equivalent: Transform2D.is_finite()

pub fn orthonormalized(self) -> Transform2D

Returns the transform with the basis orthogonal (90 degrees), and normalized axis vectors (scale of 1 or -1).

Godot equivalent: Transform2D.orthonormalized()

pub fn rotated(self, angle: f32) -> Transform2D

Returns a copy of the transform rotated by the given angle (in radians). This method is an optimized version of multiplying the given transform X with a corresponding rotation transform R from the left, i.e., R * X. This can be seen as transforming with respect to the global/parent frame.

Godot equivalent: Transform2D.rotated()

pub fn rotated_local(self, angle: f32) -> Transform2D

Returns a copy of the transform rotated by the given angle (in radians). This method is an optimized version of multiplying the given transform X with a corresponding rotation transform R from the right, i.e., X * R. This can be seen as transforming with respect to the local frame.

Godot equivalent: Transform2D.rotated_local()

pub fn scaled(self, scale: Vector2) -> Transform2D

Returns a copy of the transform scaled by the given scale factor. This method is an optimized version of multiplying the given transform X with a corresponding scaling transform S from the left, i.e., S * X. This can be seen as transforming with respect to the global/parent frame.

Godot equivalent: Transform2D.scaled()

pub fn scaled_local(self, scale: Vector2) -> Transform2D

Returns a copy of the transform scaled by the given scale factor. This method is an optimized version of multiplying the given transform X with a corresponding scaling transform S from the right, i.e., X * S. This can be seen as transforming with respect to the local frame.

Godot equivalent: Transform2D.scaled_local()

pub fn translated(self, offset: Vector2) -> Transform2D

Returns a copy of the transform translated by the given offset. This method is an optimized version of multiplying the given transform X with a corresponding translation transform T from the left, i.e., T * X. This can be seen as transforming with respect to the global/parent frame.

Godot equivalent: Transform2D.translated()

pub fn translated_local(self, offset: Vector2) -> Transform2D

Returns a copy of the transform translated by the given offset. This method is an optimized version of multiplying the given transform X with a corresponding translation transform T from the right, i.e., X * T. This can be seen as transforming with respect to the local frame.

Godot equivalent: Transform2D.translated()

pub fn basis_xform(&self, v: Vector2) -> Vector2

Returns a vector transformed (multiplied) by the basis matrix. This method does not account for translation (the origin vector).

Godot equivalent: Transform2D.basis_xform()

pub fn basis_xform_inv(&self, v: Vector2) -> Vector2

Returns a vector transformed (multiplied) by the inverse basis matrix. This method does not account for translation (the origin vector).

Godot equivalent: Transform2D.basis_xform_inv()