Trait godot::prelude::ExtensionLibrary

pub unsafe trait ExtensionLibrary {
    // Provided method
    fn load_library(handle: &mut InitHandle) -> bool { ... }
}

Defines the entry point for a GDExtension Rust library.

Every library should have exactly one implementation of this trait. It is always used in combination with the #[gdextension] proc-macro attribute.

The simplest usage is as follows. This will automatically perform the necessary init and cleanup routines, and register all classes marked with #[derive(GodotClass)], without needing to mention them in a central list. The order in which classes are registered is not specified.

// This is just a type tag without any functionality
struct MyExtension;

#[gdextension]
unsafe impl ExtensionLibrary for MyExtension {}

Safety

By using godot-rust, you accept the safety considerations as outlined in the book. Please make sure you fully understand the implications.

The library cannot enforce any safety guarantees outside Rust code, which means that you as a user are responsible to uphold them: namely in GDScript code or other GDExtension bindings loaded by the engine. Violating this may cause undefined behavior, even when invoking safe functions.

Provided Methods

fn load_library(handle: &mut InitHandle) -> bool