godot

Module register

Register/export Rust symbols to Godot: classes, methods, enums…

Modules

property

Registration support for property types.

Structs

RpcConfig

Configuration for a remote procedure call, used with #[rpc(config = ...)].

Attribute Macros

godot_api

Proc-macro attribute to be used with impl blocks of #[derive(GodotClass)] structs.

godot_dyn

Generates a Class -> dyn Trait upcasting relation.

Derive Macros

Export

Derive macro for Export on enums.

GodotClass

Derive macro for GodotClass on structs.

GodotConvert

Derive macro for GodotConvert on structs.

Var

Derive macro for Var on enums.