godot::register::property
Module export_info_functions

Functions used to translate user-provided arguments into export hints.

You are not supposed to use these functions directly. They are used by the #[export] macro to generate the correct export hint.

Each function is named the same as the equivalent Godot annotation.
For instance, @export_range in Godot is fn export_range here.

Functions

export_color_no_alpha

export_dir

export_enum

Equivalent to @export_enum in Godot.

export_exp_easing

export_file

Equivalent to @export_file in Godot.

export_file_inner

export_flags

Equivalent to @export_flags in Godot.

export_flags_2d_navigation

export_flags_2d_physics

export_flags_2d_render

export_flags_3d_navigation

export_flags_3d_physics

export_flags_3d_render

export_global_dir

export_global_file

Equivalent to @export_global_file in Godot.

export_multiline

export_placeholder

export_range

Mark an exported numerical value to use the editor’s range UI.