Trait godot::register::property::Export

pub trait Export: Var {
    // Required method
    fn default_export_info() -> PropertyHintInfo;
}

Trait implemented for types that can be used as #[export] fields.

Required Methods

fn default_export_info() -> PropertyHintInfo

The export info to use for an exported field of this type, if no other export info is specified.

Object Safety

This trait is not object safe.