Trait godot::sys::GodotFfi

pub unsafe trait GodotFfi {
    // Required methods
    unsafe fn from_sys(ptr: *mut __GdextType) -> Self;
    unsafe fn from_sys_init(init_fn: impl FnOnce(*mut __GdextType)) -> Self;
    fn sys(&self) -> *mut __GdextType;
    unsafe fn from_arg_ptr(ptr: *mut __GdextType, call_type: PtrcallType) -> Self;
    unsafe fn move_return_ptr(self, dst: *mut __GdextType, call_type: PtrcallType);

    // Provided methods
    unsafe fn from_sys_init_default(
        init_fn: impl FnOnce(*mut __GdextType)
    ) -> Self
       where Self: Sized { ... }
    fn sys_mut(&mut self) -> *mut __GdextType { ... }
    fn sys_const(&self) -> *const __GdextType { ... }
}

Adds methods to convert from and to Godot FFI pointers. See crate::ffi_methods for ergonomic implementation.

Safety

from_arg_ptr and move_return_ptr must properly initialize and clean up values given the PtrcallType provided by the caller.

Required Methods

unsafe fn from_sys(ptr: *mut __GdextType) -> Self

Construct from Godot opaque pointer.

Safety

ptr must be a valid type ptr: it must follow Godot’s convention to encode Self, which is different depending on the type. The type in ptr must not require any special consideration upon referencing. Such as incrementing a refcount.

unsafe fn from_sys_init(init_fn: impl FnOnce(*mut __GdextType)) -> Self

Construct uninitialized opaque data, then initialize it with init_fn function.

Safety

init_fn must be a function that correctly handles a (possibly-uninitialized) type ptr.

fn sys(&self) -> *mut __GdextType

Return Godot opaque pointer, for an immutable operation.

Note that this is a *mut pointer despite taking &self by shared-ref. This is because most of Godot’s rust API is not const-correct. This can still enhance user code (calling sys_mut ensures no aliasing at the time of the call).

unsafe fn from_arg_ptr(ptr: *mut __GdextType, call_type: PtrcallType) -> Self

Construct from a pointer to an argument in a call.

Safety

ptr must be a valid type ptr: it must follow Godot’s convention to encode Self, which is different depending on the type.

ptr must encode Self according to the given call_type’s encoding of argument values.

unsafe fn move_return_ptr(self, dst: *mut __GdextType, call_type: PtrcallType)

Move self into the pointer in pointer dst, dropping what is already in `dst.

Safety

dst must be a valid type ptr: it must follow Godot’s convention to encode Self, which is different depending on the type.

dst must be able to accept a value of type Self encoded according to the given call_type’s encoding of return values.

Provided Methods

unsafe fn from_sys_init_default(init_fn: impl FnOnce(*mut __GdextType)) -> Selfwhere Self: Sized,

Like Self::from_sys_init, but pre-initializes the sys pointer to a Default::default() instance before calling init_fn.

Some FFI functions in Godot expect a pre-existing instance at the destination pointer, e.g. CoW/ref-counted builtin types like Array, Dictionary, String, StringName.

If not overridden, this just calls Self::from_sys_init.

Safety

init_fn must be a function that correctly handles a (possibly-uninitialized) type ptr.

fn sys_mut(&mut self) -> *mut __GdextType

Return Godot opaque pointer, for a mutable operation.

Should usually not be overridden; behaves like sys() but ensures no aliasing at the time of the call (not necessarily during any subsequent modifications though).

fn sys_const(&self) -> *const __GdextType