Trait godot::sys::GodotFuncMarshal

pub trait GodotFuncMarshal: Sized {
    type Via: Debug;

    // Required methods
    unsafe fn try_from_arg(
        ptr: *mut __GdextType,
        call_type: PtrcallType
    ) -> Result<Self, Self::Via>;
    unsafe fn try_return(
        self,
        dst: *mut __GdextType,
        call_type: PtrcallType
    ) -> Result<(), Self>;
}

Trait implemented for all types that can be passed to and from user-defined #[func] methods through Godot’s ptrcall calling convention.

Required Associated Types

type Via: Debug

Intermediate type through which Self is converted, and which can cause failure.

Required Methods

unsafe fn try_from_arg( ptr: *mut __GdextType, call_type: PtrcallType ) -> Result<Self, Self::Via>

Used for function arguments. On failure, the argument which can’t be converted to Self is returned.

Safety

The value behind ptr must be the C FFI type that corresponds to Self. See also GodotFfi::from_arg_ptr.

unsafe fn try_return( self, dst: *mut __GdextType, call_type: PtrcallType ) -> Result<(), Self>

Used for function return values. On failure, self which can’t be converted to Via is returned.

Safety

The value behind ptr must be the C FFI type that corresponds to Self. dst must point to an initialized value of type Via. See also GodotFfi::move_return_ptr.