Module godot::tools

Higher-level additions to the Godot engine API.

Contains functionality that extends existing Godot classes and functions, to make them more versatile or better integrated with Rust.

Macros

tr

A convenience macro for using the Object::tr() and Object::tr_ex() methods.

tr_n

A convenience macro for using the Object::tr_n() and Object::tr_n_ex() methods.

Structs

GFile

Open a file for reading or writing.

Functions

load

⚠️ Loads a resource from the filesystem located at path, panicking on error.

save

⚠️ Saves a Resource-inheriting object into the file located at path.

try_load

Loads a resource from the filesystem located at path.

try_save

Saves a Resource-inheriting object into the file located at path.